Calculate Akaike weights for model averaging

Akaike weights are calculated based on the relative expected Kullback-Leibler information as specified by Burnham and Anderson (2004).

Usage

aw(object, ...)

# S3 method for class 'mkinfit'
aw(object, ...)

# S3 method for class 'mmkin'
aw(object, ...)

# S3 method for class 'mixed.mmkin'
aw(object, ...)

# S3 method for class 'multistart'
aw(object, ...)

Arguments

object

An mmkin column object, containing two or more mkinfit models that have been fitted to the same data, or an mkinfit object. In the latter case, further mkinfit objects fitted to the same data should be specified as dots arguments.

...

Not used in the method for mmkin column objects, further mkinfit objects in the method for mkinfit objects.

References

Burnham KP and Anderson DR (2004) Multimodel Inference: Understanding AIC and BIC in Model Selection. Sociological Methods & Research 33(2) 261-304

Examples

# \dontrun{
f_sfo <- mkinfit("SFO", FOCUS_2006_D, quiet = TRUE)
f_dfop <- mkinfit("DFOP", FOCUS_2006_D, quiet = TRUE)
aw_sfo_dfop <- aw(f_sfo, f_dfop)
sum(aw_sfo_dfop)
#> [1] 1
aw_sfo_dfop # SFO gets more weight as it has less parameters and a similar fit
#> [1] 0.5970258 0.4029742
f <- mmkin(c("SFO", "FOMC", "DFOP"), list("FOCUS D" = FOCUS_2006_D), cores = 1, quiet = TRUE)
aw(f)
#> [1] 0.4808722 0.1945539 0.3245740
sum(aw(f))
#> [1] 1
aw(f[c("SFO", "DFOP")])
#> [1] 0.5970258 0.4029742
# }