Confidence intervals for parameters of mkinfit objects

The default method 'profile' is based on the profile likelihood for each parameter. The method uses two nested optimisations. The speed of the method could likely be improved by using the method of Venzon and Moolgavkar (1988).

# S3 method for mkinfit
confint(object, parm, level = 0.95, alpha = 1 -
  level, cutoff, method = c("profile", "quadratic"),
  transformed = TRUE, backtransform = TRUE,
  distribution = c("student_t", "normal"), quiet = FALSE, ...)

Arguments

object
An mkinfit object

parm
A vector of names of the parameters which are to be given confidence intervals. If missing, all parameters are considered.

level
The confidence level required

alpha
The allowed error probability, overrides 'level' if specified.

cutoff
Possibility to specify an alternative cutoff for the difference in the log-likelihoods at the confidence boundary. Specifying an explicit cutoff value overrides arguments 'level' and 'alpha'

method
The 'profile' method searches the parameter space for the cutoff of the confidence intervals by means of a likelihood ratio test. The 'quadratic' method approximates the likelihood function at the optimised parameters using the second term of the Taylor expansion, using a second derivative (hessian) contained in the object.

transformed
If the quadratic approximation is used, should it be applied to the likelihood based on the transformed parameters?

backtransform
If we approximate the likelihood in terms of the transformed parameters, should we backtransform the parameters with their confidence intervals?

distribution
For the quadratic approximation, should we use the student t distribution or assume normal distribution for the parameter estimate

quiet
Should we suppress messages?

...
Not used

Value

A matrix with columns giving lower and upper confidence limits for each parameter.

References

Pawitan Y (2013) In all likelihood - Statistical modelling and inference using likelihood. Clarendon Press, Oxford. Venzon DJ and Moolgavkar SH (1988) A Method for Computing Profile-Likelihood Based Confidence Intervals, Applied Statistics, 37, 87–94.

Examples

f <- mkinfit("SFO", FOCUS_2006_C, quiet = TRUE)
confint(f, method = "quadratic")

#>                     2.5%      97.5%
#> parent_0      71.8242430 93.1600766
#> k_parent_sink  0.2109541  0.4440528
#> sigma          1.9778868  7.3681380

# \dontrun{
  confint(f, method = "profile")

#> Profiling the likelihood

#>                     2.5%      97.5%
#> parent_0      73.0641834 92.1392181
#> k_parent_sink  0.2170293  0.4235348
#> sigma          3.1307772  8.0628314

# }