Aerobic soil degradation data on dimethenamid and dimethenamid-P from the EU assessment in 2018

The datasets were extracted from the active substance evaluation dossier published by EFSA. Kinetic evaluations shown for these datasets are intended to illustrate and advance kinetic modelling. The fact that these data and some results are shown here does not imply a license to use them in the context of pesticide registrations, as the use of the data may be constrained by data protection regulations.

dimethenamid_2018

Format

An mkindsg object grouping seven datasets with some meta information

Source

Rapporteur Member State Germany, Co-Rapporteur Member State Bulgaria (2018) Renewal Assessment Report Dimethenamid-P Volume 3 - B.8 Environmental fate and behaviour Rev. 2 - November 2017 https://open.efsa.europa.eu/study-inventory/EFSA-Q-2014-00716

Details

The R code used to create this data object is installed with this package in the 'dataset_generation' directory. In the code, page numbers are given for specific pieces of information in the comments.

Examples

print(dimethenamid_2018)
#> <mkindsg> holding 7 mkinds objects
#> Title $title:  Aerobic soil degradation data on dimethenamid-P from the EU assessment in 2018 
#> Occurrence of observed compounds $observed_n:
#> DMTAP   M23   M27   M31  DMTA 
#>     3     7     7     7     4 
#> Time normalisation factors $f_time_norm:
#> [1] 1.0000000 0.9706477 1.2284784 1.2284784 0.6233856 0.7678922 0.6733938
#> Meta information $meta:
#>                      study  usda_soil_type study_moisture_ref_type rel_moisture
#> Calke        Unsworth 2014      Sandy loam                     pF2         1.00
#> Borstel  Staudenmaier 2009            Sand                     pF1         0.50
#> Elliot 1        Wendt 1997       Clay loam                   pF2.5         0.75
#> Elliot 2        Wendt 1997       Clay loam                   pF2.5         0.75
#> Flaach          König 1996 Sandy clay loam                     pF1         0.40
#> BBA 2.2         König 1995      Loamy sand                     pF1         0.40
#> BBA 2.3         König 1995      Sandy loam                     pF1         0.40
#>          study_ref_moisture temperature
#> Calke                    NA          20
#> Borstel               23.00          20
#> Elliot 1              33.37          23
#> Elliot 2              33.37          23
#> Flaach                   NA          20
#> BBA 2.2                  NA          20
#> BBA 2.3                  NA          20
dmta_ds <- lapply(1:7, function(i) {
  ds_i <- dimethenamid_2018$ds[[i]]$data
  ds_i[ds_i$name == "DMTAP", "name"] <-  "DMTA"
  ds_i$time <- ds_i$time * dimethenamid_2018$f_time_norm[i]
  ds_i
})
names(dmta_ds) <- sapply(dimethenamid_2018$ds, function(ds) ds$title)
dmta_ds[["Elliot"]] <- rbind(dmta_ds[["Elliot 1"]], dmta_ds[["Elliot 2"]])
dmta_ds[["Elliot 1"]] <- NULL
dmta_ds[["Elliot 2"]] <- NULL
# \dontrun{
dfop_sfo3_plus <- mkinmod(
  DMTA = mkinsub("DFOP", c("M23", "M27", "M31")),
  M23 = mkinsub("SFO"),
  M27 = mkinsub("SFO"),
  M31 = mkinsub("SFO", "M27", sink = FALSE),
  quiet = TRUE
)
f_dmta_mkin_tc <- mmkin(
  list("DFOP-SFO3+" = dfop_sfo3_plus),
  dmta_ds, quiet = TRUE, error_model = "tc")
nlmixr_model(f_dmta_mkin_tc)
#> Error in nlmixr_model(f_dmta_mkin_tc): could not find function "nlmixr_model"
# The focei fit takes about four minutes on my system
system.time(
  f_dmta_nlmixr_focei <- nlmixr(f_dmta_mkin_tc, est = "focei",
    control = nlmixr::foceiControl(print = 500))
)
#> Error in nlmixr(f_dmta_mkin_tc, est = "focei", control = nlmixr::foceiControl(print = 500)): could not find function "nlmixr"
#> Timing stopped at: 0 0 0
summary(f_dmta_nlmixr_focei)
#> Error in summary(f_dmta_nlmixr_focei): object 'f_dmta_nlmixr_focei' not found
plot(f_dmta_nlmixr_focei)
#> Error in plot(f_dmta_nlmixr_focei): object 'f_dmta_nlmixr_focei' not found
# Using saemix takes about 18 minutes
system.time(
  f_dmta_saemix <- saem(f_dmta_mkin_tc, test_log_parms = TRUE)
)
#> DINTDY-  T (=R1) illegal      
#> In above message, R1 = 115.507
#>  
#>       T not in interval TCUR - HU (= R1) to TCUR (=R2)      
#> In above message, R1 = 112.133, R2 = 113.577
#>  
#> DLSODA-  At T (=R1), too much accuracy requested  
#>       for precision of machine..  See TOLSF (=R2) 
#> In above message, R1 = 55.3899, R2 = nan
#>  
#> Error in out[available, var]: (subscript) logical subscript too long
#> Timing stopped at: 12.76 3.069 11.79
#> Timing stopped at: 13.77 4.719 12.37

# nlmixr with est = "saem" is pretty fast with default iteration numbers, most
# of the time (about 2.5 minutes) is spent for calculating the log likelihood at the end
# The likelihood calculated for the nlmixr fit is much lower than that found by saemix
# Also, the trace plot and the plot of the individual predictions is not
# convincing for the parent. It seems we are fitting an overparameterised
# model, so the result we get strongly depends on starting parameters and control settings.
system.time(
  f_dmta_nlmixr_saem <- nlmixr(f_dmta_mkin_tc, est = "saem",
    control = nlmixr::saemControl(print = 500, logLik = TRUE, nmc = 9))
)
#> Error in nlmixr(f_dmta_mkin_tc, est = "saem", control = nlmixr::saemControl(print = 500,     logLik = TRUE, nmc = 9)): could not find function "nlmixr"
#> Timing stopped at: 0 0 0.001
traceplot(f_dmta_nlmixr_saem$nm)
#> Error in traceplot(f_dmta_nlmixr_saem$nm): could not find function "traceplot"
summary(f_dmta_nlmixr_saem)
#> Error in summary(f_dmta_nlmixr_saem): object 'f_dmta_nlmixr_saem' not found
plot(f_dmta_nlmixr_saem)
#> Error in plot(f_dmta_nlmixr_saem): object 'f_dmta_nlmixr_saem' not found
# }