Confidence intervals for parameters in saem.mmkin objects

Confidence intervals for parameters in saem.mmkin objects

# S3 method for saem.mmkin
intervals(object, level = 0.95, backtransform = TRUE, ...)

Arguments

object

The fitted saem.mmkin object

level

The confidence level. Must be the default of 0.95 as this is what is available in the saemix object

backtransform

In case the model was fitted with mkin transformations, should we backtransform the parameters where a one to one correlation between transformed and backtransformed parameters exists?

...

For compatibility with the generic method

Value

An object with 'intervals.saem.mmkin' and 'intervals.lme' in the class attribute