Function to calculate maximum time weighted average concentrations from kinetic models fitted with mkinfit

This function calculates maximum moving window time weighted average concentrations (TWAs) for kinetic models fitted with mkinfit. Currently, only calculations for the parent are implemented for the SFO, FOMC, DFOP and HS models, using the analytical formulas given in the PEC soil section of the FOCUS guidance.

Usage

max_twa_parent(fit, windows)

max_twa_sfo(M0 = 1, k, t)

max_twa_fomc(M0 = 1, alpha, beta, t)

max_twa_dfop(M0 = 1, k1, k2, g, t)

max_twa_hs(M0 = 1, k1, k2, tb, t)

Arguments

fit

An object of class mkinfit.

windows

The width of the time windows for which the TWAs should be calculated.

M0

The initial concentration for which the maximum time weighted average over the decline curve should be calculated. The default is to use a value of 1, which means that a relative maximum time weighted average factor (f_twa) is calculated.

k

The rate constant in the case of SFO kinetics.

t

The width of the time window.

alpha

Parameter of the FOMC model.

beta

Parameter of the FOMC model.

k1

The first rate constant of the DFOP or the HS kinetics.

k2

The second rate constant of the DFOP or the HS kinetics.

g

Parameter of the DFOP model.

tb

Parameter of the HS model.

Value

For max_twa_parent, a numeric vector, named using the windows argument. For the other functions, a numeric vector of length one (also known as 'a number').

References

FOCUS (2006) “Guidance Document on Estimating Persistence and Degradation Kinetics from Environmental Fate Studies on Pesticides in EU Registration” Report of the FOCUS Work Group on Degradation Kinetics, EC Document Reference Sanco/10058/2005 version 2.0, 434 pp, http://esdac.jrc.ec.europa.eu/projects/degradation-kinetics

Author

Johannes Ranke

Examples

  fit <- mkinfit("FOMC", FOCUS_2006_C, quiet = TRUE)
  max_twa_parent(fit, c(7, 21))
#>        7       21 
#> 34.71343 18.22124