Convert a dataframe with observations over time into long format

This function simply takes a dataframe with one independent variable and several dependent variable and converts it into the long form as required by mkinfit.

mkin_wide_to_long(wide_data, time = "t")

Arguments

wide_data

The dataframe must contain one variable with the time values specified by the time argument and usually more than one column of observed values.

time

The name of the time variable.

Value

Dataframe in long format as needed for mkinfit.

Author

Johannes Ranke

Examples

wide <- data.frame(t = c(1,2,3), x = c(1,4,7), y = c(3,4,5))
mkin_wide_to_long(wide)
#>   name time value
#> 1    x    1     1
#> 2    x    2     4
#> 3    x    3     7
#> 4    y    1     3
#> 5    y    2     4
#> 6    y    3     5