Create an nlme model for an mmkin row object

This functions sets up a nonlinear mixed effects model for an mmkin row object. An mmkin row object is essentially a list of mkinfit objects that have been obtained by fitting the same model to a list of datasets.

# S3 method for mmkin
nlme(
  model,
  data = sys.frame(sys.parent()),
  fixed,
  random = fixed,
  groups,
  start,
  correlation = NULL,
  weights = NULL,
  subset,
  method = c("ML", "REML"),
  na.action = na.fail,
  naPattern,
  control = list(),
  verbose = FALSE
)

Arguments

model
An mmkin row object.

data
Ignored, data are taken from the mmkin model

fixed
Ignored, all degradation parameters fitted in the mmkin model are used as fixed parameters

random
If not specified, all fixed effects are complemented with uncorrelated random effects

groups
See the documentation of nlme

start
If not specified, mean values of the fitted degradation parameters taken from the mmkin object are used

correlation
See the documentation of nlme

weights
passed to nlme

subset
passed to nlme

method
passed to nlme

na.action
passed to nlme

naPattern
passed to nlme

control
passed to nlme

verbose
passed to nlme

Value

Upon success, a fitted nlme.mmkin object, which is an nlme object with additional elements

See also

Examples

ds <- lapply(experimental_data_for_UBA_2019[6:10],
 function(x) subset(x$data[c("name", "time", "value")], name == "parent"))
f <- mmkin("SFO", ds, quiet = TRUE, cores = 1)
library(nlme)
f_nlme <- nlme(f)
nlme(f, random = parent_0 ~ 1)

#> Nonlinear mixed-effects model fit by maximum likelihood
#>   Model: value ~ deg_func(name, time, parent_0, log_k_parent_sink) 
#>   Data: structure(list(ds = structure(c(1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,  1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 2L, 2L, 2L, 2L,  2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 3L, 3L,  3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 4L, 4L,  4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L,  4L, 4L, 5L, 5L, 5L, 5L, 5L, 5L, 5L, 5L, 5L, 5L, 5L, 5L, 5L, 5L,  5L, 5L), .Label = c("1", "2", "3", "4", "5"), class = c("ordered",  "factor")), name = c("parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent"), time = c(0, 0, 3, 3, 6, 6, 10, 10, 20, 20,  34, 34, 55, 55, 90, 90, 112, 112, 132, 132, 0, 0, 3, 3, 7, 7,  14, 14, 30, 30, 60, 60, 90, 90, 120, 120, 180, 180, 0, 0, 1,  1, 3, 3, 8, 8, 14, 14, 27, 27, 48, 48, 70, 70, 0, 0, 1, 1, 3,  3, 8, 8, 14, 14, 27, 27, 48, 48, 70, 70, 91, 91, 120, 120, 0,  0, 8, 8, 14, 14, 21, 21, 41, 41, 63, 63, 91, 91, 120, 120), value = c(97.2,  96.4, 71.1, 69.2, 58.1, 56.6, 44.4, 43.4, 33.3, 29.2, 17.6, 18,  10.5, 9.3, 4.5, 4.7, 3, 3.4, 2.3, 2.7, 93.6, 92.3, 87, 82.2,  74, 73.9, 64.2, 69.5, 54, 54.6, 41.1, 38.4, 32.5, 35.5, 28.1,  29, 26.5, 27.6, 91.9, 90.8, 64.9, 66.2, 43.5, 44.1, 18.3, 18.1,  10.2, 10.8, 4.9, 3.3, 1.6, 1.5, 1.1, 0.9, 99.8, 98.3, 77.1, 77.2,  59, 58.1, 27.4, 29.2, 19.1, 29.6, 10.1, 18.2, 4.5, 9.1, 2.3,  2.9, 2, 1.8, 2, 2.2, 96.1, 94.3, 73.9, 73.9, 69.4, 73.1, 65.6,  65.3, 55.9, 54.4, 47, 49.3, 44.7, 46.7, 42.1, 41.3)), row.names = c(NA,  -90L), class = c("nfnGroupedData", "nfGroupedData", "groupedData",  "data.frame"), formula = value ~ time | ds, FUN = function (x)  max(x, na.rm = TRUE), order.groups = FALSE) 
#>   Log-likelihood: -394.4901
#>   Fixed: list(parent_0 ~ 1, log_k_parent_sink ~ 1) 
#>          parent_0 log_k_parent_sink 
#>         73.985522         -3.869079 
#> 
#> Random effects:
#>  Formula: parent_0 ~ 1 | ds
#>         parent_0 Residual
#> StdDev:  18.6134 18.22029
#> 
#> Number of Observations: 90
#> Number of Groups: 5 

f_nlme <- nlme(f, start = c(parent_0 = 100, log_k_parent_sink = 0.1))
update(f_nlme, random = parent_0 ~ 1)

#> Nonlinear mixed-effects model fit by maximum likelihood
#>   Model: value ~ deg_func(name, time, parent_0, log_k_parent_sink) 
#>   Data: structure(list(ds = structure(c(1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L,  1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 2L, 2L, 2L, 2L,  2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 3L, 3L,  3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 4L, 4L,  4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L,  4L, 4L, 5L, 5L, 5L, 5L, 5L, 5L, 5L, 5L, 5L, 5L, 5L, 5L, 5L, 5L,  5L, 5L), .Label = c("1", "2", "3", "4", "5"), class = c("ordered",  "factor")), name = c("parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent", "parent", "parent", "parent", "parent", "parent",  "parent", "parent"), time = c(0, 0, 3, 3, 6, 6, 10, 10, 20, 20,  34, 34, 55, 55, 90, 90, 112, 112, 132, 132, 0, 0, 3, 3, 7, 7,  14, 14, 30, 30, 60, 60, 90, 90, 120, 120, 180, 180, 0, 0, 1,  1, 3, 3, 8, 8, 14, 14, 27, 27, 48, 48, 70, 70, 0, 0, 1, 1, 3,  3, 8, 8, 14, 14, 27, 27, 48, 48, 70, 70, 91, 91, 120, 120, 0,  0, 8, 8, 14, 14, 21, 21, 41, 41, 63, 63, 91, 91, 120, 120), value = c(97.2,  96.4, 71.1, 69.2, 58.1, 56.6, 44.4, 43.4, 33.3, 29.2, 17.6, 18,  10.5, 9.3, 4.5, 4.7, 3, 3.4, 2.3, 2.7, 93.6, 92.3, 87, 82.2,  74, 73.9, 64.2, 69.5, 54, 54.6, 41.1, 38.4, 32.5, 35.5, 28.1,  29, 26.5, 27.6, 91.9, 90.8, 64.9, 66.2, 43.5, 44.1, 18.3, 18.1,  10.2, 10.8, 4.9, 3.3, 1.6, 1.5, 1.1, 0.9, 99.8, 98.3, 77.1, 77.2,  59, 58.1, 27.4, 29.2, 19.1, 29.6, 10.1, 18.2, 4.5, 9.1, 2.3,  2.9, 2, 1.8, 2, 2.2, 96.1, 94.3, 73.9, 73.9, 69.4, 73.1, 65.6,  65.3, 55.9, 54.4, 47, 49.3, 44.7, 46.7, 42.1, 41.3)), row.names = c(NA,  -90L), class = c("nfnGroupedData", "nfGroupedData", "groupedData",  "data.frame"), formula = value ~ time | ds, FUN = function (x)  max(x, na.rm = TRUE), order.groups = FALSE) 
#>   Log-likelihood: -404.3729
#>   Fixed: list(parent_0 ~ 1, log_k_parent_sink ~ 1) 
#>          parent_0 log_k_parent_sink 
#>         75.933480         -3.555983 
#> 
#> Random effects:
#>  Formula: parent_0 ~ 1 | ds
#>            parent_0 Residual
#> StdDev: 0.002416792 21.63027
#> 
#> Number of Observations: 90
#> Number of Groups: 5