Print mkinmod objects

Print mkinmod objects in a way that the user finds his way to get to its components.

# S3 method for mkinmod
print(x, ...)

Arguments

x
An mkinmod object.

...
Not used.

Examples

  m_synth_SFO_lin <- mkinmod(parent = list(type = "SFO", to = "M1"),
                             M1 = list(type = "SFO", to = "M2"),
                             M2 = list(type = "SFO"), use_of_ff = "max")

#> Successfully compiled differential equation model from auto-generated C code.

  print(m_synth_SFO_lin)

#> <mkinmod> model generated with
#> Use of formation fractions $use_of_ff: max 
#> Specification $spec:
#> $parent
#> $type: SFO; $to: M1; $sink: TRUE
#> $M1
#> $type: SFO; $to: M2; $sink: TRUE
#> $M2
#> $type: SFO; $sink: TRUE
#> Coefficient matrix $coefmat available
#> Compiled model $cf available
#> Differential equations:
#> d_parent/dt = - k_parent * parent
#> d_M1/dt = + f_parent_to_M1 * k_parent * parent - k_M1 * M1
#> d_M2/dt = + f_M1_to_M2 * k_M1 * M1 - k_M2 * M2