Metabolism data set used for checking the software quality of KinGUI

This dataset was used for a comparison of KinGUI and ModelMaker to check the software quality of KinGUI in the original publication (Schäfer et al., 2007). The results from the fitting are also included.

data(schaefer07_complex_case)

Format

The data set is a data frame with 8 observations on the following 6 variables.

time

a numeric vector

parent

a numeric vector

A1

a numeric vector

B1

a numeric vector

C1

a numeric vector

A2

a numeric vector

The results are a data frame with 14 results for different parameter values

References

Schäfer D, Mikolasch B, Rainbird P and Harvey B (2007). KinGUI: a new kinetic software tool for evaluations according to FOCUS degradation kinetics. In: Del Re AAM, Capri E, Fragoulis G and Trevisan M (Eds.). Proceedings of the XIII Symposium Pesticide Chemistry, Piacenza, 2007, p. 916-923.

Examples

data <- mkin_wide_to_long(schaefer07_complex_case, time = "time")
model <- mkinmod(
  parent = list(type = "SFO", to = c("A1", "B1", "C1"), sink = FALSE),
  A1 = list(type = "SFO", to = "A2"),
  B1 = list(type = "SFO"),
  C1 = list(type = "SFO"),
  A2 = list(type = "SFO"), use_of_ff = "max")

#> Successfully compiled differential equation model from auto-generated C code.

mkinfit(model, data)

#> Model cost at call  1 :  2511.655 
#> Model cost at call  2 :  2511.655 
#> Model cost at call  11 :  1436.639 
#> Model cost at call  12 :  1436.638 
#> Model cost at call  13 :  1436.566 
#> Model cost at call  21 :  643.6583 
#> Model cost at call  22 :  643.6583 
#> Model cost at call  23 :  643.6582 
#> Model cost at call  29 :  643.6576 
#> Model cost at call  31 :  454.0244 
#> Model cost at call  32 :  454.0241 
#> Model cost at call  34 :  454.0229 
#> Model cost at call  43 :  378.1144 
#> Model cost at call  45 :  378.1143 
#> Model cost at call  53 :  357.245 
#> Model cost at call  55 :  357.2449 
#> Model cost at call  56 :  357.2447 
#> Model cost at call  63 :  354.3415 
#> Model cost at call  64 :  354.3415 
#> Model cost at call  65 :  354.3413 
#> Model cost at call  73 :  332.49 
#> Model cost at call  74 :  332.49 
#> Model cost at call  81 :  332.4899 
#> Model cost at call  83 :  315.2962 
#> Model cost at call  84 :  306.3085 
#> Model cost at call  86 :  306.3084 
#> Model cost at call  87 :  306.3084 
#> Model cost at call  92 :  306.3083 
#> Model cost at call  94 :  290.6377 
#> Model cost at call  96 :  290.6375 
#> Model cost at call  98 :  290.6375 
#> Model cost at call  101 :  290.6371 
#> Model cost at call  105 :  269.09 
#> Model cost at call  107 :  269.0899 
#> Model cost at call  115 :  259.7551 
#> Model cost at call  120 :  259.7549 
#> Model cost at call  123 :  259.7547 
#> Model cost at call  126 :  253.7973 
#> Model cost at call  128 :  253.7972 
#> Model cost at call  137 :  251.7358 
#> Model cost at call  139 :  251.7358 
#> Model cost at call  147 :  250.7394 
#> Model cost at call  149 :  250.7393 
#> Model cost at call  157 :  249.1148 
#> Model cost at call  159 :  249.1148 
#> Model cost at call  167 :  246.8768 
#> Model cost at call  169 :  246.8768 
#> Model cost at call  177 :  244.9758 
#> Model cost at call  179 :  244.9758 
#> Model cost at call  187 :  243.2914 
#> Model cost at call  189 :  243.2914 
#> Model cost at call  190 :  243.2914 
#> Model cost at call  194 :  243.2914 
#> Model cost at call  199 :  242.9202 
#> Model cost at call  201 :  242.9202 
#> Model cost at call  202 :  242.9202 
#> Model cost at call  209 :  242.7695 
#> Model cost at call  211 :  242.7695 
#> Model cost at call  216 :  242.7695 
#> Model cost at call  219 :  242.5771 
#> Model cost at call  221 :  242.5771 
#> Model cost at call  229 :  242.4402 
#> Model cost at call  231 :  242.4402 
#> Model cost at call  239 :  242.1878 
#> Model cost at call  241 :  242.1878 
#> Model cost at call  249 :  242.0553 
#> Model cost at call  251 :  242.0553 
#> Model cost at call  256 :  242.0553 
#> Model cost at call  259 :  241.8761 
#> Model cost at call  260 :  241.7412 
#> Model cost at call  261 :  241.6954 
#> Model cost at call  264 :  241.6954 
#> Model cost at call  275 :  241.5982 
#> Model cost at call  277 :  241.5982 
#> Model cost at call  285 :  241.5459 
#> Model cost at call  287 :  241.5459 
#> Model cost at call  295 :  241.4837 
#> Model cost at call  297 :  241.4837 
#> Model cost at call  305 :  241.3882 
#> Model cost at call  306 :  241.3161 
#> Model cost at call  307 :  241.2315 
#> Model cost at call  309 :  241.2315 
#> Model cost at call  314 :  241.2315 
#> Model cost at call  317 :  240.9738 
#> Model cost at call  322 :  240.9738 
#> Model cost at call  327 :  240.8244 
#> Model cost at call  329 :  240.8244 
#> Model cost at call  337 :  240.7005 
#> Model cost at call  339 :  240.7005 
#> Model cost at call  342 :  240.7005 
#> Model cost at call  347 :  240.629 
#> Model cost at call  350 :  240.629 
#> Model cost at call  357 :  240.6193 
#> Model cost at call  358 :  240.6193 
#> Model cost at call  364 :  240.6193 
#> Model cost at call  367 :  240.6193 
#> Model cost at call  369 :  240.5873 
#> Model cost at call  374 :  240.5873 
#> Model cost at call  380 :  240.578 
#> Model cost at call  382 :  240.578 
#> Model cost at call  390 :  240.5723 
#> Model cost at call  393 :  240.5723 
#> Model cost at call  403 :  240.569 
#> Model cost at call  404 :  240.569 
#> Model cost at call  413 :  240.569 
#> Model cost at call  415 :  240.5688 
#> Model cost at call  416 :  240.5688 
#> Model cost at call  417 :  240.5688 
#> Model cost at call  431 :  240.5686 
#> Model cost at call  432 :  240.5686 
#> Model cost at call  434 :  240.5686 
#> Model cost at call  443 :  240.5686 
#> Model cost at call  444 :  240.5686 
#> Model cost at call  447 :  240.5686 
#> Model cost at call  449 :  240.5686 
#> Model cost at call  450 :  240.5686 
#> Model cost at call  466 :  240.5686 
#> Model cost at call  470 :  240.5686 
#> Model cost at call  485 :  240.5686 
#> Model cost at call  509 :  240.5686 
#> Optimisation by method Port successfully terminated.

#> $par
#>       parent_0   log_k_parent       log_k_A1       log_k_B1       log_k_C1 
#>     91.9181598     -3.0020485     -4.2735924     -3.9846764     -2.7852180 
#>       log_k_A2 f_parent_ilr_1 f_parent_ilr_2     f_A1_ilr_1 
#>     -3.7166415      0.4718588     -0.3589948     -0.1477244 
#> 
#> $ssr
#> [1] 240.5686
#> 
#> $convergence
#> [1] 0
#> 
#> $iterations
#> [1] 43
#> 
#> $evaluations
#> function gradient 
#>       62      441 
#> 
#> $counts
#> [1] "relative convergence (4)"
#> 
#> $hessian
#>                   parent_0 log_k_parent      log_k_A1      log_k_B1
#> parent_0         7.3650812   -92.141920 -1.001134e+01 -2.432415e+00
#> log_k_parent   -92.1419204  6632.673492 -4.316240e+01 -1.320833e+01
#> log_k_A1       -10.0113364   -43.162398  6.071628e+02  0.000000e+00
#> log_k_B1        -2.4324147   -13.208329  0.000000e+00  1.572303e+02
#> log_k_C1        -4.7153201  -118.288037 -5.878291e-05 -3.073041e-06
#> log_k_A2        -0.4360727    -5.304259 -1.977980e+01  0.000000e+00
#> f_parent_ilr_1  10.5460899   271.145438 -5.299954e+02  1.874235e+02
#> f_parent_ilr_2  11.6409409   222.570696 -4.773816e+02 -1.159875e+02
#> f_A1_ilr_1       0.5572072    10.374810  2.850173e+01  0.000000e+00
#>                     log_k_C1      log_k_A2 f_parent_ilr_1 f_parent_ilr_2
#> parent_0       -4.715320e+00 -4.360727e-01       10.54609       11.64094
#> log_k_parent   -1.182880e+02 -5.304259e+00      271.14544      222.57070
#> log_k_A1       -5.878291e-05 -1.977980e+01     -529.99537     -477.38164
#> log_k_B1       -3.073041e-06  0.000000e+00      187.42348     -115.98754
#> log_k_C1        3.372749e+02 -2.395674e-06       56.85184      305.98862
#> log_k_A2       -2.395674e-06  2.749192e+01      -23.08549      -20.79373
#> f_parent_ilr_1  5.685184e+01 -2.308549e+01     1256.24941      632.09769
#> f_parent_ilr_2  3.059886e+02 -2.079373e+01      632.09769     1250.65147
#> f_A1_ilr_1      3.158891e-06 -3.129286e+01       29.49830       26.56991
#>                   f_A1_ilr_1
#> parent_0        5.572072e-01
#> log_k_parent    1.037481e+01
#> log_k_A1        2.850173e+01
#> log_k_B1        0.000000e+00
#> log_k_C1        3.158891e-06
#> log_k_A2       -3.129286e+01
#> f_parent_ilr_1  2.949830e+01
#> f_parent_ilr_2  2.656991e+01
#> f_A1_ilr_1      3.998554e+01
#> 
#> $residuals
#>     parent     parent     parent     parent     parent     parent     parent 
#> -1.2818402 -1.9372115 -0.5105519  3.8165318 -2.3531716  4.8043342 -2.2775432 
#>     parent         A1         A1         A1         A1         A1         A1 
#> -5.3608524  4.1967522  2.9032987 -1.3124875 -0.6021093  2.5092324 -1.8861396 
#>         B1         B1         B1         B1         B1         C1         C1 
#>  4.3801768  5.5002481 -5.7917184  1.3852658  0.5313301  1.2796458  1.7105311 
#>         C1         C1         C1         C1         C1         A2         A2 
#>  3.7116712 -0.1182953  0.5228429 -0.8570298 -3.5476556 -0.5447276 -1.3652404 
#>         A2         A2         A2         A2         A2 
#> -0.3330261 -0.5802059  0.1285850  0.2119280 -0.1381990 
#> 
#> $ms
#> [1] 7.289956
#> 
#> $var_ms
#>     parent         A1         B1         C1         A2 
#> 10.3459333  6.3301336 17.0367907  4.5639474  0.3841002 
#> 
#> $var_ms_unscaled
#>     parent         A1         B1         C1         A2 
#> 10.3459333  6.3301336 17.0367907  4.5639474  0.3841002 
#> 
#> $var_ms_unweighted
#>     parent         A1         B1         C1         A2 
#> 10.3459333  6.3301336 17.0367907  4.5639474  0.3841002 
#> 
#> $rank
#> [1] 9
#> 
#> $df.residual
#> [1] 24
#> 
#> $solution_type
#> [1] "deSolve"
#> 
#> $transform_rates
#> [1] TRUE
#> 
#> $transform_fractions
#> [1] TRUE
#> 
#> $method.modFit
#> [1] "Port"
#> 
#> $maxit.modFit
#> [1] "auto"
#> 
#> $calls
#> [1] 523
#> 
#> $time
#>    user  system elapsed 
#>   5.004   0.000   5.004 
#> 
#> $mkinmod
#> <mkinmod> model generated with
#> Use of formation fractions $use_of_ff: max 
#> Specification $spec:
#> $parent
#> $type: SFO; $to: A1, B1, C1; $sink: FALSE
#> $A1
#> $type: SFO; $to: A2; $sink: TRUE
#> $B1
#> $type: SFO; $sink: TRUE
#> $C1
#> $type: SFO; $sink: TRUE
#> $A2
#> $type: SFO; $sink: TRUE
#> Coefficient matrix $coefmat available
#> Compiled model $cf available
#> 
#> $observed
#>      name time value
#> 1  parent    0 93.20
#> 2  parent    1 89.40
#> 3  parent    3 79.70
#> 4  parent    7 61.10
#> 5  parent   14 48.20
#> 6  parent   30 15.90
#> 7  parent   62  6.50
#> 8  parent  100  6.00
#> 9      A1    0    NA
#> 10     A1    1    NA
#> 11     A1    3  0.55
#> 12     A1    7  6.87
#> 13     A1   14 17.08
#> 14     A1   30 21.68
#> 15     A1   62 15.77
#> 16     A1  100 13.63
#> 17     B1    0    NA
#> 18     B1    1    NA
#> 19     B1    3    NA
#> 20     B1    7  0.55
#> 21     B1   14  2.31
#> 22     B1   30 15.76
#> 23     B1   62  6.36
#> 24     B1  100  3.74
#> 25     C1    0    NA
#> 26     C1    1  0.55
#> 27     C1    3  3.20
#> 28     C1    7  5.46
#> 29     C1   14 12.55
#> 30     C1   30 10.45
#> 31     C1   62  4.74
#> 32     C1  100  4.33
#> 33     A2    0    NA
#> 34     A2    1  0.55
#> 35     A2    3  1.41
#> 36     A2    7  0.55
#> 37     A2   14  1.29
#> 38     A2   30  1.95
#> 39     A2   62  3.54
#> 40     A2  100  3.86
#> 
#> $obs_vars
#> [1] "parent" "A1"     "B1"     "C1"     "A2"    
#> 
#> $predicted
#>       name       time        value
#> 1   parent   0.000000 91.918159794
#> 2   parent   1.000000 87.462788491
#> 3   parent   1.010101 87.418904506
#> 4   parent   2.020202 83.139880984
#> 5   parent   3.000000 79.189448055
#> 6   parent   3.030303 79.070309209
#> 7   parent   4.040404 75.199936833
#> 8   parent   5.050505 71.519013349
#> 9   parent   6.060606 68.018265517
#> 10  parent   7.000000 64.916531757
#> 11  parent   7.070707 64.688874011
#> 12  parent   8.080808 61.522451197
#> 13  parent   9.090909 58.511020005
#> 14  parent  10.101010 55.646993828
#> 15  parent  11.111111 52.923157412
#> 16  parent  12.121212 50.332648680
#> 17  parent  13.131313 47.868941444
#> 18  parent  14.000000 45.846828365
#> 19  parent  14.141414 45.525828960
#> 20  parent  15.151515 43.297408299
#> 21  parent  16.161616 41.178065468
#> 22  parent  17.171717 39.162461272
#> 23  parent  18.181818 37.245517861
#> 24  parent  19.191919 35.422405939
#> 25  parent  20.202020 33.688532595
#> 26  parent  21.212121 32.039529737
#> 27  parent  22.222222 30.471243081
#> 28  parent  23.232323 28.979721692
#> 29  parent  24.242424 27.561208025
#> 30  parent  25.252525 26.212128463
#> 31  parent  26.262626 24.929084310
#> 32  parent  27.272727 23.708843233
#> 33  parent  28.282828 22.548331117
#> 34  parent  29.292929 21.444624318
#> 35  parent  30.000000 20.704334210
#> 36  parent  30.303030 20.394942302
#> 37  parent  31.313131 19.396640638
#> 38  parent  32.323232 18.447204335
#> 39  parent  33.333333 17.544241506
#> 40  parent  34.343434 16.685477346
#> 41  parent  35.353535 15.868748397
#> 42  parent  36.363636 15.091997098
#> 43  parent  37.373737 14.353266603
#> 44  parent  38.383838 13.650695852
#> 45  parent  39.393939 12.982514879
#> 46  parent  40.404040 12.347040357
#> 47  parent  41.414141 11.742671354
#> 48  parent  42.424242 11.167885303
#> 49  parent  43.434343 10.621234162
#> 50  parent  44.444444 10.101340770
#> 51  parent  45.454545  9.606895375
#> 52  parent  46.464646  9.136652336
#> 53  parent  47.474747  8.689426985
#> 54  parent  48.484848  8.264092640
#> 55  parent  49.494949  7.859577770
#> 56  parent  50.505051  7.474863293
#> 57  parent  51.515152  7.108980009
#> 58  parent  52.525253  6.761006160
#> 59  parent  53.535354  6.430065106
#> 60  parent  54.545455  6.115323117
#> 61  parent  55.555556  5.815987274
#> 62  parent  56.565657  5.531303470
#> 63  parent  57.575758  5.260554508
#> 64  parent  58.585859  5.003058299
#> 65  parent  59.595960  4.758166141
#> 66  parent  60.606061  4.525261085
#> 67  parent  61.616162  4.303756381
#> 68  parent  62.000000  4.222456793
#> 69  parent  62.626263  4.093093997
#> 70  parent  63.636364  3.892743220
#> 71  parent  64.646465  3.702199310
#> 72  parent  65.656566  3.520982238
#> 73  parent  66.666667  3.348635468
#> 74  parent  67.676768  3.184724813
#> 75  parent  68.686869  3.028837337
#> 76  parent  69.696970  2.880580317
#> 77  parent  70.707071  2.739580256
#> 78  parent  71.717172  2.605481934
#> 79  parent  72.727273  2.477947523
#> 80  parent  73.737374  2.356655730
#> 81  parent  74.747475  2.241300986
#> 82  parent  75.757576  2.131592683
#> 83  parent  76.767677  2.027254437
#> 84  parent  77.777778  1.928023390
#> 85  parent  78.787879  1.833649553
#> 86  parent  79.797980  1.743895173
#> 87  parent  80.808081  1.658534134
#> 88  parent  81.818182  1.577351390
#> 89  parent  82.828283  1.500142419
#> 90  parent  83.838384  1.426712710
#> 91  parent  84.848485  1.356877275
#> 92  parent  85.858586  1.290460179
#> 93  parent  86.868687  1.227294099
#> 94  parent  87.878788  1.167219904
#> 95  parent  88.888889  1.110086250
#> 96  parent  89.898990  1.055749203
#> 97  parent  90.909091  1.004071872
#> 98  parent  91.919192  0.954924068
#> 99  parent  92.929293  0.908181975
#> 100 parent  93.939394  0.863727837
#> 101 parent  94.949495  0.821449662
#> 102 parent  95.959596  0.781240940
#> 103 parent  96.969697  0.743000375
#> 104 parent  97.979798  0.706631627
#> 105 parent  98.989899  0.672043075
#> 106 parent 100.000000  0.639147580
#> 107     A1   0.000000  0.000000000
#> 108     A1   1.000000  1.685461006
#> 109     A1   1.010101  1.701940789
#> 110     A1   2.020202  3.296791533
#> 111     A1   3.000000  4.746752202
#> 112     A1   3.030303  4.790126465
#> 113     A1   4.040404  6.187242320
#> 114     A1   5.050505  7.493171988
#> 115     A1   6.060606  8.712697491
#> 116     A1   7.000000  9.773298725
#> 117     A1   7.070707  9.850362326
#> 118     A1   8.080808 10.910483202
#> 119     A1   9.090909 11.897161206
#> 120     A1  10.101010 12.814292412
#> 121     A1  11.111111 13.665577981
#> 122     A1  12.121212 14.454533757
#> 123     A1  13.131313 15.184499397
#> 124     A1  14.000000 15.767512526
#> 125     A1  14.141414 15.858647054
#> 126     A1  15.151515 16.479989628
#> 127     A1  16.161616 17.051388624
#> 128     A1  17.171717 17.575561608
#> 129     A1  18.181818 18.055089316
#> 130     A1  19.191919 18.492422399
#> 131     A1  20.202020 18.889887843
#> 132     A1  21.212121 19.249695079
#> 133     A1  22.222222 19.573941783
#> 134     A1  23.232323 19.864619397
#> 135     A1  24.242424 20.123618383
#> 136     A1  25.252525 20.352733211
#> 137     A1  26.262626 20.553667106
#> 138     A1  27.272727 20.728036563
#> 139     A1  28.282828 20.877375640
#> 140     A1  29.292929 21.003140039
#> 141     A1  30.000000 21.077890710
#> 142     A1  30.303030 21.106710984
#> 143     A1  31.313131 21.189398917
#> 144     A1  32.323232 21.252447002
#> 145     A1  33.333333 21.297034466
#> 146     A1  34.343434 21.324279770
#> 147     A1  35.353535 21.335243623
#> 148     A1  36.363636 21.330931858
#> 149     A1  37.373737 21.312298151
#> 150     A1  38.383838 21.280246621
#> 151     A1  39.393939 21.235634295
#> 152     A1  40.404040 21.179273450
#> 153     A1  41.414141 21.111933845
#> 154     A1  42.424242 21.034344838
#> 155     A1  43.434343 20.947197407
#> 156     A1  44.444444 20.851146060
#> 157     A1  45.454545 20.746810660
#> 158     A1  46.464646 20.634778158
#> 159     A1  47.474747 20.515604239
#> 160     A1  48.484848 20.389814887
#> 161     A1  49.494949 20.257907875
#> 162     A1  50.505051 20.120354180
#> 163     A1  51.515152 19.977599327
#> 164     A1  52.525253 19.830064674
#> 165     A1  53.535354 19.678148618
#> 166     A1  54.545455 19.522227762
#> 167     A1  55.555556 19.362658007
#> 168     A1  56.565657 19.199775600
#> 169     A1  57.575758 19.033898126
#> 170     A1  58.585859 18.865325451
#> 171     A1  59.595960 18.694340625
#> 172     A1  60.606061 18.521210729
#> 173     A1  61.616162 18.346187688
#> 174     A1  62.000000 18.279232408
#> 175     A1  62.626263 18.169509043
#> 176     A1  63.636364 17.991398686
#> 177     A1  64.646465 17.812067549
#> 178     A1  65.656566 17.631714275
#> 179     A1  66.666667 17.450525840
#> 180     A1  67.676768 17.268678156
#> 181     A1  68.686869 17.086336636
#> 182     A1  69.696970 16.903656738
#> 183     A1  70.707071 16.720784474
#> 184     A1  71.717172 16.537856901
#> 185     A1  72.727273 16.355002582
#> 186     A1  73.737374 16.172342031
#> 187     A1  74.747475 15.989988127
#> 188     A1  75.757576 15.808046514
#> 189     A1  76.767677 15.626615980
#> 190     A1  77.777778 15.445788814
#> 191     A1  78.787879 15.265651148
#> 192     A1  79.797980 15.086283284
#> 193     A1  80.808081 14.907759996
#> 194     A1  81.818182 14.730150830
#> 195     A1  82.828283 14.553520376
#> 196     A1  83.838384 14.377928535
#> 197     A1  84.848485 14.203430771
#> 198     A1  85.858586 14.030078345
#> 199     A1  86.868687 13.857918547
#> 200     A1  87.878788 13.686994907
#> 201     A1  88.888889 13.517347398
#> 202     A1  89.898990 13.349012635
#> 203     A1  90.909091 13.182024056
#> 204     A1  91.919192 13.016412097
#> 205     A1  92.929293 12.852204356
#> 206     A1  93.939394 12.689425755
#> 207     A1  94.949495 12.528098688
#> 208     A1  95.959596 12.368243159
#> 209     A1  96.969697 12.209876925
#> 210     A1  97.979798 12.053015616
#> 211     A1  98.989899 11.897672861
#> 212     A1 100.000000 11.743860400
#> 213     B1   0.000000  0.000000000
#> 214     B1   1.000000  0.862762059
#> 215     B1   1.010101  0.871177048
#> 216     B1   2.020202  1.683497848
#> 217     B1   3.000000  2.418226457
#> 218     B1   3.030303  2.440145075
#> 219     B1   4.040404  3.144139999
#> 220     B1   5.050505  3.798350490
#> 221     B1   6.060606  4.405498633
#> 222     B1   7.000000  4.930176837
#> 223     B1   7.070707  4.968167964
#> 224     B1   8.080808  5.488810347
#> 225     B1   9.090909  5.969752521
#> 226     B1  10.101010  6.413202316
#> 227     B1  11.111111  6.821254568
#> 228     B1  12.121212  7.195896744
#> 229     B1  13.131313  7.539014282
#> 230     B1  14.000000  7.810248132
#> 231     B1  14.141414  7.852395679
#> 232     B1  15.151515  8.137737320
#> 233     B1  16.161616  8.396648072
#> 234     B1  17.171717  8.630653651
#> 235     B1  18.181818  8.841200774
#> 236     B1  19.191919  9.029661109
#> 237     B1  20.202020  9.197335022
#> 238     B1  21.212121  9.345455150
#> 239     B1  22.222222  9.475189788
#> 240     B1  23.232323  9.587646116
#> 241     B1  24.242424  9.683873262
#> 242     B1  25.252525  9.764865214
#> 243     B1  26.262626  9.831563593
#> 244     B1  27.272727  9.884860284
#> 245     B1  28.282828  9.925599936
#> 246     B1  29.292929  9.954582344
#> 247     B1  30.000000  9.968281596
#> 248     B1  30.303030  9.972564708
#> 249     B1  31.313131  9.980263783
#> 250     B1  32.323232  9.978357919
#> 251     B1  33.333333  9.967489009
#> 252     B1  34.343434  9.948264327
#> 253     B1  35.353535  9.921258285
#> 254     B1  36.363636  9.887014102
#> 255     B1  37.373737  9.846045383
#> 256     B1  38.383838  9.798837632
#> 257     B1  39.393939  9.745849674
#> 258     B1  40.404040  9.687515023
#> 259     B1  41.414141  9.624243169
#> 260     B1  42.424242  9.556420809
#> 261     B1  43.434343  9.484413012
#> 262     B1  44.444444  9.408564328
#> 263     B1  45.454545  9.329199843
#> 264     B1  46.464646  9.246626179
#> 265     B1  47.474747  9.161132446
#> 266     B1  48.484848  9.072991146
#> 267     B1  49.494949  8.982459028
#> 268     B1  50.505051  8.889777910
#> 269     B1  51.515152  8.795175451
#> 270     B1  52.525253  8.698865886
#> 271     B1  53.535354  8.601050726
#> 272     B1  54.545455  8.501919425
#> 273     B1  55.555556  8.401650008
#> 274     B1  56.565657  8.300409672
#> 275     B1  57.575758  8.198355355
#> 276     B1  58.585859  8.095634277
#> 277     B1  59.595960  7.992384454
#> 278     B1  60.606061  7.888735183
#> 279     B1  61.616162  7.784807509
#> 280     B1  62.000000  7.745265792
#> 281     B1  62.626263  7.680714664
#> 282     B1  63.636364  7.576562482
#> 283     B1  64.646465  7.472449799
#> 284     B1  65.656566  7.368468826
#> 285     B1  66.666667  7.264705509
#> 286     B1  67.676768  7.161239868
#> 287     B1  68.686869  7.058146319
#> 288     B1  69.696970  6.955493978
#> 289     B1  70.707071  6.853346953
#> 290     B1  71.717172  6.751764620
#> 291     B1  72.727273  6.650801882
#> 292     B1  73.737374  6.550509419
#> 293     B1  74.747475  6.450933922
#> 294     B1  75.757576  6.352118318
#> 295     B1  76.767677  6.254101979
#> 296     B1  77.777778  6.156920928
#> 297     B1  78.787879  6.060608023
#> 298     B1  79.797980  5.965193142
#> 299     B1  80.808081  5.870703355
#> 300     B1  81.818182  5.777163083
#> 301     B1  82.828283  5.684594257
#> 302     B1  83.838384  5.593016458
#> 303     B1  84.848485  5.502447062
#> 304     B1  85.858586  5.412901366
#> 305     B1  86.868687  5.324392718
#> 306     B1  87.878788  5.236932630
#> 307     B1  88.888889  5.150530889
#> 308     B1  89.898990  5.065195670
#> 309     B1  90.909091  4.980933628
#> 310     B1  91.919192  4.897749999
#> 311     B1  92.929293  4.815648688
#> 312     B1  93.939394  4.734632351
#> 313     B1  94.949495  4.654702481
#> 314     B1  95.959596  4.575859481
#> 315     B1  96.969697  4.498102737
#> 316     B1  97.979798  4.421430686
#> 317     B1  98.989899  4.345840882
#> 318     B1 100.000000  4.271330056
#> 319     C1   0.000000  0.000000000
#> 320     C1   1.000000  1.829645786
#> 321     C1   1.010101  1.847087763
#> 322     C1   2.020202  3.492133303
#> 323     C1   3.000000  4.910531064
#> 324     C1   3.030303  4.951772742
#> 325     C1   4.040404  6.241420142
#> 326     C1   5.050505  7.375351980
#> 327     C1   6.060606  8.366785999
#> 328     C1   7.000000  9.171671206
#> 329     C1   7.070707  9.227954769
#> 330     C1   8.080808  9.970174354
#> 331     C1   9.090909 10.603908370
#> 332     C1  10.101010 11.138827767
#> 333     C1  11.111111 11.583866567
#> 334     C1  12.121212 11.947273869
#> 335     C1  13.131313 12.236662337
#> 336     C1  14.000000 12.431704739
#> 337     C1  14.141414 12.459053419
#> 338     C1  15.151515 12.620919488
#> 339     C1  16.161616 12.728223141
#> 340     C1  17.171717 12.786453805
#> 341     C1  18.181818 12.800661859
#> 342     C1  19.191919 12.775490422
#> 343     C1  20.202020 12.715204956
#> 344     C1  21.212121 12.623720845
#> 345     C1  22.222222 12.504629065
#> 346     C1  23.232323 12.361220091
#> 347     C1  24.242424 12.196506142
#> 348     C1  25.252525 12.013241882
#> 349     C1  26.262626 11.813943686
#> 350     C1  27.272727 11.600907551
#> 351     C1  28.282828 11.376225763
#> 352     C1  29.292929 11.141802382
#> 353     C1  30.000000 10.972842888
#> 354     C1  30.303030 10.899367648
#> 355     C1  31.313131 10.650491354
#> 356     C1  32.323232 10.396595286
#> 357     C1  33.333333 10.138964763
#> 358     C1  34.343434  9.878759358
#> 359     C1  35.353535  9.617022857
#> 360     C1  36.363636  9.354692485
#> 361     C1  37.373737  9.092607481
#> 362     C1  38.383838  8.831517041
#> 363     C1  39.393939  8.572087685
#> 364     C1  40.404040  8.314910084
#> 365     C1  41.414141  8.060505385
#> 366     C1  42.424242  7.809331068
#> 367     C1  43.434343  7.561786371
#> 368     C1  44.444444  7.318217302
#> 369     C1  45.454545  7.078921287
#> 370     C1  46.464646  6.844151456
#> 371     C1  47.474747  6.614120611
#> 372     C1  48.484848  6.389004885
#> 373     C1  49.494949  6.168947129
#> 374     C1  50.505051  5.954060026
#> 375     C1  51.515152  5.744428970
#> 376     C1  52.525253  5.540114721
#> 377     C1  53.535354  5.341155842
#> 378     C1  54.545455  5.147570951
#> 379     C1  55.555556  4.959360784
#> 380     C1  56.565657  4.776510102
#> 381     C1  57.575758  4.598989433
#> 382     C1  58.585859  4.426756673
#> 383     C1  59.595960  4.259758556
#> 384     C1  60.606061  4.097932000
#> 385     C1  61.616162  3.941205338
#> 386     C1  62.000000  3.882970158
#> 387     C1  62.626263  3.789499444
#> 388     C1  63.636364  3.642728760
#> 389     C1  64.646465  3.500802233
#> 390     C1  65.656566  3.363624171
#> 391     C1  66.666667  3.231095021
#> 392     C1  67.676768  3.103112069
#> 393     C1  68.686869  2.979570086
#> 394     C1  69.696970  2.860361903
#> 395     C1  70.707071  2.745378939
#> 396     C1  71.717172  2.634511667
#> 397     C1  72.727273  2.527650041
#> 398     C1  73.737374  2.424683880
#> 399     C1  74.747475  2.325503203
#> 400     C1  75.757576  2.229998536
#> 401     C1  76.767677  2.138061182
#> 402     C1  77.777778  2.049583458
#> 403     C1  78.787879  1.964458908
#> 404     C1  79.797980  1.882582485
#> 405     C1  80.808081  1.803850715
#> 406     C1  81.818182  1.728161832
#> 407     C1  82.828283  1.655415900
#> 408     C1  83.838384  1.585514911
#> 409     C1  84.848485  1.518362874
#> 410     C1  85.858586  1.453865880
#> 411     C1  86.868687  1.391932162
#> 412     C1  87.878788  1.332472134
#> 413     C1  88.888889  1.275398429
#> 414     C1  89.898990  1.220625918
#> 415     C1  90.909091  1.168071723
#> 416     C1  91.919192  1.117655227
#> 417     C1  92.929293  1.069298066
#> 418     C1  93.939394  1.022924125
#> 419     C1  94.949495  0.978459525
#> 420     C1  95.959596  0.935832597
#> 421     C1  96.969697  0.894973866
#> 422     C1  97.979798  0.855816021
#> 423     C1  98.989899  0.818293881
#> 424     C1 100.000000  0.782344364
#> 425     A2   0.000000  0.000000000
#> 426     A2   1.000000  0.005272357
#> 427     A2   1.010101  0.005377817
#> 428     A2   2.020202  0.020885524
#> 429     A2   3.000000  0.044759575
#> 430     A2   3.030303  0.045628064
#> 431     A2   4.040404  0.078765936
#> 432     A2   5.050505  0.119512155
#> 433     A2   6.060606  0.167129381
#> 434     A2   7.000000  0.216973934
#> 435     A2   7.070707  0.220927189
#> 436     A2   8.080808  0.280259484
#> 437     A2   9.090909  0.344522046
#> 438     A2  10.101010  0.413150206
#> 439     A2  11.111111  0.485616641
#> 440     A2  12.121212  0.561429288
#> 441     A2  13.131313  0.640129357
#> 442     A2  14.000000  0.709794102
#> 443     A2  14.141414  0.721289460
#> 444     A2  15.151515  0.804511827
#> 445     A2  16.161616  0.889426625
#> 446     A2  17.171717  0.975690359
#> 447     A2  18.181818  1.062984358
#> 448     A2  19.191919  1.151013342
#> 449     A2  20.202020  1.239504068
#> 450     A2  21.212121  1.328204041
#> 451     A2  22.222222  1.416880297
#> 452     A2  23.232323  1.505318253
#> 453     A2  24.242424  1.593320615
#> 454     A2  25.252525  1.680706344
#> 455     A2  26.262626  1.767309680
#> 456     A2  27.272727  1.852979219
#> 457     A2  28.282828  1.937577034
#> 458     A2  29.292929  2.020977853
#> 459     A2  30.000000  2.078585030
#> 460     A2  30.303030  2.103068270
#> 461     A2  31.313131  2.183746011
#> 462     A2  32.323232  2.262919231
#> 463     A2  33.333333  2.340505852
#> 464     A2  34.343434  2.416432940
#> 465     A2  35.353535  2.490636111
#> 466     A2  36.363636  2.563058979
#> 467     A2  37.373737  2.633652622
#> 468     A2  38.383838  2.702375089
#> 469     A2  39.393939  2.769190926
#> 470     A2  40.404040  2.834070737
#> 471     A2  41.414141  2.896990764
#> 472     A2  42.424242  2.957932489
#> 473     A2  43.434343  3.016882265
#> 474     A2  44.444444  3.073830964
#> 475     A2  45.454545  3.128773647
#> 476     A2  46.464646  3.181709250
#> 477     A2  47.474747  3.232640290
#> 478     A2  48.484848  3.281572591
#> 479     A2  49.494949  3.328515022
#> 480     A2  50.505051  3.373479253
#> 481     A2  51.515152  3.416479521
#> 482     A2  52.525253  3.457532417
#> 483     A2  53.535354  3.496656681
#> 484     A2  54.545455  3.533873012
#> 485     A2  55.555556  3.569203883
#> 486     A2  56.565657  3.602673379
#> 487     A2  57.575758  3.634307034
#> 488     A2  58.585859  3.664131686
#> 489     A2  59.595960  3.692175334
#> 490     A2  60.606061  3.718467012
#> 491     A2  61.616162  3.743036663
#> 492     A2  62.000000  3.751927986
#> 493     A2  62.626263  3.765915028
#> 494     A2  63.636364  3.787133539
#> 495     A2  64.646465  3.806724217
#> 496     A2  65.656566  3.824719582
#> 497     A2  66.666667  3.841152565
#> 498     A2  67.676768  3.856056426
#> 499     A2  68.686869  3.869464684
#> 500     A2  69.696970  3.881411040
#> 501     A2  70.707071  3.891929316
#> 502     A2  71.717172  3.901053396
#> 503     A2  72.727273  3.908817168
#> 504     A2  73.737374  3.915254472
#> 505     A2  74.747475  3.920399054
#> 506     A2  75.757576  3.924284521
#> 507     A2  76.767677  3.926944303
#> 508     A2  77.777778  3.928411610
#> 509     A2  78.787879  3.928719404
#> 510     A2  79.797980  3.927900364
#> 511     A2  80.808081  3.925986861
#> 512     A2  81.818182  3.923010926
#> 513     A2  82.828283  3.919004234
#> 514     A2  83.838384  3.913998077
#> 515     A2  84.848485  3.908023347
#> 516     A2  85.858586  3.901110518
#> 517     A2  86.868687  3.893289633
#> 518     A2  87.878788  3.884590288
#> 519     A2  88.888889  3.875041619
#> 520     A2  89.898990  3.864672297
#> 521     A2  90.909091  3.853510511
#> 522     A2  91.919192  3.841583970
#> 523     A2  92.929293  3.828919886
#> 524     A2  93.939394  3.815544978
#> 525     A2  94.949495  3.801485462
#> 526     A2  95.959596  3.786767051
#> 527     A2  96.969697  3.771414951
#> 528     A2  97.979798  3.755453860
#> 529     A2  98.989899  3.738907968
#> 530     A2 100.000000  3.721800959
#> 
#> $cost
#> function (P) 
#> {
#>     assign("calls", calls + 1, inherits = TRUE)
#>     if (trace_parms) 
#>         cat(P, "\n")
#>     if (length(state.ini.optim) > 0) {
#>         odeini <- c(P[1:length(state.ini.optim)], state.ini.fixed)
#>         names(odeini) <- c(state.ini.optim.boxnames, state.ini.fixed.boxnames)
#>     }
#>     else {
#>         odeini <- state.ini.fixed
#>         names(odeini) <- state.ini.fixed.boxnames
#>     }
#>     odeparms <- c(P[(length(state.ini.optim) + 1):length(P)], 
#>         transparms.fixed)
#>     parms <- backtransform_odeparms(odeparms, mkinmod, transform_rates = transform_rates, 
#>         transform_fractions = transform_fractions)
#>     out <- mkinpredict(mkinmod, parms, odeini, outtimes, solution_type = solution_type, 
#>         use_compiled = use_compiled, method.ode = method.ode, 
#>         atol = atol, rtol = rtol, ...)
#>     assign("out_predicted", out, inherits = TRUE)
#>     mC <- modCost(out, observed, y = "value", err = err, weight = weight, 
#>         scaleVar = scaleVar)
#>     if (mC$model < cost.old) {
#>         if (!quiet) 
#>             cat("Model cost at call ", calls, ": ", mC$model, 
#>                 "\n")
#>         if (plot) {
#>             outtimes_plot = seq(min(observed$time), max(observed$time), 
#>                 length.out = 100)
#>             out_plot <- mkinpredict(mkinmod, parms, odeini, outtimes_plot, 
#>                 solution_type = solution_type, use_compiled = use_compiled, 
#>                 method.ode = method.ode, atol = atol, rtol = rtol, 
#>                 ...)
#>             plot(0, type = "n", xlim = range(observed$time), 
#>                 ylim = c(0, max(observed$value, na.rm = TRUE)), 
#>                 xlab = "Time", ylab = "Observed")
#>             col_obs <- pch_obs <- 1:length(obs_vars)
#>             lty_obs <- rep(1, length(obs_vars))
#>             names(col_obs) <- names(pch_obs) <- names(lty_obs) <- obs_vars
#>             for (obs_var in obs_vars) {
#>                 points(subset(observed, name == obs_var, c(time, 
#>                   value)), pch = pch_obs[obs_var], col = col_obs[obs_var])
#>             }
#>             matlines(out_plot$time, out_plot[-1], col = col_obs, 
#>                 lty = lty_obs)
#>             legend("topright", inset = c(0.05, 0.05), legend = obs_vars, 
#>                 col = col_obs, pch = pch_obs, lty = 1:length(pch_obs))
#>         }
#>         assign("cost.old", mC$model, inherits = TRUE)
#>     }
#>     return(mC)
#> }
#> <environment: 0x36a83b0>
#> 
#> $cost_notrans
#> function (P) 
#> {
#>     if (length(state.ini.optim) > 0) {
#>         odeini <- c(P[1:length(state.ini.optim)], state.ini.fixed)
#>         names(odeini) <- c(state.ini.optim.boxnames, state.ini.fixed.boxnames)
#>     }
#>     else {
#>         odeini <- state.ini.fixed
#>         names(odeini) <- state.ini.fixed.boxnames
#>     }
#>     odeparms <- c(P[(length(state.ini.optim) + 1):length(P)], 
#>         parms.fixed)
#>     out <- mkinpredict(mkinmod, odeparms, odeini, outtimes, solution_type = solution_type, 
#>         use_compiled = use_compiled, method.ode = method.ode, 
#>         atol = atol, rtol = rtol, ...)
#>     mC <- modCost(out, observed, y = "value", err = err, weight = weight, 
#>         scaleVar = scaleVar)
#>     return(mC)
#> }
#> <environment: 0x36a83b0>
#> 
#> $hessian_notrans
#>                    parent_0     k_parent          k_A1          k_B1
#> parent_0           7.365081   -1854.5113 -7.186039e+02 -1.307858e+02
#> k_parent       -1854.511330 2686790.7676 -6.235542e+04 -1.429363e+04
#> k_A1            -718.603865  -62355.4211  3.128242e+06  0.000000e+00
#> k_B1            -130.785796  -14293.6348  0.000000e+00  4.545506e+05
#> k_C1             -76.404274  -38575.9391  1.190516e-02 -9.422820e-04
#> k_A2             -17.933942   -4390.5079 -5.838973e+04  0.000000e+00
#> f_parent_to_A1    75.150866   43257.2599 -1.733841e+05  0.000000e+00
#> f_parent_to_B1    29.265575   17940.1132  0.000000e+00 -6.150198e+04
#> f_parent_to_C1    20.661354   19692.5582 -6.146186e-05 -1.990817e-03
#> f_A1_to_A2         1.593279     597.0744  5.849840e+03  0.000000e+00
#>                         k_C1          k_A2 f_parent_to_A1 f_parent_to_B1
#> parent_0       -7.640427e+01 -1.793394e+01   7.515087e+01   2.926558e+01
#> k_parent       -3.857594e+04 -4.390508e+03   4.325726e+04   1.794011e+04
#> k_A1            1.190516e-02 -5.838973e+04  -1.733841e+05   0.000000e+00
#> k_B1           -9.422820e-04  0.000000e+00   0.000000e+00  -6.150198e+04
#> k_C1            8.855106e+04  4.105787e-04  -1.354551e-03   5.852620e-04
#> k_A2            4.105787e-04  4.649850e+04  -4.327086e+03   0.000000e+00
#> f_parent_to_A1 -1.354551e-03 -4.327086e+03   1.813234e+04   0.000000e+00
#> f_parent_to_B1  5.852620e-04  0.000000e+00   0.000000e+00   1.376213e+04
#> f_parent_to_C1 -1.658031e+04  2.903794e-04   1.946385e-03   1.325258e-03
#> f_A1_to_A2     -4.367402e-05 -3.679910e+03   3.844249e+02   0.000000e+00
#>                f_parent_to_C1    f_A1_to_A2
#> parent_0         2.066135e+01  1.593279e+00
#> k_parent         1.969256e+04  5.970744e+02
#> k_A1            -6.146186e-05  5.849840e+03
#> k_B1            -1.990817e-03  0.000000e+00
#> k_C1            -1.658031e+04 -4.367402e-05
#> k_A2             2.903794e-04 -3.679910e+03
#> f_parent_to_A1   1.946385e-03  3.844249e+02
#> f_parent_to_B1   1.325258e-03  0.000000e+00
#> f_parent_to_C1   4.483759e+03 -3.796730e-05
#> f_A1_to_A2      -3.796730e-05  3.269288e+02
#> 
#> $start
#>                     value   type
#> parent_0       93.2000000  state
#> k_parent        0.1000000 deparm
#> k_A1            0.1001000 deparm
#> k_B1            0.1002000 deparm
#> k_C1            0.1003000 deparm
#> k_A2            0.1004000 deparm
#> f_parent_to_A1  0.3333333 deparm
#> f_parent_to_B1  0.3333333 deparm
#> f_parent_to_C1  0.3333333 deparm
#> f_A1_to_A2      0.5000000 deparm
#> 
#> $start_transformed
#>                    value lower upper
#> parent_0       93.200000  -Inf   Inf
#> log_k_parent   -2.302585  -Inf   Inf
#> log_k_A1       -2.301586  -Inf   Inf
#> log_k_B1       -2.300587  -Inf   Inf
#> log_k_C1       -2.299590  -Inf   Inf
#> log_k_A2       -2.298593  -Inf   Inf
#> f_parent_ilr_1  0.000000  -Inf   Inf
#> f_parent_ilr_2  0.000000  -Inf   Inf
#> f_A1_ilr_1      0.000000  -Inf   Inf
#> 
#> $fixed
#>      value  type
#> A1_0     0 state
#> B1_0     0 state
#> C1_0     0 state
#> A2_0     0 state
#> 
#> $data
#>    time variable observed    predicted   residual
#> 1     0   parent    93.20 91.918159794  1.2818402
#> 2     1   parent    89.40 87.462788491  1.9372115
#> 3     3   parent    79.70 79.189448055  0.5105519
#> 4     7   parent    61.10 64.916531757 -3.8165318
#> 5    14   parent    48.20 45.846828365  2.3531716
#> 6    30   parent    15.90 20.704334210 -4.8043342
#> 7    62   parent     6.50  4.222456793  2.2775432
#> 8   100   parent     6.00  0.639147580  5.3608524
#> 9     0       A1       NA  0.000000000         NA
#> 10    1       A1       NA  1.685461006         NA
#> 11    3       A1     0.55  4.746752202 -4.1967522
#> 12    7       A1     6.87  9.773298725 -2.9032987
#> 13   14       A1    17.08 15.767512526  1.3124875
#> 14   30       A1    21.68 21.077890710  0.6021093
#> 15   62       A1    15.77 18.279232408 -2.5092324
#> 16  100       A1    13.63 11.743860400  1.8861396
#> 17    0       B1       NA  0.000000000         NA
#> 18    1       B1       NA  0.862762059         NA
#> 19    3       B1       NA  2.418226457         NA
#> 20    7       B1     0.55  4.930176837 -4.3801768
#> 21   14       B1     2.31  7.810248132 -5.5002481
#> 22   30       B1    15.76  9.968281596  5.7917184
#> 23   62       B1     6.36  7.745265792 -1.3852658
#> 24  100       B1     3.74  4.271330056 -0.5313301
#> 25    0       C1       NA  0.000000000         NA
#> 26    1       C1     0.55  1.829645786 -1.2796458
#> 27    3       C1     3.20  4.910531064 -1.7105311
#> 28    7       C1     5.46  9.171671206 -3.7116712
#> 29   14       C1    12.55 12.431704739  0.1182953
#> 30   30       C1    10.45 10.972842888 -0.5228429
#> 31   62       C1     4.74  3.882970158  0.8570298
#> 32  100       C1     4.33  0.782344364  3.5476556
#> 33    0       A2       NA  0.000000000         NA
#> 34    1       A2     0.55  0.005272357  0.5447276
#> 35    3       A2     1.41  0.044759575  1.3652404
#> 36    7       A2     0.55  0.216973934  0.3330261
#> 37   14       A2     1.29  0.709794102  0.5802059
#> 38   30       A2     1.95  2.078585030 -0.1285850
#> 39   62       A2     3.54  3.751927986 -0.2119280
#> 40  100       A2     3.86  3.721800959  0.1381990
#> 
#> $atol
#> [1] 1e-08
#> 
#> $rtol
#> [1] 1e-10
#> 
#> $weight.ini
#> [1] "none"
#> 
#> $reweight.tol
#> [1] 1e-08
#> 
#> $reweight.max.iter
#> [1] 10
#> 
#> $bparms.optim
#>       parent_0       k_parent           k_A1           k_B1           k_C1 
#>    91.91815979     0.04968519     0.01393165     0.01859846     0.06171564 
#>           k_A2 f_parent_to_A1 f_parent_to_B1 f_parent_to_C1     f_A1_to_A2 
#>     0.02431549     0.38096192     0.19546676     0.42357132     0.44796066 
#> 
#> $bparms.fixed
#> A1_0 B1_0 C1_0 A2_0 
#>    0    0    0    0 
#> 
#> $bparms.ode
#>       k_parent f_parent_to_A1 f_parent_to_B1 f_parent_to_C1           k_A1 
#>     0.04968519     0.38096192     0.19546676     0.42357132     0.01393165 
#>     f_A1_to_A2           k_B1           k_C1           k_A2 
#>     0.44796066     0.01859846     0.06171564     0.02431549 
#> 
#> $bparms.state
#>   parent       A1       B1       C1       A2 
#> 91.91816  0.00000  0.00000  0.00000  0.00000 
#> 
#> $date
#> [1] "Fri Nov 18 15:20:45 2016"
#> 
#> attr(,"class")
#> [1] "mkinfit" "modFit"