Summary method for class "mmkin"

Shows status information on the mkinfit objects contained in the object and gives an overview of ill-defined parameters calculated by illparms.

# S3 method for mmkin
summary(object, conf.level = 0.95, ...)

# S3 method for summary.mmkin
print(x, digits = max(3, getOption("digits") - 3), ...)

Arguments

object

an object of class mmkin

conf.level

confidence level for testing parameters

...

optional arguments passed to methods like print.

x

an object of class summary.mmkin.

digits

number of digits to use for printing

Examples

fits <- mmkin(
  c("SFO", "FOMC"),
  list("FOCUS A" = FOCUS_2006_A,
       "FOCUS C" = FOCUS_2006_C),
  quiet = TRUE, cores = 1)
  summary(fits)
#> Error model: Constant variance 
#> Fitted in 0.835 s
#> 
#> Status:
#>       dataset
#> model  FOCUS A FOCUS C
#>   SFO  OK      OK     
#>   FOMC OK      OK     
#> 
#> OK: No warnings
#> 
#> Ill-defined parameters:
#>       dataset
#> model  FOCUS A                      FOCUS C
#>   SFO                                      
#>   FOMC parent_0, alpha, beta, sigma