Summary method for class "saem.mmkin"

Lists model equations, initial parameter values, optimised parameters for fixed effects (population), random effects (deviations from the population mean) and residual error model, as well as the resulting endpoints such as formation fractions and DT50 values. Optionally (default is FALSE), the data are listed in full.

# S3 method for saem.mmkin
summary(object, data = FALSE, verbose = FALSE, distimes = TRUE, ...)

# S3 method for summary.saem.mmkin
print(x, digits = max(3, getOption("digits") - 3), verbose = x$verbose, ...)

Arguments

object

an object of class saem.mmkin

data

logical, indicating whether the full data should be included in the summary.

verbose

Should the summary be verbose?

distimes

logical, indicating whether DT50 and DT90 values should be included.

...

optional arguments passed to methods like print.

x

an object of class summary.saem.mmkin

digits

Number of digits to use for printing

Value

The summary function returns a list based on the saemix::SaemixObjectobtained in the fit, with at least the following additional components

saemixversion, mkinversion, Rversion

The saemix, mkin and R versions used

date.fit, date.summary

The dates where the fit and the summary were produced

diffs

The differential equations used in the degradation model

use_of_ff

Was maximum or minimum use made of formation fractions

data

The data

confint_trans

Transformed parameters as used in the optimisation, with confidence intervals

confint_back

Backtransformed parameters, with confidence intervals if available

confint_errmod

Error model parameters with confidence intervals

ff

The estimated formation fractions derived from the fitted model.

distimes

The DT50 and DT90 values for each observed variable.

SFORB

If applicable, eigenvalues of SFORB components of the model.

The print method is called for its side effect, i.e. printing the summary.

Author

Johannes Ranke for the mkin specific parts saemix authors for the parts inherited from saemix.

Examples

# Generate five datasets following DFOP-SFO kinetics
sampling_times = c(0, 1, 3, 7, 14, 28, 60, 90, 120)
dfop_sfo <- mkinmod(parent = mkinsub("DFOP", "m1"),
 m1 = mkinsub("SFO"), quiet = TRUE)
set.seed(1234)
k1_in <- rlnorm(5, log(0.1), 0.3)
k2_in <- rlnorm(5, log(0.02), 0.3)
g_in <- plogis(rnorm(5, qlogis(0.5), 0.3))
f_parent_to_m1_in <- plogis(rnorm(5, qlogis(0.3), 0.3))
k_m1_in <- rlnorm(5, log(0.02), 0.3)

pred_dfop_sfo <- function(k1, k2, g, f_parent_to_m1, k_m1) {
  mkinpredict(dfop_sfo,
    c(k1 = k1, k2 = k2, g = g, f_parent_to_m1 = f_parent_to_m1, k_m1 = k_m1),
    c(parent = 100, m1 = 0),
    sampling_times)
}

ds_mean_dfop_sfo <- lapply(1:5, function(i) {
  mkinpredict(dfop_sfo,
    c(k1 = k1_in[i], k2 = k2_in[i], g = g_in[i],
      f_parent_to_m1 = f_parent_to_m1_in[i], k_m1 = k_m1_in[i]),
    c(parent = 100, m1 = 0),
    sampling_times)
})
names(ds_mean_dfop_sfo) <- paste("ds", 1:5)

ds_syn_dfop_sfo <- lapply(ds_mean_dfop_sfo, function(ds) {
  add_err(ds,
    sdfunc = function(value) sqrt(1^2 + value^2 * 0.07^2),
    n = 1)[[1]]
})

# \dontrun{
# Evaluate using mmkin and saem
f_mmkin_dfop_sfo <- mmkin(list(dfop_sfo), ds_syn_dfop_sfo,
  quiet = TRUE, error_model = "tc", cores = 5)
f_saem_dfop_sfo <- saem(f_mmkin_dfop_sfo)
summary(f_saem_dfop_sfo, data = TRUE)
#> saemix version used for fitting:      3.0 
#> mkin version used for pre-fitting:  1.1.0 
#> R version used for fitting:         4.2.0 
#> Date of fit:     Wed May 18 20:42:18 2022 
#> Date of summary: Wed May 18 20:42:18 2022 
#> 
#> Equations:
#> d_parent/dt = - ((k1 * g * exp(-k1 * time) + k2 * (1 - g) * exp(-k2 *
#>            time)) / (g * exp(-k1 * time) + (1 - g) * exp(-k2 * time)))
#>            * parent
#> d_m1/dt = + f_parent_to_m1 * ((k1 * g * exp(-k1 * time) + k2 * (1 - g)
#>            * exp(-k2 * time)) / (g * exp(-k1 * time) + (1 - g) *
#>            exp(-k2 * time))) * parent - k_m1 * m1
#> 
#> Data:
#> 171 observations of 2 variable(s) grouped in 5 datasets
#> 
#> Model predictions using solution type analytical 
#> 
#> Fitted in 12.833 s
#> Using 300, 100 iterations and 10 chains
#> 
#> Variance model: Two-component variance function 
#> 
#> Mean of starting values for individual parameters:
#>        parent_0        log_k_m1 f_parent_qlogis          log_k1          log_k2 
#>       101.65645        -4.05368        -0.94311        -2.35943        -4.07006 
#>        g_qlogis 
#>        -0.01133 
#> 
#> Fixed degradation parameter values:
#> None
#> 
#> Results:
#> 
#> Likelihood computed by importance sampling
#>     AIC   BIC logLik
#>   829.3 823.8 -400.6
#> 
#> Optimised parameters:
#>                      est.  lower    upper
#> parent_0        101.26516 98.412 104.1188
#> log_k_m1         -4.07510 -4.168  -3.9819
#> f_parent_qlogis  -0.94320 -1.350  -0.5364
#> log_k1           -2.87697 -3.703  -2.0510
#> log_k2           -3.58133 -4.313  -2.8492
#> g_qlogis         -0.03174 -1.564   1.5008
#> 
#> Correlation: 
#>                 parnt_0 lg_k_m1 f_prnt_ log_k1  log_k2 
#> log_k_m1        -0.4167                                
#> f_parent_qlogis -0.2084  0.2383                        
#> log_k1           0.1324 -0.1242 -0.0500                
#> log_k2           0.0173  0.0001 -0.0067  0.1137        
#> g_qlogis        -0.0849  0.0691  0.0399 -0.3916 -0.3012
#> 
#> Random effects:
#>                       est.    lower   upper
#> SD.parent_0        0.16032 -34.0484 34.3690
#> SD.log_k_m1        0.01463  -0.3605  0.3898
#> SD.f_parent_qlogis 0.44391   0.1646  0.7232
#> SD.log_k1          0.84861   0.3044  1.3928
#> SD.log_k2          0.79185   0.2944  1.2893
#> SD.g_qlogis        0.34997  -3.0572  3.7572
#> 
#> Variance model:
#>        est.   lower  upper
#> a.1 0.87200 0.69118 1.0528
#> b.1 0.07963 0.06435 0.0949
#> 
#> Backtransformed parameters:
#>                     est.    lower     upper
#> parent_0       101.26516 98.41151 104.11881
#> k_m1             0.01699  0.01548   0.01865
#> f_parent_to_m1   0.28025  0.20588   0.36902
#> k1               0.05630  0.02465   0.12861
#> k2               0.02784  0.01339   0.05789
#> g                0.49207  0.17303   0.81770
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_m1   0.2803
#> parent_sink 0.7197
#> 
#> Estimated disappearance times:
#>         DT50   DT90 DT50back DT50_k1 DT50_k2
#> parent 17.28  63.65    19.16   12.31    24.9
#> m1     40.80 135.52       NA      NA      NA
#> 
#> Data:
#>    ds   name time observed predicted   residual    std standardized
#>  ds 1 parent    0     89.8 1.012e+02 -11.441861 8.1085    -1.411097
#>  ds 1 parent    0    104.1 1.012e+02   2.858139 8.1085     0.352487
#>  ds 1 parent    1     88.7 9.642e+01  -7.717294 7.7267    -0.998787
#>  ds 1 parent    1     95.5 9.642e+01  -0.917294 7.7267    -0.118718
#>  ds 1 parent    3     81.8 8.761e+01  -5.811045 7.0304    -0.826561
#>  ds 1 parent    3     94.5 8.761e+01   6.888955 7.0304     0.979883
#>  ds 1 parent    7     71.5 7.288e+01  -1.380912 5.8683    -0.235315
#>  ds 1 parent    7     70.3 7.288e+01  -2.580912 5.8683    -0.439802
#>  ds 1 parent   14     54.2 5.406e+01   0.139892 4.3920     0.031851
#>  ds 1 parent   14     49.6 5.406e+01  -4.460108 4.3920    -1.015504
#>  ds 1 parent   28     31.5 3.222e+01  -0.722286 2.7099    -0.266540
#>  ds 1 parent   28     28.8 3.222e+01  -3.422286 2.7099    -1.262902
#>  ds 1 parent   60     12.1 1.280e+01  -0.696320 1.3411    -0.519211
#>  ds 1 parent   60     13.6 1.280e+01   0.803680 1.3411     0.599263
#>  ds 1 parent   90      6.2 6.147e+00   0.052772 1.0000     0.052772
#>  ds 1 parent   90      8.3 6.147e+00   2.152772 1.0000     2.152793
#>  ds 1 parent  120      2.2 3.047e+00  -0.847292 0.9051    -0.936096
#>  ds 1 parent  120      2.4 3.047e+00  -0.647292 0.9051    -0.715134
#>  ds 1     m1    1      0.3 1.172e+00  -0.871912 0.8770    -0.994217
#>  ds 1     m1    1      0.2 1.172e+00  -0.971912 0.8770    -1.108244
#>  ds 1     m1    3      2.2 3.253e+00  -1.053335 0.9097    -1.157932
#>  ds 1     m1    3      3.0 3.253e+00  -0.253335 0.9097    -0.278492
#>  ds 1     m1    7      6.5 6.524e+00  -0.024120 1.0150    -0.023763
#>  ds 1     m1    7      5.0 6.524e+00  -1.524120 1.0150    -1.501571
#>  ds 1     m1   14     10.2 1.012e+01   0.076122 1.1875     0.064101
#>  ds 1     m1   14      9.5 1.012e+01  -0.623878 1.1875    -0.525359
#>  ds 1     m1   28     12.2 1.268e+01  -0.478789 1.3340    -0.358908
#>  ds 1     m1   28     13.4 1.268e+01   0.721211 1.3340     0.540631
#>  ds 1     m1   60     11.8 1.085e+01   0.954822 1.2272     0.778023
#>  ds 1     m1   60     13.2 1.085e+01   2.354822 1.2272     1.918794
#>  ds 1     m1   90      6.6 7.751e+00  -1.151363 1.0683    -1.077718
#>  ds 1     m1   90      9.3 7.751e+00   1.548637 1.0683     1.449581
#>  ds 1     m1  120      3.5 5.236e+00  -1.735804 0.9665    -1.795893
#>  ds 1     m1  120      5.4 5.236e+00   0.164196 0.9665     0.169880
#>  ds 2 parent    0    118.0 1.013e+02  16.726081 8.1110     2.062142
#>  ds 2 parent    0     99.8 1.013e+02  -1.473919 8.1110    -0.181718
#>  ds 2 parent    1     90.2 9.591e+01  -5.708953 7.6864    -0.742730
#>  ds 2 parent    1     94.6 9.591e+01  -1.308953 7.6864    -0.170294
#>  ds 2 parent    3     96.1 8.654e+01   9.564450 6.9454     1.377088
#>  ds 2 parent    3     78.4 8.654e+01  -8.135550 6.9454    -1.171355
#>  ds 2 parent    7     77.9 7.206e+01   5.840823 5.8037     1.006405
#>  ds 2 parent    7     77.7 7.206e+01   5.640823 5.8037     0.971944
#>  ds 2 parent   14     56.0 5.562e+01   0.376382 4.5141     0.083379
#>  ds 2 parent   14     54.7 5.562e+01  -0.923618 4.5141    -0.204607
#>  ds 2 parent   28     36.6 3.849e+01  -1.891356 3.1865    -0.593546
#>  ds 2 parent   28     36.8 3.849e+01  -1.691356 3.1865    -0.530782
#>  ds 2 parent   60     22.1 2.105e+01   1.052796 1.8892     0.557274
#>  ds 2 parent   60     24.7 2.105e+01   3.652796 1.8892     1.933527
#>  ds 2 parent   90     12.4 1.249e+01  -0.085726 1.3224    -0.064825
#>  ds 2 parent   90     10.8 1.249e+01  -1.685726 1.3224    -1.274728
#>  ds 2 parent  120      6.8 7.425e+00  -0.625179 1.0535    -0.593407
#>  ds 2 parent  120      7.9 7.425e+00   0.474821 1.0535     0.450690
#>  ds 2     m1    1      1.3 1.420e+00  -0.119671 0.8793    -0.136098
#>  ds 2     m1    3      3.7 3.831e+00  -0.130721 0.9238    -0.141501
#>  ds 2     m1    3      4.7 3.831e+00   0.869279 0.9238     0.940968
#>  ds 2     m1    7      8.1 7.307e+00   0.793062 1.0483     0.756531
#>  ds 2     m1    7      7.9 7.307e+00   0.593062 1.0483     0.565743
#>  ds 2     m1   14     10.1 1.060e+01  -0.500946 1.2136    -0.412765
#>  ds 2     m1   14     10.3 1.060e+01  -0.300946 1.2136    -0.247970
#>  ds 2     m1   28     10.7 1.236e+01  -1.663249 1.3151    -1.264727
#>  ds 2     m1   28     12.2 1.236e+01  -0.163249 1.3151    -0.124134
#>  ds 2     m1   60     10.7 1.063e+01   0.065269 1.2155     0.053697
#>  ds 2     m1   60     12.5 1.063e+01   1.865269 1.2155     1.534560
#>  ds 2     m1   90      9.1 8.134e+00   0.965918 1.0862     0.889243
#>  ds 2     m1   90      7.4 8.134e+00  -0.734082 1.0862    -0.675811
#>  ds 2     m1  120      6.1 5.918e+00   0.181998 0.9912     0.183617
#>  ds 2     m1  120      4.5 5.918e+00  -1.418002 0.9912    -1.430615
#>  ds 3 parent    0    106.2 1.013e+02   4.927395 8.1109     0.607501
#>  ds 3 parent    0    106.9 1.013e+02   5.627395 8.1109     0.693805
#>  ds 3 parent    1    107.4 9.362e+01  13.778324 7.5055     1.835758
#>  ds 3 parent    1     96.1 9.362e+01   2.478324 7.5055     0.330200
#>  ds 3 parent    3     79.4 8.111e+01  -1.714426 6.5174    -0.263054
#>  ds 3 parent    3     82.6 8.111e+01   1.485574 6.5174     0.227940
#>  ds 3 parent    7     63.9 6.406e+01  -0.160919 5.1749    -0.031096
#>  ds 3 parent    7     62.4 6.406e+01  -1.660919 5.1749    -0.320957
#>  ds 3 parent   14     51.0 4.811e+01   2.894697 3.9284     0.736859
#>  ds 3 parent   14     47.1 4.811e+01  -1.005303 3.9284    -0.255905
#>  ds 3 parent   28     36.1 3.433e+01   1.766324 2.8695     0.615541
#>  ds 3 parent   28     36.6 3.433e+01   2.266324 2.8695     0.789785
#>  ds 3 parent   60     20.1 1.973e+01   0.371727 1.7967     0.206897
#>  ds 3 parent   60     19.8 1.973e+01   0.071727 1.7967     0.039922
#>  ds 3 parent   90     11.3 1.194e+01  -0.639567 1.2900    -0.495771
#>  ds 3 parent   90     10.7 1.194e+01  -1.239567 1.2900    -0.960871
#>  ds 3 parent  120      8.2 7.228e+00   0.972496 1.0448     0.930805
#>  ds 3 parent  120      7.3 7.228e+00   0.072496 1.0448     0.069388
#>  ds 3     m1    0      0.8 2.046e-12   0.800000 0.8720     0.917426
#>  ds 3     m1    1      1.8 1.802e+00  -0.002337 0.8837    -0.002644
#>  ds 3     m1    1      2.3 1.802e+00   0.497663 0.8837     0.563136
#>  ds 3     m1    3      4.2 4.662e+00  -0.461759 0.9477    -0.487230
#>  ds 3     m1    3      4.1 4.662e+00  -0.561759 0.9477    -0.592746
#>  ds 3     m1    7      6.8 8.261e+00  -1.460846 1.0923    -1.337436
#>  ds 3     m1    7     10.1 8.261e+00   1.839154 1.0923     1.683785
#>  ds 3     m1   14     11.4 1.088e+01   0.520066 1.2292     0.423096
#>  ds 3     m1   14     12.8 1.088e+01   1.920066 1.2292     1.562058
#>  ds 3     m1   28     11.5 1.143e+01   0.066711 1.2606     0.052919
#>  ds 3     m1   28     10.6 1.143e+01  -0.833289 1.2606    -0.661009
#>  ds 3     m1   60      7.5 9.224e+00  -1.723819 1.1401    -1.511997
#>  ds 3     m1   60      8.6 9.224e+00  -0.623819 1.1401    -0.547165
#>  ds 3     m1   90      7.3 6.951e+00   0.348894 1.0328     0.337804
#>  ds 3     m1   90      8.1 6.951e+00   1.148894 1.0328     1.112374
#>  ds 3     m1  120      5.3 5.028e+00   0.272355 0.9595     0.283849
#>  ds 3     m1  120      3.8 5.028e+00  -1.227645 0.9595    -1.279452
#>  ds 4 parent    0    104.7 1.012e+02   3.450870 8.1091     0.425557
#>  ds 4 parent    0     88.3 1.012e+02 -12.949130 8.1091    -1.596872
#>  ds 4 parent    1     94.2 9.787e+01  -3.670543 7.8417    -0.468083
#>  ds 4 parent    1     94.6 9.787e+01  -3.270543 7.8417    -0.417073
#>  ds 4 parent    3     78.1 9.154e+01 -13.439273 7.3409    -1.830749
#>  ds 4 parent    3     96.5 9.154e+01   4.960727 7.3409     0.675769
#>  ds 4 parent    7     76.2 8.040e+01  -4.195825 6.4607    -0.649438
#>  ds 4 parent    7     77.8 8.040e+01  -2.595825 6.4607    -0.401787
#>  ds 4 parent   14     70.8 6.485e+01   5.950631 5.2368     1.136312
#>  ds 4 parent   14     67.3 6.485e+01   2.450631 5.2368     0.467964
#>  ds 4 parent   28     43.1 4.403e+01  -0.927716 3.6126    -0.256803
#>  ds 4 parent   28     45.1 4.403e+01   1.072284 3.6126     0.296821
#>  ds 4 parent   60     21.3 2.121e+01   0.085295 1.9010     0.044868
#>  ds 4 parent   60     23.5 2.121e+01   2.285295 1.9010     1.202136
#>  ds 4 parent   90     11.8 1.187e+01  -0.069302 1.2859    -0.053892
#>  ds 4 parent   90     12.1 1.187e+01   0.230698 1.2859     0.179402
#>  ds 4 parent  120      7.0 6.896e+00   0.104203 1.0305     0.101121
#>  ds 4 parent  120      6.2 6.896e+00  -0.695797 1.0305    -0.675218
#>  ds 4     m1    0      1.6 4.661e-12   1.600000 0.8720     1.834853
#>  ds 4     m1    1      0.9 6.932e-01   0.206778 0.8737     0.236656
#>  ds 4     m1    3      3.7 1.958e+00   1.742149 0.8858     1.966684
#>  ds 4     m1    3      2.0 1.958e+00   0.042149 0.8858     0.047581
#>  ds 4     m1    7      3.6 4.056e+00  -0.456030 0.9299    -0.490412
#>  ds 4     m1    7      3.8 4.056e+00  -0.256030 0.9299    -0.275333
#>  ds 4     m1   14      7.1 6.624e+00   0.475608 1.0191     0.466682
#>  ds 4     m1   14      6.6 6.624e+00  -0.024392 1.0191    -0.023934
#>  ds 4     m1   28      9.5 9.013e+00   0.487100 1.1293     0.431311
#>  ds 4     m1   28      9.3 9.013e+00   0.287100 1.1293     0.254218
#>  ds 4     m1   60      8.3 8.710e+00  -0.409941 1.1142    -0.367933
#>  ds 4     m1   60      9.0 8.710e+00   0.290059 1.1142     0.260335
#>  ds 4     m1   90      6.6 6.696e+00  -0.095953 1.0221    -0.093880
#>  ds 4     m1   90      7.7 6.696e+00   1.004047 1.0221     0.982349
#>  ds 4     m1  120      3.7 4.803e+00  -1.102994 0.9522    -1.158384
#>  ds 4     m1  120      3.5 4.803e+00  -1.302994 0.9522    -1.368427
#>  ds 5 parent    0    110.4 1.013e+02   9.136362 8.1102     1.126526
#>  ds 5 parent    0    112.1 1.013e+02  10.836362 8.1102     1.336138
#>  ds 5 parent    1     93.5 9.462e+01  -1.119415 7.5844    -0.147594
#>  ds 5 parent    1     91.0 9.462e+01  -3.619415 7.5844    -0.477216
#>  ds 5 parent    3     71.0 8.313e+01 -12.125525 6.6761    -1.816251
#>  ds 5 parent    3     89.7 8.313e+01   6.574475 6.6761     0.984774
#>  ds 5 parent    7     60.4 6.580e+01  -5.399792 5.3114    -1.016637
#>  ds 5 parent    7     59.1 6.580e+01  -6.699792 5.3114    -1.261392
#>  ds 5 parent   14     56.5 4.728e+01   9.217968 3.8645     2.385273
#>  ds 5 parent   14     47.0 4.728e+01  -0.282032 3.8645    -0.072979
#>  ds 5 parent   28     30.2 3.077e+01  -0.574817 2.6010    -0.220999
#>  ds 5 parent   28     23.9 3.077e+01  -6.874817 2.6010    -2.643147
#>  ds 5 parent   60     17.0 1.796e+01  -0.957722 1.6748    -0.571838
#>  ds 5 parent   60     18.7 1.796e+01   0.742278 1.6748     0.443201
#>  ds 5 parent   90     11.3 1.186e+01  -0.560382 1.2854    -0.435957
#>  ds 5 parent   90     11.9 1.186e+01   0.039618 1.2854     0.030821
#>  ds 5 parent  120      9.0 7.875e+00   1.125373 1.0740     1.047800
#>  ds 5 parent  120      8.1 7.875e+00   0.225373 1.0740     0.209838
#>  ds 5     m1    0      0.7 7.105e-15   0.700000 0.8720     0.802748
#>  ds 5     m1    1      3.0 3.133e+00  -0.132776 0.9070    -0.146393
#>  ds 5     m1    1      2.6 3.133e+00  -0.532776 0.9070    -0.587416
#>  ds 5     m1    3      5.1 8.401e+00  -3.300544 1.0990    -3.003203
#>  ds 5     m1    3      7.5 8.401e+00  -0.900544 1.0990    -0.819415
#>  ds 5     m1    7     16.5 1.580e+01   0.697935 1.5309     0.455905
#>  ds 5     m1    7     19.0 1.580e+01   3.197935 1.5309     2.088955
#>  ds 5     m1   14     22.9 2.230e+01   0.604922 1.9779     0.305846
#>  ds 5     m1   14     23.2 2.230e+01   0.904922 1.9779     0.457525
#>  ds 5     m1   28     22.2 2.446e+01  -2.255019 2.1336    -1.056917
#>  ds 5     m1   28     24.4 2.446e+01  -0.055019 2.1336    -0.025787
#>  ds 5     m1   60     15.5 1.874e+01  -3.242459 1.7285    -1.875916
#>  ds 5     m1   60     19.8 1.874e+01   1.057541 1.7285     0.611838
#>  ds 5     m1   90     14.9 1.352e+01   1.376176 1.3856     0.993173
#>  ds 5     m1   90     14.2 1.352e+01   0.676176 1.3856     0.487989
#>  ds 5     m1  120     10.9 9.609e+00   1.291013 1.1601     1.112858
#>  ds 5     m1  120     10.4 9.609e+00   0.791013 1.1601     0.681856
# }