Transform formation fractions as in the first published mkin version

The transformed fractions can be restricted between 0 and 1 in model optimisations. Therefore this transformation was used originally in mkin. It was later replaced by the ilr transformation because the ilr transformed fractions can assumed to follow normal distribution. As the ilr transformation is not available in RxODE and can therefore not be used in the nlmixr modelling language, this transformation is currently used for translating mkin models with formation fractions to more than one target compartment for fitting with nlmixr in nlmixr_model. However, this implementation cannot be used there, as it is not accessible from RxODE.

tffm0(ff)

invtffm0(ff_trans)

Arguments

ff
Vector of untransformed formation fractions. The sum must be smaller or equal to one

ff_trans
Vector of transformed formation fractions that can be restricted to the interval from 0 to 1

Value

A vector of the transformed formation fractions

A vector of backtransformed formation fractions for natural use in degradation models

Examples

ff_example <- c(
  0.10983681, 0.09035905, 0.08399383
)
ff_example_trans <- tffm0(ff_example)
invtffm0(ff_example_trans)

#> [1] 0.10983681 0.09035905 0.08399383