Function to calculate maximum time weighted average concentrations from kinetic models fitted with mkinfit

This function calculates maximum moving window time weighted average concentrations (TWAs) for kinetic models fitted with mkinfit. Currently, only calculations for the parent are implemented for the SFO, FOMC and DFOP models, using the analytical formulas given in the PEC soil section of the FOCUS guidance.

twa(fit, windows)

Arguments

fit
An object of class mkinfit.

windows
The width of the time windows for which the TWAs should be calculated.

Value

A numeric vector, named using the windows argument.

References

FOCUS (2006) “Guidance Document on Estimating Persistence and Degradation Kinetics from Environmental Fate Studies on Pesticides in EU Registration” Report of the FOCUS Work Group on Degradation Kinetics, EC Document Reference Sanco/10058/2005 version 2.0, 434 pp, http://esdac.jrc.ec.europa.eu/projects/degradation-kinetics

Examples

  fit <- mkinfit("FOMC", FOCUS_2006_C, quiet = TRUE)
  twa(fit, c(7, 21))

#>        7       21 
#> 34.71343 18.22124