Aerobic soil degradation data on 2,4-D from the EU assessment in 2014

The five datasets were extracted from the active substance evaluation dossier published by EFSA. Kinetic evaluations shown for these datasets are intended to illustrate and advance kinetic modelling. The fact that these data and some results are shown here does not imply a license to use them in the context of pesticide registrations, as the use of the data may be constrained by data protection regulations.

D24_2014

Format

An mkindsg object grouping five datasets

Source

Hellenic Ministry of Rural Development and Agriculture (2014) Final addendum to the Renewal Assessment Report - public version - 2,4-D Volume 3 Annex B.8 Fate and behaviour in the environment http://registerofquestions.efsa.europa.eu/roqFrontend/outputLoader?output=ON-3812

Details

Data for the first dataset are from p. 685. Data for the other four datasets were used in the preprocessed versions given in the kinetics section (p. 761ff.), with the exception of residues smaller than 1 for DCP in the soil from Site I2, where the values given on p. 694 were used.

The R code used to create this data object is installed with this package in the 'dataset_generation' directory. In the code, page numbers are given for specific pieces of information in the comments.

Examples

print(D24_2014)

#> <mkindsg> holding 5 mkinds objects
#> Title $title:  Aerobic soil degradation data on 2,4-D from the EU assessment in 2014 
#> Occurrene of observed compounds $observed_n:
#> D24 DCP DCA 
#>   5   4   4 
#> Meta information $meta:
#>                       study usda_soil_type study_moisture_ref_type rel_moisture
#> 1                Cohen 1991      Silt loam                    <NA>           NA
#> 2 Liu and Adelfinskaya 2011      Silt loam                     pF1          0.5
#> 3 Liu and Adelfinskaya 2011           Loam                     pF1          0.5
#> 4 Liu and Adelfinskaya 2011           Loam                     pF1          0.5
#> 5 Liu and Adelfinskaya 2011     Loamy sand                     pF1          0.5
#>   temperature
#> 1          25
#> 2          20
#> 3          20
#> 4          20
#> 5          20

print(D24_2014$ds[[1]], data = TRUE)

#> <mkinds> with $title:  Mississippi 
#> Observed compounds $observed:  D24 
#> Sampling times $sampling_times:
#> 0, 2, 4, 7, 15, 24, 35, 56, 71, 114, 183, 273, 365 
#> With a maximum of  1  replicates
#>    time  D24
#> 1     0 96.8
#> 2     2 81.0
#> 3     4 81.7
#> 4     7 88.2
#> 5    15 66.3
#> 6    24 72.9
#> 7    35 62.6
#> 8    56 54.6
#> 9    71 35.2
#> 10  114 18.0
#> 11  183 11.3
#> 12  273  9.9
#> 13  365  6.3

m_D24 = mkinmod(D24 = mkinsub("SFO", to = "DCP"),
  DCP = mkinsub("SFO", to = "DCA"),
  DCA = mkinsub("SFO"))

#> Temporary DLL for differentials generated and loaded

print(D24_2014)

#> <mkindsg> holding 5 mkinds objects
#> Title $title:  Aerobic soil degradation data on 2,4-D from the EU assessment in 2014 
#> Occurrene of observed compounds $observed_n:
#> D24 DCP DCA 
#>   5   4   4 
#> Meta information $meta:
#>                       study usda_soil_type study_moisture_ref_type rel_moisture
#> 1                Cohen 1991      Silt loam                    <NA>           NA
#> 2 Liu and Adelfinskaya 2011      Silt loam                     pF1          0.5
#> 3 Liu and Adelfinskaya 2011           Loam                     pF1          0.5
#> 4 Liu and Adelfinskaya 2011           Loam                     pF1          0.5
#> 5 Liu and Adelfinskaya 2011     Loamy sand                     pF1          0.5
#>   temperature
#> 1          25
#> 2          20
#> 3          20
#> 4          20
#> 5          20