Results of fitting the SFO model to Datasets A to F of FOCUS (2006)

A table with the fitted parameters and the resulting DT50 and DT90 values generated with different software packages. Taken directly from FOCUS (2006). The results from fitting the data with the Topfit software was removed, as the initial concentration of the parent compound was fixed to a value of 100 in this fit.

Usage

FOCUS_2006_SFO_ref_A_to_F

Format

A data frame containing the following variables.

package

a factor giving the name of the software package

M0

The fitted initial concentration of the parent compound

k

The fitted first-order degradation rate constant

DT50

The resulting half-life of the parent compound

DT90

The resulting DT90 of the parent compound

dataset

The FOCUS dataset that was used

Source

FOCUS (2006) “Guidance Document on Estimating Persistence and Degradation Kinetics from Environmental Fate Studies on Pesticides in EU Registration” Report of the FOCUS Work Group on Degradation Kinetics, EC Document Reference Sanco/10058/2005 version 2.0, 434 pp, http://esdac.jrc.ec.europa.eu/projects/degradation-kinetics

Examples

data(FOCUS_2006_SFO_ref_A_to_F)