Datasets A to F from the FOCUS Kinetics report from 2006

Data taken from FOCUS (2006), p. 258.

FOCUS_2006_datasets

Format

6 datasets with observations on the following variables.

name

a factor containing the name of the observed variable

time

a numeric vector containing time points

value

a numeric vector containing concentrations in percent of applied radioactivity

Source

FOCUS (2006) “Guidance Document on Estimating Persistence and Degradation Kinetics from Environmental Fate Studies on Pesticides in EU Registration” Report of the FOCUS Work Group on Degradation Kinetics, EC Document Reference Sanco/10058/2005 version 2.0, 434 pp, http://focus.jrc.ec.europa.eu/dk

Examples

FOCUS_2006_C

#>     name time value
#> 1 parent    0  85.1
#> 2 parent    1  57.9
#> 3 parent    3  29.9
#> 4 parent    7  14.6
#> 5 parent   14   9.7
#> 6 parent   28   6.6
#> 7 parent   63   4.0
#> 8 parent   91   3.9
#> 9 parent  119   0.6