Function to calculate endpoints for further use from kinetic models fitted with mkinfit

This function calculates DT50 and DT90 values as well as formation fractions from kinetic models fitted with mkinfit. If the SFORB model was specified for one of the parents or metabolites, the Eigenvalues are returned. These are equivalent to the rate constants of the DFOP model, but with the advantage that the SFORB model can also be used for metabolites.

endpoints(fit)

Arguments

fit
An object of class mkinfit or nlme.mmkin or another object that has list components mkinmod containing an mkinmod degradation model, and two numeric vectors, bparms.optim and bparms.fixed, that contain parameter values for that model.

Value

A list with a matrix of dissipation times named distimes, and, if applicable, a vector of formation fractions named ff and, if the SFORB model was in use, a vector of eigenvalues of these SFORB models, equivalent to DFOP rate constants

Details

Additional DT50 values are calculated from the FOMC DT90 and k1 and k2 from HS and DFOP, as well as from Eigenvalues b1 and b2 of any SFORB models

Note

The function is used internally by summary.mkinfit and summary.nlme.mmkin

Author

Johannes Ranke

Examples

  fit <- mkinfit("FOMC", FOCUS_2006_C, quiet = TRUE)
  endpoints(fit)

#> $distimes
#>            DT50    DT90 DT50back
#> parent 1.785233 15.1479 4.559973
#> 

  # \dontrun{
    fit_2 <- mkinfit("DFOP", FOCUS_2006_C, quiet = TRUE)
    endpoints(fit_2)

#> $distimes
#>            DT50     DT90 DT50back  DT50_k1  DT50_k2
#> parent 1.886925 21.25106 6.397207 1.508293 38.83438
#> 

    fit_3 <- mkinfit("SFORB", FOCUS_2006_C, quiet = TRUE)
    endpoints(fit_3)

#> $ff
#> parent_free 
#>           1 
#> 
#> $SFORB
#> parent_b1 parent_b2 
#> 0.4595574 0.0178488 
#> 
#> $distimes
#>            DT50     DT90 DT50back DT50_parent_b1 DT50_parent_b2
#> parent 1.886925 21.25106 6.397208       1.508293       38.83438
#> 

  # }