Function to calculate endpoints for further use from kinetic models fitted with mkinfit

This function calculates DT50 and DT90 values as well as formation fractions from kinetic models fitted with mkinfit. If the SFORB model was specified for one of the parents or metabolites, the Eigenvalues are returned. These are equivalent to the rate constantes of the DFOP model, but with the advantage that the SFORB model can also be used for metabolites.

endpoints(fit)

Arguments

fit

An object of class mkinfit.

Note

The function is used internally by summary.mkinfit.

Value

A list with the components mentioned above.