Normalisation factors for aerobic soil degradation according to FOCUS guidance

Time step normalisation factors for aerobic soil degradation as described in Appendix 8 to the FOCUS kinetics guidance (FOCUS 2014, p. 369).

f_time_norm_focus(object, ...)

# S3 method for numeric
f_time_norm_focus(
  object,
  moisture = NA,
  field_moisture = NA,
  temperature = object,
  Q10 = 2.58,
  walker = 0.7,
  f_na = NA,
  ...
)

# S3 method for mkindsg
f_time_norm_focus(
  object,
  study_moisture_ref_source = c("auto", "meta", "focus"),
  Q10 = 2.58,
  walker = 0.7,
  f_na = NA,
  ...
)

Arguments

object

An object containing information used for the calculations

...

Currently not used

moisture

Numeric vector of moisture contents in \% w/w

field_moisture

Numeric vector of moisture contents at field capacity (pF2) in \% w/w

temperature

Numeric vector of temperatures in °C

Q10

The Q10 value used for temperature normalisation

walker

The Walker exponent used for moisture normalisation

f_na

The factor to use for NA values. If set to NA, only factors for complete cases will be returned.

study_moisture_ref_source

Source for the reference value used to calculate the study moisture. If 'auto', preference is given to a reference moisture given in the meta information, otherwise the focus soil moisture for the soil class is used

References

FOCUS (2006) “Guidance Document on Estimating Persistence and Degradation Kinetics from Environmental Fate Studies on Pesticides in EU Registration” Report of the FOCUS Work Group on Degradation Kinetics, EC Document Reference Sanco/10058/2005 version 2.0, 434 pp, http://esdac.jrc.ec.europa.eu/projects/degradation-kinetics FOCUS (2014) “Generic guidance for Estimating Persistence and Degradation Kinetics from Environmental Fate Studies on Pesticides in EU Registration” Report of the FOCUS Work Group on Degradation Kinetics, Version 1.1, 18 December 2014 http://esdac.jrc.ec.europa.eu/projects/degradation-kinetics

See also

Examples

f_time_norm_focus(25, 20, 25) # 1.37, compare FOCUS 2014 p. 184
#> [1] 1.373956

D24_2014$meta
#>                                 study usda_soil_type study_moisture_ref_type
#> Mississippi                Cohen 1991      Silt loam                    <NA>
#> Fayette     Liu and Adelfinskaya 2011      Silt loam                     pF1
#> RefSol 03-G Liu and Adelfinskaya 2011           Loam                     pF1
#> Site E1     Liu and Adelfinskaya 2011           Loam                     pF1
#> Site I2     Liu and Adelfinskaya 2011     Loamy sand                     pF1
#>             rel_moisture temperature
#> Mississippi           NA          25
#> Fayette              0.5          20
#> RefSol 03-G          0.5          20
#> Site E1              0.5          20
#> Site I2              0.5          20
# No moisture normalisation in the first dataset, so we use f_na = 1 to get
# temperature only normalisation as in the EU evaluation
f_time_norm_focus(D24_2014, study_moisture_ref_source = "focus", f_na = 1)
#> $f_time_norm was (re)set to normalised values