FOCUS default values for soil moisture contents at field capacity, MWHC and 1/3 bar

The value were transcribed from p. 36. The table assumes field capacity corresponds to pF2, MWHC to pF 1 and 1/3 bar to pF 2.5.

focus_soil_moisture

Format

A matrix with upper case USDA soil classes as row names, and water tension ('pF1', 'pF2', 'pF 2.5') as column names

Source

Anonymous (2014) Generic Guidance for Tier 1 FOCUS Ground Water Assessment Version 2.2, May 2014 https://esdac.jrc.ec.europa.eu/projects/ground-water

Examples

focus_soil_moisture

#>                 pF1 pF2 pF2.5
#> Sand             24  12     7
#> Loamy sand       24  14     9
#> Sandy loam       27  19    15
#> Sandy clay loam  28  22    18
#> Clay loam        32  28    25
#> Loam             31  25    21
#> Silt loam        32  26    21
#> Silty clay loam  34  30    27
#> Silt             31  27    21
#> Sandy clay       41  35    31
#> Silty clay       44  40    36
#> Clay             53  48    43