Method to get the names of ill-defined parameters

The method for generalised nonlinear regression fits as obtained with mkinfit and mmkin checks if the degradation parameters pass the Wald test (in degradation kinetics often simply called t-test) for significant difference from zero. For this test, the parameterisation without parameter transformations is used.

illparms(object, ...)

# S3 method for mkinfit
illparms(object, conf.level = 0.95, ...)

# S3 method for illparms.mkinfit
print(x, ...)

# S3 method for mmkin
illparms(object, conf.level = 0.95, ...)

# S3 method for illparms.mmkin
print(x, ...)

# S3 method for saem.mmkin
illparms(
  object,
  conf.level = 0.95,
  random = TRUE,
  errmod = TRUE,
  slopes = TRUE,
  ...
)

# S3 method for illparms.saem.mmkin
print(x, ...)

# S3 method for mhmkin
illparms(object, conf.level = 0.95, random = TRUE, errmod = TRUE, ...)

# S3 method for illparms.mhmkin
print(x, ...)

Arguments

object

The object to investigate

...

For potential future extensions

conf.level

The confidence level for checking p values

x

The object to be printed

random

For hierarchical fits, should random effects be tested?

errmod

For hierarchical fits, should error model parameters be tested?

slopes

For hierarchical saem fits using saemix as backend, should slope parameters in the covariate model(starting with 'beta_') be tested?

Value

For mkinfit or saem objects, a character vector of parameter names. For mmkin or mhmkin objects, a matrix like object of class 'illparms.mmkin' or 'illparms.mhmkin'.

Details

The method for hierarchical model fits, also known as nonlinear mixed-effects model fits as obtained with saem and mhmkin checks if any of the confidence intervals for the random effects expressed as standard deviations include zero, and if the confidence intervals for the error model parameters include zero.

Note

All return objects have printing methods. For the single fits, printing does not output anything in the case no ill-defined parameters are found.

Examples

fit <- mkinfit("FOMC", FOCUS_2006_A, quiet = TRUE)
illparms(fit)
#> [1] "parent_0" "alpha"    "beta"     "sigma"   
# \dontrun{
fits <- mmkin(
  c("SFO", "FOMC"),
  list("FOCUS A" = FOCUS_2006_A,
       "FOCUS C" = FOCUS_2006_C),
  quiet = TRUE)
illparms(fits)
#>       dataset
#> model  FOCUS A                      FOCUS C
#>   SFO                                      
#>   FOMC parent_0, alpha, beta, sigma        
# }