Calculated the log-likelihood of a fitted mkinfit object

This function returns the product of the likelihood densities of each observed value, as calculated as part of the fitting procedure using dnorm, i.e. assuming normal distribution, and with the means predicted by the degradation model, and the standard deviations predicted by the error model.

# S3 method for mkinfit
logLik(object, ...)

Arguments

object

An object of class mkinfit.

...

For compatibility with the generic method

Value

An object of class logLik with the number of estimated parameters (degradation model parameters plus variance model parameters) as attribute.

Details

The total number of estimated parameters returned with the value of the likelihood is calculated as the sum of fitted degradation model parameters and the fitted error model parameters.

See also

Author

Johannes Ranke

Examples

  # \dontrun{
  sfo_sfo <- mkinmod(
    parent = mkinsub("SFO", to = "m1"),
    m1 = mkinsub("SFO")
  )
#> Temporary DLL for differentials generated and loaded
  d_t <- subset(FOCUS_2006_D, value != 0)
  f_nw <- mkinfit(sfo_sfo, d_t, quiet = TRUE) # no weighting (weights are unity)
  f_obs <- update(f_nw, error_model = "obs")
  f_tc <- update(f_nw, error_model = "tc")
  AIC(f_nw, f_obs, f_tc)
#>       df      AIC
#> f_nw   5 204.4486
#> f_obs  6 205.8727
#> f_tc   6 141.9656
  # }