Datasets on aerobic soil metabolism of 2,4,5-T in six soils

Time course of 2,4,5-trichlorophenoxyacetic acid, and the corresponding 2,4,5-trichlorophenol and 2,4,5-trichloroanisole as recovered in diethylether extracts.

mccall81_245T

Format

A dataframe containing the following variables.

name

the name of the compound observed. Note that T245 is used as an acronym for 2,4,5-T. T245 is a legitimate object name in R, which is necessary for specifying models using mkinmod.

time

a numeric vector containing sampling times in days after treatment

value

a numeric vector containing concentrations in percent of applied radioactivity

soil

a factor containing the name of the soil

Source

McCall P, Vrona SA, Kelley SS (1981) Fate of uniformly carbon-14 ring labeled 2,4,5-Trichlorophenoxyacetic acid and 2,4-dichlorophenoxyacetic acid. J Agric Chem 29, 100-107 http://dx.doi.org/10.1021/jf00103a026

Examples

  SFO_SFO_SFO <- mkinmod(T245 = list(type = "SFO", to = "phenol"),
                         phenol = list(type = "SFO", to = "anisole"),
                         anisole = list(type = "SFO"))

#> Successfully compiled differential equation model from auto-generated C code.

    fit.1 <- mkinfit(SFO_SFO_SFO, subset(mccall81_245T, soil == "Commerce"))

#> Model cost at call  1 :  13655.2 
#> Model cost at call  2 :  13655.2 
#> Model cost at call  8 :  9191.59 
#> Model cost at call  16 :  5663.584 
#> Model cost at call  17 :  5663.582 
#> Model cost at call  18 :  5663.566 
#> Model cost at call  19 :  5663.55 
#> Model cost at call  23 :  3035.567 
#> Model cost at call  24 :  3035.566 
#> Model cost at call  26 :  3035.561 
#> Model cost at call  27 :  3035.559 
#> Model cost at call  28 :  3035.558 
#> Model cost at call  30 :  2252.97 
#> Model cost at call  31 :  2252.965 
#> Model cost at call  32 :  2252.965 
#> Model cost at call  36 :  2252.958 
#> Model cost at call  37 :  1220.201 
#> Model cost at call  38 :  1220.2 
#> Model cost at call  39 :  1220.199 
#> Model cost at call  41 :  1220.183 
#> Model cost at call  45 :  507.1358 
#> Model cost at call  48 :  507.1353 
#> Model cost at call  52 :  198.5695 
#> Model cost at call  55 :  198.5694 
#> Model cost at call  60 :  175.8239 
#> Model cost at call  63 :  175.8239 
#> Model cost at call  67 :  170.6316 
#> Model cost at call  69 :  170.6316 
#> Model cost at call  74 :  165.881 
#> Model cost at call  77 :  165.881 
#> Model cost at call  81 :  161.6423 
#> Model cost at call  84 :  161.6423 
#> Model cost at call  88 :  158.2953 
#> Model cost at call  91 :  158.2953 
#> Model cost at call  95 :  156.7311 
#> Model cost at call  97 :  156.731 
#> Model cost at call  103 :  155.998 
#> Model cost at call  106 :  155.998 
#> Model cost at call  110 :  155.6263 
#> Model cost at call  113 :  155.6262 
#> Model cost at call  116 :  155.6262 
#> Model cost at call  117 :  155.1202 
#> Model cost at call  120 :  155.1202 
#> Model cost at call  124 :  154.4133 
#> Model cost at call  125 :  153.8489 
#> Model cost at call  126 :  152.7413 
#> Model cost at call  129 :  152.7413 
#> Model cost at call  132 :  152.7413 
#> Model cost at call  133 :  150.3995 
#> Model cost at call  134 :  148.7256 
#> Model cost at call  135 :  147.2618 
#> Model cost at call  137 :  147.2618 
#> Model cost at call  139 :  147.2618 
#> Model cost at call  142 :  145.115 
#> Model cost at call  144 :  145.115 
#> Model cost at call  145 :  145.115 
#> Model cost at call  146 :  145.115 
#> Model cost at call  147 :  145.115 
#> Model cost at call  150 :  143.7878 
#> Model cost at call  157 :  142.8531 
#> Model cost at call  160 :  142.8531 
#> Model cost at call  164 :  142.7325 
#> Model cost at call  167 :  142.7325 
#> Model cost at call  169 :  142.7325 
#> Model cost at call  172 :  141.8393 
#> Model cost at call  173 :  141.8393 
#> Model cost at call  174 :  141.8393 
#> Model cost at call  179 :  141.3446 
#> Model cost at call  180 :  141.3446 
#> Model cost at call  181 :  141.3446 
#> Model cost at call  186 :  141.1111 
#> Model cost at call  193 :  140.7786 
#> Model cost at call  200 :  140.451 
#> Model cost at call  207 :  140.1395 
#> Model cost at call  214 :  139.7903 
#> Model cost at call  221 :  139.5476 
#> Model cost at call  228 :  139.4441 
#> Model cost at call  235 :  139.3204 
#> Model cost at call  242 :  139.2508 
#> Model cost at call  249 :  139.1891 
#> Model cost at call  256 :  139.1561 
#> Model cost at call  263 :  139.137 
#> Model cost at call  270 :  139.1278 
#> Model cost at call  271 :  139.1278 
#> Model cost at call  279 :  139.1211 
#> Model cost at call  280 :  139.1211 
#> Model cost at call  281 :  139.1211 
#> Model cost at call  282 :  139.1211 
#> Model cost at call  286 :  139.1179 
#> Model cost at call  287 :  139.1179 
#> Model cost at call  291 :  139.1179 
#> Model cost at call  294 :  139.1179 
#> Model cost at call  298 :  139.1158 
#> Model cost at call  299 :  139.1158 
#> Model cost at call  302 :  139.1158 
#> Model cost at call  310 :  139.1148 
#> Model cost at call  311 :  139.1148 
#> Model cost at call  315 :  139.1148 
#> Model cost at call  322 :  139.1143 
#> Model cost at call  323 :  139.1143 
#> Model cost at call  326 :  139.1143 
#> Model cost at call  334 :  139.114 
#> Model cost at call  335 :  139.114 
#> Model cost at call  338 :  139.114 
#> Model cost at call  339 :  139.114 
#> Model cost at call  346 :  139.1139 
#> Model cost at call  347 :  139.1139 
#> Model cost at call  350 :  139.1139 
#> Model cost at call  351 :  139.1139 
#> Model cost at call  358 :  139.1139 
#> Model cost at call  359 :  139.1139 
#> Model cost at call  362 :  139.1139 
#> Model cost at call  363 :  139.1139 
#> Model cost at call  370 :  139.1138 
#> Model cost at call  382 :  139.1138 
#> Model cost at call  390 :  139.1138 
#> Model cost at call  395 :  139.1138 
#> Model cost at call  403 :  139.1138 
#> Model cost at call  408 :  139.1138 
#> Model cost at call  416 :  139.1138 
#> Model cost at call  421 :  139.1138 
#> Model cost at call  429 :  139.1138 
#> Model cost at call  434 :  139.1138 
#> Model cost at call  442 :  139.1138 
#> Model cost at call  447 :  139.1138 
#> Model cost at call  455 :  139.1138 
#> Model cost at call  460 :  139.1138 
#> Model cost at call  468 :  139.1138 
#> Model cost at call  473 :  139.1138 
#> Model cost at call  481 :  139.1138 
#> Model cost at call  486 :  139.1138 
#> Model cost at call  494 :  139.1138 
#> Model cost at call  499 :  139.1138 
#> Model cost at call  507 :  139.1138 
#> Model cost at call  512 :  139.1138 
#> Model cost at call  519 :  139.1138 
#> Model cost at call  520 :  139.1138 
#> Model cost at call  525 :  139.1138 
#> Model cost at call  538 :  139.1138 
#> Model cost at call  545 :  139.1138 
#> Model cost at call  546 :  139.1138 
#> Model cost at call  551 :  139.1138 
#> Model cost at call  558 :  139.1138 
#> Model cost at call  559 :  139.1138 
#> Model cost at call  564 :  139.1138 
#> Model cost at call  571 :  139.1138 
#> Model cost at call  580 :  139.1138 
#> Model cost at call  588 :  139.1138 

#> Warning: Optimisation by method Port did not converge.
#> Convergence code is 1

    summary(fit.1, data = FALSE)

#> mkin version:    0.9.44.9000 
#> R version:       3.3.2 
#> Date of fit:     Fri Nov 18 15:19:33 2016 
#> Date of summary: Fri Nov 18 15:19:33 2016 
#> 
#> 
#> Warning: Optimisation by method Port did not converge.
#> Convergence code is 1 
#> 
#> 
#> Equations:
#> d_T245/dt = - k_T245_sink * T245 - k_T245_phenol * T245
#> d_phenol/dt = + k_T245_phenol * T245 - k_phenol_sink * phenol -
#>            k_phenol_anisole * phenol
#> d_anisole/dt = + k_phenol_anisole * phenol - k_anisole_sink * anisole
#> 
#> Model predictions using solution type deSolve 
#> 
#> Fitted with method Port using 590 model solutions performed in 3.45 s
#> 
#> Weighting: none
#> 
#> Starting values for parameters to be optimised:
#>                     value   type
#> T245_0           100.9000  state
#> k_T245_sink        0.1000 deparm
#> k_T245_phenol      0.1001 deparm
#> k_phenol_sink      0.1002 deparm
#> k_phenol_anisole   0.1003 deparm
#> k_anisole_sink     0.1004 deparm
#> 
#> Starting values for the transformed parameters actually optimised:
#>                           value lower upper
#> T245_0               100.900000  -Inf   Inf
#> log_k_T245_sink       -2.302585  -Inf   Inf
#> log_k_T245_phenol     -2.301586  -Inf   Inf
#> log_k_phenol_sink     -2.300587  -Inf   Inf
#> log_k_phenol_anisole  -2.299590  -Inf   Inf
#> log_k_anisole_sink    -2.298593  -Inf   Inf
#> 
#> Fixed parameter values:
#>           value  type
#> phenol_0      0 state
#> anisole_0     0 state
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                      Estimate Std. Error Lower Upper
#> T245_0               103.9000         NA    NA    NA
#> log_k_T245_sink       -4.1130         NA    NA    NA
#> log_k_T245_phenol     -3.6120         NA    NA    NA
#> log_k_phenol_sink    -26.3900         NA    NA    NA
#> log_k_phenol_anisole  -0.9037         NA    NA    NA
#> log_k_anisole_sink    -5.0090         NA    NA    NA
#> 
#> Parameter correlation:

#> Warning: Could not estimate covariance matrix; singular system:

#> Could not estimate covariance matrix; singular system:
#> 
#> Residual standard error: 2.78 on 18 degrees of freedom
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                   Estimate   t value    Pr(>t) Lower Upper
#> T245_0           1.039e+02 4.282e+01 7.236e-20    NA    NA
#> k_T245_sink      1.636e-02 8.901e-01 1.926e-01    NA    NA
#> k_T245_phenol    2.701e-02 1.504e+00 7.498e-02    NA    NA
#> k_phenol_sink    3.457e-12 1.230e-11 5.000e-01    NA    NA
#> k_phenol_anisole 4.051e-01 2.518e+00 1.075e-02    NA    NA
#> k_anisole_sink   6.679e-03 8.146e+00 9.469e-08    NA    NA
#> 
#> Chi2 error levels in percent:
#>          err.min n.optim df
#> All data  10.070       6 16
#> T245       7.908       3  5
#> phenol   106.445       2  5
#> anisole    5.379       1  6
#> 
#> Resulting formation fractions:
#>                       ff
#> T245_sink      3.772e-01
#> T245_phenol    6.228e-01
#> phenol_sink    8.534e-12
#> phenol_anisole 1.000e+00
#> anisole_sink   1.000e+00
#> 
#> Estimated disappearance times:
#>            DT50    DT90
#> T245     15.982  53.091
#> phenol    1.711   5.685
#> anisole 103.784 344.763

  # No covariance matrix and k_phenol_sink is really small, therefore fix it to zero
  fit.2 <- mkinfit(SFO_SFO_SFO, subset(mccall81_245T, soil == "Commerce"),
                 parms.ini = c(k_phenol_sink = 0),
                 fixed_parms = "k_phenol_sink", quiet = TRUE)
  summary(fit.2, data = FALSE)

#> mkin version:    0.9.44.9000 
#> R version:       3.3.2 
#> Date of fit:     Fri Nov 18 15:19:35 2016 
#> Date of summary: Fri Nov 18 15:19:35 2016 
#> 
#> Equations:
#> d_T245/dt = - k_T245_sink * T245 - k_T245_phenol * T245
#> d_phenol/dt = + k_T245_phenol * T245 - k_phenol_sink * phenol -
#>            k_phenol_anisole * phenol
#> d_anisole/dt = + k_phenol_anisole * phenol - k_anisole_sink * anisole
#> 
#> Model predictions using solution type deSolve 
#> 
#> Fitted with method Port using 246 model solutions performed in 1.477 s
#> 
#> Weighting: none
#> 
#> Starting values for parameters to be optimised:
#>                     value   type
#> T245_0           100.9000  state
#> k_T245_sink        0.1000 deparm
#> k_T245_phenol      0.1001 deparm
#> k_phenol_anisole   0.1002 deparm
#> k_anisole_sink     0.1003 deparm
#> 
#> Starting values for the transformed parameters actually optimised:
#>                           value lower upper
#> T245_0               100.900000  -Inf   Inf
#> log_k_T245_sink       -2.302585  -Inf   Inf
#> log_k_T245_phenol     -2.301586  -Inf   Inf
#> log_k_phenol_anisole  -2.300587  -Inf   Inf
#> log_k_anisole_sink    -2.299590  -Inf   Inf
#> 
#> Fixed parameter values:
#>               value   type
#> phenol_0          0  state
#> anisole_0         0  state
#> k_phenol_sink     0 deparm
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                      Estimate Std. Error  Lower    Upper
#> T245_0               103.9000    2.35200 98.930 108.8000
#> log_k_T245_sink       -4.1130    0.13250 -4.390  -3.8350
#> log_k_T245_phenol     -3.6120    0.05002 -3.716  -3.5070
#> log_k_phenol_anisole  -0.9037    0.30580 -1.544  -0.2637
#> log_k_anisole_sink    -5.0090    0.11180 -5.243  -4.7750
#> 
#> Parameter correlation:
#>                        T245_0 log_k_T245_sink log_k_T245_phenol
#> T245_0                1.00000         0.63761           -0.1742
#> log_k_T245_sink       0.63761         1.00000           -0.3831
#> log_k_T245_phenol    -0.17416        -0.38313            1.0000
#> log_k_phenol_anisole -0.05948         0.08745           -0.3047
#> log_k_anisole_sink   -0.16208        -0.60469            0.5227
#>                      log_k_phenol_anisole log_k_anisole_sink
#> T245_0                           -0.05948            -0.1621
#> log_k_T245_sink                   0.08745            -0.6047
#> log_k_T245_phenol                -0.30470             0.5227
#> log_k_phenol_anisole              1.00000            -0.1774
#> log_k_anisole_sink               -0.17744             1.0000
#> 
#> Residual standard error: 2.706 on 19 degrees of freedom
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                   Estimate t value    Pr(>t)     Lower     Upper
#> T245_0           1.039e+02  44.160 6.462e-21 98.930000 108.80000
#> k_T245_sink      1.636e-02   7.545 1.978e-07  0.012400   0.02159
#> k_T245_phenol    2.701e-02  19.990 1.607e-14  0.024320   0.02999
#> k_phenol_anisole 4.051e-01   3.270 2.014e-03  0.213600   0.76820
#> k_anisole_sink   6.679e-03   8.942 1.544e-08  0.005285   0.00844
#> 
#> Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   9.831       5 17
#> T245       7.908       3  5
#> phenol    99.808       1  6
#> anisole    5.379       1  6
#> 
#> Resulting formation fractions:
#>                    ff
#> T245_sink      0.3772
#> T245_phenol    0.6228
#> phenol_anisole 1.0000
#> phenol_sink    0.0000
#> anisole_sink   1.0000
#> 
#> Estimated disappearance times:
#>            DT50    DT90
#> T245     15.982  53.091
#> phenol    1.711   5.685
#> anisole 103.784 344.763