Convert a dataframe from long to wide format

This function takes a dataframe in the long form, i.e. with a row for each observed value, and converts it into a dataframe with one independent variable and several dependent variables as columns.

mkin_long_to_wide(long_data, time = "time", outtime = "time")

Arguments

long_data
The dataframe must contain one variable called "time" with the time values specified by the time argument, one column called "name" with the grouping of the observed values, and finally one column of observed values called "value".

time
The name of the time variable in the long input data.

outtime
The name of the time variable in the wide output data.

Value

Dataframe in wide format.

Author

Johannes Ranke

Examples

mkin_long_to_wide(FOCUS_2006_D)

#>    time parent    m1
#> 1     0  99.46  0.00
#> 2     0 102.04  0.00
#> 3     1  93.50  4.84
#> 4     1  92.50  5.64
#> 5     3  63.23 12.91
#> 6     3  68.99 12.96
#> 7     7  52.32 22.97
#> 8     7  55.13 24.47
#> 9    14  27.27 41.69
#> 10   14  26.64 33.21
#> 11   21  11.50 44.37
#> 12   21  11.64 46.44
#> 13   35   2.85 41.22
#> 14   35   2.91 37.95
#> 15   50   0.69 41.19
#> 16   50   0.63 40.01
#> 17   75   0.05 40.09
#> 18   75   0.06 33.85
#> 19  100     NA 31.04
#> 20  100     NA 33.13
#> 21  120     NA 25.15
#> 22  120     NA 33.31