A dataset class for mkin

At the moment this dataset class is hardly used in mkin. For example, mkinfit does not take mkinds datasets as argument, but works with dataframes such as the on contained in the data field of mkinds objects. Some datasets provided by this package come as mkinds objects nevertheless.

See also

Public fields

title

A full title for the dataset

sampling_times

The sampling times

time_unit

The time unit

observed

Names of the observed variables

unit

The unit of the observations

replicates

The maximum number of replicates per sampling time

data

A data frame with at least the columns name, time and value in order to be compatible with mkinfit

Methods

Public methods

mkinds$new()

mkinds$clone()

Method new()

Create a new mkinds object

Usage

mkinds$new(title = "", data, time_unit = NA, unit = NA)

Arguments

title

The dataset title

data

The data

time_unit

The time unit

unit

The unit of the observations

Method clone()

The objects of this class are cloneable with this method.

Usage

mkinds$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

mds <- mkinds$new("FOCUS A", FOCUS_2006_A)
print(mds)

#> <mkinds> with $title:  FOCUS A 
#> Observed compounds $observed:  parent 
#> Sampling times $sampling_times:  0, 3, 7, 14, 30, 62, 90, 118 
#> With a maximum of  1  replicates