A class for dataset groups for mkin

A container for working with datasets that share at least one compound, so that combined evaluations are desirable.

Time normalisation factors are initialised with a value of 1 for each dataset if no data are supplied.

# S3 method for mkindsg
print(x, data = FALSE, verbose = data, ...)

Arguments

x

An mkindsg object.

data

Should the mkinds objects be printed with their data?

verbose

Should the mkinds objects be printed?

...

Not used.

Public fields

title

A title for the dataset group

ds

A list of mkinds objects

observed_n

Occurrence counts of compounds in datasets

f_time_norm

Time normalisation factors

meta

A data frame with a row for each dataset, containing additional information in the form of categorical data (factors) or numerical data (e.g. temperature, moisture, or covariates like soil pH).

Methods

Public methods

mkindsg$new()

mkindsg$clone()

Method new()

Create a new mkindsg object

Usage

mkindsg$new(title = "", ds, f_time_norm = rep(1, length(ds)), meta)

Arguments

title

The title

ds

A list of mkinds objects

f_time_norm

Time normalisation factors

meta

The meta data

Method clone()

The objects of this class are cloneable with this method.

Usage

mkindsg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

mdsg <- mkindsg$new("Experimental X", experimental_data_for_UBA_2019[6:10])
print(mdsg)
#> <mkindsg> holding 5 mkinds objects
#> Title $title:  Experimental X 
#> Occurrence of observed compounds $observed_n:
#> parent     A1 
#>      5      5 
print(mdsg, verbose = TRUE)
#> <mkindsg> holding 5 mkinds objects
#> Title $title:  Experimental X 
#> Occurrence of observed compounds $observed_n:
#> parent     A1 
#>      5      5 
#> 
#> Datasets $ds:
#> <mkinds> with $title:  Soil 6 
#> Observed compounds $observed:  parent, A1 
#> Sampling times $sampling_times:
#> 0, 3, 6, 10, 20, 34, 55, 90, 112, 132 
#> With a maximum of  2  replicates
#> Time unit:  days 
#> Observation unit:  \%AR 
#> 
#> <mkinds> with $title:  Soil 7 
#> Observed compounds $observed:  parent, A1 
#> Sampling times $sampling_times:
#> 0, 3, 7, 14, 30, 60, 90, 120, 180 
#> With a maximum of  2  replicates
#> Time unit:  days 
#> Observation unit:  \%AR 
#> 
#> <mkinds> with $title:  Soil 8 
#> Observed compounds $observed:  parent, A1 
#> Sampling times $sampling_times:
#> 0, 1, 3, 8, 14, 27, 48, 70 
#> With a maximum of  2  replicates
#> Time unit:  days 
#> Observation unit:  \%AR 
#> 
#> <mkinds> with $title:  Soil 9 
#> Observed compounds $observed:  parent, A1 
#> Sampling times $sampling_times:
#> 0, 1, 3, 8, 14, 27, 48, 70, 91, 120 
#> With a maximum of  2  replicates
#> Time unit:  days 
#> Observation unit:  \%AR 
#> 
#> <mkinds> with $title:  Soil 10 
#> Observed compounds $observed:  parent, A1 
#> Sampling times $sampling_times:
#> 0, 8, 14, 21, 41, 63, 91, 120 
#> With a maximum of  2  replicates
#> Time unit:  days 
#> Observation unit:  \%AR 
print(mdsg, verbose = TRUE, data = TRUE)
#> <mkindsg> holding 5 mkinds objects
#> Title $title:  Experimental X 
#> Occurrence of observed compounds $observed_n:
#> parent     A1 
#>      5      5 
#> 
#> Datasets $ds:
#> <mkinds> with $title:  Soil 6 
#> Observed compounds $observed:  parent, A1 
#> Sampling times $sampling_times:
#> 0, 3, 6, 10, 20, 34, 55, 90, 112, 132 
#> With a maximum of  2  replicates
#> Time unit:  days 
#> Observation unit:  \%AR 
#>    time parent   A1
#> 1     0   97.2   NA
#> 2     0   96.4   NA
#> 3     3   71.1  4.3
#> 4     3   69.2  4.6
#> 5     6   58.1  7.0
#> 6     6   56.6  7.2
#> 7    10   44.4  8.2
#> 8    10   43.4  8.0
#> 9    20   33.3 11.0
#> 10   20   29.2 13.7
#> 11   34   17.6 11.5
#> 12   34   18.0 12.7
#> 13   55   10.5 14.9
#> 14   55    9.3 14.5
#> 15   90    4.5 12.1
#> 16   90    4.7 12.3
#> 17  112    3.0  9.9
#> 18  112    3.4 10.2
#> 19  132    2.3  8.8
#> 20  132    2.7  7.8
#> 
#> <mkinds> with $title:  Soil 7 
#> Observed compounds $observed:  parent, A1 
#> Sampling times $sampling_times:
#> 0, 3, 7, 14, 30, 60, 90, 120, 180 
#> With a maximum of  2  replicates
#> Time unit:  days 
#> Observation unit:  \%AR 
#>    time parent   A1
#> 1     0   93.6   NA
#> 2     0   92.3   NA
#> 3     3   87.0  3.9
#> 4     3   82.2  3.1
#> 5     7   74.0  6.9
#> 6     7   73.9  6.6
#> 7    14   64.2 10.4
#> 8    14   69.5  8.3
#> 9    30   54.0 14.4
#> 10   30   54.6 13.7
#> 11   60   41.1 22.1
#> 12   60   38.4 22.3
#> 13   90   32.5 27.5
#> 14   90   35.5 25.4
#> 15  120   28.1 28.0
#> 16  120   29.0 26.6
#> 17  180   26.5 25.8
#> 18  180   27.6 25.3
#> 
#> <mkinds> with $title:  Soil 8 
#> Observed compounds $observed:  parent, A1 
#> Sampling times $sampling_times:
#> 0, 1, 3, 8, 14, 27, 48, 70 
#> With a maximum of  2  replicates
#> Time unit:  days 
#> Observation unit:  \%AR 
#>    time parent   A1
#> 1     0   91.9   NA
#> 2     0   90.8   NA
#> 3     1   64.9  9.6
#> 4     1   66.2  7.7
#> 5     3   43.5 15.0
#> 6     3   44.1 15.1
#> 7     8   18.3 21.2
#> 8     8   18.1 21.1
#> 9    14   10.2 19.7
#> 10   14   10.8 18.9
#> 11   27    4.9 17.5
#> 12   27    3.3 15.9
#> 13   48    1.6  9.5
#> 14   48    1.5  9.8
#> 15   70    1.1  6.2
#> 16   70    0.9  6.1
#> 
#> <mkinds> with $title:  Soil 9 
#> Observed compounds $observed:  parent, A1 
#> Sampling times $sampling_times:
#> 0, 1, 3, 8, 14, 27, 48, 70, 91, 120 
#> With a maximum of  2  replicates
#> Time unit:  days 
#> Observation unit:  \%AR 
#>    time parent   A1
#> 1     0   99.8   NA
#> 2     0   98.3   NA
#> 3     1   77.1  4.2
#> 4     1   77.2  3.9
#> 5     3   59.0  7.4
#> 6     3   58.1  7.9
#> 7     8   27.4 14.5
#> 8     8   29.2 13.7
#> 9    14   19.1 14.2
#> 10   14   29.6 12.2
#> 11   27   10.1 13.7
#> 12   27   18.2 13.2
#> 13   48    4.5 13.6
#> 14   48    9.1 15.4
#> 15   70    2.3 10.4
#> 16   70    2.9 11.6
#> 17   91    2.0 10.0
#> 18   91    1.8  9.5
#> 19  120    2.0  9.1
#> 20  120    2.2  9.0
#> 
#> <mkinds> with $title:  Soil 10 
#> Observed compounds $observed:  parent, A1 
#> Sampling times $sampling_times:
#> 0, 8, 14, 21, 41, 63, 91, 120 
#> With a maximum of  2  replicates
#> Time unit:  days 
#> Observation unit:  \%AR 
#>    time parent   A1
#> 1     0   96.1   NA
#> 2     0   94.3   NA
#> 3     8   73.9  3.3
#> 4     8   73.9  3.4
#> 5    14   69.4  3.9
#> 6    14   73.1  2.9
#> 7    21   65.6  6.4
#> 8    21   65.3  7.2
#> 9    41   55.9  9.1
#> 10   41   54.4  8.5
#> 11   63   47.0 11.7
#> 12   63   49.3 12.0
#> 13   91   44.7 13.3
#> 14   91   46.7 13.2
#> 15  120   42.1 14.3
#> 16  120   41.3 12.1