Fit a kinetic model to data with one or more state variables

This function maximises the likelihood of the observed data using the Port algorithm nlminb, and the specified initial or fixed parameters and starting values. In each step of the optimsation, the kinetic model is solved using the function mkinpredict. The parameters of the selected error model are fitted simultaneously with the degradation model parameters, as both of them are arguments of the likelihood function.

Per default, parameters in the kinetic models are internally transformed in order to better satisfy the assumption of a normal distribution of their estimators.

mkinfit(mkinmod, observed,
  parms.ini = "auto",
  state.ini = "auto",
  err.ini = "auto",
  fixed_parms = NULL, fixed_initials = names(mkinmod$diffs)[-1],
  from_max_mean = FALSE,
  solution_type = c("auto", "analytical", "eigen", "deSolve"),
  method.ode = "lsoda",
  use_compiled = "auto",
  control = list(eval.max = 300, iter.max = 200),
  transform_rates = TRUE,
  transform_fractions = TRUE,
  quiet = FALSE,
  atol = 1e-8, rtol = 1e-10, n.outtimes = 100,
  error_model = c("const", "obs", "tc"),
  trace_parms = FALSE, ...)

Arguments

mkinmod
A list of class mkinmod, containing the kinetic model to be fitted to the data, or one of the shorthand names ("SFO", "FOMC", "DFOP", "HS", "SFORB", "IORE"). If a shorthand name is given, a parent only degradation model is generated for the variable with the highest value in observed.

observed
A dataframe with the observed data. The first column called "name" must contain the name of the observed variable for each data point. The second column must contain the times of observation, named "time". The third column must be named "value" and contain the observed values. Zero values in the "value" column will be removed, with a warning, in order to avoid problems with fitting the two-component error model. This is not expected to be a problem, because in general, values of zero are not observed in degradation data, because there is a lower limit of detection.

parms.ini
A named vector of initial values for the parameters, including parameters to be optimised and potentially also fixed parameters as indicated by fixed_parms. If set to "auto", initial values for rate constants are set to default values. Using parameter names that are not in the model gives an error.

It is possible to only specify a subset of the parameters that the model needs. You can use the parameter lists "bparms.ode" from a previously fitted model, which contains the differential equation parameters from this model. This works nicely if the models are nested. An example is given below.

state.ini
A named vector of initial values for the state variables of the model. In case the observed variables are represented by more than one model variable, the names will differ from the names of the observed variables (see map component of mkinmod). The default is to set the initial value of the first model variable to the mean of the time zero values for the variable with the maximum observed value, and all others to 0. If this variable has no time zero observations, its initial value is set to 100.

err.ini
A named vector of initial values for the error model parameters to be optimised. If set to "auto", initial values are set to default values. Otherwise, inital values for all error model parameters must be given.

fixed_parms
The names of parameters that should not be optimised but rather kept at the values specified in parms.ini.

fixed_initials
The names of model variables for which the initial state at time 0 should be excluded from the optimisation. Defaults to all state variables except for the first one.

from_max_mean
If this is set to TRUE, and the model has only one observed variable, then data before the time of the maximum observed value (after averaging for each sampling time) are discarded, and this time is subtracted from all remaining time values, so the time of the maximum observed mean value is the new time zero.

solution_type
If set to "eigen", the solution of the system of differential equations is based on the spectral decomposition of the coefficient matrix in cases that this is possible. If set to "deSolve", a numerical ode solver from package deSolve is used. If set to "analytical", an analytical solution of the model is used. This is only implemented for simple degradation experiments with only one state variable, i.e. with no metabolites. The default is "auto", which uses "analytical" if possible, otherwise "deSolve" if a compiler is present, and "eigen" if no compiler is present and the model can be expressed using eigenvalues and eigenvectors. This argument is passed on to the helper function mkinpredict.

method.ode
The solution method passed via mkinpredict to ode in case the solution type is "deSolve". The default "lsoda" is performant, but sometimes fails to converge.

use_compiled
If set to FALSE, no compiled version of the mkinmod model is used in the calls to mkinpredict even if a compiled version is present.

control
A list of control arguments passed to nlminb.

transform_rates
Boolean specifying if kinetic rate constants should be transformed in the model specification used in the fitting for better compliance with the assumption of normal distribution of the estimator. If TRUE, also alpha and beta parameters of the FOMC model are log-transformed, as well as k1 and k2 rate constants for the DFOP and HS models and the break point tb of the HS model. If FALSE, zero is used as a lower bound for the rates in the optimisation.

transform_fractions
Boolean specifying if formation fractions constants should be transformed in the model specification used in the fitting for better compliance with the assumption of normal distribution of the estimator. The default (TRUE) is to do transformations. If TRUE, the g parameter of the DFOP and HS models are also transformed, as they can also be seen as compositional data. The transformation used for these transformations is the ilr transformation.

quiet
Suppress printing out the current value of the negative log-likelihood after each improvement?

atol
Absolute error tolerance, passed to ode. Default is 1e-8, lower than in lsoda.

rtol
Absolute error tolerance, passed to ode. Default is 1e-10, much lower than in lsoda.

n.outtimes
The length of the dataseries that is produced by the model prediction function mkinpredict. This impacts the accuracy of the numerical solver if that is used (see solution_type argument. The default value is 100.

error_model
If the error model is "const", a constant standard deviation is assumed.

If the error model is "obs", each observed variable is assumed to have its own variance.

If the error model is "tc" (two-component error model), a two component error model similar to the one described by Rocke and Lorenzato (1995) is used for setting up the likelihood function. Note that this model deviates from the model by Rocke and Lorenzato, as their model implies that the errors follow a lognormal distribution for large values, not a normal distribution as assumed by this method.

trace_parms
Should a trace of the parameter values be listed?

…
Further arguments that will be passed on to deSolve.

Value

A list with "mkinfit" in the class attribute. A summary can be obtained by summary.mkinfit.

See also

Note

When using the "IORE" submodel for metabolites, fitting with "transform_rates = TRUE" (the default) often leads to failures of the numerical ODE solver. In this situation it may help to switch off the internal rate transformation.

Source

Rocke, David M. und Lorenzato, Stefan (1995) A two-component model for measurement error in analytical chemistry. Technometrics 37(2), 176-184.

Examples

# Use shorthand notation for parent only degradation
fit <- mkinfit("FOMC", FOCUS_2006_C, quiet = TRUE)
summary(fit)

#> mkin version used for fitting:    0.9.49.4 
#> R version used for fitting:       3.6.0 
#> Date of fit:     Fri May  3 19:07:19 2019 
#> Date of summary: Fri May  3 19:07:19 2019 
#> 
#> Equations:
#> d_parent/dt = - (alpha/beta) * 1/((time/beta) + 1) * parent
#> 
#> Model predictions using solution type analytical 
#> 
#> Fitted using 222 model solutions performed in 0.463 s
#> 
#> Error model:
#> Constant variance 
#> 
#> Starting values for parameters to be optimised:
#>              value   type
#> parent_0 85.100000  state
#> alpha     1.000000 deparm
#> beta     10.000000 deparm
#> sigma     1.857444  error
#> 
#> Starting values for the transformed parameters actually optimised:
#>               value lower upper
#> parent_0  85.100000  -Inf   Inf
#> log_alpha  0.000000  -Inf   Inf
#> log_beta   2.302585  -Inf   Inf
#> sigma      1.857444     0   Inf
#> 
#> Fixed parameter values:
#> None
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>           Estimate Std. Error    Lower   Upper
#> parent_0  85.87000     1.8070 81.23000 90.5200
#> log_alpha  0.05192     0.1353 -0.29580  0.3996
#> log_beta   0.65100     0.2287  0.06315  1.2390
#> sigma      1.85700     0.4378  0.73200  2.9830
#> 
#> Parameter correlation:
#>             parent_0  log_alpha   log_beta     sigma
#> parent_0   1.000e+00 -1.565e-01 -3.142e-01 4.770e-08
#> log_alpha -1.565e-01  1.000e+00  9.564e-01 9.974e-08
#> log_beta  -3.142e-01  9.564e-01  1.000e+00 8.468e-08
#> sigma      4.770e-08  9.974e-08  8.468e-08 1.000e+00
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>          Estimate t value    Pr(>t)   Lower  Upper
#> parent_0   85.870  47.530 3.893e-08 81.2300 90.520
#> alpha       1.053   7.393 3.562e-04  0.7439  1.491
#> beta        1.917   4.373 3.601e-03  1.0650  3.451
#> sigma       1.857   4.243 4.074e-03  0.7320  2.983
#> 
#> FOCUS Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   6.657       3  6
#> parent     6.657       3  6
#> 
#> Estimated disappearance times:
#>         DT50  DT90 DT50back
#> parent 1.785 15.15     4.56
#> 
#> Data:
#>  time variable observed predicted residual
#>     0   parent     85.1    85.875  -0.7749
#>     1   parent     57.9    55.191   2.7091
#>     3   parent     29.9    31.845  -1.9452
#>     7   parent     14.6    17.012  -2.4124
#>    14   parent      9.7     9.241   0.4590
#>    28   parent      6.6     4.754   1.8460
#>    63   parent      4.0     2.102   1.8977
#>    91   parent      3.9     1.441   2.4590
#>   119   parent      0.6     1.092  -0.4919

# One parent compound, one metabolite, both single first order.
# Use mkinsub for convenience in model formulation. Pathway to sink included per default.
SFO_SFO <- mkinmod(
  parent = mkinsub("SFO", "m1"),
  m1 = mkinsub("SFO"))

#> Successfully compiled differential equation model from auto-generated C code.

# Fit the model to the FOCUS example dataset D using defaults
print(system.time(fit <- mkinfit(SFO_SFO, FOCUS_2006_D,
                           solution_type = "eigen", quiet = TRUE)))

#> Warning: Observations with value of zero were removed from the data

#>        User      System verstrichen 
#>       1.493       0.000       1.494 

coef(fit)

#> NULL

endpoints(fit)

#> $ff
#> parent_sink   parent_m1     m1_sink 
#>    0.485524    0.514476    1.000000 
#> 
#> $SFORB
#> logical(0)
#> 
#> $distimes
#>              DT50      DT90
#> parent   7.022929  23.32967
#> m1     131.760712 437.69961
#> 

# deSolve is slower when no C compiler (gcc) was available during model generation
print(system.time(fit.deSolve <- mkinfit(SFO_SFO, FOCUS_2006_D,
                           solution_type = "deSolve")))

#> Warning: Observations with value of zero were removed from the data

#> Ordinary least squares optimisation

#> Sum of squared residuals at call 1: 18915.53
#> Sum of squared residuals at call 2: 18915.53
#> Sum of squared residuals at call 6: 11424.02
#> Sum of squared residuals at call 10: 11424
#> Sum of squared residuals at call 12: 4094.396
#> Sum of squared residuals at call 16: 4094.396
#> Sum of squared residuals at call 19: 1340.595
#> Sum of squared residuals at call 20: 1340.593
#> Sum of squared residuals at call 25: 1072.239
#> Sum of squared residuals at call 28: 1072.236
#> Sum of squared residuals at call 30: 874.2615
#> Sum of squared residuals at call 33: 874.2611
#> Sum of squared residuals at call 35: 616.2375
#> Sum of squared residuals at call 37: 616.237
#> Sum of squared residuals at call 40: 467.4386
#> Sum of squared residuals at call 42: 467.438
#> Sum of squared residuals at call 46: 398.2913
#> Sum of squared residuals at call 48: 398.2913
#> Sum of squared residuals at call 49: 398.2912
#> Sum of squared residuals at call 51: 395.0711
#> Sum of squared residuals at call 54: 395.071
#> Sum of squared residuals at call 56: 378.3298
#> Sum of squared residuals at call 59: 378.3298
#> Sum of squared residuals at call 62: 376.9812
#> Sum of squared residuals at call 64: 376.9811
#> Sum of squared residuals at call 67: 375.2085
#> Sum of squared residuals at call 69: 375.2085
#> Sum of squared residuals at call 70: 375.2085
#> Sum of squared residuals at call 71: 375.2085
#> Sum of squared residuals at call 72: 374.5723
#> Sum of squared residuals at call 74: 374.5723
#> Sum of squared residuals at call 77: 374.0075
#> Sum of squared residuals at call 79: 374.0075
#> Sum of squared residuals at call 80: 374.0075
#> Sum of squared residuals at call 82: 373.1711
#> Sum of squared residuals at call 84: 373.1711
#> Sum of squared residuals at call 87: 372.6445
#> Sum of squared residuals at call 88: 372.1615
#> Sum of squared residuals at call 90: 372.1615
#> Sum of squared residuals at call 91: 372.1615
#> Sum of squared residuals at call 94: 371.6464
#> Sum of squared residuals at call 99: 371.4299
#> Sum of squared residuals at call 101: 371.4299
#> Sum of squared residuals at call 104: 371.4071
#> Sum of squared residuals at call 106: 371.4071
#> Sum of squared residuals at call 107: 371.4071
#> Sum of squared residuals at call 109: 371.2524
#> Sum of squared residuals at call 113: 371.2524
#> Sum of squared residuals at call 114: 371.2136
#> Sum of squared residuals at call 115: 371.2136
#> Sum of squared residuals at call 116: 371.2136
#> Sum of squared residuals at call 119: 371.2134
#> Sum of squared residuals at call 120: 371.2134
#> Sum of squared residuals at call 122: 371.2134
#> Sum of squared residuals at call 123: 371.2134
#> Sum of squared residuals at call 125: 371.2134
#> Sum of squared residuals at call 126: 371.2134
#> Sum of squared residuals at call 135: 371.2134
#> Negative log-likelihood at call 145: 97.22429
#> Optimisation successfully terminated.
#>        User      System verstrichen 
#>       1.082       0.000       1.083 

coef(fit.deSolve)

#> NULL

endpoints(fit.deSolve)

#> $ff
#> parent_sink   parent_m1     m1_sink 
#>    0.485524    0.514476    1.000000 
#> 
#> $SFORB
#> logical(0)
#> 
#> $distimes
#>              DT50      DT90
#> parent   7.022929  23.32967
#> m1     131.760712 437.69961
#> 

# Use stepwise fitting, using optimised parameters from parent only fit, FOMC

FOMC_SFO <- mkinmod(
  parent = mkinsub("FOMC", "m1"),
  m1 = mkinsub("SFO"))

#> Successfully compiled differential equation model from auto-generated C code.

# Fit the model to the FOCUS example dataset D using defaults
fit.FOMC_SFO <- mkinfit(FOMC_SFO, FOCUS_2006_D, quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

# Use starting parameters from parent only FOMC fit
fit.FOMC = mkinfit("FOMC", FOCUS_2006_D, quiet = TRUE)
fit.FOMC_SFO <- mkinfit(FOMC_SFO, FOCUS_2006_D, quiet = TRUE,
  parms.ini = fit.FOMC$bparms.ode)

#> Warning: Observations with value of zero were removed from the data

# Use stepwise fitting, using optimised parameters from parent only fit, SFORB
SFORB_SFO <- mkinmod(
  parent = list(type = "SFORB", to = "m1", sink = TRUE),
  m1 = list(type = "SFO"))

#> Successfully compiled differential equation model from auto-generated C code.

# Fit the model to the FOCUS example dataset D using defaults
fit.SFORB_SFO <- mkinfit(SFORB_SFO, FOCUS_2006_D, quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

fit.SFORB_SFO.deSolve <- mkinfit(SFORB_SFO, FOCUS_2006_D, solution_type = "deSolve",
                                 quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

# Use starting parameters from parent only SFORB fit (not really needed in this case)
fit.SFORB = mkinfit("SFORB", FOCUS_2006_D, quiet = TRUE)
fit.SFORB_SFO <- mkinfit(SFORB_SFO, FOCUS_2006_D, parms.ini = fit.SFORB$bparms.ode, quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

# Weighted fits, including IRLS
SFO_SFO.ff <- mkinmod(parent = mkinsub("SFO", "m1"),
                      m1 = mkinsub("SFO"), use_of_ff = "max")

#> Successfully compiled differential equation model from auto-generated C code.

f.noweight <- mkinfit(SFO_SFO.ff, FOCUS_2006_D, quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

summary(f.noweight)

#> mkin version used for fitting:    0.9.49.4 
#> R version used for fitting:       3.6.0 
#> Date of fit:     Fri May  3 19:07:35 2019 
#> Date of summary: Fri May  3 19:07:35 2019 
#> 
#> Equations:
#> d_parent/dt = - k_parent * parent
#> d_m1/dt = + f_parent_to_m1 * k_parent * parent - k_m1 * m1
#> 
#> Model predictions using solution type deSolve 
#> 
#> Fitted using 421 model solutions performed in 1.099 s
#> 
#> Error model:
#> Constant variance 
#> 
#> Starting values for parameters to be optimised:
#>                     value   type
#> parent_0       100.750000  state
#> k_parent         0.100000 deparm
#> k_m1             0.100100 deparm
#> f_parent_to_m1   0.500000 deparm
#> sigma            3.125504  error
#> 
#> Starting values for the transformed parameters actually optimised:
#>                     value lower upper
#> parent_0       100.750000  -Inf   Inf
#> log_k_parent    -2.302585  -Inf   Inf
#> log_k_m1        -2.301586  -Inf   Inf
#> f_parent_ilr_1   0.000000  -Inf   Inf
#> sigma            3.125504     0   Inf
#> 
#> Fixed parameter values:
#>      value  type
#> m1_0     0 state
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                Estimate Std. Error    Lower    Upper
#> parent_0       99.60000    1.57000 96.40000 102.8000
#> log_k_parent   -2.31600    0.04087 -2.39900  -2.2330
#> log_k_m1       -5.24800    0.13320 -5.51800  -4.9770
#> f_parent_ilr_1  0.04096    0.06312 -0.08746   0.1694
#> sigma           3.12600    0.35850  2.39600   3.8550
#> 
#> Parameter correlation:
#>                  parent_0 log_k_parent   log_k_m1 f_parent_ilr_1      sigma
#> parent_0        1.000e+00    5.174e-01 -1.688e-01     -5.471e-01 -2.265e-07
#> log_k_parent    5.174e-01    1.000e+00 -3.263e-01     -5.426e-01  3.785e-07
#> log_k_m1       -1.688e-01   -3.263e-01  1.000e+00      7.478e-01 -1.386e-07
#> f_parent_ilr_1 -5.471e-01   -5.426e-01  7.478e-01      1.000e+00 -3.641e-08
#> sigma          -2.265e-07    3.785e-07 -1.386e-07     -3.641e-08  1.000e+00
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                 Estimate t value    Pr(>t)     Lower     Upper
#> parent_0       99.600000  63.430 2.298e-36 96.400000 1.028e+02
#> k_parent        0.098700  24.470 4.955e-23  0.090820 1.073e-01
#> k_m1            0.005261   7.510 6.165e-09  0.004012 6.898e-03
#> f_parent_to_m1  0.514500  23.070 3.104e-22  0.469100 5.596e-01
#> sigma           3.126000   8.718 2.235e-10  2.396000 3.855e+00
#> 
#> FOCUS Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   6.398       4 15
#> parent     6.459       2  7
#> m1         4.690       2  8
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_m1   0.5145
#> parent_sink 0.4855
#> 
#> Estimated disappearance times:
#>           DT50   DT90
#> parent   7.023  23.33
#> m1     131.761 437.70
#> 
#> Data:
#>  time variable observed predicted   residual
#>     0   parent    99.46  99.59848 -1.385e-01
#>     0   parent   102.04  99.59848  2.442e+00
#>     1   parent    93.50  90.23787  3.262e+00
#>     1   parent    92.50  90.23787  2.262e+00
#>     3   parent    63.23  74.07319 -1.084e+01
#>     3   parent    68.99  74.07319 -5.083e+00
#>     7   parent    52.32  49.91206  2.408e+00
#>     7   parent    55.13  49.91206  5.218e+00
#>    14   parent    27.27  25.01257  2.257e+00
#>    14   parent    26.64  25.01257  1.627e+00
#>    21   parent    11.50  12.53462 -1.035e+00
#>    21   parent    11.64  12.53462 -8.946e-01
#>    35   parent     2.85   3.14787 -2.979e-01
#>    35   parent     2.91   3.14787 -2.379e-01
#>    50   parent     0.69   0.71624 -2.624e-02
#>    50   parent     0.63   0.71624 -8.624e-02
#>    75   parent     0.05   0.06074 -1.074e-02
#>    75   parent     0.06   0.06074 -7.381e-04
#>     1       m1     4.84   4.80296  3.704e-02
#>     1       m1     5.64   4.80296  8.370e-01
#>     3       m1    12.91  13.02400 -1.140e-01
#>     3       m1    12.96  13.02400 -6.400e-02
#>     7       m1    22.97  25.04476 -2.075e+00
#>     7       m1    24.47  25.04476 -5.748e-01
#>    14       m1    41.69  36.69002  5.000e+00
#>    14       m1    33.21  36.69002 -3.480e+00
#>    21       m1    44.37  41.65310  2.717e+00
#>    21       m1    46.44  41.65310  4.787e+00
#>    35       m1    41.22  43.31312 -2.093e+00
#>    35       m1    37.95  43.31312 -5.363e+00
#>    50       m1    41.19  41.21831 -2.831e-02
#>    50       m1    40.01  41.21831 -1.208e+00
#>    75       m1    40.09  36.44703  3.643e+00
#>    75       m1    33.85  36.44703 -2.597e+00
#>   100       m1    31.04  31.98163 -9.416e-01
#>   100       m1    33.13  31.98163  1.148e+00
#>   120       m1    25.15  28.78984 -3.640e+00
#>   120       m1    33.31  28.78984  4.520e+00

f.obs <- mkinfit(SFO_SFO.ff, FOCUS_2006_D, error_model = "obs", quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

summary(f.obs)

#> mkin version used for fitting:    0.9.49.4 
#> R version used for fitting:       3.6.0 
#> Date of fit:     Fri May  3 19:07:37 2019 
#> Date of summary: Fri May  3 19:07:37 2019 
#> 
#> Equations:
#> d_parent/dt = - k_parent * parent
#> d_m1/dt = + f_parent_to_m1 * k_parent * parent - k_m1 * m1
#> 
#> Model predictions using solution type deSolve 
#> 
#> Fitted using 756 model solutions performed in 1.973 s
#> 
#> Error model:
#> Variance unique to each observed variable 
#> 
#> Starting values for parameters to be optimised:
#>                   value   type
#> parent_0       100.7500  state
#> k_parent         0.1000 deparm
#> k_m1             0.1001 deparm
#> f_parent_to_m1   0.5000 deparm
#> sigma_parent     3.0000  error
#> sigma_m1         3.0000  error
#> 
#> Starting values for the transformed parameters actually optimised:
#>                     value lower upper
#> parent_0       100.750000  -Inf   Inf
#> log_k_parent    -2.302585  -Inf   Inf
#> log_k_m1        -2.301586  -Inf   Inf
#> f_parent_ilr_1   0.000000  -Inf   Inf
#> sigma_parent     3.000000     0   Inf
#> sigma_m1         3.000000     0   Inf
#> 
#> Fixed parameter values:
#>      value  type
#> m1_0     0 state
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                Estimate Std. Error    Lower    Upper
#> parent_0       99.65000    1.70200 96.19000 103.1000
#> log_k_parent   -2.31300    0.04376 -2.40200  -2.2240
#> log_k_m1       -5.25000    0.12430 -5.50400  -4.9970
#> f_parent_ilr_1  0.03861    0.06171 -0.08708   0.1643
#> sigma_parent    3.40100    0.56820  2.24400   4.5590
#> sigma_m1        2.85500    0.45240  1.93400   3.7770
#> 
#> Parameter correlation:
#>                parent_0 log_k_parent log_k_m1 f_parent_ilr_1 sigma_parent
#> parent_0        1.00000      0.51078 -0.19133       -0.59997     0.035685
#> log_k_parent    0.51078      1.00000 -0.37458       -0.59239     0.069840
#> log_k_m1       -0.19133     -0.37458  1.00000        0.74398    -0.026160
#> f_parent_ilr_1 -0.59997     -0.59239  0.74398        1.00000    -0.041377
#> sigma_parent    0.03569      0.06984 -0.02616       -0.04138     1.000000
#> sigma_m1       -0.03385     -0.06626  0.02482        0.03925    -0.004628
#>                 sigma_m1
#> parent_0       -0.033847
#> log_k_parent   -0.066264
#> log_k_m1        0.024822
#> f_parent_ilr_1  0.039254
#> sigma_parent   -0.004628
#> sigma_m1        1.000000
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                 Estimate t value    Pr(>t)     Lower     Upper
#> parent_0       99.650000  58.560 2.004e-34 96.190000 1.031e+02
#> k_parent        0.098970  22.850 1.099e-21  0.090530 1.082e-01
#> k_m1            0.005245   8.046 1.732e-09  0.004072 6.756e-03
#> f_parent_to_m1  0.513600  23.560 4.352e-22  0.469300 5.578e-01
#> sigma_parent    3.401000   5.985 5.662e-07  2.244000 4.559e+00
#> sigma_m1        2.855000   6.311 2.215e-07  1.934000 3.777e+00
#> 
#> FOCUS Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   6.398       4 15
#> parent     6.464       2  7
#> m1         4.682       2  8
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_m1   0.5136
#> parent_sink 0.4864
#> 
#> Estimated disappearance times:
#>           DT50   DT90
#> parent   7.003  23.26
#> m1     132.154 439.01
#> 
#> Data:
#>  time variable observed predicted   residual
#>     0   parent    99.46  99.65425 -1.942e-01
#>     0   parent   102.04  99.65425  2.386e+00
#>     1   parent    93.50  90.26338  3.237e+00
#>     1   parent    92.50  90.26338  2.237e+00
#>     3   parent    63.23  74.05309 -1.082e+01
#>     3   parent    68.99  74.05309 -5.063e+00
#>     7   parent    52.32  49.84326  2.477e+00
#>     7   parent    55.13  49.84326  5.287e+00
#>    14   parent    27.27  24.92970  2.340e+00
#>    14   parent    26.64  24.92970  1.710e+00
#>    21   parent    11.50  12.46888 -9.689e-01
#>    21   parent    11.64  12.46888 -8.289e-01
#>    35   parent     2.85   3.11924 -2.692e-01
#>    35   parent     2.91   3.11924 -2.092e-01
#>    50   parent     0.69   0.70678 -1.678e-02
#>    50   parent     0.63   0.70678 -7.678e-02
#>    75   parent     0.05   0.05952 -9.522e-03
#>    75   parent     0.06   0.05952  4.776e-04
#>     1       m1     4.84   4.81075  2.925e-02
#>     1       m1     5.64   4.81075  8.292e-01
#>     3       m1    12.91  13.04197 -1.320e-01
#>     3       m1    12.96  13.04197 -8.197e-02
#>     7       m1    22.97  25.06848 -2.098e+00
#>     7       m1    24.47  25.06848 -5.985e-01
#>    14       m1    41.69  36.70308  4.987e+00
#>    14       m1    33.21  36.70308 -3.493e+00
#>    21       m1    44.37  41.65115  2.719e+00
#>    21       m1    46.44  41.65115  4.789e+00
#>    35       m1    41.22  43.29464 -2.075e+00
#>    35       m1    37.95  43.29464 -5.345e+00
#>    50       m1    41.19  41.19947 -9.473e-03
#>    50       m1    40.01  41.19947 -1.189e+00
#>    75       m1    40.09  36.44035  3.650e+00
#>    75       m1    33.85  36.44035 -2.590e+00
#>   100       m1    31.04  31.98773 -9.477e-01
#>   100       m1    33.13  31.98773  1.142e+00
#>   120       m1    25.15  28.80430 -3.654e+00
#>   120       m1    33.31  28.80430  4.506e+00

f.tc <- mkinfit(SFO_SFO.ff, FOCUS_2006_D, error_model = "tc", quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

summary(f.tc)

#> Warning: NaNs wurden erzeugt

#> Warning: NaNs wurden erzeugt

#> Warning: NaNs wurden erzeugt

#> Warning: diag(.) had 0 or NA entries; non-finite result is doubtful

#> mkin version used for fitting:    0.9.49.4 
#> R version used for fitting:       3.6.0 
#> Date of fit:     Fri May  3 19:07:40 2019 
#> Date of summary: Fri May  3 19:07:40 2019 
#> 
#> Equations:
#> d_parent/dt = - k_parent * parent
#> d_m1/dt = + f_parent_to_m1 * k_parent * parent - k_m1 * m1
#> 
#> Model predictions using solution type deSolve 
#> 
#> Fitted using 888 model solutions performed in 3.579 s
#> 
#> Error model:
#> Two-component variance function 
#> 
#> Starting values for parameters to be optimised:
#>                   value   type
#> parent_0       100.7500  state
#> k_parent         0.1000 deparm
#> k_m1             0.1001 deparm
#> f_parent_to_m1   0.5000 deparm
#> sigma_low        0.1000  error
#> rsd_high         0.1000  error
#> 
#> Starting values for the transformed parameters actually optimised:
#>                     value lower upper
#> parent_0       100.750000  -Inf   Inf
#> log_k_parent    -2.302585  -Inf   Inf
#> log_k_m1        -2.301586  -Inf   Inf
#> f_parent_ilr_1   0.000000  -Inf   Inf
#> sigma_low        0.100000     0   Inf
#> rsd_high         0.100000     0   Inf
#> 
#> Fixed parameter values:
#>      value  type
#> m1_0     0 state
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                 Estimate Std. Error    Lower     Upper
#> parent_0       100.30000    3.06400 94.08000 106.60000
#> log_k_parent    -2.31100    0.02913 -2.37100  -2.25200
#> log_k_m1        -5.27800    0.10890 -5.49900  -5.05600
#> f_parent_ilr_1   0.02128    0.07223 -0.12590   0.16840
#> sigma_low        0.99650        NaN      NaN       NaN
#> rsd_high         0.07560    0.01059  0.05402   0.09717
#> 
#> Parameter correlation:
#>                parent_0 log_k_parent  log_k_m1 f_parent_ilr_1 sigma_low
#> parent_0        1.00000      0.64703 -0.220160       -0.79035       NaN
#> log_k_parent    0.64703      1.00000 -0.339633       -0.67982       NaN
#> log_k_m1       -0.22016     -0.33963  1.000000        0.62815       NaN
#> f_parent_ilr_1 -0.79035     -0.67982  0.628148        1.00000       NaN
#> sigma_low           NaN          NaN       NaN            NaN         1
#> rsd_high       -0.05745     -0.01123  0.003502        0.00568       NaN
#>                 rsd_high
#> parent_0       -0.057450
#> log_k_parent   -0.011230
#> log_k_m1        0.003502
#> f_parent_ilr_1  0.005680
#> sigma_low            NaN
#> rsd_high        1.000000
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                 Estimate t value    Pr(>t)     Lower     Upper
#> parent_0       1.003e+02  32.740 1.759e-26 94.080000 1.066e+02
#> k_parent       9.914e-02  34.330 4.045e-27  0.093430 1.052e-01
#> k_m1           5.105e-03   9.186 8.682e-11  0.004089 6.372e-03
#> f_parent_to_m1 5.075e-01  19.880 7.143e-20  0.455600 5.593e-01
#> sigma_low      9.965e-01     NaN       NaN       NaN       NaN
#> rsd_high       7.560e-02   7.137 2.114e-08  0.054020 9.717e-02
#> 
#> FOCUS Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   6.433       4 15
#> parent     6.506       2  7
#> m1         4.692       2  8
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_m1   0.5075
#> parent_sink 0.4925
#> 
#> Estimated disappearance times:
#>           DT50   DT90
#> parent   6.992  23.23
#> m1     135.787 451.08
#> 
#> Data:
#>  time variable observed predicted   residual
#>     0   parent    99.46 100.32122  -0.861220
#>     0   parent   102.04 100.32122   1.718780
#>     1   parent    93.50  90.85266   2.647340
#>     1   parent    92.50  90.85266   1.647340
#>     3   parent    63.23  74.51219 -11.282190
#>     3   parent    68.99  74.51219  -5.522190
#>     7   parent    52.32  50.11950   2.200504
#>     7   parent    55.13  50.11950   5.010504
#>    14   parent    27.27  25.03921   2.230792
#>    14   parent    26.64  25.03921   1.600792
#>    21   parent    11.50  12.50934  -1.009342
#>    21   parent    11.64  12.50934  -0.869342
#>    35   parent     2.85   3.12221  -0.272211
#>    35   parent     2.91   3.12221  -0.212211
#>    50   parent     0.69   0.70572  -0.015724
#>    50   parent     0.63   0.70572  -0.075724
#>    75   parent     0.05   0.05919  -0.009191
#>    75   parent     0.06   0.05919   0.000809
#>     1       m1     4.84   4.79307   0.046928
#>     1       m1     5.64   4.79307   0.846928
#>     3       m1    12.91  12.99398  -0.083980
#>     3       m1    12.96  12.99398  -0.033980
#>     7       m1    22.97  24.97744  -2.007441
#>     7       m1    24.47  24.97744  -0.507441
#>    14       m1    41.69  36.57917   5.110827
#>    14       m1    33.21  36.57917  -3.369173
#>    21       m1    44.37  41.52931   2.840692
#>    21       m1    46.44  41.52931   4.910692
#>    35       m1    41.22  43.22614  -2.006138
#>    35       m1    37.95  43.22614  -5.276138
#>    50       m1    41.19  41.20972  -0.019717
#>    50       m1    40.01  41.20972  -1.199717
#>    75       m1    40.09  36.57312   3.516882
#>    75       m1    33.85  36.57312  -2.723118
#>   100       m1    31.04  32.21655  -1.176546
#>   100       m1    33.13  32.21655   0.913454
#>   120       m1    25.15  29.09181  -3.941814
#>   120       m1    33.31  29.09181   4.218186