Fit a kinetic model to data with one or more state variables

This function maximises the likelihood of the observed data using the Port algorithm stats::nlminb(), and the specified initial or fixed parameters and starting values. In each step of the optimisation, the kinetic model is solved using the function mkinpredict(), except if an analytical solution is implemented, in which case the model is solved using the degradation function in the mkinmod object. The parameters of the selected error model are fitted simultaneously with the degradation model parameters, as both of them are arguments of the likelihood function.

mkinfit(
  mkinmod,
  observed,
  parms.ini = "auto",
  state.ini = "auto",
  err.ini = "auto",
  fixed_parms = NULL,
  fixed_initials = names(mkinmod$diffs)[-1],
  from_max_mean = FALSE,
  solution_type = c("auto", "analytical", "eigen", "deSolve"),
  method.ode = "lsoda",
  use_compiled = "auto",
  control = list(eval.max = 300, iter.max = 200),
  transform_rates = TRUE,
  transform_fractions = TRUE,
  quiet = FALSE,
  atol = 1e-08,
  rtol = 1e-10,
  error_model = c("const", "obs", "tc"),
  error_model_algorithm = c("auto", "d_3", "direct", "twostep", "threestep",
    "fourstep", "IRLS", "OLS"),
  reweight.tol = 1e-08,
  reweight.max.iter = 10,
  trace_parms = FALSE,
  ...
)

Arguments

mkinmod
A list of class mkinmod, containing the kinetic model to be fitted to the data, or one of the shorthand names ("SFO", "FOMC", "DFOP", "HS", "SFORB", "IORE"). If a shorthand name is given, a parent only degradation model is generated for the variable with the highest value in observed.

observed
A dataframe with the observed data. The first column called "name" must contain the name of the observed variable for each data point. The second column must contain the times of observation, named "time". The third column must be named "value" and contain the observed values. Zero values in the "value" column will be removed, with a warning, in order to avoid problems with fitting the two-component error model. This is not expected to be a problem, because in general, values of zero are not observed in degradation data, because there is a lower limit of detection.

parms.ini
A named vector of initial values for the parameters, including parameters to be optimised and potentially also fixed parameters as indicated by fixed_parms. If set to "auto", initial values for rate constants are set to default values. Using parameter names that are not in the model gives an error.

It is possible to only specify a subset of the parameters that the model needs. You can use the parameter lists "bparms.ode" from a previously fitted model, which contains the differential equation parameters from this model. This works nicely if the models are nested. An example is given below.

state.ini
A named vector of initial values for the state variables of the model. In case the observed variables are represented by more than one model variable, the names will differ from the names of the observed variables (see map component of mkinmod). The default is to set the initial value of the first model variable to the mean of the time zero values for the variable with the maximum observed value, and all others to 0. If this variable has no time zero observations, its initial value is set to 100.

err.ini
A named vector of initial values for the error model parameters to be optimised. If set to "auto", initial values are set to default values. Otherwise, inital values for all error model parameters must be given.

fixed_parms
The names of parameters that should not be optimised but rather kept at the values specified in parms.ini. Alternatively, a named numeric vector of parameters to be fixed, regardless of the values in parms.ini.

fixed_initials
The names of model variables for which the initial state at time 0 should be excluded from the optimisation. Defaults to all state variables except for the first one.

from_max_mean
If this is set to TRUE, and the model has only one observed variable, then data before the time of the maximum observed value (after averaging for each sampling time) are discarded, and this time is subtracted from all remaining time values, so the time of the maximum observed mean value is the new time zero.

solution_type
If set to "eigen", the solution of the system of differential equations is based on the spectral decomposition of the coefficient matrix in cases that this is possible. If set to "deSolve", a numerical ode solver from package deSolve is used. If set to "analytical", an analytical solution of the model is used. This is only implemented for relatively simple degradation models. The default is "auto", which uses "analytical" if possible, otherwise "deSolve" if a compiler is present, and "eigen" if no compiler is present and the model can be expressed using eigenvalues and eigenvectors.

method.ode
The solution method passed via mkinpredict() to deSolve::ode() in case the solution type is "deSolve". The default "lsoda" is performant, but sometimes fails to converge.

use_compiled
If set to FALSE, no compiled version of the mkinmod model is used in the calls to mkinpredict() even if a compiled version is present.

control
A list of control arguments passed to stats::nlminb().

transform_rates
Boolean specifying if kinetic rate constants should be transformed in the model specification used in the fitting for better compliance with the assumption of normal distribution of the estimator. If TRUE, also alpha and beta parameters of the FOMC model are log-transformed, as well as k1 and k2 rate constants for the DFOP and HS models and the break point tb of the HS model. If FALSE, zero is used as a lower bound for the rates in the optimisation.

transform_fractions
Boolean specifying if formation fractions constants should be transformed in the model specification used in the fitting for better compliance with the assumption of normal distribution of the estimator. The default (TRUE) is to do transformations. If TRUE, the g parameter of the DFOP and HS models are also transformed, as they can also be seen as compositional data. The transformation used for these transformations is the ilr() transformation.

quiet
Suppress printing out the current value of the negative log-likelihood after each improvement?

atol
Absolute error tolerance, passed to deSolve::ode(). Default is 1e-8, which is lower than the default in the deSolve::lsoda() function which is used per default.

rtol
Absolute error tolerance, passed to deSolve::ode(). Default is 1e-10, much lower than in deSolve::lsoda().

error_model
If the error model is "const", a constant standard deviation is assumed.

If the error model is "obs", each observed variable is assumed to have its own variance.

If the error model is "tc" (two-component error model), a two component error model similar to the one described by Rocke and Lorenzato (1995) is used for setting up the likelihood function. Note that this model deviates from the model by Rocke and Lorenzato, as their model implies that the errors follow a lognormal distribution for large values, not a normal distribution as assumed by this method.

error_model_algorithm
If "auto", the selected algorithm depends on the error model. If the error model is "const", unweighted nonlinear least squares fitting ("OLS") is selected. If the error model is "obs", or "tc", the "d_3" algorithm is selected.

The algorithm "d_3" will directly minimize the negative log-likelihood and independently also use the three step algorithm described below. The fit with the higher likelihood is returned.

The algorithm "direct" will directly minimize the negative log-likelihood.

The algorithm "twostep" will minimize the negative log-likelihood after an initial unweighted least squares optimisation step.

The algorithm "threestep" starts with unweighted least squares, then optimizes only the error model using the degradation model parameters found, and then minimizes the negative log-likelihood with free degradation and error model parameters.

The algorithm "fourstep" starts with unweighted least squares, then optimizes only the error model using the degradation model parameters found, then optimizes the degradation model again with fixed error model parameters, and finally minimizes the negative log-likelihood with free degradation and error model parameters.

The algorithm "IRLS" (Iteratively Reweighted Least Squares) starts with unweighted least squares, and then iterates optimization of the error model parameters and subsequent optimization of the degradation model using those error model parameters, until the error model parameters converge.

reweight.tol
Tolerance for the convergence criterion calculated from the error model parameters in IRLS fits.

reweight.max.iter
Maximum number of iterations in IRLS fits.

trace_parms
Should a trace of the parameter values be listed?

...
Further arguments that will be passed on to deSolve::ode().

Value

A list with "mkinfit" in the class attribute.

Details

Per default, parameters in the kinetic models are internally transformed in order to better satisfy the assumption of a normal distribution of their estimators.

Note

When using the "IORE" submodel for metabolites, fitting with "transform_rates = TRUE" (the default) often leads to failures of the numerical ODE solver. In this situation it may help to switch off the internal rate transformation.

References

Rocke DM and Lorenzato S (1995) A two-component model for measurement error in analytical chemistry. Technometrics 37(2), 176-184.

Ranke J and Meinecke S (2019) Error Models for the Kinetic Evaluation of Chemical Degradation Data. Environments 6(12) 124 doi:10.3390/environments6120124.

See also

Examples

# Use shorthand notation for parent only degradation
fit <- mkinfit("FOMC", FOCUS_2006_C, quiet = TRUE)
summary(fit)

#> mkin version used for fitting:    0.9.50.3 
#> R version used for fitting:       4.0.0 
#> Date of fit:     Fri May 15 07:30:32 2020 
#> Date of summary: Fri May 15 07:30:32 2020 
#> 
#> Equations:
#> d_parent/dt = - (alpha/beta) * 1/((time/beta) + 1) * parent
#> 
#> Model predictions using solution type analytical 
#> 
#> Fitted using 222 model solutions performed in 0.048 s
#> 
#> Error model: Constant variance 
#> 
#> Error model algorithm: OLS 
#> 
#> Starting values for parameters to be optimised:
#>          value   type
#> parent_0  85.1  state
#> alpha      1.0 deparm
#> beta      10.0 deparm
#> 
#> Starting values for the transformed parameters actually optimised:
#>               value lower upper
#> parent_0  85.100000  -Inf   Inf
#> log_alpha  0.000000  -Inf   Inf
#> log_beta   2.302585  -Inf   Inf
#> 
#> Fixed parameter values:
#> None
#> 
#> Results:
#> 
#>        AIC      BIC    logLik
#>   44.68652 45.47542 -18.34326
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>           Estimate Std. Error    Lower   Upper
#> parent_0  85.87000     1.8070 81.23000 90.5200
#> log_alpha  0.05192     0.1353 -0.29580  0.3996
#> log_beta   0.65100     0.2287  0.06315  1.2390
#> sigma      1.85700     0.4378  0.73200  2.9830
#> 
#> Parameter correlation:
#>             parent_0  log_alpha   log_beta     sigma
#> parent_0   1.000e+00 -1.565e-01 -3.142e-01 4.758e-08
#> log_alpha -1.565e-01  1.000e+00  9.564e-01 1.007e-07
#> log_beta  -3.142e-01  9.564e-01  1.000e+00 8.568e-08
#> sigma      4.758e-08  1.007e-07  8.568e-08 1.000e+00
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>          Estimate t value    Pr(>t)   Lower  Upper
#> parent_0   85.870  47.530 3.893e-08 81.2300 90.520
#> alpha       1.053   7.393 3.562e-04  0.7439  1.491
#> beta        1.917   4.373 3.601e-03  1.0650  3.451
#> sigma       1.857   4.243 4.074e-03  0.7320  2.983
#> 
#> FOCUS Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   6.657       3  6
#> parent     6.657       3  6
#> 
#> Estimated disappearance times:
#>         DT50  DT90 DT50back
#> parent 1.785 15.15     4.56
#> 
#> Data:
#>  time variable observed predicted residual
#>     0   parent     85.1    85.875  -0.7749
#>     1   parent     57.9    55.191   2.7091
#>     3   parent     29.9    31.845  -1.9452
#>     7   parent     14.6    17.012  -2.4124
#>    14   parent      9.7     9.241   0.4590
#>    28   parent      6.6     4.754   1.8460
#>    63   parent      4.0     2.102   1.8977
#>    91   parent      3.9     1.441   2.4590
#>   119   parent      0.6     1.092  -0.4919

# One parent compound, one metabolite, both single first order.
# Use mkinsub for convenience in model formulation. Pathway to sink included per default.
SFO_SFO <- mkinmod(
  parent = mkinsub("SFO", "m1"),
  m1 = mkinsub("SFO"))

#> Successfully compiled differential equation model from auto-generated C code.

# Fit the model to the FOCUS example dataset D using defaults
print(system.time(fit <- mkinfit(SFO_SFO, FOCUS_2006_D,
                           solution_type = "eigen", quiet = TRUE)))

#> Warning: Observations with value of zero were removed from the data

#>    user  system elapsed 
#>   0.413   0.005   0.419 

parms(fit)

#>       parent_0       k_parent           k_m1 f_parent_to_m1          sigma 
#>   99.598483222    0.098697734    0.005260651    0.514475962    3.125503875 

endpoints(fit)

#> $ff
#>   parent_m1 parent_sink 
#>    0.514476    0.485524 
#> 
#> $distimes
#>              DT50      DT90
#> parent   7.022929  23.32967
#> m1     131.760715 437.69962
#> 

# \dontrun{
# deSolve is slower when no C compiler (gcc) was available during model generation
print(system.time(fit.deSolve <- mkinfit(SFO_SFO, FOCUS_2006_D,
                           solution_type = "deSolve")))

#> Warning: Observations with value of zero were removed from the data

#> Ordinary least squares optimisation

#> Sum of squared residuals at call 1: 15156.12
#> Sum of squared residuals at call 2: 15156.12
#> Sum of squared residuals at call 6: 8243.645
#> Sum of squared residuals at call 12: 6290.712
#> Sum of squared residuals at call 13: 6290.683
#> Sum of squared residuals at call 15: 6290.452
#> Sum of squared residuals at call 18: 1700.749
#> Sum of squared residuals at call 20: 1700.611
#> Sum of squared residuals at call 24: 1190.923
#> Sum of squared residuals at call 26: 1190.922
#> Sum of squared residuals at call 29: 1017.417
#> Sum of squared residuals at call 31: 1017.417
#> Sum of squared residuals at call 33: 1017.416
#> Sum of squared residuals at call 34: 644.0471
#> Sum of squared residuals at call 36: 644.0469
#> Sum of squared residuals at call 38: 644.0469
#> Sum of squared residuals at call 39: 590.5025
#> Sum of squared residuals at call 41: 590.5022
#> Sum of squared residuals at call 43: 590.5016
#> Sum of squared residuals at call 44: 543.219
#> Sum of squared residuals at call 45: 543.2187
#> Sum of squared residuals at call 46: 543.2186
#> Sum of squared residuals at call 50: 391.348
#> Sum of squared residuals at call 51: 391.3479
#> Sum of squared residuals at call 56: 386.4789
#> Sum of squared residuals at call 58: 386.4789
#> Sum of squared residuals at call 60: 386.4779
#> Sum of squared residuals at call 61: 384.0686
#> Sum of squared residuals at call 63: 384.0686
#> Sum of squared residuals at call 66: 382.7813
#> Sum of squared residuals at call 68: 382.7813
#> Sum of squared residuals at call 70: 382.7813
#> Sum of squared residuals at call 71: 378.9273
#> Sum of squared residuals at call 73: 378.9273
#> Sum of squared residuals at call 75: 378.9272
#> Sum of squared residuals at call 76: 377.4847
#> Sum of squared residuals at call 78: 377.4846
#> Sum of squared residuals at call 81: 375.9738
#> Sum of squared residuals at call 83: 375.9738
#> Sum of squared residuals at call 86: 375.3387
#> Sum of squared residuals at call 88: 375.3387
#> Sum of squared residuals at call 91: 374.5774
#> Sum of squared residuals at call 93: 374.5774
#> Sum of squared residuals at call 95: 374.5774
#> Sum of squared residuals at call 96: 373.5433
#> Sum of squared residuals at call 100: 373.5433
#> Sum of squared residuals at call 102: 373.2654
#> Sum of squared residuals at call 104: 373.2654
#> Sum of squared residuals at call 107: 372.6841
#> Sum of squared residuals at call 111: 372.684
#> Sum of squared residuals at call 114: 372.6374
#> Sum of squared residuals at call 116: 372.6374
#> Sum of squared residuals at call 119: 372.6223
#> Sum of squared residuals at call 121: 372.6223
#> Sum of squared residuals at call 123: 372.6223
#> Sum of squared residuals at call 124: 372.5903
#> Sum of squared residuals at call 126: 372.5903
#> Sum of squared residuals at call 129: 372.5445
#> Sum of squared residuals at call 130: 372.4921
#> Sum of squared residuals at call 131: 372.2377
#> Sum of squared residuals at call 132: 371.5434
#> Sum of squared residuals at call 134: 371.5434
#> Sum of squared residuals at call 137: 371.2857
#> Sum of squared residuals at call 139: 371.2857
#> Sum of squared residuals at call 143: 371.2247
#> Sum of squared residuals at call 144: 371.2247
#> Sum of squared residuals at call 149: 371.2189
#> Sum of squared residuals at call 150: 371.2145
#> Sum of squared residuals at call 153: 371.2145
#> Sum of squared residuals at call 155: 371.2138
#> Sum of squared residuals at call 156: 371.2138
#> Sum of squared residuals at call 157: 371.2138
#> Sum of squared residuals at call 161: 371.2134
#> Sum of squared residuals at call 162: 371.2134
#> Sum of squared residuals at call 165: 371.2134
#> Sum of squared residuals at call 166: 371.2134
#> Sum of squared residuals at call 168: 371.2134
#> Negative log-likelihood at call 178: 97.22429

#> Optimisation successfully terminated.

#>    user  system elapsed 
#>   0.354   0.000   0.354 

parms(fit.deSolve)

#>       parent_0       k_parent           k_m1 f_parent_to_m1          sigma 
#>   99.598480759    0.098697739    0.005260651    0.514475958    3.125503874 

endpoints(fit.deSolve)

#> $ff
#>   parent_m1 parent_sink 
#>    0.514476    0.485524 
#> 
#> $distimes
#>              DT50      DT90
#> parent   7.022929  23.32966
#> m1     131.760731 437.69967
#> 

# }

# Use stepwise fitting, using optimised parameters from parent only fit, FOMC
# \dontrun{
FOMC_SFO <- mkinmod(
  parent = mkinsub("FOMC", "m1"),
  m1 = mkinsub("SFO"))

#> Successfully compiled differential equation model from auto-generated C code.

# Fit the model to the FOCUS example dataset D using defaults
fit.FOMC_SFO <- mkinfit(FOMC_SFO, FOCUS_2006_D, quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

# Use starting parameters from parent only FOMC fit
fit.FOMC = mkinfit("FOMC", FOCUS_2006_D, quiet = TRUE)
fit.FOMC_SFO <- mkinfit(FOMC_SFO, FOCUS_2006_D, quiet = TRUE,
  parms.ini = fit.FOMC$bparms.ode)

#> Warning: Observations with value of zero were removed from the data

# Use stepwise fitting, using optimised parameters from parent only fit, SFORB
SFORB_SFO <- mkinmod(
  parent = list(type = "SFORB", to = "m1", sink = TRUE),
  m1 = list(type = "SFO"))

#> Successfully compiled differential equation model from auto-generated C code.

# Fit the model to the FOCUS example dataset D using defaults
fit.SFORB_SFO <- mkinfit(SFORB_SFO, FOCUS_2006_D, quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

fit.SFORB_SFO.deSolve <- mkinfit(SFORB_SFO, FOCUS_2006_D, solution_type = "deSolve",
                                 quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

# Use starting parameters from parent only SFORB fit (not really needed in this case)
fit.SFORB = mkinfit("SFORB", FOCUS_2006_D, quiet = TRUE)
fit.SFORB_SFO <- mkinfit(SFORB_SFO, FOCUS_2006_D, parms.ini = fit.SFORB$bparms.ode, quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

#> Warning: Initial parameter(s) k_parent_free_sink not used in the model

# }

# \dontrun{
# Weighted fits, including IRLS (error_model = "obs")
SFO_SFO.ff <- mkinmod(parent = mkinsub("SFO", "m1"),
                      m1 = mkinsub("SFO"), use_of_ff = "max")

#> Successfully compiled differential equation model from auto-generated C code.

f.noweight <- mkinfit(SFO_SFO.ff, FOCUS_2006_D, quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

summary(f.noweight)

#> mkin version used for fitting:    0.9.50.3 
#> R version used for fitting:       4.0.0 
#> Date of fit:     Fri May 15 07:30:37 2020 
#> Date of summary: Fri May 15 07:30:37 2020 
#> 
#> Equations:
#> d_parent/dt = - k_parent * parent
#> d_m1/dt = + f_parent_to_m1 * k_parent * parent - k_m1 * m1
#> 
#> Model predictions using solution type analytical 
#> 
#> Fitted using 421 model solutions performed in 0.147 s
#> 
#> Error model: Constant variance 
#> 
#> Error model algorithm: OLS 
#> 
#> Starting values for parameters to be optimised:
#>                   value   type
#> parent_0       100.7500  state
#> k_parent         0.1000 deparm
#> k_m1             0.1001 deparm
#> f_parent_to_m1   0.5000 deparm
#> 
#> Starting values for the transformed parameters actually optimised:
#>                     value lower upper
#> parent_0       100.750000  -Inf   Inf
#> log_k_parent    -2.302585  -Inf   Inf
#> log_k_m1        -2.301586  -Inf   Inf
#> f_parent_ilr_1   0.000000  -Inf   Inf
#> 
#> Fixed parameter values:
#>      value  type
#> m1_0     0 state
#> 
#> Results:
#> 
#>        AIC      BIC    logLik
#>   204.4486 212.6365 -97.22429
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                Estimate Std. Error    Lower    Upper
#> parent_0       99.60000    1.57000 96.40000 102.8000
#> log_k_parent   -2.31600    0.04087 -2.39900  -2.2330
#> log_k_m1       -5.24800    0.13320 -5.51800  -4.9770
#> f_parent_ilr_1  0.04096    0.06312 -0.08746   0.1694
#> sigma           3.12600    0.35850  2.39600   3.8550
#> 
#> Parameter correlation:
#>                  parent_0 log_k_parent   log_k_m1 f_parent_ilr_1      sigma
#> parent_0        1.000e+00    5.174e-01 -1.688e-01     -5.471e-01 -3.214e-07
#> log_k_parent    5.174e-01    1.000e+00 -3.263e-01     -5.426e-01  3.168e-07
#> log_k_m1       -1.688e-01   -3.263e-01  1.000e+00      7.478e-01 -1.410e-07
#> f_parent_ilr_1 -5.471e-01   -5.426e-01  7.478e-01      1.000e+00  5.093e-10
#> sigma          -3.214e-07    3.168e-07 -1.410e-07      5.093e-10  1.000e+00
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                 Estimate t value    Pr(>t)     Lower     Upper
#> parent_0       99.600000  63.430 2.298e-36 96.400000 1.028e+02
#> k_parent        0.098700  24.470 4.955e-23  0.090820 1.073e-01
#> k_m1            0.005261   7.510 6.165e-09  0.004012 6.898e-03
#> f_parent_to_m1  0.514500  23.070 3.104e-22  0.469100 5.596e-01
#> sigma           3.126000   8.718 2.235e-10  2.396000 3.855e+00
#> 
#> FOCUS Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   6.398       4 15
#> parent     6.459       2  7
#> m1         4.690       2  8
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_m1   0.5145
#> parent_sink 0.4855
#> 
#> Estimated disappearance times:
#>           DT50   DT90
#> parent   7.023  23.33
#> m1     131.761 437.70
#> 
#> Data:
#>  time variable observed predicted   residual
#>     0   parent    99.46  99.59848 -1.385e-01
#>     0   parent   102.04  99.59848  2.442e+00
#>     1   parent    93.50  90.23787  3.262e+00
#>     1   parent    92.50  90.23787  2.262e+00
#>     3   parent    63.23  74.07319 -1.084e+01
#>     3   parent    68.99  74.07319 -5.083e+00
#>     7   parent    52.32  49.91206  2.408e+00
#>     7   parent    55.13  49.91206  5.218e+00
#>    14   parent    27.27  25.01257  2.257e+00
#>    14   parent    26.64  25.01257  1.627e+00
#>    21   parent    11.50  12.53462 -1.035e+00
#>    21   parent    11.64  12.53462 -8.946e-01
#>    35   parent     2.85   3.14787 -2.979e-01
#>    35   parent     2.91   3.14787 -2.379e-01
#>    50   parent     0.69   0.71624 -2.624e-02
#>    50   parent     0.63   0.71624 -8.624e-02
#>    75   parent     0.05   0.06074 -1.074e-02
#>    75   parent     0.06   0.06074 -7.381e-04
#>     1       m1     4.84   4.80296  3.704e-02
#>     1       m1     5.64   4.80296  8.370e-01
#>     3       m1    12.91  13.02400 -1.140e-01
#>     3       m1    12.96  13.02400 -6.400e-02
#>     7       m1    22.97  25.04476 -2.075e+00
#>     7       m1    24.47  25.04476 -5.748e-01
#>    14       m1    41.69  36.69002  5.000e+00
#>    14       m1    33.21  36.69002 -3.480e+00
#>    21       m1    44.37  41.65310  2.717e+00
#>    21       m1    46.44  41.65310  4.787e+00
#>    35       m1    41.22  43.31312 -2.093e+00
#>    35       m1    37.95  43.31312 -5.363e+00
#>    50       m1    41.19  41.21831 -2.831e-02
#>    50       m1    40.01  41.21831 -1.208e+00
#>    75       m1    40.09  36.44703  3.643e+00
#>    75       m1    33.85  36.44703 -2.597e+00
#>   100       m1    31.04  31.98163 -9.416e-01
#>   100       m1    33.13  31.98163  1.148e+00
#>   120       m1    25.15  28.78984 -3.640e+00
#>   120       m1    33.31  28.78984  4.520e+00

f.obs <- mkinfit(SFO_SFO.ff, FOCUS_2006_D, error_model = "obs", quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

summary(f.obs)

#> mkin version used for fitting:    0.9.50.3 
#> R version used for fitting:       4.0.0 
#> Date of fit:     Fri May 15 07:30:38 2020 
#> Date of summary: Fri May 15 07:30:38 2020 
#> 
#> Equations:
#> d_parent/dt = - k_parent * parent
#> d_m1/dt = + f_parent_to_m1 * k_parent * parent - k_m1 * m1
#> 
#> Model predictions using solution type analytical 
#> 
#> Fitted using 978 model solutions performed in 0.337 s
#> 
#> Error model: Variance unique to each observed variable 
#> 
#> Error model algorithm: d_3 
#> Direct fitting and three-step fitting yield approximately the same likelihood 
#> 
#> Starting values for parameters to be optimised:
#>                   value   type
#> parent_0       100.7500  state
#> k_parent         0.1000 deparm
#> k_m1             0.1001 deparm
#> f_parent_to_m1   0.5000 deparm
#> sigma_parent     3.0000  error
#> sigma_m1         3.0000  error
#> 
#> Starting values for the transformed parameters actually optimised:
#>                     value lower upper
#> parent_0       100.750000  -Inf   Inf
#> log_k_parent    -2.302585  -Inf   Inf
#> log_k_m1        -2.301586  -Inf   Inf
#> f_parent_ilr_1   0.000000  -Inf   Inf
#> sigma_parent     3.000000     0   Inf
#> sigma_m1         3.000000     0   Inf
#> 
#> Fixed parameter values:
#>      value  type
#> m1_0     0 state
#> 
#> Results:
#> 
#>        AIC      BIC    logLik
#>   205.8727 215.6982 -96.93634
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                Estimate Std. Error    Lower    Upper
#> parent_0       99.65000    1.70200 96.19000 103.1000
#> log_k_parent   -2.31300    0.04376 -2.40200  -2.2240
#> log_k_m1       -5.25000    0.12430 -5.50400  -4.9970
#> f_parent_ilr_1  0.03861    0.06171 -0.08708   0.1643
#> sigma_parent    3.40100    0.56820  2.24400   4.5590
#> sigma_m1        2.85500    0.45240  1.93400   3.7770
#> 
#> Parameter correlation:
#>                parent_0 log_k_parent log_k_m1 f_parent_ilr_1 sigma_parent
#> parent_0        1.00000      0.51078 -0.19133       -0.59997     0.035670
#> log_k_parent    0.51078      1.00000 -0.37458       -0.59239     0.069833
#> log_k_m1       -0.19133     -0.37458  1.00000        0.74398    -0.026158
#> f_parent_ilr_1 -0.59997     -0.59239  0.74398        1.00000    -0.041369
#> sigma_parent    0.03567      0.06983 -0.02616       -0.04137     1.000000
#> sigma_m1       -0.03385     -0.06627  0.02482        0.03926    -0.004628
#>                 sigma_m1
#> parent_0       -0.033847
#> log_k_parent   -0.066265
#> log_k_m1        0.024823
#> f_parent_ilr_1  0.039256
#> sigma_parent   -0.004628
#> sigma_m1        1.000000
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                 Estimate t value    Pr(>t)     Lower     Upper
#> parent_0       99.650000  58.560 2.004e-34 96.190000 1.031e+02
#> k_parent        0.098970  22.850 1.099e-21  0.090530 1.082e-01
#> k_m1            0.005245   8.046 1.732e-09  0.004072 6.756e-03
#> f_parent_to_m1  0.513600  23.560 4.352e-22  0.469300 5.578e-01
#> sigma_parent    3.401000   5.985 5.662e-07  2.244000 4.559e+00
#> sigma_m1        2.855000   6.311 2.215e-07  1.934000 3.777e+00
#> 
#> FOCUS Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   6.398       4 15
#> parent     6.464       2  7
#> m1         4.682       2  8
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_m1   0.5136
#> parent_sink 0.4864
#> 
#> Estimated disappearance times:
#>           DT50   DT90
#> parent   7.003  23.26
#> m1     132.154 439.01
#> 
#> Data:
#>  time variable observed predicted   residual
#>     0   parent    99.46  99.65417 -1.942e-01
#>     0   parent   102.04  99.65417  2.386e+00
#>     1   parent    93.50  90.26332  3.237e+00
#>     1   parent    92.50  90.26332  2.237e+00
#>     3   parent    63.23  74.05306 -1.082e+01
#>     3   parent    68.99  74.05306 -5.063e+00
#>     7   parent    52.32  49.84325  2.477e+00
#>     7   parent    55.13  49.84325  5.287e+00
#>    14   parent    27.27  24.92971  2.340e+00
#>    14   parent    26.64  24.92971  1.710e+00
#>    21   parent    11.50  12.46890 -9.689e-01
#>    21   parent    11.64  12.46890 -8.289e-01
#>    35   parent     2.85   3.11925 -2.692e-01
#>    35   parent     2.91   3.11925 -2.092e-01
#>    50   parent     0.69   0.70679 -1.679e-02
#>    50   parent     0.63   0.70679 -7.679e-02
#>    75   parent     0.05   0.05952 -9.523e-03
#>    75   parent     0.06   0.05952  4.772e-04
#>     1       m1     4.84   4.81075  2.925e-02
#>     1       m1     5.64   4.81075  8.292e-01
#>     3       m1    12.91  13.04196 -1.320e-01
#>     3       m1    12.96  13.04196 -8.196e-02
#>     7       m1    22.97  25.06847 -2.098e+00
#>     7       m1    24.47  25.06847 -5.985e-01
#>    14       m1    41.69  36.70308  4.987e+00
#>    14       m1    33.21  36.70308 -3.493e+00
#>    21       m1    44.37  41.65115  2.719e+00
#>    21       m1    46.44  41.65115  4.789e+00
#>    35       m1    41.22  43.29465 -2.075e+00
#>    35       m1    37.95  43.29465 -5.345e+00
#>    50       m1    41.19  41.19948 -9.479e-03
#>    50       m1    40.01  41.19948 -1.189e+00
#>    75       m1    40.09  36.44035  3.650e+00
#>    75       m1    33.85  36.44035 -2.590e+00
#>   100       m1    31.04  31.98773 -9.477e-01
#>   100       m1    33.13  31.98773  1.142e+00
#>   120       m1    25.15  28.80429 -3.654e+00
#>   120       m1    33.31  28.80429  4.506e+00

f.tc <- mkinfit(SFO_SFO.ff, FOCUS_2006_D, error_model = "tc", quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

summary(f.tc)

#> mkin version used for fitting:    0.9.50.3 
#> R version used for fitting:       4.0.0 
#> Date of fit:     Fri May 15 07:30:39 2020 
#> Date of summary: Fri May 15 07:30:39 2020 
#> 
#> Equations:
#> d_parent/dt = - k_parent * parent
#> d_m1/dt = + f_parent_to_m1 * k_parent * parent - k_m1 * m1
#> 
#> Model predictions using solution type analytical 
#> 
#> Fitted using 1875 model solutions performed in 0.658 s
#> 
#> Error model: Two-component variance function 
#> 
#> Error model algorithm: d_3 
#> Direct fitting and three-step fitting yield approximately the same likelihood 
#> 
#> Starting values for parameters to be optimised:
#>                   value   type
#> parent_0       100.7500  state
#> k_parent         0.1000 deparm
#> k_m1             0.1001 deparm
#> f_parent_to_m1   0.5000 deparm
#> sigma_low        0.1000  error
#> rsd_high         0.1000  error
#> 
#> Starting values for the transformed parameters actually optimised:
#>                     value lower upper
#> parent_0       100.750000  -Inf   Inf
#> log_k_parent    -2.302585  -Inf   Inf
#> log_k_m1        -2.301586  -Inf   Inf
#> f_parent_ilr_1   0.000000  -Inf   Inf
#> sigma_low        0.100000     0   Inf
#> rsd_high         0.100000     0   Inf
#> 
#> Fixed parameter values:
#>      value  type
#> m1_0     0 state
#> 
#> Results:
#> 
#>        AIC      BIC    logLik
#>   141.9656 151.7911 -64.98278
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                 Estimate Std. Error     Lower     Upper
#> parent_0       100.70000   2.621000 95.400000 106.10000
#> log_k_parent    -2.29700   0.008862 -2.315000  -2.27900
#> log_k_m1        -5.26600   0.091310 -5.452000  -5.08000
#> f_parent_ilr_1   0.02374   0.055300 -0.088900   0.13640
#> sigma_low        0.00305   0.004829 -0.006786   0.01289
#> rsd_high         0.07928   0.009418  0.060100   0.09847
#> 
#> Parameter correlation:
#>                parent_0 log_k_parent log_k_m1 f_parent_ilr_1 sigma_low rsd_high
#> parent_0        1.00000      0.67644 -0.10215       -0.76822   0.14294 -0.08783
#> log_k_parent    0.67644      1.00000 -0.15102       -0.59491   0.34611 -0.08125
#> log_k_m1       -0.10215     -0.15102  1.00000        0.51808  -0.05236  0.01240
#> f_parent_ilr_1 -0.76822     -0.59491  0.51808        1.00000  -0.13900  0.03248
#> sigma_low       0.14294      0.34611 -0.05236       -0.13900   1.00000 -0.16546
#> rsd_high       -0.08783     -0.08125  0.01240        0.03248  -0.16546  1.00000
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                 Estimate  t value    Pr(>t)     Lower     Upper
#> parent_0       1.007e+02  38.4300 1.180e-28 95.400000 1.061e+02
#> k_parent       1.006e-01 112.8000 1.718e-43  0.098760 1.024e-01
#> k_m1           5.167e-03  10.9500 1.171e-12  0.004290 6.223e-03
#> f_parent_to_m1 5.084e-01  26.0100 2.146e-23  0.468600 5.481e-01
#> sigma_low      3.050e-03   0.6314 2.661e-01 -0.006786 1.289e-02
#> rsd_high       7.928e-02   8.4170 6.418e-10  0.060100 9.847e-02
#> 
#> FOCUS Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   6.475       4 15
#> parent     6.573       2  7
#> m1         4.671       2  8
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_m1   0.5084
#> parent_sink 0.4916
#> 
#> Estimated disappearance times:
#>           DT50  DT90
#> parent   6.893  22.9
#> m1     134.156 445.7
#> 
#> Data:
#>  time variable observed predicted   residual
#>     0   parent    99.46 100.73434  -1.274340
#>     0   parent   102.04 100.73434   1.305660
#>     1   parent    93.50  91.09751   2.402486
#>     1   parent    92.50  91.09751   1.402486
#>     3   parent    63.23  74.50141 -11.271410
#>     3   parent    68.99  74.50141  -5.511410
#>     7   parent    52.32  49.82880   2.491200
#>     7   parent    55.13  49.82880   5.301200
#>    14   parent    27.27  24.64809   2.621908
#>    14   parent    26.64  24.64809   1.991908
#>    21   parent    11.50  12.19232  -0.692315
#>    21   parent    11.64  12.19232  -0.552315
#>    35   parent     2.85   2.98327  -0.133266
#>    35   parent     2.91   2.98327  -0.073266
#>    50   parent     0.69   0.66013   0.029874
#>    50   parent     0.63   0.66013  -0.030126
#>    75   parent     0.05   0.05344  -0.003438
#>    75   parent     0.06   0.05344   0.006562
#>     1       m1     4.84   4.88645  -0.046451
#>     1       m1     5.64   4.88645   0.753549
#>     3       m1    12.91  13.22867  -0.318669
#>     3       m1    12.96  13.22867  -0.268669
#>     7       m1    22.97  25.36417  -2.394166
#>     7       m1    24.47  25.36417  -0.894166
#>    14       m1    41.69  37.00974   4.680263
#>    14       m1    33.21  37.00974  -3.799737
#>    21       m1    44.37  41.90133   2.468669
#>    21       m1    46.44  41.90133   4.538669
#>    35       m1    41.22  43.45691  -2.236913
#>    35       m1    37.95  43.45691  -5.506913
#>    50       m1    41.19  41.34199  -0.151985
#>    50       m1    40.01  41.34199  -1.331985
#>    75       m1    40.09  36.61471   3.475295
#>    75       m1    33.85  36.61471  -2.764705
#>   100       m1    31.04  32.20082  -1.160823
#>   100       m1    33.13  32.20082   0.929177
#>   120       m1    25.15  29.04130  -3.891304
#>   120       m1    33.31  29.04130   4.268696

# }