Extract model parameters from mkinfit models

This function always returns degradation model parameters as well as error model parameters, in order to avoid working with a fitted model without considering the error structure that was assumed for the fit.

parms(object, ...)

# S3 method for mkinfit
parms(object, transformed = FALSE, ...)

# S3 method for mmkin
parms(object, transformed = FALSE, ...)

Arguments

object
A fitted model object. Methods are implemented for mkinfit() objects and for mmkin() objects.

...
Not used

transformed
Should the parameters be returned as used internally during the optimisation?

Value

For mkinfit objects, a numeric vector of fitted model parameters. For mmkin row objects, a matrix with the parameters with a row for each dataset. If the mmkin object has more than one row, a list of such matrices is returned.

Examples

# mkinfit objects
fit <- mkinfit("SFO", FOCUS_2006_C, quiet = TRUE)
parms(fit)

#>   parent_0   k_parent      sigma 
#> 82.4921598  0.3060633  4.6730124 

parms(fit, transformed = TRUE)

#>     parent_0 log_k_parent        sigma 
#>    82.492160    -1.183963     4.673012 

# mmkin objects
ds <- lapply(experimental_data_for_UBA_2019[6:10],
 function(x) subset(x$data[c("name", "time", "value")]))
names(ds) <- paste("Dataset", 6:10)
# \dontrun{
fits <- mmkin(c("SFO", "FOMC", "DFOP"), ds, quiet = TRUE, cores = 1)
parms(fits["SFO", ])

#>            Dataset 6    Dataset 7  Dataset 8  Dataset 9  Dataset 10
#> parent_0 88.52275400 82.666781678 86.8547308 91.7779306 82.14809450
#> k_parent  0.05794659  0.009647805  0.2102974  0.1232258  0.00720421
#> sigma     5.15274487  7.040168584  3.6769645  6.4669234  6.50457673

parms(fits[, 2])

#> $SFO
#>             Dataset 7
#> parent_0 82.666781678
#> k_parent  0.009647805
#> sigma     7.040168584
#> 
#> $FOMC
#>           Dataset 7
#> parent_0 92.6837649
#> alpha     0.4967832
#> beta     14.1451255
#> sigma     1.9167519
#> 
#> $DFOP
#>             Dataset 7
#> parent_0 91.058971597
#> k1        0.044946770
#> k2        0.002868336
#> g         0.526942414
#> sigma     2.221302196
#> 

parms(fits)

#> $SFO
#>            Dataset 6    Dataset 7  Dataset 8  Dataset 9  Dataset 10
#> parent_0 88.52275400 82.666781678 86.8547308 91.7779306 82.14809450
#> k_parent  0.05794659  0.009647805  0.2102974  0.1232258  0.00720421
#> sigma     5.15274487  7.040168584  3.6769645  6.4669234  6.50457673
#> 
#> $FOMC
#>          Dataset 6  Dataset 7 Dataset 8 Dataset 9 Dataset 10
#> parent_0 95.558575 92.6837649 90.719787 98.383939 94.8481458
#> alpha     1.338667  0.4967832  1.639099  1.074460  0.2805272
#> beta     13.033315 14.1451255  5.007077  4.397126  6.9052224
#> sigma     1.847671  1.9167519  1.066063  3.146056  1.6222778
#> 
#> $DFOP
#>            Dataset 6    Dataset 7   Dataset 8   Dataset 9   Dataset 10
#> parent_0 96.55213663 91.058971597 90.34509493 98.14858820 94.311323733
#> k1        0.21954588  0.044946770  0.41232288  0.31697588  0.080663857
#> k2        0.02957934  0.002868336  0.07581766  0.03260384  0.003425417
#> g         0.44845068  0.526942414  0.66091967  0.65322767  0.342652880
#> sigma     1.35690468  2.221302196  1.34169076  2.87159846  1.942067831
#> 

parms(fits, transformed = TRUE)

#> $SFO
#>              Dataset 6 Dataset 7 Dataset 8 Dataset 9 Dataset 10
#> parent_0     88.522754 82.666782 86.854731 91.777931  82.148094
#> log_k_parent -2.848234 -4.641025 -1.559232 -2.093737  -4.933090
#> sigma         5.152745  7.040169  3.676964  6.466923   6.504577
#> 
#> $FOMC
#>            Dataset 6  Dataset 7  Dataset 8   Dataset 9 Dataset 10
#> parent_0  95.5585751 92.6837649 90.7197870 98.38393897  94.848146
#> log_alpha  0.2916741 -0.6996015  0.4941466  0.07181816  -1.271085
#> log_beta   2.5675088  2.6493701  1.6108523  1.48095106   1.932278
#> sigma      1.8476712  1.9167519  1.0660627  3.14605557   1.622278
#> 
#> $DFOP
#>           Dataset 6  Dataset 7  Dataset 8  Dataset 9 Dataset 10
#> parent_0 96.5521366 91.0589716 90.3450949 98.1485882 94.3113237
#> log_k1   -1.5161940 -3.1022764 -0.8859486 -1.1489296 -2.5174647
#> log_k2   -3.5206791 -5.8540232 -2.5794240 -3.4233253 -5.6765322
#> g_qlogis -0.2069326  0.1078741  0.6673953  0.6332573 -0.6514943
#> sigma     1.3569047  2.2213022  1.3416908  2.8715985  1.9420678
#> 

# }