Summary method for class "mkinfit"

Lists model equations, the summary as returned by summary.modFit, the chi2 error levels calculated according to FOCUS guidance (2006) as far as defined therein, and optionally the data, consisting of observed, predicted and residual values.

# S3 method for mkinfit
summary(object, data = TRUE, distimes = TRUE, alpha = 0.05, ...)
# S3 method for summary.mkinfit
print(x, digits = max(3, getOption("digits") - 3), ...)

Arguments

object
an object of class mkinfit.

x
an object of class summary.mkinfit.

data
logical, indicating whether the data should be included in the summary.

distimes
logical, indicating whether DT50 and DT90 values should be included.

alpha
error level for confidence interval estimation from t distribution

digits
Number of digits to use for printing

…
optional arguments passed to methods like print.

Value

The summary function returns a list derived from summary.modFit, with components, among others

version, Rversion

The mkin and R versions used

date.fit, date.summary

The dates where the fit and the summary were produced

use_of_ff

Was maximum or minimum use made of formation fractions

residuals, residualVariance, sigma, modVariance, df

As in summary.modFit

cov.unscaled, cov.scaled, info, niter, stopmess, par

As in summary.modFit

bpar

Optimised and backtransformed parameters

diffs

The differential equations used in the model

data

The data (see Description above).

start

The starting values and bounds, if applicable, for optimised parameters.

fixed

The values of fixed parameters.

errmin

The chi2 error levels for each observed variable.

bparms.ode

All backtransformed ODE parameters, for use as starting parameters for related models.

ff

The estimated formation fractions derived from the fitted model.

distimes

The DT50 and DT90 values for each observed variable.

SFORB

If applicable, eigenvalues of SFORB components of the model.

The print method is called for its side effect, i.e. printing the summary.
References

FOCUS (2006) “Guidance Document on Estimating Persistence and Degradation Kinetics from Environmental Fate Studies on Pesticides in EU Registration” Report of the FOCUS Work Group on Degradation Kinetics, EC Document Reference Sanco/10058/2005 version 2.0, 434 pp, http://esdac.jrc.ec.europa.eu/projects/degradation-kinetics

Examples

  summary(mkinfit(mkinmod(parent = mkinsub("SFO")), FOCUS_2006_A, quiet = TRUE))

#> mkin version:    0.9.46 
#> R version:       3.4.1 
#> Date of fit:     Sat Jul 29 15:15:30 2017 
#> Date of summary: Sat Jul 29 15:15:30 2017 
#> 
#> Equations:
#> d_parent/dt = - k_parent_sink * parent
#> 
#> Model predictions using solution type analytical 
#> 
#> Fitted with method Port using 35 model solutions performed in 0.084 s
#> 
#> Weighting: none
#> 
#> Starting values for parameters to be optimised:
#>                value   type
#> parent_0      101.24  state
#> k_parent_sink   0.10 deparm
#> 
#> Starting values for the transformed parameters actually optimised:
#>                        value lower upper
#> parent_0          101.240000  -Inf   Inf
#> log_k_parent_sink  -2.302585  -Inf   Inf
#> 
#> Fixed parameter values:
#> None
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                   Estimate Std. Error  Lower   Upper
#> parent_0           109.200     4.3910 98.410 119.900
#> log_k_parent_sink   -3.291     0.1152 -3.573  -3.009
#> 
#> Parameter correlation:
#>                   parent_0 log_k_parent_sink
#> parent_0             1.000             0.575
#> log_k_parent_sink    0.575             1.000
#> 
#> Residual standard error: 6.08 on 6 degrees of freedom
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                Estimate t value    Pr(>t)    Lower     Upper
#> parent_0      109.20000  24.860 1.394e-07 98.41000 119.90000
#> k_parent_sink   0.03722   8.679 6.457e-05  0.02807   0.04934
#> 
#> Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   8.385       2  6
#> parent     8.385       2  6
#> 
#> Resulting formation fractions:
#>             ff
#> parent_sink  1
#> 
#> Estimated disappearance times:
#>         DT50  DT90
#> parent 18.62 61.87
#> 
#> Data:
#>  time variable observed predicted residual
#>     0   parent   101.24   109.153  -7.9132
#>     3   parent    99.27    97.622   1.6484
#>     7   parent    90.11    84.119   5.9913
#>    14   parent    72.19    64.826   7.3641
#>    30   parent    29.71    35.738  -6.0283
#>    62   parent     5.98    10.862  -4.8818
#>    90   parent     1.54     3.831  -2.2911
#>   118   parent     0.39     1.351  -0.9613