Summary method for class "mkinfit"

Lists model equations, initial parameter values, optimised parameters with some uncertainty statistics, the chi2 error levels calculated according to FOCUS guidance (2006) as defined therein, formation fractions, DT50 values and optionally the data, consisting of observed, predicted and residual values.

# S3 method for mkinfit
summary(object, data = TRUE, distimes = TRUE, alpha = 0.05, ...)

# S3 method for summary.mkinfit
print(x, digits = max(3, getOption("digits") - 3), ...)

Arguments

object
an object of class mkinfit.

data
logical, indicating whether the data should be included in the summary.

distimes
logical, indicating whether DT50 and DT90 values should be included.

alpha
error level for confidence interval estimation from t distribution

...
optional arguments passed to methods like print.

x
an object of class summary.mkinfit.

digits
Number of digits to use for printing

Value

The summary function returns a list with components, among others

version, Rversion

The mkin and R versions used

date.fit, date.summary

The dates where the fit and the summary were produced

diffs

The differential equations used in the model

use_of_ff

Was maximum or minimum use made of formation fractions

bpar

Optimised and backtransformed parameters

data

The data (see Description above).

start

The starting values and bounds, if applicable, for optimised parameters.

fixed

The values of fixed parameters.

errmin

The chi2 error levels for each observed variable.

bparms.ode

All backtransformed ODE parameters, for use as starting parameters for related models.

errparms

Error model parameters.

ff

The estimated formation fractions derived from the fitted model.

distimes

The DT50 and DT90 values for each observed variable.

SFORB

If applicable, eigenvalues of SFORB components of the model.

The print method is called for its side effect, i.e. printing the summary.
References

FOCUS (2006) “Guidance Document on Estimating Persistence and Degradation Kinetics from Environmental Fate Studies on Pesticides in EU Registration” Report of the FOCUS Work Group on Degradation Kinetics, EC Document Reference Sanco/10058/2005 version 2.0, 434 pp, http://esdac.jrc.ec.europa.eu/projects/degradation-kinetics

Author

Johannes Ranke

Examples

  summary(mkinfit(mkinmod(parent = mkinsub("SFO")), FOCUS_2006_A, quiet = TRUE))

#> mkin version used for fitting:    0.9.50.4 
#> R version used for fitting:       4.0.3 
#> Date of fit:     Sat Nov  7 13:28:04 2020 
#> Date of summary: Sat Nov  7 13:28:04 2020 
#> 
#> Equations:
#> d_parent/dt = - k_parent * parent
#> 
#> Model predictions using solution type analytical 
#> 
#> Fitted using 131 model solutions performed in 0.028 s
#> 
#> Error model: Constant variance 
#> 
#> Error model algorithm: OLS 
#> 
#> Starting values for parameters to be optimised:
#>           value   type
#> parent_0 101.24  state
#> k_parent   0.10 deparm
#> 
#> Starting values for the transformed parameters actually optimised:
#>                   value lower upper
#> parent_0     101.240000  -Inf   Inf
#> log_k_parent  -2.302585  -Inf   Inf
#> 
#> Fixed parameter values:
#> None
#> 
#> Results:
#> 
#>        AIC     BIC    logLik
#>   55.28197 55.5203 -24.64099
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>              Estimate Std. Error  Lower   Upper
#> parent_0      109.200    3.70400 99.630 118.700
#> log_k_parent   -3.291    0.09176 -3.527  -3.055
#> sigma           5.266    1.31600  1.882   8.649
#> 
#> Parameter correlation:
#>               parent_0 log_k_parent     sigma
#> parent_0     1.000e+00    5.428e-01 1.642e-07
#> log_k_parent 5.428e-01    1.000e+00 2.507e-07
#> sigma        1.642e-07    2.507e-07 1.000e+00
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>           Estimate t value    Pr(>t)   Lower     Upper
#> parent_0 109.20000   29.47 4.218e-07 99.6300 118.70000
#> k_parent   0.03722   10.90 5.650e-05  0.0294   0.04712
#> sigma      5.26600    4.00 5.162e-03  1.8820   8.64900
#> 
#> FOCUS Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   8.385       2  6
#> parent     8.385       2  6
#> 
#> Estimated disappearance times:
#>         DT50  DT90
#> parent 18.62 61.87
#> 
#> Data:
#>  time variable observed predicted residual
#>     0   parent   101.24   109.153  -7.9132
#>     3   parent    99.27    97.622   1.6484
#>     7   parent    90.11    84.119   5.9913
#>    14   parent    72.19    64.826   7.3641
#>    30   parent    29.71    35.738  -6.0283
#>    62   parent     5.98    10.862  -4.8818
#>    90   parent     1.54     3.831  -2.2911
#>   118   parent     0.39     1.351  -0.9613