Summary method for class "nlme.mmkin"

Lists model equations, initial parameter values, optimised parameters for fixed effects (population), random effects (deviations from the population mean) and residual error model, as well as the resulting endpoints such as formation fractions and DT50 values. Optionally (default is FALSE), the data are listed in full.

# S3 method for nlme.mmkin
summary(
  object,
  data = FALSE,
  verbose = FALSE,
  distimes = TRUE,
  alpha = 0.05,
  ...
)

# S3 method for summary.nlme.mmkin
print(x, digits = max(3, getOption("digits") - 3), verbose = x$verbose, ...)

Arguments

object
an object of class nlme.mmkin

data
logical, indicating whether the full data should be included in the summary.

verbose
Should the summary be verbose?

distimes
logical, indicating whether DT50 and DT90 values should be included.

alpha
error level for confidence interval estimation from the t distribution

...
optional arguments passed to methods like print.

x
an object of class summary.nlme.mmkin

digits
Number of digits to use for printing

Value

The summary function returns a list based on the nlme object obtained in the fit, with at least the following additional components

nlmeversion, mkinversion, Rversion

The nlme, mkin and R versions used

date.fit, date.summary

The dates where the fit and the summary were produced

diffs

The differential equations used in the degradation model

use_of_ff

Was maximum or minimum use made of formation fractions

data

The data

confint_trans

Transformed parameters as used in the optimisation, with confidence intervals

confint_back

Backtransformed parameters, with confidence intervals if available

ff

The estimated formation fractions derived from the fitted model.

distimes

The DT50 and DT90 values for each observed variable.

SFORB

If applicable, eigenvalues of SFORB components of the model.

The print method is called for its side effect, i.e. printing the summary.
Author

Johannes Ranke for the mkin specific parts José Pinheiro and Douglas Bates for the components inherited from nlme

Examples

# Generate five datasets following SFO kinetics
sampling_times = c(0, 1, 3, 7, 14, 28, 60, 90, 120)
dt50_sfo_in_pop <- 50
k_in_pop <- log(2) / dt50_sfo_in_pop
set.seed(1234)
k_in <- rlnorm(5, log(k_in_pop), 0.5)
SFO <- mkinmod(parent = mkinsub("SFO"))

pred_sfo <- function(k) {
  mkinpredict(SFO,
    c(k_parent = k),
    c(parent = 100),
    sampling_times)
}

ds_sfo_mean <- lapply(k_in, pred_sfo)
names(ds_sfo_mean) <- paste("ds", 1:5)

set.seed(12345)
ds_sfo_syn <- lapply(ds_sfo_mean, function(ds) {
  add_err(ds,
    sdfunc = function(value) sqrt(1^2 + value^2 * 0.07^2),
    n = 1)[[1]]
})

# Evaluate using mmkin and nlme
library(nlme)
f_mmkin <- mmkin("SFO", ds_sfo_syn, quiet = TRUE, error_model = "tc", cores = 1)

#> Warning: Optimisation did not converge:
#> iteration limit reached without convergence (10)

f_nlme <- nlme(f_mmkin)

#> Warning: Iteration 1, LME step: nlminb() did not converge (code = 1). Do increase 'msMaxIter'!

summary(f_nlme, data = TRUE)

#> nlme version used for fitting:      3.1.150.1 
#> mkin version used for pre-fitting:  0.9.50.4 
#> R version used for fitting:         4.0.3 
#> Date of fit:     Mon Nov 30 16:01:23 2020 
#> Date of summary: Mon Nov 30 16:01:23 2020 
#> 
#> Equations:
#> d_parent/dt = - k_parent * parent
#> 
#> Data:
#> 90 observations of 1 variable(s) grouped in 5 datasets
#> 
#> Model predictions using solution type analytical 
#> 
#> Fitted in 0.996 s using 6 iterations
#> 
#> Variance model: Two-component variance function 
#> 
#> Mean of starting values for individual parameters:
#>     parent_0 log_k_parent 
#>   101.568773    -4.454103 
#> 
#> Fixed degradation parameter values:
#> None
#> 
#> Results:
#> 
#>        AIC      BIC    logLik
#>   586.4159 603.9145 -286.2079
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>               lower    est.   upper
#> parent_0     99.360 101.593 103.825
#> log_k_parent -4.979  -4.451  -3.924
#> 
#> Correlation: 
#>              prnt_0
#> log_k_parent 0.155 
#> 
#> Random effects:
#>  Formula: list(parent_0 ~ 1, log_k_parent ~ 1)
#>  Level: ds
#>  Structure: General positive-definite, Log-Cholesky parametrization
#>              StdDev    Corr  
#> parent_0     0.2624248 prnt_0
#> log_k_parent 0.5907146 0.999 
#> Residual     1.0000000       
#> 
#> Variance function:
#>  Structure: Constant plus proportion of variance covariate
#>  Formula: ~fitted(.) 
#>  Parameter estimates:
#>       const        prop 
#> 6.56706e-24 7.89583e-02 
#> 
#> Backtransformed parameters with asymmetric confidence intervals:
#>              lower      est.     upper
#> parent_0 99.360213 101.59280 103.82539
#> k_parent  0.006883   0.01166   0.01976
#> 
#> Estimated disappearance times:
#>         DT50  DT90
#> parent 59.44 197.4
#> 
#> Data:
#>    ds   name time observed predicted  residual    std standardized
#>  ds 1 parent    0    104.1   101.417   2.68283 8.0077     0.335030
#>  ds 1 parent    0    105.0   101.417   3.58283 8.0077     0.447422
#>  ds 1 parent    1     98.5   100.624  -2.12400 7.9451    -0.267334
#>  ds 1 parent    1     96.1   100.624  -4.52400 7.9451    -0.569407
#>  ds 1 parent    3    101.9    99.056   2.84379 7.8213     0.363595
#>  ds 1 parent    3     85.2    99.056 -13.85621 7.8213    -1.771597
#>  ds 1 parent    7     99.1    95.994   3.10646 7.5795     0.409850
#>  ds 1 parent    7     93.0    95.994  -2.99354 7.5795    -0.394953
#>  ds 1 parent   14     88.1    90.860  -2.75997 7.1741    -0.384710
#>  ds 1 parent   14     84.1    90.860  -6.75997 7.1741    -0.942268
#>  ds 1 parent   28     80.2    81.402  -1.20174 6.4273    -0.186973
#>  ds 1 parent   28     91.3    81.402   9.89826 6.4273     1.540024
#>  ds 1 parent   60     65.1    63.316   1.78361 4.9994     0.356767
#>  ds 1 parent   60     65.8    63.316   2.48361 4.9994     0.496785
#>  ds 1 parent   90     47.8    50.029  -2.22862 3.9502    -0.564182
#>  ds 1 parent   90     53.5    50.029   3.47138 3.9502     0.878792
#>  ds 1 parent  120     37.6    39.529  -1.92946 3.1212    -0.618182
#>  ds 1 parent  120     39.3    39.529  -0.22946 3.1212    -0.073516
#>  ds 2 parent    0    107.9   101.711   6.18875 8.0309     0.770613
#>  ds 2 parent    0    102.1   101.711   0.38875 8.0309     0.048406
#>  ds 2 parent    1    103.8   100.174   3.62597 7.9096     0.458428
#>  ds 2 parent    1    108.6   100.174   8.42597 7.9096     1.065287
#>  ds 2 parent    3     91.0    97.169  -6.16895 7.6723    -0.804055
#>  ds 2 parent    3     84.9    97.169 -12.26895 7.6723    -1.599124
#>  ds 2 parent    7     79.3    91.427 -12.12652 7.2189    -1.679833
#>  ds 2 parent    7    100.9    91.427   9.47348 7.2189     1.312320
#>  ds 2 parent   14     77.3    82.182  -4.88174 6.4889    -0.752318
#>  ds 2 parent   14     83.5    82.182   1.31826 6.4889     0.203155
#>  ds 2 parent   28     66.8    66.402   0.39792 5.2430     0.075895
#>  ds 2 parent   28     63.3    66.402  -3.10208 5.2430    -0.591662
#>  ds 2 parent   60     40.8    40.789   0.01138 3.2206     0.003532
#>  ds 2 parent   60     44.8    40.789   4.01138 3.2206     1.245537
#>  ds 2 parent   90     27.8    25.830   1.97002 2.0395     0.965937
#>  ds 2 parent   90     27.0    25.830   1.17002 2.0395     0.573682
#>  ds 2 parent  120     15.2    16.357  -1.15721 1.2915    -0.895991
#>  ds 2 parent  120     15.5    16.357  -0.85721 1.2915    -0.663710
#>  ds 3 parent    0     97.7   101.907  -4.20726 8.0464    -0.522873
#>  ds 3 parent    0     88.2   101.907 -13.70726 8.0464    -1.703521
#>  ds 3 parent    1    109.9    99.522  10.37805 7.8581     1.320685
#>  ds 3 parent    1     97.8    99.522  -1.72195 7.8581    -0.219130
#>  ds 3 parent    3    100.5    94.918   5.58248 7.4945     0.744875
#>  ds 3 parent    3     77.4    94.918 -17.51752 7.4945    -2.337375
#>  ds 3 parent    7     78.3    86.338  -8.03788 6.8171    -1.179077
#>  ds 3 parent    7     90.3    86.338   3.96212 6.8171     0.581204
#>  ds 3 parent   14     76.0    73.147   2.85281 5.7756     0.493944
#>  ds 3 parent   14     79.1    73.147   5.95281 5.7756     1.030687
#>  ds 3 parent   28     46.0    52.504  -6.50373 4.1456    -1.568826
#>  ds 3 parent   28     53.4    52.504   0.89627 4.1456     0.216197
#>  ds 3 parent   60     25.1    24.605   0.49461 1.9428     0.254586
#>  ds 3 parent   60     21.4    24.605  -3.20539 1.9428    -1.649882
#>  ds 3 parent   90     11.0    12.090  -1.09046 0.9546    -1.142271
#>  ds 3 parent   90     14.2    12.090   2.10954 0.9546     2.209770
#>  ds 3 parent  120      5.8     5.941  -0.14094 0.4691    -0.300464
#>  ds 3 parent  120      6.1     5.941   0.15906 0.4691     0.339077
#>  ds 4 parent    0     95.3   101.177  -5.87672 7.9887    -0.735625
#>  ds 4 parent    0    102.0   101.177   0.82328 7.9887     0.103056
#>  ds 4 parent    1    104.4   100.716   3.68438 7.9523     0.463307
#>  ds 4 parent    1    105.4   100.716   4.68438 7.9523     0.589057
#>  ds 4 parent    3    113.7    99.800  13.90026 7.8800     1.763989
#>  ds 4 parent    3     82.3    99.800 -17.49974 7.8800    -2.220774
#>  ds 4 parent    7     98.1    97.993   0.10713 7.7374     0.013846
#>  ds 4 parent    7     87.8    97.993 -10.19287 7.7374    -1.317359
#>  ds 4 parent   14     97.9    94.909   2.99079 7.4939     0.399098
#>  ds 4 parent   14    104.8    94.909   9.89079 7.4939     1.319851
#>  ds 4 parent   28     85.0    89.030  -4.02995 7.0297    -0.573279
#>  ds 4 parent   28     77.2    89.030 -11.82995 7.0297    -1.682864
#>  ds 4 parent   60     82.2    76.923   5.27690 6.0737     0.868808
#>  ds 4 parent   60     86.1    76.923   9.17690 6.0737     1.510919
#>  ds 4 parent   90     70.5    67.073   3.42743 5.2959     0.647182
#>  ds 4 parent   90     61.7    67.073  -5.37257 5.2959    -1.014470
#>  ds 4 parent  120     60.0    58.483   1.51654 4.6178     0.328416
#>  ds 4 parent  120     56.4    58.483  -2.08346 4.6178    -0.451184
#>  ds 5 parent    0     92.6   101.752  -9.15161 8.0341    -1.139091
#>  ds 5 parent    0    116.5   101.752  14.74839 8.0341     1.835716
#>  ds 5 parent    1    108.0   100.069   7.93112 7.9013     1.003778
#>  ds 5 parent    1    104.9   100.069   4.83112 7.9013     0.611436
#>  ds 5 parent    3    100.5    96.786   3.71355 7.6421     0.485934
#>  ds 5 parent    3     89.5    96.786  -7.28645 7.6421    -0.953462
#>  ds 5 parent    7     91.7    90.541   1.15895 7.1490     0.162114
#>  ds 5 parent    7     95.1    90.541   4.55895 7.1490     0.637707
#>  ds 5 parent   14     82.2    80.566   1.63437 6.3613     0.256923
#>  ds 5 parent   14     84.5    80.566   3.93437 6.3613     0.618483
#>  ds 5 parent   28     60.5    63.791  -3.29084 5.0368    -0.653357
#>  ds 5 parent   28     72.8    63.791   9.00916 5.0368     1.788662
#>  ds 5 parent   60     38.3    37.412   0.88840 2.9540     0.300748
#>  ds 5 parent   60     40.7    37.412   3.28840 2.9540     1.113217
#>  ds 5 parent   90     22.5    22.685  -0.18500 1.7912    -0.103287
#>  ds 5 parent   90     20.8    22.685  -1.88500 1.7912    -1.052387
#>  ds 5 parent  120     13.4    13.755  -0.35534 1.0861    -0.327173
#>  ds 5 parent  120     13.8    13.755   0.04466 1.0861     0.041118