Summary method for class "saem.mmkin"

Lists model equations, initial parameter values, optimised parameters for fixed effects (population), random effects (deviations from the population mean) and residual error model, as well as the resulting endpoints such as formation fractions and DT50 values. Optionally (default is FALSE), the data are listed in full.

Usage

# S3 method for class 'saem.mmkin'
summary(
  object,
  data = FALSE,
  verbose = FALSE,
  covariates = NULL,
  covariate_quantile = 0.5,
  distimes = TRUE,
  ...
)

# S3 method for class 'summary.saem.mmkin'
print(x, digits = max(3, getOption("digits") - 3), verbose = x$verbose, ...)

Arguments

object

an object of class saem.mmkin

data

logical, indicating whether the full data should be included in the summary.

verbose

Should the summary be verbose?

covariates

Numeric vector with covariate values for all variables in any covariate models in the object. If given, it overrides 'covariate_quantile'.

covariate_quantile

This argument only has an effect if the fitted object has covariate models. If so, the default is to show endpoints for the median of the covariate values (50th percentile).

distimes

logical, indicating whether DT50 and DT90 values should be included.

...

optional arguments passed to methods like print.

x

an object of class summary.saem.mmkin

digits

Number of digits to use for printing

Value

The summary function returns a list based on the saemix::SaemixObject obtained in the fit, with at least the following additional components

saemixversion, mkinversion, Rversion

The saemix, mkin and R versions used

date.fit, date.summary

The dates where the fit and the summary were produced

diffs

The differential equations used in the degradation model

use_of_ff

Was maximum or minimum use made of formation fractions

data

The data

confint_trans

Transformed parameters as used in the optimisation, with confidence intervals

confint_back

Backtransformed parameters, with confidence intervals if available

confint_errmod

Error model parameters with confidence intervals

ff

The estimated formation fractions derived from the fitted model.

distimes

The DT50 and DT90 values for each observed variable.

SFORB

If applicable, eigenvalues of SFORB components of the model.

The print method is called for its side effect, i.e. printing the summary.

Author

Johannes Ranke for the mkin specific parts saemix authors for the parts inherited from saemix.

Examples

# Generate five datasets following DFOP-SFO kinetics
sampling_times = c(0, 1, 3, 7, 14, 28, 60, 90, 120)
dfop_sfo <- mkinmod(parent = mkinsub("DFOP", "m1"),
 m1 = mkinsub("SFO"), quiet = TRUE)
set.seed(1234)
k1_in <- rlnorm(5, log(0.1), 0.3)
k2_in <- rlnorm(5, log(0.02), 0.3)
g_in <- plogis(rnorm(5, qlogis(0.5), 0.3))
f_parent_to_m1_in <- plogis(rnorm(5, qlogis(0.3), 0.3))
k_m1_in <- rlnorm(5, log(0.02), 0.3)

pred_dfop_sfo <- function(k1, k2, g, f_parent_to_m1, k_m1) {
  mkinpredict(dfop_sfo,
    c(k1 = k1, k2 = k2, g = g, f_parent_to_m1 = f_parent_to_m1, k_m1 = k_m1),
    c(parent = 100, m1 = 0),
    sampling_times)
}

ds_mean_dfop_sfo <- lapply(1:5, function(i) {
  mkinpredict(dfop_sfo,
    c(k1 = k1_in[i], k2 = k2_in[i], g = g_in[i],
      f_parent_to_m1 = f_parent_to_m1_in[i], k_m1 = k_m1_in[i]),
    c(parent = 100, m1 = 0),
    sampling_times)
})
names(ds_mean_dfop_sfo) <- paste("ds", 1:5)

ds_syn_dfop_sfo <- lapply(ds_mean_dfop_sfo, function(ds) {
  add_err(ds,
    sdfunc = function(value) sqrt(1^2 + value^2 * 0.07^2),
    n = 1)[[1]]
})

# \dontrun{
# Evaluate using mmkin and saem
f_mmkin_dfop_sfo <- mmkin(list(dfop_sfo), ds_syn_dfop_sfo,
  quiet = TRUE, error_model = "tc", cores = 5)
f_saem_dfop_sfo <- saem(f_mmkin_dfop_sfo)
print(f_saem_dfop_sfo)
#> Kinetic nonlinear mixed-effects model fit by SAEM
#> Structural model:
#> d_parent/dt = - ((k1 * g * exp(-k1 * time) + k2 * (1 - g) * exp(-k2 *
#>            time)) / (g * exp(-k1 * time) + (1 - g) * exp(-k2 * time)))
#>            * parent
#> d_m1/dt = + f_parent_to_m1 * ((k1 * g * exp(-k1 * time) + k2 * (1 - g)
#>            * exp(-k2 * time)) / (g * exp(-k1 * time) + (1 - g) *
#>            exp(-k2 * time))) * parent - k_m1 * m1
#> 
#> Data:
#> 171 observations of 2 variable(s) grouped in 5 datasets
#> 
#> Likelihood computed by importance sampling
#>     AIC   BIC logLik
#>   810.8 805.4 -391.4
#> 
#> Fitted parameters:
#>                      estimate     lower     upper
#> parent_0           100.966822  97.90584 104.02780
#> log_k_m1            -4.076164  -4.17485  -3.97748
#> f_parent_qlogis     -0.940902  -1.35358  -0.52823
#> log_k1              -2.363988  -2.71690  -2.01107
#> log_k2              -4.060016  -4.21743  -3.90260
#> g_qlogis            -0.029999  -0.44766   0.38766
#> a.1                  0.876272   0.67790   1.07464
#> b.1                  0.079594   0.06521   0.09398
#> SD.parent_0          0.076322 -76.45825  76.61089
#> SD.log_k_m1          0.005052  -1.08943   1.09953
#> SD.f_parent_qlogis   0.446968   0.16577   0.72816
#> SD.log_k1            0.348786   0.09502   0.60255
#> SD.log_k2            0.147456   0.03111   0.26380
#> SD.g_qlogis          0.348244   0.02794   0.66854
illparms(f_saem_dfop_sfo)
#> [1] "sd(parent_0)" "sd(log_k_m1)"
f_saem_dfop_sfo_2 <- update(f_saem_dfop_sfo,
  no_random_effect = c("parent_0", "log_k_m1"))
illparms(f_saem_dfop_sfo_2)
intervals(f_saem_dfop_sfo_2)
#> Approximate 95% confidence intervals
#> 
#>  Fixed effects:
#>                      lower         est.        upper
#> parent_0       98.04247057 101.09950884 104.15654711
#> k_m1            0.01528983   0.01687734   0.01862969
#> f_parent_to_m1  0.20447650   0.27932896   0.36887691
#> k1              0.06779844   0.09638524   0.13702550
#> k2              0.01495629   0.01741775   0.02028431
#> g               0.37669311   0.48368409   0.59219202
#> 
#>  Random effects:
#>                          lower      est.     upper
#> sd(f_parent_qlogis) 0.16515113 0.4448330 0.7245148
#> sd(log_k1)          0.08982399 0.3447403 0.5996565
#> sd(log_k2)          0.02806780 0.1419560 0.2558443
#> sd(g_qlogis)        0.04908644 0.3801993 0.7113121
#> 
#>  
#>          lower       est.      upper
#> a.1 0.67993373 0.87630147 1.07266921
#> b.1 0.06522297 0.07920531 0.09318766
summary(f_saem_dfop_sfo_2, data = TRUE)
#> saemix version used for fitting:      3.3 
#> mkin version used for pre-fitting:  1.2.10 
#> R version used for fitting:         4.5.0 
#> Date of fit:     Mon May 12 20:27:02 2025 
#> Date of summary: Mon May 12 20:27:02 2025 
#> 
#> Equations:
#> d_parent/dt = - ((k1 * g * exp(-k1 * time) + k2 * (1 - g) * exp(-k2 *
#>            time)) / (g * exp(-k1 * time) + (1 - g) * exp(-k2 * time)))
#>            * parent
#> d_m1/dt = + f_parent_to_m1 * ((k1 * g * exp(-k1 * time) + k2 * (1 - g)
#>            * exp(-k2 * time)) / (g * exp(-k1 * time) + (1 - g) *
#>            exp(-k2 * time))) * parent - k_m1 * m1
#> 
#> Data:
#> 171 observations of 2 variable(s) grouped in 5 datasets
#> 
#> Model predictions using solution type analytical 
#> 
#> Fitted in 9.206 s
#> Using 300, 100 iterations and 10 chains
#> 
#> Variance model: Two-component variance function 
#> 
#> Starting values for degradation parameters:
#>        parent_0        log_k_m1 f_parent_qlogis          log_k1          log_k2 
#>       101.65645        -4.05368        -0.94311        -2.35943        -4.07006 
#>        g_qlogis 
#>        -0.01133 
#> 
#> Fixed degradation parameter values:
#> None
#> 
#> Starting values for random effects (square root of initial entries in omega):
#>                 parent_0 log_k_m1 f_parent_qlogis log_k1 log_k2 g_qlogis
#> parent_0           6.742   0.0000          0.0000 0.0000 0.0000    0.000
#> log_k_m1           0.000   0.2236          0.0000 0.0000 0.0000    0.000
#> f_parent_qlogis    0.000   0.0000          0.5572 0.0000 0.0000    0.000
#> log_k1             0.000   0.0000          0.0000 0.8031 0.0000    0.000
#> log_k2             0.000   0.0000          0.0000 0.0000 0.2931    0.000
#> g_qlogis           0.000   0.0000          0.0000 0.0000 0.0000    0.807
#> 
#> Starting values for error model parameters:
#> a.1 b.1 
#>   1   1 
#> 
#> Results:
#> 
#> Likelihood computed by importance sampling
#>     AIC   BIC logLik
#>   806.9 802.2 -391.5
#> 
#> Optimised parameters:
#>                         est.    lower     upper
#> parent_0           101.09951 98.04247 104.15655
#> log_k_m1            -4.08178 -4.18057  -3.98300
#> f_parent_qlogis     -0.94779 -1.35855  -0.53704
#> log_k1              -2.33940 -2.69122  -1.98759
#> log_k2              -4.05027 -4.20262  -3.89791
#> g_qlogis            -0.06529 -0.50361   0.37303
#> a.1                  0.87630  0.67993   1.07267
#> b.1                  0.07921  0.06522   0.09319
#> SD.f_parent_qlogis   0.44483  0.16515   0.72451
#> SD.log_k1            0.34474  0.08982   0.59966
#> SD.log_k2            0.14196  0.02807   0.25584
#> SD.g_qlogis          0.38020  0.04909   0.71131
#> 
#> Correlation: 
#>                 parnt_0 lg_k_m1 f_prnt_ log_k1  log_k2 
#> log_k_m1        -0.4716                                
#> f_parent_qlogis -0.2394  0.2617                        
#> log_k1           0.1677 -0.1566 -0.0659                
#> log_k2           0.0165  0.0638  0.0045  0.2013        
#> g_qlogis         0.1118 -0.1118 -0.0340 -0.2324 -0.3419
#> 
#> Random effects:
#>                      est.   lower  upper
#> SD.f_parent_qlogis 0.4448 0.16515 0.7245
#> SD.log_k1          0.3447 0.08982 0.5997
#> SD.log_k2          0.1420 0.02807 0.2558
#> SD.g_qlogis        0.3802 0.04909 0.7113
#> 
#> Variance model:
#>        est.   lower   upper
#> a.1 0.87630 0.67993 1.07267
#> b.1 0.07921 0.06522 0.09319
#> 
#> Backtransformed parameters:
#>                     est.    lower     upper
#> parent_0       101.09951 98.04247 104.15655
#> k_m1             0.01688  0.01529   0.01863
#> f_parent_to_m1   0.27933  0.20448   0.36888
#> k1               0.09639  0.06780   0.13703
#> k2               0.01742  0.01496   0.02028
#> g                0.48368  0.37669   0.59219
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_m1   0.2793
#> parent_sink 0.7207
#> 
#> Estimated disappearance times:
#>         DT50   DT90 DT50back DT50_k1 DT50_k2
#> parent 15.66  94.28    28.38   7.191    39.8
#> m1     41.07 136.43       NA      NA      NA
#> 
#> Data:
#>    ds   name time observed  predicted  residual    std standardized
#>  ds 1 parent    0     89.8  1.011e+02 -11.29951 8.0554    -1.402721
#>  ds 1 parent    0    104.1  1.011e+02   3.00049 8.0554     0.372481
#>  ds 1 parent    1     88.7  9.624e+01  -7.53600 7.6726    -0.982195
#>  ds 1 parent    1     95.5  9.624e+01  -0.73600 7.6726    -0.095925
#>  ds 1 parent    3     81.8  8.736e+01  -5.55672 6.9744    -0.796732
#>  ds 1 parent    3     94.5  8.736e+01   7.14328 6.9744     1.024217
#>  ds 1 parent    7     71.5  7.251e+01  -1.00511 5.8093    -0.173019
#>  ds 1 parent    7     70.3  7.251e+01  -2.20511 5.8093    -0.379585
#>  ds 1 parent   14     54.2  5.356e+01   0.63921 4.3319     0.147560
#>  ds 1 parent   14     49.6  5.356e+01  -3.96079 4.3319    -0.914340
#>  ds 1 parent   28     31.5  3.175e+01  -0.25429 2.6634    -0.095475
#>  ds 1 parent   28     28.8  3.175e+01  -2.95429 2.6634    -1.109218
#>  ds 1 parent   60     12.1  1.281e+01  -0.71388 1.3409    -0.532390
#>  ds 1 parent   60     13.6  1.281e+01   0.78612 1.3409     0.586271
#>  ds 1 parent   90      6.2  6.405e+00  -0.20462 1.0125    -0.202083
#>  ds 1 parent   90      8.3  6.405e+00   1.89538 1.0125     1.871910
#>  ds 1 parent  120      2.2  3.329e+00  -1.12941 0.9151    -1.234165
#>  ds 1 parent  120      2.4  3.329e+00  -0.92941 0.9151    -1.015615
#>  ds 1     m1    1      0.3  1.177e+00  -0.87699 0.8812    -0.995168
#>  ds 1     m1    1      0.2  1.177e+00  -0.97699 0.8812    -1.108644
#>  ds 1     m1    3      2.2  3.268e+00  -1.06821 0.9137    -1.169063
#>  ds 1     m1    3      3.0  3.268e+00  -0.26821 0.9137    -0.293536
#>  ds 1     m1    7      6.5  6.555e+00  -0.05539 1.0186    -0.054377
#>  ds 1     m1    7      5.0  6.555e+00  -1.55539 1.0186    -1.527022
#>  ds 1     m1   14     10.2  1.017e+01   0.03108 1.1902     0.026117
#>  ds 1     m1   14      9.5  1.017e+01  -0.66892 1.1902    -0.562010
#>  ds 1     m1   28     12.2  1.270e+01  -0.50262 1.3342    -0.376708
#>  ds 1     m1   28     13.4  1.270e+01   0.69738 1.3342     0.522686
#>  ds 1     m1   60     11.8  1.078e+01   1.01734 1.2236     0.831403
#>  ds 1     m1   60     13.2  1.078e+01   2.41734 1.2236     1.975530
#>  ds 1     m1   90      6.6  7.686e+00  -1.08586 1.0670    -1.017675
#>  ds 1     m1   90      9.3  7.686e+00   1.61414 1.0670     1.512779
#>  ds 1     m1  120      3.5  5.205e+00  -1.70467 0.9684    -1.760250
#>  ds 1     m1  120      5.4  5.205e+00   0.19533 0.9684     0.201701
#>  ds 2 parent    0    118.0  1.011e+02  16.90049 8.0554     2.098026
#>  ds 2 parent    0     99.8  1.011e+02  -1.29951 8.0554    -0.161321
#>  ds 2 parent    1     90.2  9.574e+01  -5.53784 7.6334    -0.725473
#>  ds 2 parent    1     94.6  9.574e+01  -1.13784 7.6334    -0.149060
#>  ds 2 parent    3     96.1  8.638e+01   9.72233 6.8975     1.409551
#>  ds 2 parent    3     78.4  8.638e+01  -7.97767 6.8975    -1.156610
#>  ds 2 parent    7     77.9  7.194e+01   5.95854 5.7651     1.033547
#>  ds 2 parent    7     77.7  7.194e+01   5.75854 5.7651     0.998856
#>  ds 2 parent   14     56.0  5.558e+01   0.42141 4.4885     0.093888
#>  ds 2 parent   14     54.7  5.558e+01  -0.87859 4.4885    -0.195742
#>  ds 2 parent   28     36.6  3.852e+01  -1.92382 3.1746    -0.605999
#>  ds 2 parent   28     36.8  3.852e+01  -1.72382 3.1746    -0.543000
#>  ds 2 parent   60     22.1  2.108e+01   1.02043 1.8856     0.541168
#>  ds 2 parent   60     24.7  2.108e+01   3.62043 1.8856     1.920034
#>  ds 2 parent   90     12.4  1.250e+01  -0.09675 1.3220    -0.073184
#>  ds 2 parent   90     10.8  1.250e+01  -1.69675 1.3220    -1.283492
#>  ds 2 parent  120      6.8  7.426e+00  -0.62587 1.0554    -0.593027
#>  ds 2 parent  120      7.9  7.426e+00   0.47413 1.0554     0.449242
#>  ds 2     m1    1      1.3  1.417e+00  -0.11735 0.8835    -0.132825
#>  ds 2     m1    3      3.7  3.823e+00  -0.12301 0.9271    -0.132673
#>  ds 2     m1    3      4.7  3.823e+00   0.87699 0.9271     0.945909
#>  ds 2     m1    7      8.1  7.288e+00   0.81180 1.0494     0.773619
#>  ds 2     m1    7      7.9  7.288e+00   0.61180 1.0494     0.583025
#>  ds 2     m1   14     10.1  1.057e+01  -0.46957 1.2119    -0.387459
#>  ds 2     m1   14     10.3  1.057e+01  -0.26957 1.2119    -0.222432
#>  ds 2     m1   28     10.7  1.234e+01  -1.63555 1.3124    -1.246185
#>  ds 2     m1   28     12.2  1.234e+01  -0.13555 1.3124    -0.103281
#>  ds 2     m1   60     10.7  1.065e+01   0.04641 1.2165     0.038151
#>  ds 2     m1   60     12.5  1.065e+01   1.84641 1.2165     1.517773
#>  ds 2     m1   90      9.1  8.177e+00   0.92337 1.0896     0.847403
#>  ds 2     m1   90      7.4  8.177e+00  -0.77663 1.0896    -0.712734
#>  ds 2     m1  120      6.1  5.966e+00   0.13404 0.9956     0.134631
#>  ds 2     m1  120      4.5  5.966e+00  -1.46596 0.9956    -1.472460
#>  ds 3 parent    0    106.2  1.011e+02   5.10049 8.0554     0.633175
#>  ds 3 parent    0    106.9  1.011e+02   5.80049 8.0554     0.720073
#>  ds 3 parent    1    107.4  9.365e+01  13.74627 7.4695     1.840332
#>  ds 3 parent    1     96.1  9.365e+01   2.44627 7.4695     0.327504
#>  ds 3 parent    3     79.4  8.139e+01  -1.99118 6.5059    -0.306058
#>  ds 3 parent    3     82.6  8.139e+01   1.20882 6.5059     0.185803
#>  ds 3 parent    7     63.9  6.445e+01  -0.54666 5.1792    -0.105549
#>  ds 3 parent    7     62.4  6.445e+01  -2.04666 5.1792    -0.395170
#>  ds 3 parent   14     51.0  4.830e+01   2.69944 3.9247     0.687800
#>  ds 3 parent   14     47.1  4.830e+01  -1.20056 3.9247    -0.305896
#>  ds 3 parent   28     36.1  3.426e+01   1.83885 2.8516     0.644839
#>  ds 3 parent   28     36.6  3.426e+01   2.33885 2.8516     0.820177
#>  ds 3 parent   60     20.1  1.968e+01   0.42208 1.7881     0.236053
#>  ds 3 parent   60     19.8  1.968e+01   0.12208 1.7881     0.068273
#>  ds 3 parent   90     11.3  1.194e+01  -0.64013 1.2893    -0.496496
#>  ds 3 parent   90     10.7  1.194e+01  -1.24013 1.2893    -0.961865
#>  ds 3 parent  120      8.2  7.247e+00   0.95264 1.0476     0.909381
#>  ds 3 parent  120      7.3  7.247e+00   0.05264 1.0476     0.050254
#>  ds 3     m1    0      0.8 -2.956e-12   0.80000 0.8763     0.912928
#>  ds 3     m1    1      1.8  1.757e+00   0.04318 0.8873     0.048666
#>  ds 3     m1    1      2.3  1.757e+00   0.54318 0.8873     0.612186
#>  ds 3     m1    3      4.2  4.566e+00  -0.36607 0.9480    -0.386149
#>  ds 3     m1    3      4.1  4.566e+00  -0.46607 0.9480    -0.491634
#>  ds 3     m1    7      6.8  8.157e+00  -1.35680 1.0887    -1.246241
#>  ds 3     m1    7     10.1  8.157e+00   1.94320 1.0887     1.784855
#>  ds 3     m1   14     11.4  1.085e+01   0.55367 1.2272     0.451182
#>  ds 3     m1   14     12.8  1.085e+01   1.95367 1.2272     1.592023
#>  ds 3     m1   28     11.5  1.149e+01   0.01098 1.2633     0.008689
#>  ds 3     m1   28     10.6  1.149e+01  -0.88902 1.2633    -0.703717
#>  ds 3     m1   60      7.5  9.295e+00  -1.79500 1.1445    -1.568351
#>  ds 3     m1   60      8.6  9.295e+00  -0.69500 1.1445    -0.607245
#>  ds 3     m1   90      7.3  7.017e+00   0.28305 1.0377     0.272775
#>  ds 3     m1   90      8.1  7.017e+00   1.08305 1.0377     1.043720
#>  ds 3     m1  120      5.3  5.087e+00   0.21272 0.9645     0.220547
#>  ds 3     m1  120      3.8  5.087e+00  -1.28728 0.9645    -1.334660
#>  ds 4 parent    0    104.7  1.011e+02   3.60049 8.0554     0.446965
#>  ds 4 parent    0     88.3  1.011e+02 -12.79951 8.0554    -1.588930
#>  ds 4 parent    1     94.2  9.755e+01  -3.35176 7.7762    -0.431030
#>  ds 4 parent    1     94.6  9.755e+01  -2.95176 7.7762    -0.379591
#>  ds 4 parent    3     78.1  9.095e+01 -12.85198 7.2570    -1.770981
#>  ds 4 parent    3     96.5  9.095e+01   5.54802 7.2570     0.764508
#>  ds 4 parent    7     76.2  7.949e+01  -3.29267 6.3569    -0.517966
#>  ds 4 parent    7     77.8  7.949e+01  -1.69267 6.3569    -0.266272
#>  ds 4 parent   14     70.8  6.384e+01   6.95621 5.1321     1.355423
#>  ds 4 parent   14     67.3  6.384e+01   3.45621 5.1321     0.673445
#>  ds 4 parent   28     43.1  4.345e+01  -0.35291 3.5515    -0.099370
#>  ds 4 parent   28     45.1  4.345e+01   1.64709 3.5515     0.463771
#>  ds 4 parent   60     21.3  2.137e+01  -0.07478 1.9063    -0.039229
#>  ds 4 parent   60     23.5  2.137e+01   2.12522 1.9063     1.114813
#>  ds 4 parent   90     11.8  1.205e+01  -0.24925 1.2957    -0.192375
#>  ds 4 parent   90     12.1  1.205e+01   0.05075 1.2957     0.039168
#>  ds 4 parent  120      7.0  6.967e+00   0.03315 1.0356     0.032013
#>  ds 4 parent  120      6.2  6.967e+00  -0.76685 1.0356    -0.740510
#>  ds 4     m1    0      1.6  1.421e-13   1.60000 0.8763     1.825856
#>  ds 4     m1    1      0.9  7.250e-01   0.17503 0.8782     0.199310
#>  ds 4     m1    3      3.7  2.038e+00   1.66201 0.8910     1.865236
#>  ds 4     m1    3      2.0  2.038e+00  -0.03799 0.8910    -0.042637
#>  ds 4     m1    7      3.6  4.186e+00  -0.58623 0.9369    -0.625692
#>  ds 4     m1    7      3.8  4.186e+00  -0.38623 0.9369    -0.412230
#>  ds 4     m1   14      7.1  6.752e+00   0.34768 1.0266     0.338666
#>  ds 4     m1   14      6.6  6.752e+00  -0.15232 1.0266    -0.148372
#>  ds 4     m1   28      9.5  9.034e+00   0.46628 1.1313     0.412159
#>  ds 4     m1   28      9.3  9.034e+00   0.26628 1.1313     0.235373
#>  ds 4     m1   60      8.3  8.634e+00  -0.33359 1.1115    -0.300112
#>  ds 4     m1   60      9.0  8.634e+00   0.36641 1.1115     0.329645
#>  ds 4     m1   90      6.6  6.671e+00  -0.07091 1.0233    -0.069295
#>  ds 4     m1   90      7.7  6.671e+00   1.02909 1.0233     1.005691
#>  ds 4     m1  120      3.7  4.823e+00  -1.12301 0.9559    -1.174763
#>  ds 4     m1  120      3.5  4.823e+00  -1.32301 0.9559    -1.383979
#>  ds 5 parent    0    110.4  1.011e+02   9.30049 8.0554     1.154563
#>  ds 5 parent    0    112.1  1.011e+02  11.00049 8.0554     1.365601
#>  ds 5 parent    1     93.5  9.440e+01  -0.90098 7.5282    -0.119681
#>  ds 5 parent    1     91.0  9.440e+01  -3.40098 7.5282    -0.451764
#>  ds 5 parent    3     71.0  8.287e+01 -11.86698 6.6217    -1.792122
#>  ds 5 parent    3     89.7  8.287e+01   6.83302 6.6217     1.031907
#>  ds 5 parent    7     60.4  6.562e+01  -5.22329 5.2711    -0.990936
#>  ds 5 parent    7     59.1  6.562e+01  -6.52329 5.2711    -1.237566
#>  ds 5 parent   14     56.5  4.739e+01   9.10588 3.8548     2.362225
#>  ds 5 parent   14     47.0  4.739e+01  -0.39412 3.8548    -0.102240
#>  ds 5 parent   28     30.2  3.118e+01  -0.98128 2.6206    -0.374451
#>  ds 5 parent   28     23.9  3.118e+01  -7.28128 2.6206    -2.778500
#>  ds 5 parent   60     17.0  1.804e+01  -1.03959 1.6761    -0.620224
#>  ds 5 parent   60     18.7  1.804e+01   0.66041 1.6761     0.394008
#>  ds 5 parent   90     11.3  1.165e+01  -0.35248 1.2727    -0.276958
#>  ds 5 parent   90     11.9  1.165e+01   0.24752 1.2727     0.194488
#>  ds 5 parent  120      9.0  7.556e+00   1.44368 1.0612     1.360449
#>  ds 5 parent  120      8.1  7.556e+00   0.54368 1.0612     0.512338
#>  ds 5     m1    0      0.7 -1.421e-14   0.70000 0.8763     0.798812
#>  ds 5     m1    1      3.0  3.160e+00  -0.15979 0.9113    -0.175340
#>  ds 5     m1    1      2.6  3.160e+00  -0.55979 0.9113    -0.614254
#>  ds 5     m1    3      5.1  8.448e+00  -3.34789 1.1026    -3.036487
#>  ds 5     m1    3      7.5  8.448e+00  -0.94789 1.1026    -0.859720
#>  ds 5     m1    7     16.5  1.581e+01   0.68760 1.5286     0.449839
#>  ds 5     m1    7     19.0  1.581e+01   3.18760 1.5286     2.085373
#>  ds 5     m1   14     22.9  2.218e+01   0.71983 1.9632     0.366658
#>  ds 5     m1   14     23.2  2.218e+01   1.01983 1.9632     0.519469
#>  ds 5     m1   28     22.2  2.425e+01  -2.05105 2.1113    -0.971479
#>  ds 5     m1   28     24.4  2.425e+01   0.14895 2.1113     0.070552
#>  ds 5     m1   60     15.5  1.876e+01  -3.25968 1.7250    -1.889646
#>  ds 5     m1   60     19.8  1.876e+01   1.04032 1.7250     0.603074
#>  ds 5     m1   90     14.9  1.365e+01   1.25477 1.3914     0.901806
#>  ds 5     m1   90     14.2  1.365e+01   0.55477 1.3914     0.398714
#>  ds 5     m1  120     10.9  9.726e+00   1.17443 1.1667     1.006587
#>  ds 5     m1  120     10.4  9.726e+00   0.67443 1.1667     0.578044
# Add a correlation between random effects of g and k2
cov_model_3 <- f_saem_dfop_sfo_2$so@model@covariance.model
cov_model_3["log_k2", "g_qlogis"] <- 1
cov_model_3["g_qlogis", "log_k2"] <- 1
f_saem_dfop_sfo_3 <- update(f_saem_dfop_sfo,
  covariance.model = cov_model_3)
intervals(f_saem_dfop_sfo_3)
#> Approximate 95% confidence intervals
#> 
#>  Fixed effects:
#>                      lower         est.        upper
#> parent_0       98.42519529 101.51623115 104.60726702
#> k_m1            0.01505059   0.01662123   0.01835577
#> f_parent_to_m1  0.20100222   0.27477835   0.36332008
#> k1              0.07347479   0.10139028   0.13991179
#> k2              0.01469861   0.01771120   0.02134125
#> g               0.35506898   0.46263682   0.57379888
#> 
#>  Random effects:
#>                             lower       est.     upper
#> sd(f_parent_qlogis)    0.16472883  0.4435866 0.7224443
#> sd(log_k1)             0.05323856  0.2981783 0.5431180
#> sd(log_k2)             0.05013379  0.1912531 0.3323723
#> sd(g_qlogis)           0.04710647  0.3997298 0.7523531
#> corr(log_k2,g_qlogis) -1.31087397 -0.5845703 0.1417334
#> 
#>  
#>          lower       est.      upper
#> a.1 0.67769608 0.87421677 1.07073746
#> b.1 0.06525119 0.07925135 0.09325151
# The correlation does not improve the fit judged by AIC and BIC, although
# the likelihood is higher with the additional parameter
anova(f_saem_dfop_sfo, f_saem_dfop_sfo_2, f_saem_dfop_sfo_3)
#> Data: 171 observations of 2 variable(s) grouped in 5 datasets
#> 
#>                   npar    AIC    BIC     Lik
#> f_saem_dfop_sfo_2   12 806.91 802.23 -391.46
#> f_saem_dfop_sfo_3   13 807.96 802.88 -390.98
#> f_saem_dfop_sfo     14 810.83 805.36 -391.41
# }