Summary method for class "saem.mmkin"

Lists model equations, initial parameter values, optimised parameters for fixed effects (population), random effects (deviations from the population mean) and residual error model, as well as the resulting endpoints such as formation fractions and DT50 values. Optionally (default is FALSE), the data are listed in full.

# S3 method for saem.mmkin
summary(object, data = FALSE, verbose = FALSE, distimes = TRUE, ...)

# S3 method for summary.saem.mmkin
print(x, digits = max(3, getOption("digits") - 3), verbose = x$verbose, ...)

Arguments

object
an object of class saem.mmkin

data
logical, indicating whether the full data should be included in the summary.

verbose
Should the summary be verbose?

distimes
logical, indicating whether DT50 and DT90 values should be included.

...
optional arguments passed to methods like print.

x
an object of class summary.saem.mmkin

digits
Number of digits to use for printing

Value

The summary function returns a list based on the saemix::SaemixObject obtained in the fit, with at least the following additional components

saemixversion, mkinversion, Rversion

The saemix, mkin and R versions used

date.fit, date.summary

The dates where the fit and the summary were produced

diffs

The differential equations used in the degradation model

use_of_ff

Was maximum or minimum use made of formation fractions

data

The data

confint_trans

Transformed parameters as used in the optimisation, with confidence intervals

confint_back

Backtransformed parameters, with confidence intervals if available

ff

The estimated formation fractions derived from the fitted model.

distimes

The DT50 and DT90 values for each observed variable.

SFORB

If applicable, eigenvalues of SFORB components of the model.

The print method is called for its side effect, i.e. printing the summary.
Author

Johannes Ranke for the mkin specific parts saemix authors for the parts inherited from saemix.

Examples

# Generate five datasets following SFO kinetics
sampling_times = c(0, 1, 3, 7, 14, 28, 60, 90, 120)
dt50_sfo_in_pop <- 50
k_in_pop <- log(2) / dt50_sfo_in_pop
set.seed(1234)
k_in <- rlnorm(5, log(k_in_pop), 0.5)
SFO <- mkinmod(parent = mkinsub("SFO"))

pred_sfo <- function(k) {
  mkinpredict(SFO,
    c(k_parent = k),
    c(parent = 100),
    sampling_times)
}

ds_sfo_mean <- lapply(k_in, pred_sfo)
names(ds_sfo_mean) <- paste("ds", 1:5)

ds_sfo_syn <- lapply(ds_sfo_mean, function(ds) {
  add_err(ds,
    sdfunc = function(value) sqrt(1^2 + value^2 * 0.07^2),
    n = 1)[[1]]
})

# \dontrun{
# Evaluate using mmkin and saem
f_mmkin <- mmkin("SFO", ds_sfo_syn, quiet = TRUE, error_model = "tc", cores = 1)
f_saem <- saem(f_mmkin)

#> Running main SAEM algorithm
#> [1] "Sun Nov  8 02:58:03 2020"
#> ....
#>     Minimisation finished
#> [1] "Sun Nov  8 02:58:07 2020"

summary(f_saem, data = TRUE)

#> saemix version used for fitting:      3.1.9000 
#> mkin version used for pre-fitting:  0.9.50.4 
#> R version used for fitting:         4.0.3 
#> Date of fit:     Sun Nov  8 02:58:07 2020 
#> Date of summary: Sun Nov  8 02:58:07 2020 
#> 
#> Equations:
#> d_parent/dt = - k_parent * parent
#> 
#> Data:
#> 90 observations of 1 variable(s) grouped in 5 datasets
#> 
#> Model predictions using solution type analytical 
#> 
#> Fitted in 4.175 s using 300, 100 iterations
#> 
#> Variance model: Two-component variance function 
#> 
#> Mean of starting values for individual parameters:
#>     parent_0 log_k_parent 
#>    97.849556    -4.455036 
#> 
#> Fixed degradation parameter values:
#> None
#> 
#> Results:
#> 
#> Likelihood computed by importance sampling
#>        AIC      BIC    logLik
#>   555.6718 553.3284 -271.8359
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                   est.     lower      upper
#> parent_0     97.840291 94.701895 100.978686
#> log_k_parent -4.464246 -5.001413  -3.927078
#> 
#> Correlation: 
#>              prnt_0
#> log_k_parent 0.022 
#> 
#> Random effects:
#>                      est.     lower     upper
#> SD.parent_0     3.0805465 0.4960170 5.6650760
#> SD.log_k_parent 0.6109508 0.2301398 0.9917617
#> 
#> Variance model:
#>           est.      lower      upper
#> a.1 1.06249961 0.23012265 1.89487657
#> b.1 0.04649661 0.03060849 0.06238473
#> 
#> Backtransformed parameters with asymmetric confidence intervals:
#>                 est.        lower        upper
#> parent_0 97.84029088 94.701895315 100.97868644
#> k_parent  0.01151338  0.006728434   0.01970115
#> 
#> Estimated disappearance times:
#>        DT50 DT90
#> parent 60.2  200
#> 
#> Data:
#>    ds   name time observed predicted   residual standardized
#>  ds 1 parent    0    103.6    97.303   6.297361     -1.12720
#>  ds 1 parent    0     95.9    97.303  -1.402639      0.25107
#>  ds 1 parent    1     95.4    96.599  -1.198705      0.21583
#>  ds 1 parent    1     95.3    96.599  -1.298705      0.23383
#>  ds 1 parent    3     91.6    95.206  -3.606076      0.65693
#>  ds 1 parent    3     94.5    95.206  -0.706076      0.12863
#>  ds 1 parent    7     88.1    92.481  -4.380761      0.81692
#>  ds 1 parent    7     89.6    92.481  -2.880761      0.53720
#>  ds 1 parent   14     90.3    87.898   2.402166     -0.46649
#>  ds 1 parent   14     96.0    87.898   8.102166     -1.57340
#>  ds 1 parent   28     80.2    79.402   0.797948     -0.16783
#>  ds 1 parent   28     77.9    79.402  -1.502052      0.31593
#>  ds 1 parent   60     59.3    62.940  -3.639829      0.91247
#>  ds 1 parent   60     59.6    62.940  -3.339829      0.83726
#>  ds 1 parent   90     59.4    50.620   8.779547     -2.56999
#>  ds 1 parent   90     51.0    50.620   0.379547     -0.11110
#>  ds 1 parent  120     38.8    40.712  -1.912380      0.64706
#>  ds 1 parent  120     38.9    40.712  -1.812380      0.61323
#>  ds 2 parent    0    103.2    97.185   6.015158     -1.07774
#>  ds 2 parent    0     95.1    97.185  -2.084842      0.37354
#>  ds 2 parent    1     88.3    95.609  -7.309466      1.32706
#>  ds 2 parent    1    102.4    95.609   6.790534     -1.23285
#>  ds 2 parent    3     88.4    92.535  -4.134911      0.77071
#>  ds 2 parent    3     95.2    92.535   2.665089     -0.49675
#>  ds 2 parent    7     83.5    86.679  -3.179231      0.62426
#>  ds 2 parent    7     96.4    86.679   9.720769     -1.90873
#>  ds 2 parent   14     77.3    77.309  -0.009268      0.00199
#>  ds 2 parent   14     76.0    77.309  -1.309268      0.28113
#>  ds 2 parent   28     61.7    61.499   0.201491     -0.05137
#>  ds 2 parent   28     56.5    61.499  -4.998509      1.27449
#>  ds 2 parent   60     35.1    36.454  -1.353571      0.49087
#>  ds 2 parent   60     32.2    36.454  -4.253571      1.54256
#>  ds 2 parent   90     21.2    22.326  -1.125996      0.53604
#>  ds 2 parent   90     23.3    22.326   0.974004     -0.46368
#>  ds 2 parent  120     14.1    13.674   0.426440     -0.25110
#>  ds 2 parent  120     16.9    13.674   3.226440     -1.89983
#>  ds 3 parent    0     92.4    93.838  -1.437848      0.26501
#>  ds 3 parent    0     94.0    93.838   0.162152     -0.02989
#>  ds 3 parent    1     95.7    91.709   3.991384     -0.74932
#>  ds 3 parent    1     90.8    91.709  -0.908616      0.17058
#>  ds 3 parent    3     86.7    87.594  -0.893997      0.17409
#>  ds 3 parent    3     85.8    87.594  -1.793997      0.34934
#>  ds 3 parent    7     77.1    79.910  -2.810299      0.58817
#>  ds 3 parent    7     81.5    79.910   1.589701     -0.33271
#>  ds 3 parent   14     69.1    68.050   1.050102     -0.24845
#>  ds 3 parent   14     62.4    68.050  -5.649898      1.33675
#>  ds 3 parent   28     49.1    49.349  -0.248837      0.07412
#>  ds 3 parent   28     47.2    49.349  -2.148837      0.64010
#>  ds 3 parent   60     21.9    23.676  -1.775776      0.82085
#>  ds 3 parent   60     23.6    23.676  -0.075776      0.03503
#>  ds 3 parent   90     12.4    11.892   0.507660     -0.31425
#>  ds 3 parent   90     13.8    11.892   1.907660     -1.18088
#>  ds 3 parent  120      4.9     5.974  -1.073521      0.80099
#>  ds 3 parent  120      7.5     5.974   1.526479     -1.13895
#>  ds 4 parent    0     91.8   102.087 -10.287197      1.77084
#>  ds 4 parent    0    104.6   102.087   2.512803     -0.43256
#>  ds 4 parent    1    117.5   101.632  15.867689     -2.74145
#>  ds 4 parent    1     99.3   101.632  -2.332311      0.40295
#>  ds 4 parent    3     94.0   100.729  -6.728610      1.17100
#>  ds 4 parent    3     98.7   100.729  -2.028610      0.35304
#>  ds 4 parent    7    109.2    98.945  10.254756     -1.81080
#>  ds 4 parent    7     89.2    98.945  -9.745244      1.72083
#>  ds 4 parent   14    103.3    95.900   7.400009     -1.34021
#>  ds 4 parent   14    103.0    95.900   7.100009     -1.28588
#>  ds 4 parent   28     90.8    90.088   0.712227     -0.13563
#>  ds 4 parent   28     88.7    90.088  -1.387773      0.26427
#>  ds 4 parent   60     74.8    78.091  -3.291273      0.70124
#>  ds 4 parent   60     75.3    78.091  -2.791273      0.59471
#>  ds 4 parent   90     71.1    68.300   2.800407     -0.66075
#>  ds 4 parent   90     78.0    68.300   9.700407     -2.28880
#>  ds 4 parent  120     59.1    59.736  -0.635668      0.16554
#>  ds 4 parent  120     53.8    59.736  -5.935668      1.54574
#>  ds 5 parent    0     94.9    98.353  -3.452694      0.61266
#>  ds 5 parent    0    101.8    98.353   3.447306     -0.61171
#>  ds 5 parent    1     96.1    96.722  -0.621996      0.11187
#>  ds 5 parent    1     97.1    96.722   0.378004     -0.06799
#>  ds 5 parent    3     93.8    93.541   0.258738     -0.04781
#>  ds 5 parent    3     85.8    93.541  -7.741262      1.43043
#>  ds 5 parent    7     87.6    87.490   0.109848     -0.02141
#>  ds 5 parent    7     94.0    87.490   6.509848     -1.26885
#>  ds 5 parent   14     82.5    77.827   4.672679     -0.99818
#>  ds 5 parent   14     81.7    77.827   3.872679     -0.82728
#>  ds 5 parent   28     60.0    61.585  -1.585419      0.40382
#>  ds 5 parent   28     61.0    61.585  -0.585419      0.14911
#>  ds 5 parent   60     32.5    36.068  -3.568284      1.30251
#>  ds 5 parent   60     35.5    36.068  -0.568284      0.20744
#>  ds 5 parent   90     21.8    21.842  -0.042136      0.02028
#>  ds 5 parent   90     24.4    21.842   2.557864     -1.23088
#>  ds 5 parent  120     14.1    13.227   0.872898     -0.52035
#>  ds 5 parent  120     12.1    13.227  -1.127102      0.67189

# }