Summary method for class "saem.mmkin"

Lists model equations, initial parameter values, optimised parameters for fixed effects (population), random effects (deviations from the population mean) and residual error model, as well as the resulting endpoints such as formation fractions and DT50 values. Optionally (default is FALSE), the data are listed in full.

# S3 method for saem.mmkin
summary(object, data = FALSE, verbose = FALSE, distimes = TRUE, ...)

# S3 method for summary.saem.mmkin
print(x, digits = max(3, getOption("digits") - 3), verbose = x$verbose, ...)

Arguments

object

an object of class saem.mmkin

data

logical, indicating whether the full data should be included in the summary.

verbose

Should the summary be verbose?

distimes

logical, indicating whether DT50 and DT90 values should be included.

...

optional arguments passed to methods like print.

x

an object of class summary.saem.mmkin

digits

Number of digits to use for printing

Value

The summary function returns a list based on the saemix::SaemixObject

obtained in the fit, with at least the following additional components

saemixversion, mkinversion, Rversion

The saemix, mkin and R versions used

date.fit, date.summary

The dates where the fit and the summary were produced

diffs

The differential equations used in the degradation model

use_of_ff

Was maximum or minimum use made of formation fractions

data

The data

confint_trans

Transformed parameters as used in the optimisation, with confidence intervals

confint_back

Backtransformed parameters, with confidence intervals if available

confint_errmod

Error model parameters with confidence intervals

ff

The estimated formation fractions derived from the fitted model.

distimes

The DT50 and DT90 values for each observed variable.

SFORB

If applicable, eigenvalues of SFORB components of the model.

The print method is called for its side effect, i.e. printing the summary.

Author

Johannes Ranke for the mkin specific parts saemix authors for the parts inherited from saemix.

Examples

# Generate five datasets following DFOP-SFO kinetics
sampling_times = c(0, 1, 3, 7, 14, 28, 60, 90, 120)
dfop_sfo <- mkinmod(parent = mkinsub("DFOP", "m1"),
 m1 = mkinsub("SFO"), quiet = TRUE)
set.seed(1234)
k1_in <- rlnorm(5, log(0.1), 0.3)
k2_in <- rlnorm(5, log(0.02), 0.3)
g_in <- plogis(rnorm(5, qlogis(0.5), 0.3))
f_parent_to_m1_in <- plogis(rnorm(5, qlogis(0.3), 0.3))
k_m1_in <- rlnorm(5, log(0.02), 0.3)

pred_dfop_sfo <- function(k1, k2, g, f_parent_to_m1, k_m1) {
  mkinpredict(dfop_sfo,
    c(k1 = k1, k2 = k2, g = g, f_parent_to_m1 = f_parent_to_m1, k_m1 = k_m1),
    c(parent = 100, m1 = 0),
    sampling_times)
}

ds_mean_dfop_sfo <- lapply(1:5, function(i) {
  mkinpredict(dfop_sfo,
    c(k1 = k1_in[i], k2 = k2_in[i], g = g_in[i],
      f_parent_to_m1 = f_parent_to_m1_in[i], k_m1 = k_m1_in[i]),
    c(parent = 100, m1 = 0),
    sampling_times)
})
names(ds_mean_dfop_sfo) <- paste("ds", 1:5)

ds_syn_dfop_sfo <- lapply(ds_mean_dfop_sfo, function(ds) {
  add_err(ds,
    sdfunc = function(value) sqrt(1^2 + value^2 * 0.07^2),
    n = 1)[[1]]
})

# \dontrun{
# Evaluate using mmkin and saem
f_mmkin_dfop_sfo <- mmkin(list(dfop_sfo), ds_syn_dfop_sfo,
  quiet = TRUE, error_model = "tc", cores = 5)
f_saem_dfop_sfo <- saem(f_mmkin_dfop_sfo)
print(f_saem_dfop_sfo)
#> Kinetic nonlinear mixed-effects model fit by SAEM
#> Structural model:
#> d_parent/dt = - ((k1 * g * exp(-k1 * time) + k2 * (1 - g) * exp(-k2 *
#>            time)) / (g * exp(-k1 * time) + (1 - g) * exp(-k2 * time)))
#>            * parent
#> d_m1/dt = + f_parent_to_m1 * ((k1 * g * exp(-k1 * time) + k2 * (1 - g)
#>            * exp(-k2 * time)) / (g * exp(-k1 * time) + (1 - g) *
#>            exp(-k2 * time))) * parent - k_m1 * m1
#> 
#> Data:
#> 171 observations of 2 variable(s) grouped in 5 datasets
#> 
#> Likelihood computed by importance sampling
#>     AIC   BIC logLik
#>   810.8 805.4 -391.4
#> 
#> Fitted parameters:
#>                     estimate     lower     upper
#> parent_0           100.86947  97.81542 103.92353
#> log_k_m1            -4.06947  -4.16944  -3.96950
#> f_parent_qlogis     -0.93256  -1.34200  -0.52312
#> log_k1              -2.37017  -2.72660  -2.01375
#> log_k2              -4.06264  -4.21344  -3.91184
#> g_qlogis            -0.02174  -0.45898   0.41549
#> a.1                  0.87598   0.67275   1.07922
#> b.1                  0.07949   0.06389   0.09509
#> SD.parent_0          0.19170 -30.36286  30.74626
#> SD.log_k_m1          0.01883  -0.28736   0.32502
#> SD.f_parent_qlogis   0.44300   0.16391   0.72209
#> SD.log_k1            0.35320   0.09661   0.60978
#> SD.log_k2            0.13707   0.02359   0.25056
#> SD.g_qlogis          0.37478   0.04490   0.70467
illparms(f_saem_dfop_sfo)
#> [1] "sd(parent_0)" "sd(log_k_m1)"
f_saem_dfop_sfo_2 <- update(f_saem_dfop_sfo,
  no_random_effect = c("parent_0", "log_k_m1"))
illparms(f_saem_dfop_sfo_2)
intervals(f_saem_dfop_sfo_2)
#> Approximate 95% confidence intervals
#> 
#>  Fixed effects:
#>                      lower         est.        upper
#> parent_0       98.36731429 101.42508066 104.48284703
#> k_m1            0.01513234   0.01670094   0.01843214
#> f_parent_to_m1  0.20221431   0.27608850   0.36461630
#> k1              0.06915073   0.09759718   0.13774560
#> k2              0.01487068   0.01740389   0.02036863
#> g               0.37365671   0.48384821   0.59563299
#> 
#>  Random effects:
#>                          lower      est.     upper
#> sd(f_parent_qlogis) 0.16439770 0.4427585 0.7211193
#> sd(log_k1)          0.08304243 0.3345213 0.5860002
#> sd(log_k2)          0.03146410 0.1490210 0.2665779
#> sd(g_qlogis)        0.06216385 0.4023430 0.7425221
#> 
#>  
#>          lower       est.      upper
#> a.1 0.67696663 0.87777355 1.07858048
#> b.1 0.06363957 0.07878001 0.09392044
summary(f_saem_dfop_sfo_2, data = TRUE)
#> saemix version used for fitting:      3.2 
#> mkin version used for pre-fitting:  1.2.2 
#> R version used for fitting:         4.2.2 
#> Date of fit:     Thu Dec 15 14:47:14 2022 
#> Date of summary: Thu Dec 15 14:47:14 2022 
#> 
#> Equations:
#> d_parent/dt = - ((k1 * g * exp(-k1 * time) + k2 * (1 - g) * exp(-k2 *
#>            time)) / (g * exp(-k1 * time) + (1 - g) * exp(-k2 * time)))
#>            * parent
#> d_m1/dt = + f_parent_to_m1 * ((k1 * g * exp(-k1 * time) + k2 * (1 - g)
#>            * exp(-k2 * time)) / (g * exp(-k1 * time) + (1 - g) *
#>            exp(-k2 * time))) * parent - k_m1 * m1
#> 
#> Data:
#> 171 observations of 2 variable(s) grouped in 5 datasets
#> 
#> Model predictions using solution type analytical 
#> 
#> Fitted in 9.623 s
#> Using 300, 100 iterations and 10 chains
#> 
#> Variance model: Two-component variance function 
#> 
#> Starting values for degradation parameters:
#>        parent_0        log_k_m1 f_parent_qlogis          log_k1          log_k2 
#>       101.65645        -4.05368        -0.94311        -2.35943        -4.07006 
#>        g_qlogis 
#>        -0.01132 
#> 
#> Fixed degradation parameter values:
#> None
#> 
#> Starting values for random effects (square root of initial entries in omega):
#>                 parent_0 log_k_m1 f_parent_qlogis log_k1 log_k2 g_qlogis
#> parent_0           6.742   0.0000          0.0000 0.0000 0.0000    0.000
#> log_k_m1           0.000   0.2236          0.0000 0.0000 0.0000    0.000
#> f_parent_qlogis    0.000   0.0000          0.5572 0.0000 0.0000    0.000
#> log_k1             0.000   0.0000          0.0000 0.8031 0.0000    0.000
#> log_k2             0.000   0.0000          0.0000 0.0000 0.2931    0.000
#> g_qlogis           0.000   0.0000          0.0000 0.0000 0.0000    0.807
#> 
#> Starting values for error model parameters:
#> a.1 b.1 
#>   1   1 
#> 
#> Results:
#> 
#> Likelihood computed by importance sampling
#>   AIC   BIC logLik
#>   807 802.3 -391.5
#> 
#> Optimised parameters:
#>                         est.    lower     upper
#> parent_0           101.42508 98.36731 104.48285
#> log_k_m1            -4.09229 -4.19092  -3.99366
#> f_parent_qlogis     -0.96395 -1.37251  -0.55538
#> log_k1              -2.32691 -2.67147  -1.98235
#> log_k2              -4.05106 -4.20836  -3.89376
#> g_qlogis            -0.06463 -0.51656   0.38730
#> a.1                  0.87777  0.67697   1.07858
#> b.1                  0.07878  0.06364   0.09392
#> SD.f_parent_qlogis   0.44276  0.16440   0.72112
#> SD.log_k1            0.33452  0.08304   0.58600
#> SD.log_k2            0.14902  0.03146   0.26658
#> SD.g_qlogis          0.40234  0.06216   0.74252
#> 
#> Correlation: 
#>                 parnt_0 lg_k_m1 f_prnt_ log_k1  log_k2 
#> log_k_m1        -0.4693                                
#> f_parent_qlogis -0.2378  0.2595                        
#> log_k1           0.1720 -0.1593 -0.0669                
#> log_k2           0.0179  0.0594  0.0035  0.1995        
#> g_qlogis         0.1073 -0.1060 -0.0322 -0.2299 -0.3168
#> 
#> Random effects:
#>                      est.   lower  upper
#> SD.f_parent_qlogis 0.4428 0.16440 0.7211
#> SD.log_k1          0.3345 0.08304 0.5860
#> SD.log_k2          0.1490 0.03146 0.2666
#> SD.g_qlogis        0.4023 0.06216 0.7425
#> 
#> Variance model:
#>        est.   lower   upper
#> a.1 0.87777 0.67697 1.07858
#> b.1 0.07878 0.06364 0.09392
#> 
#> Backtransformed parameters:
#>                    est.    lower     upper
#> parent_0       101.4251 98.36731 104.48285
#> k_m1             0.0167  0.01513   0.01843
#> f_parent_to_m1   0.2761  0.20221   0.36462
#> k1               0.0976  0.06915   0.13775
#> k2               0.0174  0.01487   0.02037
#> g                0.4838  0.37366   0.59563
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_m1   0.2761
#> parent_sink 0.7239
#> 
#> Estimated disappearance times:
#>         DT50   DT90 DT50back DT50_k1 DT50_k2
#> parent 15.54  94.33     28.4   7.102   39.83
#> m1     41.50 137.87       NA      NA      NA
#> 
#> Data:
#>    ds   name time observed predicted  residual    std standardized
#>  ds 1 parent    0     89.8 1.014e+02 -11.62508 8.0383     -1.44620
#>  ds 1 parent    0    104.1 1.014e+02   2.67492 8.0383      0.33277
#>  ds 1 parent    1     88.7 9.650e+01  -7.80311 7.6530     -1.01961
#>  ds 1 parent    1     95.5 9.650e+01  -1.00311 7.6530     -0.13107
#>  ds 1 parent    3     81.8 8.753e+01  -5.72638 6.9510     -0.82382
#>  ds 1 parent    3     94.5 8.753e+01   6.97362 6.9510      1.00326
#>  ds 1 parent    7     71.5 7.254e+01  -1.04133 5.7818     -0.18010
#>  ds 1 parent    7     70.3 7.254e+01  -2.24133 5.7818     -0.38765
#>  ds 1 parent   14     54.2 5.349e+01   0.71029 4.3044      0.16502
#>  ds 1 parent   14     49.6 5.349e+01  -3.88971 4.3044     -0.90366
#>  ds 1 parent   28     31.5 3.167e+01  -0.16616 2.6446     -0.06283
#>  ds 1 parent   28     28.8 3.167e+01  -2.86616 2.6446     -1.08379
#>  ds 1 parent   60     12.1 1.279e+01  -0.69287 1.3365     -0.51843
#>  ds 1 parent   60     13.6 1.279e+01   0.80713 1.3365      0.60392
#>  ds 1 parent   90      6.2 6.397e+00  -0.19718 1.0122     -0.19481
#>  ds 1 parent   90      8.3 6.397e+00   1.90282 1.0122      1.87996
#>  ds 1 parent  120      2.2 3.323e+00  -1.12320 0.9160     -1.22623
#>  ds 1 parent  120      2.4 3.323e+00  -0.92320 0.9160     -1.00788
#>  ds 1     m1    1      0.3 1.179e+00  -0.87919 0.8827     -0.99605
#>  ds 1     m1    1      0.2 1.179e+00  -0.97919 0.8827     -1.10935
#>  ds 1     m1    3      2.2 3.273e+00  -1.07272 0.9149     -1.17256
#>  ds 1     m1    3      3.0 3.273e+00  -0.27272 0.9149     -0.29811
#>  ds 1     m1    7      6.5 6.559e+00  -0.05872 1.0186     -0.05765
#>  ds 1     m1    7      5.0 6.559e+00  -1.55872 1.0186     -1.53032
#>  ds 1     m1   14     10.2 1.016e+01   0.03787 1.1880      0.03188
#>  ds 1     m1   14      9.5 1.016e+01  -0.66213 1.1880     -0.55734
#>  ds 1     m1   28     12.2 1.268e+01  -0.47913 1.3297     -0.36032
#>  ds 1     m1   28     13.4 1.268e+01   0.72087 1.3297      0.54211
#>  ds 1     m1   60     11.8 1.078e+01   1.02493 1.2211      0.83936
#>  ds 1     m1   60     13.2 1.078e+01   2.42493 1.2211      1.98588
#>  ds 1     m1   90      6.6 7.705e+00  -1.10464 1.0672     -1.03509
#>  ds 1     m1   90      9.3 7.705e+00   1.59536 1.0672      1.49491
#>  ds 1     m1  120      3.5 5.236e+00  -1.73617 0.9699     -1.79010
#>  ds 1     m1  120      5.4 5.236e+00   0.16383 0.9699      0.16892
#>  ds 2 parent    0    118.0 1.014e+02  16.57492 8.0383      2.06198
#>  ds 2 parent    0     99.8 1.014e+02  -1.62508 8.0383     -0.20217
#>  ds 2 parent    1     90.2 9.599e+01  -5.79045 7.6129     -0.76061
#>  ds 2 parent    1     94.6 9.599e+01  -1.39045 7.6129     -0.18264
#>  ds 2 parent    3     96.1 8.652e+01   9.57931 6.8724      1.39388
#>  ds 2 parent    3     78.4 8.652e+01  -8.12069 6.8724     -1.18164
#>  ds 2 parent    7     77.9 7.197e+01   5.93429 5.7370      1.03439
#>  ds 2 parent    7     77.7 7.197e+01   5.73429 5.7370      0.99953
#>  ds 2 parent   14     56.0 5.555e+01   0.44657 4.4637      0.10005
#>  ds 2 parent   14     54.7 5.555e+01  -0.85343 4.4637     -0.19120
#>  ds 2 parent   28     36.6 3.853e+01  -1.93170 3.1599     -0.61132
#>  ds 2 parent   28     36.8 3.853e+01  -1.73170 3.1599     -0.54803
#>  ds 2 parent   60     22.1 2.110e+01   1.00360 1.8795      0.53396
#>  ds 2 parent   60     24.7 2.110e+01   3.60360 1.8795      1.91728
#>  ds 2 parent   90     12.4 1.250e+01  -0.09712 1.3190     -0.07363
#>  ds 2 parent   90     10.8 1.250e+01  -1.69712 1.3190     -1.28667
#>  ds 2 parent  120      6.8 7.419e+00  -0.61913 1.0546     -0.58709
#>  ds 2 parent  120      7.9 7.419e+00   0.48087 1.0546      0.45599
#>  ds 2     m1    1      1.3 1.422e+00  -0.12194 0.8849     -0.13781
#>  ds 2     m1    3      3.7 3.831e+00  -0.13149 0.9282     -0.14166
#>  ds 2     m1    3      4.7 3.831e+00   0.86851 0.9282      0.93567
#>  ds 2     m1    7      8.1 7.292e+00   0.80812 1.0490      0.77034
#>  ds 2     m1    7      7.9 7.292e+00   0.60812 1.0490      0.57969
#>  ds 2     m1   14     10.1 1.055e+01  -0.45332 1.2090     -0.37495
#>  ds 2     m1   14     10.3 1.055e+01  -0.25332 1.2090     -0.20953
#>  ds 2     m1   28     10.7 1.230e+01  -1.59960 1.3074     -1.22347
#>  ds 2     m1   28     12.2 1.230e+01  -0.09960 1.3074     -0.07618
#>  ds 2     m1   60     10.7 1.065e+01   0.05342 1.2141      0.04400
#>  ds 2     m1   60     12.5 1.065e+01   1.85342 1.2141      1.52661
#>  ds 2     m1   90      9.1 8.196e+00   0.90368 1.0897      0.82930
#>  ds 2     m1   90      7.4 8.196e+00  -0.79632 1.0897     -0.73078
#>  ds 2     m1  120      6.1 5.997e+00   0.10252 0.9969      0.10284
#>  ds 2     m1  120      4.5 5.997e+00  -1.49748 0.9969     -1.50220
#>  ds 3 parent    0    106.2 1.014e+02   4.77492 8.0383      0.59402
#>  ds 3 parent    0    106.9 1.014e+02   5.47492 8.0383      0.68110
#>  ds 3 parent    1    107.4 9.390e+01  13.49935 7.4494      1.81214
#>  ds 3 parent    1     96.1 9.390e+01   2.19935 7.4494      0.29524
#>  ds 3 parent    3     79.4 8.152e+01  -2.12307 6.4821     -0.32753
#>  ds 3 parent    3     82.6 8.152e+01   1.07693 6.4821      0.16614
#>  ds 3 parent    7     63.9 6.446e+01  -0.55834 5.1533     -0.10834
#>  ds 3 parent    7     62.4 6.446e+01  -2.05834 5.1533     -0.39942
#>  ds 3 parent   14     51.0 4.826e+01   2.74073 3.9019      0.70241
#>  ds 3 parent   14     47.1 4.826e+01  -1.15927 3.9019     -0.29711
#>  ds 3 parent   28     36.1 3.424e+01   1.86399 2.8364      0.65718
#>  ds 3 parent   28     36.6 3.424e+01   2.36399 2.8364      0.83346
#>  ds 3 parent   60     20.1 1.968e+01   0.42172 1.7815      0.23672
#>  ds 3 parent   60     19.8 1.968e+01   0.12172 1.7815      0.06833
#>  ds 3 parent   90     11.3 1.195e+01  -0.64633 1.2869     -0.50222
#>  ds 3 parent   90     10.7 1.195e+01  -1.24633 1.2869     -0.96844
#>  ds 3 parent  120      8.2 7.255e+00   0.94532 1.0474      0.90251
#>  ds 3 parent  120      7.3 7.255e+00   0.04532 1.0474      0.04327
#>  ds 3     m1    0      0.8 2.956e-11   0.80000 0.8778      0.91140
#>  ds 3     m1    1      1.8 1.758e+00   0.04187 0.8886      0.04712
#>  ds 3     m1    1      2.3 1.758e+00   0.54187 0.8886      0.60978
#>  ds 3     m1    3      4.2 4.567e+00  -0.36697 0.9486     -0.38683
#>  ds 3     m1    3      4.1 4.567e+00  -0.46697 0.9486     -0.49224
#>  ds 3     m1    7      6.8 8.151e+00  -1.35124 1.0876     -1.24242
#>  ds 3     m1    7     10.1 8.151e+00   1.94876 1.0876      1.79182
#>  ds 3     m1   14     11.4 1.083e+01   0.57098 1.2240      0.46647
#>  ds 3     m1   14     12.8 1.083e+01   1.97098 1.2240      1.61022
#>  ds 3     m1   28     11.5 1.147e+01   0.03175 1.2597      0.02520
#>  ds 3     m1   28     10.6 1.147e+01  -0.86825 1.2597     -0.68928
#>  ds 3     m1   60      7.5 9.298e+00  -1.79834 1.1433     -1.57298
#>  ds 3     m1   60      8.6 9.298e+00  -0.69834 1.1433     -0.61083
#>  ds 3     m1   90      7.3 7.038e+00   0.26249 1.0382      0.25283
#>  ds 3     m1   90      8.1 7.038e+00   1.06249 1.0382      1.02340
#>  ds 3     m1  120      5.3 5.116e+00   0.18417 0.9659      0.19068
#>  ds 3     m1  120      3.8 5.116e+00  -1.31583 0.9659     -1.36232
#>  ds 4 parent    0    104.7 1.014e+02   3.27492 8.0383      0.40741
#>  ds 4 parent    0     88.3 1.014e+02 -13.12508 8.0383     -1.63281
#>  ds 4 parent    1     94.2 9.781e+01  -3.61183 7.7555     -0.46572
#>  ds 4 parent    1     94.6 9.781e+01  -3.21183 7.7555     -0.41414
#>  ds 4 parent    3     78.1 9.110e+01 -13.00467 7.2307     -1.79853
#>  ds 4 parent    3     96.5 9.110e+01   5.39533 7.2307      0.74617
#>  ds 4 parent    7     76.2 7.951e+01  -3.30511 6.3246     -0.52258
#>  ds 4 parent    7     77.8 7.951e+01  -1.70511 6.3246     -0.26960
#>  ds 4 parent   14     70.8 6.376e+01   7.03783 5.0993      1.38016
#>  ds 4 parent   14     67.3 6.376e+01   3.53783 5.0993      0.69379
#>  ds 4 parent   28     43.1 4.340e+01  -0.30456 3.5303     -0.08627
#>  ds 4 parent   28     45.1 4.340e+01   1.69544 3.5303      0.48026
#>  ds 4 parent   60     21.3 2.142e+01  -0.12077 1.9022     -0.06349
#>  ds 4 parent   60     23.5 2.142e+01   2.07923 1.9022      1.09308
#>  ds 4 parent   90     11.8 1.207e+01  -0.26813 1.2940     -0.20721
#>  ds 4 parent   90     12.1 1.207e+01   0.03187 1.2940      0.02463
#>  ds 4 parent  120      7.0 6.954e+00   0.04554 1.0347      0.04402
#>  ds 4 parent  120      6.2 6.954e+00  -0.75446 1.0347     -0.72914
#>  ds 4     m1    0      1.6 1.990e-13   1.60000 0.8778      1.82279
#>  ds 4     m1    1      0.9 7.305e-01   0.16949 0.8797      0.19267
#>  ds 4     m1    3      3.7 2.051e+00   1.64896 0.8925      1.84753
#>  ds 4     m1    3      2.0 2.051e+00  -0.05104 0.8925     -0.05719
#>  ds 4     m1    7      3.6 4.204e+00  -0.60375 0.9382     -0.64354
#>  ds 4     m1    7      3.8 4.204e+00  -0.40375 0.9382     -0.43036
#>  ds 4     m1   14      7.1 6.760e+00   0.34021 1.0267      0.33137
#>  ds 4     m1   14      6.6 6.760e+00  -0.15979 1.0267     -0.15563
#>  ds 4     m1   28      9.5 9.011e+00   0.48856 1.1289      0.43277
#>  ds 4     m1   28      9.3 9.011e+00   0.28856 1.1289      0.25561
#>  ds 4     m1   60      8.3 8.611e+00  -0.31077 1.1093     -0.28014
#>  ds 4     m1   60      9.0 8.611e+00   0.38923 1.1093      0.35086
#>  ds 4     m1   90      6.6 6.678e+00  -0.07753 1.0233     -0.07576
#>  ds 4     m1   90      7.7 6.678e+00   1.02247 1.0233      0.99915
#>  ds 4     m1  120      3.7 4.847e+00  -1.14679 0.9572     -1.19804
#>  ds 4     m1  120      3.5 4.847e+00  -1.34679 0.9572     -1.40698
#>  ds 5 parent    0    110.4 1.014e+02   8.97492 8.0383      1.11651
#>  ds 5 parent    0    112.1 1.014e+02  10.67492 8.0383      1.32800
#>  ds 5 parent    1     93.5 9.466e+01  -1.16118 7.5089     -0.15464
#>  ds 5 parent    1     91.0 9.466e+01  -3.66118 7.5089     -0.48758
#>  ds 5 parent    3     71.0 8.302e+01 -12.01844 6.5988     -1.82130
#>  ds 5 parent    3     89.7 8.302e+01   6.68156 6.5988      1.01254
#>  ds 5 parent    7     60.4 6.563e+01  -5.22574 5.2440     -0.99652
#>  ds 5 parent    7     59.1 6.563e+01  -6.52574 5.2440     -1.24442
#>  ds 5 parent   14     56.5 4.727e+01   9.22621 3.8263      2.41128
#>  ds 5 parent   14     47.0 4.727e+01  -0.27379 3.8263     -0.07156
#>  ds 5 parent   28     30.2 3.103e+01  -0.83405 2.5977     -0.32108
#>  ds 5 parent   28     23.9 3.103e+01  -7.13405 2.5977     -2.74634
#>  ds 5 parent   60     17.0 1.800e+01  -0.99696 1.6675     -0.59787
#>  ds 5 parent   60     18.7 1.800e+01   0.70304 1.6675      0.42161
#>  ds 5 parent   90     11.3 1.167e+01  -0.36809 1.2710     -0.28961
#>  ds 5 parent   90     11.9 1.167e+01   0.23191 1.2710      0.18246
#>  ds 5 parent  120      9.0 7.595e+00   1.40496 1.0623      1.32256
#>  ds 5 parent  120      8.1 7.595e+00   0.50496 1.0623      0.47535
#>  ds 5     m1    0      0.7 0.000e+00   0.70000 0.8778      0.79747
#>  ds 5     m1    1      3.0 3.158e+00  -0.15799 0.9123     -0.17317
#>  ds 5     m1    1      2.6 3.158e+00  -0.55799 0.9123     -0.61160
#>  ds 5     m1    3      5.1 8.443e+00  -3.34286 1.1013     -3.03535
#>  ds 5     m1    3      7.5 8.443e+00  -0.94286 1.1013     -0.85613
#>  ds 5     m1    7     16.5 1.580e+01   0.69781 1.5232      0.45811
#>  ds 5     m1    7     19.0 1.580e+01   3.19781 1.5232      2.09935
#>  ds 5     m1   14     22.9 2.216e+01   0.73604 1.9543      0.37663
#>  ds 5     m1   14     23.2 2.216e+01   1.03604 1.9543      0.53014
#>  ds 5     m1   28     22.2 2.423e+01  -2.03128 2.1011     -0.96678
#>  ds 5     m1   28     24.4 2.423e+01   0.16872 2.1011      0.08030
#>  ds 5     m1   60     15.5 1.876e+01  -3.25610 1.7187     -1.89455
#>  ds 5     m1   60     19.8 1.876e+01   1.04390 1.7187      0.60739
#>  ds 5     m1   90     14.9 1.366e+01   1.23585 1.3890      0.88976
#>  ds 5     m1   90     14.2 1.366e+01   0.53585 1.3890      0.38579
#>  ds 5     m1  120     10.9 9.761e+00   1.13911 1.1670      0.97613
#>  ds 5     m1  120     10.4 9.761e+00   0.63911 1.1670      0.54767
# Add a correlation between random effects of g and k2
cov_model_3 <- f_saem_dfop_sfo_2$so@model@covariance.model
cov_model_3["log_k2", "g_qlogis"] <- 1
cov_model_3["g_qlogis", "log_k2"] <- 1
f_saem_dfop_sfo_3 <- update(f_saem_dfop_sfo,
  covariance.model = cov_model_3)
intervals(f_saem_dfop_sfo_3)
#> Approximate 95% confidence intervals
#> 
#>  Fixed effects:
#>                      lower         est.        upper
#> parent_0       98.39888363 101.48951337 104.58014311
#> k_m1            0.01508704   0.01665986   0.01839665
#> f_parent_to_m1  0.20141557   0.27540583   0.36418131
#> k1              0.07708759   0.10430866   0.14114200
#> k2              0.01476621   0.01786384   0.02161129
#> g               0.33679867   0.45083525   0.57028162
#> 
#>  Random effects:
#>                             lower       est.      upper
#> sd(f_parent_qlogis)    0.38085375  0.4441841  0.5075145
#> sd(log_k1)             0.04774819  0.2660384  0.4843286
#> sd(log_k2)            -0.63842736  0.1977024  1.0338321
#> sd(g_qlogis)           0.22711289  0.4502227  0.6733326
#> corr(log_k2,g_qlogis) -0.83271473 -0.6176939 -0.4026730
#> 
#>  
#>          lower       est.      upper
#> a.1 0.67347568 0.87437392 1.07527216
#> b.1 0.06393032 0.07912417 0.09431802
# The correlation does not improve the fit judged by AIC and BIC, although
# the likelihood is higher with the additional parameter
anova(f_saem_dfop_sfo, f_saem_dfop_sfo_2, f_saem_dfop_sfo_3)
#> Data: 171 observations of 2 variable(s) grouped in 5 datasets
#> 
#>                   npar    AIC    BIC     Lik
#> f_saem_dfop_sfo_2   12 806.96 802.27 -391.48
#> f_saem_dfop_sfo_3   13 807.99 802.91 -391.00
#> f_saem_dfop_sfo     14 810.83 805.36 -391.42
# }