Synthetic datasets for one parent compound with two metabolites

The 12 datasets were generated using four different models and three different variance components. The four models are either the SFO or the DFOP model with either two sequential or two parallel metabolites.

Variance component 'a' is based on a normal distribution with standard deviation of 3, Variance component 'b' is also based on a normal distribution, but with a standard deviation of 7. Variance component 'c' is based on the error model from Rocke and Lorenzato (1995), with the minimum standard deviation (for small y values) of 0.5, and a proportionality constant of 0.07 for the increase of the standard deviation with y.

Initial concentrations for metabolites and all values where adding the variance component resulted in a value below the assumed limit of detection of 0.1 were set to NA.

As an example, the first dataset has the title SFO_lin_a and is based on the SFO model with two sequential metabolites (linear pathway), with added variance component 'a'.

Compare also the code in the example section to see the degradation models.

synthetic_data_for_UBA_2014

Format

A list containing datasets in the form internally used by the 'gmkin' package. The list has twelfe components. Each of the components is one dataset that has, among others, the following components

title

The name of the dataset, e.g. SFO_lin_a

data

A data frame with the data in the form expected by mkinfit

Source

Ranke (2014) Prüfung und Validierung von Modellierungssoftware als Alternative zu ModelMaker 4.0, Umweltbundesamt Projektnummer 27452

Rocke, David M. und Lorenzato, Stefan (1995) A two-component model for measurement error in analytical chemistry. Technometrics 37(2), 176-184.

Examples

m_synth_SFO_lin <- mkinmod(parent = list(type = "SFO", to = "M1"),
                           M1 = list(type = "SFO", to = "M2"),
                           M2 = list(type = "SFO"), use_of_ff = "max")

#> Successfully compiled differential equation model from auto-generated C code.

m_synth_SFO_par <- mkinmod(parent = list(type = "SFO", to = c("M1", "M2"),
                                         sink = FALSE),
                           M1 = list(type = "SFO"),
                           M2 = list(type = "SFO"), use_of_ff = "max")

#> Successfully compiled differential equation model from auto-generated C code.

m_synth_DFOP_lin <- mkinmod(parent = list(type = "DFOP", to = "M1"),
                            M1 = list(type = "SFO", to = "M2"),
                            M2 = list(type = "SFO"), use_of_ff = "max")

#> Successfully compiled differential equation model from auto-generated C code.

m_synth_DFOP_par <- mkinmod(parent = list(type = "DFOP", to = c("M1", "M2"),
                                          sink = FALSE),
                            M1 = list(type = "SFO"),
                            M2 = list(type = "SFO"), use_of_ff = "max")

#> Successfully compiled differential equation model from auto-generated C code.

mkinfit(m_synth_SFO_lin, synthetic_data_for_UBA_2014[[1]]$data)

#> Model cost at call  1 :  31054.59 
#> Model cost at call  3 :  31054.59 
#> Model cost at call  8 :  15089.57 
#> Model cost at call  9 :  11464.3 
#> Model cost at call  11 :  11464.1 
#> Model cost at call  16 :  5723.32 
#> Model cost at call  17 :  5723.318 
#> Model cost at call  19 :  5723.304 
#> Model cost at call  21 :  5723.304 
#> Model cost at call  24 :  3968.126 
#> Model cost at call  25 :  3968.124 
#> Model cost at call  28 :  3968.119 
#> Model cost at call  31 :  3416.421 
#> Model cost at call  32 :  3416.42 
#> Model cost at call  36 :  3416.418 
#> Model cost at call  38 :  866.5564 
#> Model cost at call  42 :  866.5557 
#> Model cost at call  45 :  670.4833 
#> Model cost at call  47 :  670.476 
#> Model cost at call  53 :  312.9905 
#> Model cost at call  57 :  312.9904 
#> Model cost at call  58 :  312.9904 
#> Model cost at call  61 :  287.8916 
#> Model cost at call  63 :  287.8916 
#> Model cost at call  66 :  287.8916 
#> Model cost at call  69 :  284.5441 
#> Model cost at call  71 :  284.5441 
#> Model cost at call  73 :  284.5441 
#> Model cost at call  76 :  283.4533 
#> Model cost at call  78 :  283.4533 
#> Model cost at call  83 :  282.1356 
#> Model cost at call  85 :  282.1356 
#> Model cost at call  88 :  282.1356 
#> Model cost at call  90 :  280.7846 
#> Model cost at call  92 :  280.7846 
#> Model cost at call  95 :  280.7846 
#> Model cost at call  97 :  278.4856 
#> Model cost at call  98 :  274.5025 
#> Model cost at call  99 :  269.2866 
#> Model cost at call  101 :  269.2866 
#> Model cost at call  102 :  269.2866 
#> Model cost at call  103 :  269.2866 
#> Model cost at call  106 :  254.1284 
#> Model cost at call  108 :  254.1283 
#> Model cost at call  109 :  254.1283 
#> Model cost at call  112 :  254.128 
#> Model cost at call  114 :  233.1376 
#> Model cost at call  116 :  233.1376 
#> Model cost at call  118 :  233.1375 
#> Model cost at call  121 :  227.5879 
#> Model cost at call  124 :  227.5879 
#> Model cost at call  125 :  227.5878 
#> Model cost at call  129 :  217.0041 
#> Model cost at call  133 :  217.0041 
#> Model cost at call  135 :  217.0041 
#> Model cost at call  136 :  215.1367 
#> Model cost at call  138 :  215.1367 
#> Model cost at call  143 :  213.3794 
#> Model cost at call  145 :  213.3794 
#> Model cost at call  150 :  211.0201 
#> Model cost at call  152 :  211.0201 
#> Model cost at call  154 :  211.0201 
#> Model cost at call  155 :  211.0201 
#> Model cost at call  157 :  210.6426 
#> Model cost at call  159 :  210.6426 
#> Model cost at call  160 :  210.6425 
#> Model cost at call  164 :  207.6331 
#> Model cost at call  167 :  207.6331 
#> Model cost at call  171 :  206.2366 
#> Model cost at call  173 :  206.2366 
#> Model cost at call  174 :  206.2366 
#> Model cost at call  178 :  204.8117 
#> Model cost at call  180 :  204.8117 
#> Model cost at call  185 :  204.7988 
#> Model cost at call  187 :  204.7988 
#> Model cost at call  190 :  204.7988 
#> Model cost at call  192 :  203.5122 
#> Model cost at call  194 :  203.5122 
#> Model cost at call  197 :  203.5122 
#> Model cost at call  198 :  203.5122 
#> Model cost at call  199 :  203.354 
#> Model cost at call  201 :  203.354 
#> Model cost at call  204 :  203.354 
#> Model cost at call  206 :  202.6825 
#> Model cost at call  208 :  202.6825 
#> Model cost at call  209 :  202.6825 
#> Model cost at call  212 :  202.6825 
#> Model cost at call  213 :  202.4582 
#> Model cost at call  215 :  202.4582 
#> Model cost at call  220 :  202.3261 
#> Model cost at call  222 :  202.3261 
#> Model cost at call  227 :  202.2306 
#> Model cost at call  229 :  202.2306 
#> Model cost at call  231 :  202.2306 
#> Model cost at call  234 :  202.115 
#> Model cost at call  236 :  202.115 
#> Model cost at call  238 :  202.115 
#> Model cost at call  241 :  202.0397 
#> Model cost at call  243 :  202.0397 
#> Model cost at call  248 :  201.8989 
#> Model cost at call  249 :  201.8551 
#> Model cost at call  252 :  201.8551 
#> Model cost at call  257 :  201.676 
#> Model cost at call  259 :  201.676 
#> Model cost at call  264 :  201.6285 
#> Model cost at call  266 :  201.6285 
#> Model cost at call  270 :  201.6284 
#> Model cost at call  271 :  201.5876 
#> Model cost at call  272 :  201.5876 
#> Model cost at call  278 :  201.5317 
#> Model cost at call  279 :  201.5317 
#> Model cost at call  286 :  201.5207 
#> Model cost at call  287 :  201.5207 
#> Model cost at call  289 :  201.5207 
#> Model cost at call  293 :  201.5207 
#> Model cost at call  294 :  201.5207 
#> Model cost at call  296 :  201.5174 
#> Model cost at call  301 :  201.5174 
#> Model cost at call  304 :  201.5169 
#> Model cost at call  305 :  201.5169 
#> Model cost at call  306 :  201.5169 
#> Model cost at call  309 :  201.5169 
#> Model cost at call  312 :  201.5169 
#> Model cost at call  314 :  201.5169 
#> Model cost at call  322 :  201.5169 
#> Model cost at call  325 :  201.5169 
#> Model cost at call  340 :  201.5169 
#> Optimisation by method Port successfully terminated.

#> $par
#>       parent_0   log_k_parent       log_k_M1       log_k_M2 f_parent_ilr_1 
#>    102.0624835     -0.3020316     -1.2067882     -3.9007519      0.8491684 
#>     f_M1_ilr_1 
#>      0.6780411 
#> 
#> $ssr
#> [1] 201.5169
#> 
#> $convergence
#> [1] 0
#> 
#> $iterations
#> [1] 43
#> 
#> $evaluations
#> function gradient 
#>       56      281 
#> 
#> $counts
#> [1] "relative convergence (4)"
#> 
#> $hessian
#>                  parent_0 log_k_parent   log_k_M1    log_k_M2 f_parent_ilr_1
#> parent_0         8.433594    -29.66715  -18.40708   -68.90161       115.9976
#> log_k_parent   -29.667146  10561.33531  675.33998    55.94284      1666.8940
#> log_k_M1       -18.407082    675.33998 6274.11801    44.01714      -614.5674
#> log_k_M2       -68.901614     55.94284   44.01714  5021.66991     -2300.4467
#> f_parent_ilr_1 115.997604   1666.89403 -614.56735 -2300.44667      3872.8569
#> f_M1_ilr_1      92.819176    604.06870 1483.45826 -2755.79082      3098.9947
#>                 f_M1_ilr_1
#> parent_0          92.81918
#> log_k_parent     604.06870
#> log_k_M1        1483.45826
#> log_k_M2       -2755.79082
#> f_parent_ilr_1  3098.99466
#> f_M1_ilr_1      3712.39824
#> 
#> $residuals
#>      parent      parent      parent      parent      parent      parent 
#>  0.56248353  0.86248353 -5.17118695  1.22881305  0.70772795  3.50772795 
#>      parent      parent      parent      parent      parent      parent 
#> -0.52282962  0.27717038 -3.49673606 -3.19999990 -0.60000000 -3.50000000 
#>          M1          M1          M1          M1          M1          M1 
#> -1.61088639 -2.61088639  5.07026619 -0.42973381  0.38714436 -2.31285564 
#>          M1          M1          M1          M1          M1          M1 
#> -3.80468869  0.79531131 -0.49999789 -3.20000000 -1.50000000 -0.60000000 
#>          M2          M2          M2          M2          M2          M2 
#> -0.34517017  0.62526794  2.22526794 -0.07941701 -1.17941701 -3.83353798 
#>          M2          M2          M2          M2          M2          M2 
#>  1.26646202  0.87274743  2.47274743 -0.21837410  0.98162590 -0.47130583 
#>          M2          M2          M2 
#> -0.67130583 -4.27893112  2.22106888 
#> 
#> $ms
#> [1] 5.1671
#> 
#> $var_ms
#>   parent       M1       M2 
#> 6.461983 5.750942 3.664121 
#> 
#> $var_ms_unscaled
#>   parent       M1       M2 
#> 6.461983 5.750942 3.664121 
#> 
#> $var_ms_unweighted
#>   parent       M1       M2 
#> 6.461983 5.750942 3.664121 
#> 
#> $rank
#> [1] 6
#> 
#> $df.residual
#> [1] 33
#> 
#> $solution_type
#> [1] "deSolve"
#> 
#> $transform_rates
#> [1] TRUE
#> 
#> $transform_fractions
#> [1] TRUE
#> 
#> $method.modFit
#> [1] "Port"
#> 
#> $maxit.modFit
#> [1] "auto"
#> 
#> $calls
#> [1] 351
#> 
#> $time
#>    user  system elapsed 
#>   2.116   0.000   2.113 
#> 
#> $mkinmod
#> <mkinmod> model generated with
#> Use of formation fractions $use_of_ff: max 
#> Specification $spec:
#> $parent
#> $type: SFO; $to: M1; $sink: TRUE
#> $M1
#> $type: SFO; $to: M2; $sink: TRUE
#> $M2
#> $type: SFO; $sink: TRUE
#> Coefficient matrix $coefmat available
#> Compiled model $cf available
#> 
#> $observed
#>      name time value override err
#> 1  parent    0 101.5       NA   1
#> 2  parent    0 101.2       NA   1
#> 3  parent    1  53.9       NA   1
#> 4  parent    1  47.5       NA   1
#> 5  parent    3  10.4       NA   1
#> 6  parent    3   7.6       NA   1
#> 7  parent    7   1.1       NA   1
#> 8  parent    7   0.3       NA   1
#> 9  parent   14    NA       NA   1
#> 10 parent   14   3.5       NA   1
#> 11 parent   28    NA       NA   1
#> 12 parent   28   3.2       NA   1
#> 13 parent   60    NA       NA   1
#> 14 parent   60    NA       NA   1
#> 15 parent   90   0.6       NA   1
#> 16 parent   90    NA       NA   1
#> 17 parent  120    NA       NA   1
#> 18 parent  120   3.5       NA   1
#> 19     M1    0    NA       NA   1
#> 20     M1    0    NA       NA   1
#> 21     M1    1  36.4       NA   1
#> 22     M1    1  37.4       NA   1
#> 23     M1    3  34.3       NA   1
#> 24     M1    3  39.8       NA   1
#> 25     M1    7  15.1       NA   1
#> 26     M1    7  17.8       NA   1
#> 27     M1   14   5.8       NA   1
#> 28     M1   14   1.2       NA   1
#> 29     M1   28    NA       NA   1
#> 30     M1   28    NA       NA   1
#> 31     M1   60   0.5       NA   1
#> 32     M1   60    NA       NA   1
#> 33     M1   90    NA       NA   1
#> 34     M1   90   3.2       NA   1
#> 35     M1  120   1.5       NA   1
#> 36     M1  120   0.6       NA   1
#> 37     M2    0    NA       NA   1
#> 38     M2    0    NA       NA   1
#> 39     M2    1    NA       NA   1
#> 40     M2    1   4.8       NA   1
#> 41     M2    3  20.9       NA   1
#> 42     M2    3  19.3       NA   1
#> 43     M2    7  42.0       NA   1
#> 44     M2    7  43.1       NA   1
#> 45     M2   14  49.4       NA   1
#> 46     M2   14  44.3       NA   1
#> 47     M2   28  34.6       NA   1
#> 48     M2   28  33.0       NA   1
#> 49     M2   60  18.8       NA   1
#> 50     M2   60  17.6       NA   1
#> 51     M2   90  10.6       NA   1
#> 52     M2   90  10.8       NA   1
#> 53     M2  120   9.8       NA   1
#> 54     M2  120   3.3       NA   1
#> 
#> $obs_vars
#> [1] "parent" "M1"     "M2"    
#> 
#> $predicted
#>       name       time         value
#> 1   parent   0.000000  1.020625e+02
#> 2   parent   1.000000  4.872881e+01
#> 3   parent   1.212121  4.165603e+01
#> 4   parent   2.424242  1.700159e+01
#> 5   parent   3.000000  1.110773e+01
#> 6   parent   3.636364  6.939072e+00
#> 7   parent   4.848485  2.832130e+00
#> 8   parent   6.060606  1.155912e+00
#> 9   parent   7.000000  5.771704e-01
#> 10  parent   7.272727  4.717769e-01
#> 11  parent   8.484848  1.925522e-01
#> 12  parent   9.696970  7.858872e-02
#> 13  parent  10.909091  3.207539e-02
#> 14  parent  12.121212  1.309133e-02
#> 15  parent  13.333333  5.343128e-03
#> 16  parent  14.000000  3.263939e-03
#> 17  parent  14.545455  2.180757e-03
#> 18  parent  15.757576  8.900590e-04
#> 19  parent  16.969697  3.632705e-04
#> 20  parent  18.181818  1.482660e-04
#> 21  parent  19.393939  6.051327e-05
#> 22  parent  20.606061  2.469808e-05
#> 23  parent  21.818182  1.008035e-05
#> 24  parent  23.030303  4.114467e-06
#> 25  parent  24.242424  1.679140e-06
#> 26  parent  25.454545  6.853728e-07
#> 27  parent  26.666667  2.797450e-07
#> 28  parent  27.878788  1.142138e-07
#> 29  parent  28.000000  1.044512e-07
#> 30  parent  29.090909  4.657425e-08
#> 31  parent  30.303030  1.900245e-08
#> 32  parent  31.515152  7.760238e-09
#> 33  parent  32.727273  3.164577e-09
#> 34  parent  33.939394  1.291779e-09
#> 35  parent  35.151515  5.261577e-10
#> 36  parent  36.363636  2.132915e-10
#> 37  parent  37.575758  8.767818e-11
#> 38  parent  38.787879  3.442792e-11
#> 39  parent  40.000000  1.827291e-11
#> 40  parent  41.212121  3.771071e-12
#> 41  parent  42.424242  6.084856e-12
#> 42  parent  43.636364 -3.377858e-12
#> 43  parent  44.848485  5.870338e-12
#> 44  parent  46.060606 -6.263257e-12
#> 45  parent  47.272727  8.743492e-12
#> 46  parent  48.484848 -9.381771e-12
#> 47  parent  49.696970  1.403389e-11
#> 48  parent  50.909091 -3.592528e-11
#> 49  parent  52.121212 -8.487459e-11
#> 50  parent  53.333333 -3.309153e-12
#> 51  parent  54.545455 -2.966799e-11
#> 52  parent  55.757576 -4.723329e-11
#> 53  parent  56.969697  7.635833e-11
#> 54  parent  58.181818 -1.887064e-11
#> 55  parent  59.393939 -1.548352e-10
#> 56  parent  60.000000 -1.053819e-10
#> 57  parent  60.606061  5.780435e-12
#> 58  parent  61.818182  9.056244e-11
#> 59  parent  63.030303 -8.889581e-11
#> 60  parent  64.242424 -6.653389e-11
#> 61  parent  65.454545  1.181114e-10
#> 62  parent  66.666667 -9.226329e-12
#> 63  parent  67.878788 -8.897326e-11
#> 64  parent  69.090909  1.984998e-10
#> 65  parent  70.303030  3.255550e-11
#> 66  parent  71.515152 -2.991002e-10
#> 67  parent  72.727273  2.254268e-10
#> 68  parent  73.939394  2.696039e-10
#> 69  parent  75.151515  1.226806e-10
#> 70  parent  76.363636  3.447399e-11
#> 71  parent  77.575758  2.048902e-11
#> 72  parent  78.787879  6.830755e-12
#> 73  parent  80.000000  8.242171e-13
#> 74  parent  81.212121 -5.357740e-12
#> 75  parent  82.424242  2.198907e-11
#> 76  parent  83.636364  3.739511e-11
#> 77  parent  84.848485 -6.616091e-12
#> 78  parent  86.060606 -2.562689e-12
#> 79  parent  87.272727  4.089395e-11
#> 80  parent  88.484848 -2.042159e-11
#> 81  parent  89.696970 -4.088127e-11
#> 82  parent  90.000000 -1.874889e-11
#> 83  parent  90.909091  4.225747e-11
#> 84  parent  92.121212  8.054402e-12
#> 85  parent  93.333333  3.917595e-12
#> 86  parent  94.545455  6.591454e-12
#> 87  parent  95.757576  2.790958e-11
#> 88  parent  96.969697  2.720721e-12
#> 89  parent  98.181818 -1.304470e-12
#> 90  parent  99.393939  1.345055e-11
#> 91  parent 100.606061 -9.662077e-12
#> 92  parent 101.818182 -2.086798e-11
#> 93  parent 103.030303  9.332507e-12
#> 94  parent 104.242424 -6.752606e-12
#> 95  parent 105.454545 -3.326620e-11
#> 96  parent 106.666667  2.500680e-11
#> 97  parent 107.878788  2.184148e-11
#> 98  parent 109.090909 -5.985657e-11
#> 99  parent 110.303030 -8.750836e-14
#> 100 parent 111.515152  1.820588e-12
#> 101 parent 112.727273 -1.261472e-11
#> 102 parent 113.939394  1.455439e-11
#> 103 parent 115.151515  1.945812e-12
#> 104 parent 116.363636  9.598249e-13
#> 105 parent 117.575758  1.724679e-12
#> 106 parent 118.787879 -1.334504e-12
#> 107 parent 120.000000 -2.804801e-11
#> 108     M1   0.000000  0.000000e+00
#> 109     M1   1.000000  3.478911e+01
#> 110     M1   1.212121  3.791354e+01
#> 111     M1   2.424242  4.185645e+01
#> 112     M1   3.000000  3.937027e+01
#> 113     M1   3.636364  3.544167e+01
#> 114     M1   4.848485  2.723995e+01
#> 115     M1   6.060606  2.000711e+01
#> 116     M1   7.000000  1.548714e+01
#> 117     M1   7.272727  1.435144e+01
#> 118     M1   8.484848  1.016177e+01
#> 119     M1   9.696970  7.142649e+00
#> 120     M1  10.909091  4.999441e+00
#> 121     M1  12.121212  3.490801e+00
#> 122     M1  13.333333  2.433954e+00
#> 123     M1  14.000000  1.995311e+00
#> 124     M1  14.545455  1.695664e+00
#> 125     M1  15.757576  1.180746e+00
#> 126     M1  16.969697  8.219589e-01
#> 127     M1  18.181818  5.720991e-01
#> 128     M1  19.393939  3.981531e-01
#> 129     M1  20.606061  2.770793e-01
#> 130     M1  21.818182  1.928162e-01
#> 131     M1  23.030303  1.341758e-01
#> 132     M1  24.242424  9.336844e-02
#> 133     M1  25.454545  6.497152e-02
#> 134     M1  26.666667  4.521101e-02
#> 135     M1  27.878788  3.146041e-02
#> 136     M1  28.000000  3.034005e-02
#> 137     M1  29.090909  2.189192e-02
#> 138     M1  30.303030  1.523362e-02
#> 139     M1  31.515152  1.060040e-02
#> 140     M1  32.727273  7.376345e-03
#> 141     M1  33.939394  5.132870e-03
#> 142     M1  35.151515  3.571730e-03
#> 143     M1  36.363636  2.485406e-03
#> 144     M1  37.575758  1.729482e-03
#> 145     M1  38.787879  1.203467e-03
#> 146     M1  40.000000  8.374380e-04
#> 147     M1  41.212121  5.827347e-04
#> 148     M1  42.424242  4.054989e-04
#> 149     M1  43.636364  2.821681e-04
#> 150     M1  44.848485  1.963481e-04
#> 151     M1  46.060606  1.366297e-04
#> 152     M1  47.272727  9.507439e-05
#> 153     M1  48.484848  6.615797e-05
#> 154     M1  49.696970  4.603629e-05
#> 155     M1  50.909091  3.203434e-05
#> 156     M1  52.121212  2.229196e-05
#> 157     M1  53.333333  1.551223e-05
#> 158     M1  54.545455  1.079420e-05
#> 159     M1  55.757576  7.511255e-06
#> 160     M1  56.969697  5.226640e-06
#> 161     M1  58.181818  3.636450e-06
#> 162     M1  59.393939  2.530191e-06
#> 163     M1  60.000000  2.110651e-06
#> 164     M1  60.606061  1.760625e-06
#> 165     M1  61.818182  1.225095e-06
#> 166     M1  63.030303  8.527010e-07
#> 167     M1  64.242424  5.934161e-07
#> 168     M1  65.454545  4.127474e-07
#> 169     M1  66.666667  2.874114e-07
#> 170     M1  67.878788  2.001921e-07
#> 171     M1  69.090909  1.389331e-07
#> 172     M1  70.303030  9.678549e-08
#> 173     M1  71.515152  6.777214e-08
#> 174     M1  72.727273  4.658761e-08
#> 175     M1  73.939394  3.226837e-08
#> 176     M1  75.151515  2.253752e-08
#> 177     M1  76.363636  1.574843e-08
#> 178     M1  77.575758  1.096303e-08
#> 179     M1  78.787879  7.638209e-09
#> 180     M1  80.000000  5.319996e-09
#> 181     M1  81.212121  3.709993e-09
#> 182     M1  82.424242  2.548810e-09
#> 183     M1  83.636364  1.744629e-09
#> 184     M1  84.848485  1.256081e-09
#> 185     M1  86.060606  8.714672e-10
#> 186     M1  87.272727  5.511830e-10
#> 187     M1  88.484848  4.466725e-10
#> 188     M1  89.696970  3.452654e-10
#> 189     M1  90.000000  2.913252e-10
#> 190     M1  90.909091  1.489262e-10
#> 191     M1  92.121212  1.311985e-10
#> 192     M1  93.333333  9.347248e-11
#> 193     M1  94.545455  6.004640e-11
#> 194     M1  95.757576  1.166926e-11
#> 195     M1  96.969697  2.968203e-11
#> 196     M1  98.181818  2.478228e-11
#> 197     M1  99.393939 -1.291838e-12
#> 198     M1 100.606061  2.366481e-11
#> 199     M1 101.818182  3.472871e-11
#> 200     M1 103.030303 -6.633877e-12
#> 201     M1 104.242424  1.248743e-11
#> 202     M1 105.454545  4.557313e-11
#> 203     M1 106.666667 -3.046261e-11
#> 204     M1 107.878788 -2.693037e-11
#> 205     M1 109.090909  7.816593e-11
#> 206     M1 110.303030  7.276098e-13
#> 207     M1 111.515152 -1.922924e-12
#> 208     M1 112.727273  1.658481e-11
#> 209     M1 113.939394 -1.858452e-11
#> 210     M1 115.151515 -2.368198e-12
#> 211     M1 116.363636 -1.138989e-12
#> 212     M1 117.575758 -2.157011e-12
#> 213     M1 118.787879  1.771568e-12
#> 214     M1 120.000000  3.624738e-11
#> 215     M2   0.000000  0.000000e+00
#> 216     M2   1.000000  4.454830e+00
#> 217     M2   1.212121  6.103803e+00
#> 218     M2   2.424242  1.667567e+01
#> 219     M2   3.000000  2.152527e+01
#> 220     M2   3.636364  2.637280e+01
#> 221     M2   4.848485  3.384106e+01
#> 222     M2   6.060606  3.910279e+01
#> 223     M2   7.000000  4.192058e+01
#> 224     M2   7.272727  4.256708e+01
#> 225     M2   8.484848  4.467909e+01
#> 226     M2   9.696970  4.581396e+01
#> 227     M2  10.909091  4.625927e+01
#> 228     M2  12.121212  4.622588e+01
#> 229     M2  13.333333  4.586473e+01
#> 230     M2  14.000000  4.556646e+01
#> 231     M2  14.545455  4.528249e+01
#> 232     M2  15.757576  4.455394e+01
#> 233     M2  16.969697  4.373119e+01
#> 234     M2  18.181818  4.285048e+01
#> 235     M2  19.393939  4.193685e+01
#> 236     M2  20.606061  4.100759e+01
#> 237     M2  21.818182  4.007456e+01
#> 238     M2  23.030303  3.914584e+01
#> 239     M2  24.242424  3.822688e+01
#> 240     M2  25.454545  3.732133e+01
#> 241     M2  26.666667  3.643154e+01
#> 242     M2  27.878788  3.555901e+01
#> 243     M2  28.000000  3.547275e+01
#> 244     M2  29.090909  3.470463e+01
#> 245     M2  30.303030  3.386887e+01
#> 246     M2  31.515152  3.305190e+01
#> 247     M2  32.727273  3.225371e+01
#> 248     M2  33.939394  3.147416e+01
#> 249     M2  35.151515  3.071300e+01
#> 250     M2  36.363636  2.996993e+01
#> 251     M2  37.575758  2.924463e+01
#> 252     M2  38.787879  2.853672e+01
#> 253     M2  40.000000  2.784585e+01
#> 254     M2  41.212121  2.717163e+01
#> 255     M2  42.424242  2.651368e+01
#> 256     M2  43.636364  2.587163e+01
#> 257     M2  44.848485  2.524511e+01
#> 258     M2  46.060606  2.463374e+01
#> 259     M2  47.272727  2.403716e+01
#> 260     M2  48.484848  2.345502e+01
#> 261     M2  49.696970  2.288698e+01
#> 262     M2  50.909091  2.233268e+01
#> 263     M2  52.121212  2.179181e+01
#> 264     M2  53.333333  2.126404e+01
#> 265     M2  54.545455  2.074905e+01
#> 266     M2  55.757576  2.024653e+01
#> 267     M2  56.969697  1.975618e+01
#> 268     M2  58.181818  1.927770e+01
#> 269     M2  59.393939  1.881081e+01
#> 270     M2  60.000000  1.858163e+01
#> 271     M2  60.606061  1.835523e+01
#> 272     M2  61.818182  1.791068e+01
#> 273     M2  63.030303  1.747690e+01
#> 274     M2  64.242424  1.705363e+01
#> 275     M2  65.454545  1.664061e+01
#> 276     M2  66.666667  1.623759e+01
#> 277     M2  67.878788  1.584433e+01
#> 278     M2  69.090909  1.546059e+01
#> 279     M2  70.303030  1.508615e+01
#> 280     M2  71.515152  1.472078e+01
#> 281     M2  72.727273  1.436425e+01
#> 282     M2  73.939394  1.401636e+01
#> 283     M2  75.151515  1.367690e+01
#> 284     M2  76.363636  1.334566e+01
#> 285     M2  77.575758  1.302244e+01
#> 286     M2  78.787879  1.270705e+01
#> 287     M2  80.000000  1.239929e+01
#> 288     M2  81.212121  1.209899e+01
#> 289     M2  82.424242  1.180597e+01
#> 290     M2  83.636364  1.152004e+01
#> 291     M2  84.848485  1.124103e+01
#> 292     M2  86.060606  1.096878e+01
#> 293     M2  87.272727  1.070313e+01
#> 294     M2  88.484848  1.044391e+01
#> 295     M2  89.696970  1.019097e+01
#> 296     M2  90.000000  1.012869e+01
#> 297     M2  90.909091  9.944151e+00
#> 298     M2  92.121212  9.703312e+00
#> 299     M2  93.333333  9.468307e+00
#> 300     M2  94.545455  9.238993e+00
#> 301     M2  95.757576  9.015233e+00
#> 302     M2  96.969697  8.796892e+00
#> 303     M2  98.181818  8.583839e+00
#> 304     M2  99.393939  8.375946e+00
#> 305     M2 100.606061  8.173088e+00
#> 306     M2 101.818182  7.975143e+00
#> 307     M2 103.030303  7.781992e+00
#> 308     M2 104.242424  7.593520e+00
#> 309     M2 105.454545  7.409611e+00
#> 310     M2 106.666667  7.230157e+00
#> 311     M2 107.878788  7.055049e+00
#> 312     M2 109.090909  6.884182e+00
#> 313     M2 110.303030  6.717454e+00
#> 314     M2 111.515152  6.554763e+00
#> 315     M2 112.727273  6.396012e+00
#> 316     M2 113.939394  6.241107e+00
#> 317     M2 115.151515  6.089953e+00
#> 318     M2 116.363636  5.942460e+00
#> 319     M2 117.575758  5.798538e+00
#> 320     M2 118.787879  5.658103e+00
#> 321     M2 120.000000  5.521069e+00
#> 
#> $cost
#> function (P) 
#> {
#>     assign("calls", calls + 1, inherits = TRUE)
#>     if (trace_parms) 
#>         cat(P, "\n")
#>     if (length(state.ini.optim) > 0) {
#>         odeini <- c(P[1:length(state.ini.optim)], state.ini.fixed)
#>         names(odeini) <- c(state.ini.optim.boxnames, state.ini.fixed.boxnames)
#>     }
#>     else {
#>         odeini <- state.ini.fixed
#>         names(odeini) <- state.ini.fixed.boxnames
#>     }
#>     odeparms <- c(P[(length(state.ini.optim) + 1):length(P)], 
#>         transparms.fixed)
#>     parms <- backtransform_odeparms(odeparms, mkinmod, transform_rates = transform_rates, 
#>         transform_fractions = transform_fractions)
#>     out <- mkinpredict(mkinmod, parms, odeini, outtimes, solution_type = solution_type, 
#>         use_compiled = use_compiled, method.ode = method.ode, 
#>         atol = atol, rtol = rtol, ...)
#>     assign("out_predicted", out, inherits = TRUE)
#>     mC <- modCost(out, observed, y = "value", err = err, weight = weight, 
#>         scaleVar = scaleVar)
#>     if (mC$model < cost.old) {
#>         if (!quiet) 
#>             cat("Model cost at call ", calls, ": ", mC$model, 
#>                 "\n")
#>         if (plot) {
#>             outtimes_plot = seq(min(observed$time), max(observed$time), 
#>                 length.out = 100)
#>             out_plot <- mkinpredict(mkinmod, parms, odeini, outtimes_plot, 
#>                 solution_type = solution_type, use_compiled = use_compiled, 
#>                 method.ode = method.ode, atol = atol, rtol = rtol, 
#>                 ...)
#>             plot(0, type = "n", xlim = range(observed$time), 
#>                 ylim = c(0, max(observed$value, na.rm = TRUE)), 
#>                 xlab = "Time", ylab = "Observed")
#>             col_obs <- pch_obs <- 1:length(obs_vars)
#>             lty_obs <- rep(1, length(obs_vars))
#>             names(col_obs) <- names(pch_obs) <- names(lty_obs) <- obs_vars
#>             for (obs_var in obs_vars) {
#>                 points(subset(observed, name == obs_var, c(time, 
#>                   value)), pch = pch_obs[obs_var], col = col_obs[obs_var])
#>             }
#>             matlines(out_plot$time, out_plot[-1], col = col_obs, 
#>                 lty = lty_obs)
#>             legend("topright", inset = c(0.05, 0.05), legend = obs_vars, 
#>                 col = col_obs, pch = pch_obs, lty = 1:length(pch_obs))
#>         }
#>         assign("cost.old", mC$model, inherits = TRUE)
#>     }
#>     return(mC)
#> }
#> <environment: 0x3ff8420>
#> 
#> $cost_notrans
#> function (P) 
#> {
#>     if (length(state.ini.optim) > 0) {
#>         odeini <- c(P[1:length(state.ini.optim)], state.ini.fixed)
#>         names(odeini) <- c(state.ini.optim.boxnames, state.ini.fixed.boxnames)
#>     }
#>     else {
#>         odeini <- state.ini.fixed
#>         names(odeini) <- state.ini.fixed.boxnames
#>     }
#>     odeparms <- c(P[(length(state.ini.optim) + 1):length(P)], 
#>         parms.fixed)
#>     out <- mkinpredict(mkinmod, odeparms, odeini, outtimes, solution_type = solution_type, 
#>         use_compiled = use_compiled, method.ode = method.ode, 
#>         atol = atol, rtol = rtol, ...)
#>     mC <- modCost(out, observed, y = "value", err = err, weight = weight, 
#>         scaleVar = scaleVar)
#>     return(mC)
#> }
#> <environment: 0x3ff8420>
#> 
#> $hessian_notrans
#>                    parent_0    k_parent        k_M1         k_M2 f_parent_to_M1
#> parent_0           8.433594   -40.12785   -61.53042    -3406.469       461.2995
#> k_parent         -40.127847 19322.43697  3053.54654     3740.691      8966.4055
#> k_M1             -61.530424  3053.54654 70106.05907     7274.316     -8169.6841
#> k_M2           -3406.468786  3740.69112  7274.31610 12274341.595   -452294.7998
#> f_parent_to_M1   461.299501  8966.40549 -8169.68407  -452294.800     61249.1755
#> f_M1_to_M2       327.648696  2884.22668 17504.38651  -480941.198     43503.6440
#>                  f_M1_to_M2
#> parent_0           327.6487
#> k_parent          2884.2267
#> k_M1             17504.3865
#> k_M2           -480941.1983
#> f_parent_to_M1   43503.6440
#> f_M1_to_M2       46258.9775
#> 
#> $start
#>                   value   type
#> parent_0       101.3500  state
#> k_parent         0.1000 deparm
#> k_M1             0.1001 deparm
#> k_M2             0.1002 deparm
#> f_parent_to_M1   0.5000 deparm
#> f_M1_to_M2       0.5000 deparm
#> 
#> $start_transformed
#>                     value lower upper
#> parent_0       101.350000  -Inf   Inf
#> log_k_parent    -2.302585  -Inf   Inf
#> log_k_M1        -2.301586  -Inf   Inf
#> log_k_M2        -2.300587  -Inf   Inf
#> f_parent_ilr_1   0.000000  -Inf   Inf
#> f_M1_ilr_1       0.000000  -Inf   Inf
#> 
#> $fixed
#>      value  type
#> M1_0     0 state
#> M2_0     0 state
#> 
#> $data
#>    time variable observed     predicted    residual
#> 1     0   parent    101.5  1.020625e+02 -0.56248353
#> 2     0   parent    101.2  1.020625e+02 -0.86248353
#> 3     1   parent     53.9  4.872881e+01  5.17118695
#> 4     1   parent     47.5  4.872881e+01 -1.22881305
#> 5     3   parent     10.4  1.110773e+01 -0.70772795
#> 6     3   parent      7.6  1.110773e+01 -3.50772795
#> 7     7   parent      1.1  5.771704e-01  0.52282962
#> 8     7   parent      0.3  5.771704e-01 -0.27717038
#> 9    14   parent       NA  3.263939e-03          NA
#> 10   14   parent      3.5  3.263939e-03  3.49673606
#> 11   28   parent       NA  1.044512e-07          NA
#> 12   28   parent      3.2  1.044512e-07  3.19999990
#> 13   60   parent       NA -1.053819e-10          NA
#> 14   60   parent       NA -1.053819e-10          NA
#> 15   90   parent      0.6 -1.874889e-11  0.60000000
#> 16   90   parent       NA -1.874889e-11          NA
#> 17  120   parent       NA -2.804801e-11          NA
#> 18  120   parent      3.5 -2.804801e-11  3.50000000
#> 19    0       M1       NA  0.000000e+00          NA
#> 20    0       M1       NA  0.000000e+00          NA
#> 21    1       M1     36.4  3.478911e+01  1.61088639
#> 22    1       M1     37.4  3.478911e+01  2.61088639
#> 23    3       M1     34.3  3.937027e+01 -5.07026619
#> 24    3       M1     39.8  3.937027e+01  0.42973381
#> 25    7       M1     15.1  1.548714e+01 -0.38714436
#> 26    7       M1     17.8  1.548714e+01  2.31285564
#> 27   14       M1      5.8  1.995311e+00  3.80468869
#> 28   14       M1      1.2  1.995311e+00 -0.79531131
#> 29   28       M1       NA  3.034005e-02          NA
#> 30   28       M1       NA  3.034005e-02          NA
#> 31   60       M1      0.5  2.110651e-06  0.49999789
#> 32   60       M1       NA  2.110651e-06          NA
#> 33   90       M1       NA  2.913252e-10          NA
#> 34   90       M1      3.2  2.913252e-10  3.20000000
#> 35  120       M1      1.5  3.624738e-11  1.50000000
#> 36  120       M1      0.6  3.624738e-11  0.60000000
#> 37    0       M2       NA  0.000000e+00          NA
#> 38    0       M2       NA  0.000000e+00          NA
#> 39    1       M2       NA  4.454830e+00          NA
#> 40    1       M2      4.8  4.454830e+00  0.34517017
#> 41    3       M2     20.9  2.152527e+01 -0.62526794
#> 42    3       M2     19.3  2.152527e+01 -2.22526794
#> 43    7       M2     42.0  4.192058e+01  0.07941701
#> 44    7       M2     43.1  4.192058e+01  1.17941701
#> 45   14       M2     49.4  4.556646e+01  3.83353798
#> 46   14       M2     44.3  4.556646e+01 -1.26646202
#> 47   28       M2     34.6  3.547275e+01 -0.87274743
#> 48   28       M2     33.0  3.547275e+01 -2.47274743
#> 49   60       M2     18.8  1.858163e+01  0.21837410
#> 50   60       M2     17.6  1.858163e+01 -0.98162590
#> 51   90       M2     10.6  1.012869e+01  0.47130583
#> 52   90       M2     10.8  1.012869e+01  0.67130583
#> 53  120       M2      9.8  5.521069e+00  4.27893112
#> 54  120       M2      3.3  5.521069e+00 -2.22106888
#> 
#> $atol
#> [1] 1e-08
#> 
#> $rtol
#> [1] 1e-10
#> 
#> $weight.ini
#> [1] "none"
#> 
#> $reweight.tol
#> [1] 1e-08
#> 
#> $reweight.max.iter
#> [1] 10
#> 
#> $bparms.optim
#>       parent_0       k_parent           k_M1           k_M2 f_parent_to_M1 
#>    102.0624835      0.7393147      0.2991566      0.0202267      0.7686858 
#>     f_M1_to_M2 
#>      0.7229005 
#> 
#> $bparms.fixed
#> M1_0 M2_0 
#>    0    0 
#> 
#> $bparms.ode
#>       k_parent f_parent_to_M1           k_M1     f_M1_to_M2           k_M2 
#>      0.7393147      0.7686858      0.2991566      0.7229005      0.0202267 
#> 
#> $bparms.state
#>   parent       M1       M2 
#> 102.0625   0.0000   0.0000 
#> 
#> $date
#> [1] "Fri Nov 18 15:20:48 2016"
#> 
#> attr(,"class")
#> [1] "mkinfit" "modFit"