Functions to transform and backtransform kinetic parameters for fitting

The transformations are intended to map parameters that should only take on restricted values to the full scale of real numbers. For kinetic rate constants and other parameters that can only take on positive values, a simple log transformation is used. For compositional parameters, such as the formations fractions that should always sum up to 1 and can not be negative, the ilr transformation is used.

transform_odeparms(
  parms,
  mkinmod,
  transform_rates = TRUE,
  transform_fractions = TRUE
)

backtransform_odeparms(
  transparms,
  mkinmod,
  transform_rates = TRUE,
  transform_fractions = TRUE
)

Arguments

parms
Parameters of kinetic models as used in the differential equations.

mkinmod
The kinetic model of class mkinmod, containing the names of the model variables that are needed for grouping the formation fractions before ilr transformation, the parameter names and the information if the pathway to sink is included in the model.

transform_rates
Boolean specifying if kinetic rate constants should be transformed in the model specification used in the fitting for better compliance with the assumption of normal distribution of the estimator. If TRUE, also alpha and beta parameters of the FOMC model are log-transformed, as well as k1 and k2 rate constants for the DFOP and HS models and the break point tb of the HS model.

transform_fractions
Boolean specifying if formation fractions constants should be transformed in the model specification used in the fitting for better compliance with the assumption of normal distribution of the estimator. The default (TRUE) is to do transformations. The g parameter of the DFOP and HS models are also transformed, as they can also be seen as compositional data. The transformation used for these transformations is the ilr transformation.

transparms
Transformed parameters of kinetic models as used in the fitting procedure.

Value

A vector of transformed or backtransformed parameters with the same names as the original parameters.

Details

The transformation of sets of formation fractions is fragile, as it supposes the same ordering of the components in forward and backward transformation. This is no problem for the internal use in mkinfit.

Functions

backtransform_odeparms: Backtransform the set of transformed parameters

Examples

SFO_SFO <- mkinmod(
  parent = list(type = "SFO", to = "m1", sink = TRUE),
  m1 = list(type = "SFO"))

#> Successfully compiled differential equation model from auto-generated C code.

# Fit the model to the FOCUS example dataset D using defaults
fit <- mkinfit(SFO_SFO, FOCUS_2006_D, quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

fit.s <- summary(fit)
# Transformed and backtransformed parameters
print(fit.s$par, 3)

#>                Estimate Std. Error   Lower   Upper
#> parent_0         99.598     1.5702 96.4038 102.793
#> log_k_parent     -2.316     0.0409 -2.3988  -2.233
#> log_k_m1         -5.248     0.1332 -5.5184  -4.977
#> f_parent_ilr_1    0.041     0.0631 -0.0875   0.169
#> sigma             3.126     0.3585  2.3961   3.855

print(fit.s$bpar, 3)

#>                Estimate se_notrans t value   Pr(>t)    Lower    Upper
#> parent_0       99.59848    1.57022   63.43 2.30e-36 96.40384 102.7931
#> k_parent        0.09870    0.00403   24.47 4.96e-23  0.09082   0.1073
#> k_m1            0.00526    0.00070    7.51 6.16e-09  0.00401   0.0069
#> f_parent_to_m1  0.51448    0.02230   23.07 3.10e-22  0.46912   0.5596
#> sigma           3.12550    0.35852    8.72 2.24e-10  2.39609   3.8549

# \dontrun{
# Compare to the version without transforming rate parameters
fit.2 <- mkinfit(SFO_SFO, FOCUS_2006_D, transform_rates = FALSE, quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

#> Error in if (cost < cost.current) {    assign("cost.current", cost, inherits = TRUE)    if (!quiet)         cat(ifelse(OLS, "Sum of squared residuals", "Negative log-likelihood"),             " at call ", calls, ": ", cost.current, "\n", sep = "")}: Fehlender Wert, wo TRUE/FALSE nötig ist

#> Timing stopped at: 0.002 0 0.003

fit.2.s <- summary(fit.2)

#> Error in summary(fit.2): Objekt 'fit.2' nicht gefunden

print(fit.2.s$par, 3)

#> Error in print(fit.2.s$par, 3): Objekt 'fit.2.s' nicht gefunden

print(fit.2.s$bpar, 3)

#> Error in print(fit.2.s$bpar, 3): Objekt 'fit.2.s' nicht gefunden

# }

initials <- fit$start$value
names(initials) <- rownames(fit$start)
transformed <- fit$start_transformed$value
names(transformed) <- rownames(fit$start_transformed)
transform_odeparms(initials, SFO_SFO)

#>       parent_0   log_k_parent       log_k_m1 f_parent_ilr_1 
#>     100.750000      -2.302585      -2.301586       0.000000 

backtransform_odeparms(transformed, SFO_SFO)

#>       parent_0       k_parent           k_m1 f_parent_to_m1 
#>       100.7500         0.1000         0.1001         0.5000 

# \dontrun{
# The case of formation fractions
SFO_SFO.ff <- mkinmod(
  parent = list(type = "SFO", to = "m1", sink = TRUE),
  m1 = list(type = "SFO"),
  use_of_ff = "max")

#> Successfully compiled differential equation model from auto-generated C code.

fit.ff <- mkinfit(SFO_SFO.ff, FOCUS_2006_D, quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

fit.ff.s <- summary(fit.ff)
print(fit.ff.s$par, 3)

#>                Estimate Std. Error   Lower   Upper
#> parent_0         99.598     1.5702 96.4038 102.793
#> log_k_parent     -2.316     0.0409 -2.3988  -2.233
#> log_k_m1         -5.248     0.1332 -5.5184  -4.977
#> f_parent_ilr_1    0.041     0.0631 -0.0875   0.169
#> sigma             3.126     0.3585  2.3961   3.855

print(fit.ff.s$bpar, 3)

#>                Estimate se_notrans t value   Pr(>t)    Lower    Upper
#> parent_0       99.59848    1.57022   63.43 2.30e-36 96.40384 102.7931
#> k_parent        0.09870    0.00403   24.47 4.96e-23  0.09082   0.1073
#> k_m1            0.00526    0.00070    7.51 6.16e-09  0.00401   0.0069
#> f_parent_to_m1  0.51448    0.02230   23.07 3.10e-22  0.46912   0.5596
#> sigma           3.12550    0.35852    8.72 2.24e-10  2.39609   3.8549

initials <- c("f_parent_to_m1" = 0.5)
transformed <- transform_odeparms(initials, SFO_SFO.ff)
backtransform_odeparms(transformed, SFO_SFO.ff)

#> f_parent_to_m1 
#>            0.5 

# And without sink
SFO_SFO.ff.2 <- mkinmod(
  parent = list(type = "SFO", to = "m1", sink = FALSE),
  m1 = list(type = "SFO"),
  use_of_ff = "max")

#> Successfully compiled differential equation model from auto-generated C code.

fit.ff.2 <- mkinfit(SFO_SFO.ff.2, FOCUS_2006_D, quiet = TRUE)

#> Warning: Observations with value of zero were removed from the data

fit.ff.2.s <- summary(fit.ff.2)
print(fit.ff.2.s$par, 3)

#>              Estimate Std. Error Lower Upper
#> parent_0        84.79      3.012 78.67 90.91
#> log_k_parent    -2.76      0.082 -2.92 -2.59
#> log_k_m1        -4.21      0.123 -4.46 -3.96
#> sigma            8.22      0.943  6.31 10.14

print(fit.ff.2.s$bpar, 3)

#>          Estimate se_notrans t value   Pr(>t)   Lower  Upper
#> parent_0  84.7916    3.01203   28.15 1.92e-25 78.6704 90.913
#> k_parent   0.0635    0.00521   12.19 2.91e-14  0.0538  0.075
#> k_m1       0.0148    0.00182    8.13 8.81e-10  0.0115  0.019
#> sigma      8.2229    0.94323    8.72 1.73e-10  6.3060 10.140

# }