Functions to transform and backtransform kinetic parameters for fitting

The transformations are intended to map parameters that should only take on restricted values to the full scale of real numbers. For kinetic rate constants and other paramters that can only take on positive values, a simple log transformation is used. For compositional parameters, such as the formations fractions that should always sum up to 1 and can not be negative, the ilr transformation is used.

The transformation of sets of formation fractions is fragile, as it supposes the same ordering of the components in forward and backward transformation. This is no problem for the internal use in mkinfit.

transform_odeparms(parms, mkinmod,
                   transform_rates = TRUE, transform_fractions = TRUE)
backtransform_odeparms(transparms, mkinmod,
                       transform_rates = TRUE, transform_fractions = TRUE)

Arguments

parms

Parameters of kinetic models as used in the differential equations.

transparms

Transformed parameters of kinetic models as used in the fitting procedure.

mkinmod

The kinetic model of class mkinmod, containing the names of the model variables that are needed for grouping the formation fractions before ilr transformation, the parameter names and the information if the pathway to sink is included in the model.

transform_rates

Boolean specifying if kinetic rate constants should be transformed in the model specification used in the fitting for better compliance with the assumption of normal distribution of the estimator. If TRUE, also alpha and beta parameters of the FOMC model are log-transformed, as well as k1 and k2 rate constants for the DFOP and HS models and the break point tb of the HS model.

transform_fractions

Boolean specifying if formation fractions constants should be transformed in the model specification used in the fitting for better compliance with the assumption of normal distribution of the estimator. The default (TRUE) is to do transformations. The g parameter of the DFOP and HS models are also transformed, as they can also be seen as compositional data. The transformation used for these transformations is the ilr transformation.

Value

A vector of transformed or backtransformed parameters with the same names as the original parameters.

Examples

SFO_SFO <- mkinmod(
  parent = list(type = "SFO", to = "m1", sink = TRUE),
  m1 = list(type = "SFO"))

#> Successfully compiled differential equation model from auto-generated C code.

# Fit the model to the FOCUS example dataset D using defaults
fit <- mkinfit(SFO_SFO, FOCUS_2006_D, quiet = TRUE)
summary(fit, data=FALSE) # See transformed and backtransformed parameters

#> mkin version:    0.9.44.9000 
#> R version:       3.3.2 
#> Date of fit:     Fri Nov 18 15:20:49 2016 
#> Date of summary: Fri Nov 18 15:20:49 2016 
#> 
#> Equations:
#> d_parent/dt = - k_parent_sink * parent - k_parent_m1 * parent
#> d_m1/dt = + k_parent_m1 * parent - k_m1_sink * m1
#> 
#> Model predictions using solution type deSolve 
#> 
#> Fitted with method Port using 153 model solutions performed in 0.681 s
#> 
#> Weighting: none
#> 
#> Starting values for parameters to be optimised:
#>                  value   type
#> parent_0      100.7500  state
#> k_parent_sink   0.1000 deparm
#> k_parent_m1     0.1001 deparm
#> k_m1_sink       0.1002 deparm
#> 
#> Starting values for the transformed parameters actually optimised:
#>                        value lower upper
#> parent_0          100.750000  -Inf   Inf
#> log_k_parent_sink  -2.302585  -Inf   Inf
#> log_k_parent_m1    -2.301586  -Inf   Inf
#> log_k_m1_sink      -2.300587  -Inf   Inf
#> 
#> Fixed parameter values:
#>      value  type
#> m1_0     0 state
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                   Estimate Std. Error  Lower   Upper
#> parent_0            99.600    1.61400 96.330 102.900
#> log_k_parent_sink   -3.038    0.07826 -3.197  -2.879
#> log_k_parent_m1     -2.980    0.04124 -3.064  -2.897
#> log_k_m1_sink       -5.248    0.13610 -5.523  -4.972
#> 
#> Parameter correlation:
#>                   parent_0 log_k_parent_sink log_k_parent_m1 log_k_m1_sink
#> parent_0           1.00000            0.6075        -0.06625       -0.1701
#> log_k_parent_sink  0.60752            1.0000        -0.08740       -0.6253
#> log_k_parent_m1   -0.06625           -0.0874         1.00000        0.4716
#> log_k_m1_sink     -0.17006           -0.6253         0.47163        1.0000
#> 
#> Residual standard error: 3.211 on 36 degrees of freedom
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                Estimate t value    Pr(>t)     Lower     Upper
#> parent_0      99.600000  61.720 2.024e-38 96.330000 1.029e+02
#> k_parent_sink  0.047920  12.780 3.050e-15  0.040890 5.616e-02
#> k_parent_m1    0.050780  24.250 3.407e-24  0.046700 5.521e-02
#> k_m1_sink      0.005261   7.349 5.758e-09  0.003992 6.933e-03
#> 
#> Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   6.398       4 15
#> parent     6.827       3  6
#> m1         4.490       1  9
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_sink 0.4855
#> parent_m1   0.5145
#> m1_sink     1.0000
#> 
#> Estimated disappearance times:
#>           DT50   DT90
#> parent   7.023  23.33
#> m1     131.761 437.70

fit.2 <- mkinfit(SFO_SFO, FOCUS_2006_D, transform_rates = FALSE)

#> Model cost at call  1 :  18915.53 
#> Model cost at call  2 :  18915.53 
#> Model cost at call  7 :  10205.88 
#> Model cost at call  10 :  10205.05 
#> Model cost at call  13 :  8136.609 
#> Model cost at call  18 :  2504.352 
#> Model cost at call  20 :  2504.35 
#> Model cost at call  22 :  2504.285 
#> Model cost at call  25 :  1747.542 
#> Model cost at call  27 :  1745.941 
#> Model cost at call  29 :  1745.431 
#> Model cost at call  30 :  1341.034 
#> Model cost at call  34 :  1341.034 
#> Model cost at call  35 :  1032.65 
#> Model cost at call  39 :  1032.649 
#> Model cost at call  40 :  919.9522 
#> Model cost at call  42 :  919.952 
#> Model cost at call  44 :  919.9518 
#> Model cost at call  45 :  903.8272 
#> Model cost at call  47 :  903.827 
#> Model cost at call  49 :  903.8268 
#> Model cost at call  50 :  780.8699 
#> Model cost at call  52 :  780.8698 
#> Model cost at call  54 :  780.8697 
#> Model cost at call  55 :  734.3043 
#> Model cost at call  57 :  734.3036 
#> Model cost at call  60 :  717.8438 
#> Model cost at call  67 :  676.3908 
#> Model cost at call  68 :  676.3907 
#> Model cost at call  69 :  676.3906 
#> Model cost at call  71 :  676.3885 
#> Model cost at call  72 :  642.2738 
#> Model cost at call  73 :  642.2738 
#> Model cost at call  76 :  642.2738 
#> Model cost at call  77 :  604.7128 
#> Model cost at call  78 :  604.7126 
#> Model cost at call  82 :  560.1285 
#> Model cost at call  83 :  560.1285 
#> Model cost at call  86 :  560.1285 
#> Model cost at call  87 :  521.6932 
#> Model cost at call  88 :  521.6932 
#> Model cost at call  91 :  521.6931 
#> Model cost at call  92 :  453.6483 
#> Model cost at call  93 :  453.6483 
#> Model cost at call  96 :  453.6483 
#> Model cost at call  98 :  422.5498 
#> Model cost at call  102 :  422.5497 
#> Model cost at call  106 :  413.6426 
#> Model cost at call  108 :  413.6416 
#> Model cost at call  111 :  407.4639 
#> Model cost at call  113 :  407.4639 
#> Model cost at call  115 :  407.4639 
#> Model cost at call  116 :  403.6974 
#> Model cost at call  118 :  403.6973 
#> Model cost at call  120 :  403.6973 
#> Model cost at call  121 :  396.004 
#> Model cost at call  123 :  396.004 
#> Model cost at call  126 :  391.2533 
#> Model cost at call  128 :  391.2533 
#> Model cost at call  130 :  391.2533 
#> Model cost at call  131 :  388.4056 
#> Model cost at call  133 :  388.4056 
#> Model cost at call  135 :  388.4055 
#> Model cost at call  137 :  384.0739 
#> Model cost at call  139 :  384.0734 
#> Model cost at call  143 :  383.6352 
#> Model cost at call  145 :  383.635 
#> Model cost at call  148 :  381.1342 
#> Model cost at call  152 :  381.1342 
#> Model cost at call  153 :  380.1443 
#> Model cost at call  157 :  380.1443 
#> Model cost at call  158 :  378.5615 
#> Model cost at call  162 :  378.5614 
#> Model cost at call  164 :  378.4583 
#> Model cost at call  168 :  378.4583 
#> Model cost at call  169 :  377.7047 
#> Model cost at call  173 :  377.7047 
#> Model cost at call  175 :  377.7042 
#> Model cost at call  178 :  377.7042 
#> Model cost at call  179 :  377.7042 
#> Model cost at call  180 :  377.1553 
#> Model cost at call  182 :  377.1553 
#> Model cost at call  183 :  377.1553 
#> Model cost at call  185 :  377.0866 
#> Model cost at call  187 :  377.0866 
#> Model cost at call  190 :  377.0288 
#> Model cost at call  194 :  377.0288 
#> Model cost at call  196 :  376.9814 
#> Model cost at call  200 :  376.9814 
#> Model cost at call  201 :  376.9488 
#> Model cost at call  203 :  376.9488 
#> Model cost at call  205 :  376.9488 
#> Model cost at call  206 :  376.9184 
#> Model cost at call  208 :  376.9184 
#> Model cost at call  211 :  376.9095 
#> Model cost at call  213 :  376.9095 
#> Model cost at call  215 :  376.9095 
#> Model cost at call  216 :  376.901 
#> Model cost at call  218 :  376.901 
#> Model cost at call  221 :  376.8936 
#> Model cost at call  225 :  376.8936 
#> Model cost at call  226 :  376.8848 
#> Model cost at call  227 :  376.8608 
#> Model cost at call  228 :  376.7665 
#> Model cost at call  229 :  376.4162 
#> Model cost at call  230 :  375.4439 
#> Model cost at call  235 :  375.4439 
#> Model cost at call  237 :  375.2281 
#> Model cost at call  241 :  375.2281 
#> Model cost at call  242 :  374.3381 
#> Model cost at call  244 :  374.3381 
#> Model cost at call  251 :  374.2632 
#> Model cost at call  253 :  374.2632 
#> Model cost at call  256 :  374.2391 
#> Model cost at call  257 :  374.2247 
#> Model cost at call  259 :  374.2247 
#> Model cost at call  261 :  374.2247 
#> Model cost at call  263 :  374.2024 
#> Model cost at call  266 :  374.2024 
#> Model cost at call  270 :  374.1952 
#> Model cost at call  274 :  374.1952 
#> Model cost at call  275 :  374.186 
#> Model cost at call  279 :  374.186 
#> Model cost at call  280 :  374.1724 
#> Model cost at call  281 :  374.156 
#> Model cost at call  282 :  374.0763 
#> Model cost at call  283 :  373.7692 
#> Model cost at call  284 :  372.729 
#> Model cost at call  285 :  371.4251 
#> Model cost at call  287 :  371.4251 
#> Model cost at call  290 :  371.2142 
#> Model cost at call  291 :  371.2142 
#> Model cost at call  292 :  371.2142 
#> Model cost at call  295 :  371.2134 
#> Model cost at call  298 :  371.2134 
#> Model cost at call  300 :  371.2134 
#> Model cost at call  309 :  371.2134 
#> Model cost at call  320 :  371.2134 
#> Optimisation by method Port successfully terminated.

summary(fit.2, data=FALSE)

#> mkin version:    0.9.44.9000 
#> R version:       3.3.2 
#> Date of fit:     Fri Nov 18 15:20:51 2016 
#> Date of summary: Fri Nov 18 15:20:51 2016 
#> 
#> Equations:
#> d_parent/dt = - k_parent_sink * parent - k_parent_m1 * parent
#> d_m1/dt = + k_parent_m1 * parent - k_m1_sink * m1
#> 
#> Model predictions using solution type deSolve 
#> 
#> Fitted with method Port using 327 model solutions performed in 1.34 s
#> 
#> Weighting: none
#> 
#> Starting values for parameters to be optimised:
#>                  value   type
#> parent_0      100.7500  state
#> k_parent_sink   0.1000 deparm
#> k_parent_m1     0.1001 deparm
#> k_m1_sink       0.1002 deparm
#> 
#> Starting values for the transformed parameters actually optimised:
#>                  value lower upper
#> parent_0      100.7500  -Inf   Inf
#> k_parent_sink   0.1000     0   Inf
#> k_parent_m1     0.1001     0   Inf
#> k_m1_sink       0.1002     0   Inf
#> 
#> Fixed parameter values:
#>      value  type
#> m1_0     0 state
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                Estimate Std. Error     Lower     Upper
#> parent_0      99.600000  1.6140000 96.330000 1.029e+02
#> k_parent_sink  0.047920  0.0037500  0.040310 5.553e-02
#> k_parent_m1    0.050780  0.0020940  0.046530 5.502e-02
#> k_m1_sink      0.005261  0.0007159  0.003809 6.713e-03
#> 
#> Parameter correlation:
#>               parent_0 k_parent_sink k_parent_m1 k_m1_sink
#> parent_0       1.00000        0.6075    -0.06625   -0.1701
#> k_parent_sink  0.60752        1.0000    -0.08740   -0.6253
#> k_parent_m1   -0.06625       -0.0874     1.00000    0.4716
#> k_m1_sink     -0.17006       -0.6253     0.47164    1.0000
#> 
#> Residual standard error: 3.211 on 36 degrees of freedom
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                Estimate t value    Pr(>t)     Lower     Upper
#> parent_0      99.600000  61.720 2.024e-38 96.330000 1.029e+02
#> k_parent_sink  0.047920  12.780 3.050e-15  0.040310 5.553e-02
#> k_parent_m1    0.050780  24.250 3.407e-24  0.046530 5.502e-02
#> k_m1_sink      0.005261   7.349 5.758e-09  0.003809 6.713e-03
#> 
#> Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   6.398       4 15
#> parent     6.827       3  6
#> m1         4.490       1  9
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_sink 0.4855
#> parent_m1   0.5145
#> m1_sink     1.0000
#> 
#> Estimated disappearance times:
#>           DT50   DT90
#> parent   7.023  23.33
#> m1     131.761 437.70

initials <- fit$start$value
names(initials) <- rownames(fit$start)
transformed <- fit$start_transformed$value
names(transformed) <- rownames(fit$start_transformed)
transform_odeparms(initials, SFO_SFO)

#>          parent_0 log_k_parent_sink   log_k_parent_m1     log_k_m1_sink 
#>        100.750000         -2.302585         -2.301586         -2.300587 

backtransform_odeparms(transformed, SFO_SFO)

#>      parent_0 k_parent_sink   k_parent_m1     k_m1_sink 
#>      100.7500        0.1000        0.1001        0.1002 

# The case of formation fractions
SFO_SFO.ff <- mkinmod(
  parent = list(type = "SFO", to = "m1", sink = TRUE),
  m1 = list(type = "SFO"),
  use_of_ff = "max")

#> Successfully compiled differential equation model from auto-generated C code.

fit.ff <- mkinfit(SFO_SFO.ff, FOCUS_2006_D)

#> Model cost at call  1 :  15156.12 
#> Model cost at call  2 :  15156.12 
#> Model cost at call  6 :  8243.644 
#> Model cost at call  12 :  6290.714 
#> Model cost at call  13 :  6290.684 
#> Model cost at call  15 :  6290.453 
#> Model cost at call  18 :  1700.75 
#> Model cost at call  20 :  1700.612 
#> Model cost at call  24 :  1190.923 
#> Model cost at call  26 :  1190.922 
#> Model cost at call  29 :  1017.417 
#> Model cost at call  31 :  1017.417 
#> Model cost at call  33 :  1017.416 
#> Model cost at call  34 :  644.0471 
#> Model cost at call  36 :  644.0469 
#> Model cost at call  38 :  644.0468 
#> Model cost at call  39 :  590.5024 
#> Model cost at call  41 :  590.5021 
#> Model cost at call  43 :  590.5015 
#> Model cost at call  44 :  543.2187 
#> Model cost at call  45 :  543.2183 
#> Model cost at call  46 :  543.2182 
#> Model cost at call  50 :  391.348 
#> Model cost at call  51 :  391.3479 
#> Model cost at call  56 :  386.4789 
#> Model cost at call  58 :  386.4789 
#> Model cost at call  60 :  386.4779 
#> Model cost at call  61 :  384.0686 
#> Model cost at call  63 :  384.0686 
#> Model cost at call  66 :  382.7812 
#> Model cost at call  68 :  382.7812 
#> Model cost at call  70 :  382.7812 
#> Model cost at call  71 :  378.9272 
#> Model cost at call  73 :  378.9272 
#> Model cost at call  75 :  378.9272 
#> Model cost at call  76 :  377.4846 
#> Model cost at call  78 :  377.4846 
#> Model cost at call  81 :  375.9738 
#> Model cost at call  83 :  375.9738 
#> Model cost at call  86 :  375.3387 
#> Model cost at call  88 :  375.3387 
#> Model cost at call  91 :  374.5774 
#> Model cost at call  93 :  374.5774 
#> Model cost at call  95 :  374.5774 
#> Model cost at call  96 :  373.5447 
#> Model cost at call  100 :  373.5446 
#> Model cost at call  102 :  373.2643 
#> Model cost at call  104 :  373.2643 
#> Model cost at call  107 :  372.6799 
#> Model cost at call  111 :  372.6798 
#> Model cost at call  114 :  372.6325 
#> Model cost at call  116 :  372.6325 
#> Model cost at call  119 :  372.6159 
#> Model cost at call  121 :  372.6159 
#> Model cost at call  123 :  372.6159 
#> Model cost at call  124 :  372.5845 
#> Model cost at call  126 :  372.5845 
#> Model cost at call  129 :  372.5375 
#> Model cost at call  130 :  372.4771 
#> Model cost at call  131 :  372.2008 
#> Model cost at call  132 :  371.4923 
#> Model cost at call  134 :  371.4923 
#> Model cost at call  137 :  371.3022 
#> Model cost at call  139 :  371.3022 
#> Model cost at call  143 :  371.2271 
#> Model cost at call  144 :  371.2271 
#> Model cost at call  148 :  371.2202 
#> Model cost at call  149 :  371.215 
#> Model cost at call  152 :  371.215 
#> Model cost at call  154 :  371.2136 
#> Model cost at call  155 :  371.2136 
#> Model cost at call  156 :  371.2136 
#> Model cost at call  160 :  371.2134 
#> Model cost at call  164 :  371.2134 
#> Model cost at call  167 :  371.2134 
#> Optimisation by method Port successfully terminated.

summary(fit.ff, data = FALSE)

#> mkin version:    0.9.44.9000 
#> R version:       3.3.2 
#> Date of fit:     Fri Nov 18 15:20:52 2016 
#> Date of summary: Fri Nov 18 15:20:52 2016 
#> 
#> Equations:
#> d_parent/dt = - k_parent * parent
#> d_m1/dt = + f_parent_to_m1 * k_parent * parent - k_m1 * m1
#> 
#> Model predictions using solution type deSolve 
#> 
#> Fitted with method Port using 185 model solutions performed in 0.764 s
#> 
#> Weighting: none
#> 
#> Starting values for parameters to be optimised:
#>                   value   type
#> parent_0       100.7500  state
#> k_parent         0.1000 deparm
#> k_m1             0.1001 deparm
#> f_parent_to_m1   0.5000 deparm
#> 
#> Starting values for the transformed parameters actually optimised:
#>                     value lower upper
#> parent_0       100.750000  -Inf   Inf
#> log_k_parent    -2.302585  -Inf   Inf
#> log_k_m1        -2.301586  -Inf   Inf
#> f_parent_ilr_1   0.000000  -Inf   Inf
#> 
#> Fixed parameter values:
#>      value  type
#> m1_0     0 state
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                Estimate Std. Error   Lower    Upper
#> parent_0       99.60000    1.61400 96.3300 102.9000
#> log_k_parent   -2.31600    0.04187 -2.4010  -2.2310
#> log_k_m1       -5.24800    0.13610 -5.5230  -4.9720
#> f_parent_ilr_1  0.04096    0.06477 -0.0904   0.1723
#> 
#> Parameter correlation:
#>                parent_0 log_k_parent log_k_m1 f_parent_ilr_1
#> parent_0         1.0000       0.5178  -0.1701        -0.5489
#> log_k_parent     0.5178       1.0000  -0.3285        -0.5451
#> log_k_m1        -0.1701      -0.3285   1.0000         0.7466
#> f_parent_ilr_1  -0.5489      -0.5451   0.7466         1.0000
#> 
#> Residual standard error: 3.211 on 36 degrees of freedom
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                 Estimate t value    Pr(>t)     Lower     Upper
#> parent_0       99.600000  61.720 2.024e-38 96.330000 1.029e+02
#> k_parent        0.098700  23.880 5.701e-24  0.090660 1.074e-01
#> k_m1            0.005261   7.349 5.758e-09  0.003992 6.933e-03
#> f_parent_to_m1  0.514500  22.490 4.374e-23  0.468100 5.606e-01
#> 
#> Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   6.398       4 15
#> parent     6.459       2  7
#> m1         4.690       2  8
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_m1   0.5145
#> parent_sink 0.4855
#> 
#> Estimated disappearance times:
#>           DT50   DT90
#> parent   7.023  23.33
#> m1     131.761 437.70

initials <- c("f_parent_to_m1" = 0.5)
transformed <- transform_odeparms(initials, SFO_SFO.ff)
backtransform_odeparms(transformed, SFO_SFO.ff)

#> f_parent_to_m1 
#>            0.5 

# And without sink
SFO_SFO.ff.2 <- mkinmod(
  parent = list(type = "SFO", to = "m1", sink = FALSE),
  m1 = list(type = "SFO"),
  use_of_ff = "max")

#> Successfully compiled differential equation model from auto-generated C code.

fit.ff.2 <- mkinfit(SFO_SFO.ff.2, FOCUS_2006_D)

#> Model cost at call  1 :  12435.14 
#> Model cost at call  2 :  12435.14 
#> Model cost at call  5 :  8276.306 
#> Model cost at call  6 :  8276.294 
#> Model cost at call  7 :  8275.676 
#> Model cost at call  9 :  5256.953 
#> Model cost at call  11 :  5256.951 
#> Model cost at call  12 :  5256.943 
#> Model cost at call  14 :  4469.745 
#> Model cost at call  18 :  4462.927 
#> Model cost at call  21 :  4462.925 
#> Model cost at call  22 :  4376.059 
#> Model cost at call  24 :  4376.058 
#> Model cost at call  27 :  4366.956 
#> Model cost at call  29 :  4366.956 
#> Model cost at call  31 :  4365.275 
#> Model cost at call  33 :  4365.275 
#> Model cost at call  35 :  4351.877 
#> Model cost at call  37 :  4351.877 
#> Model cost at call  39 :  4338.109 
#> Model cost at call  41 :  4338.109 
#> Model cost at call  43 :  4297.053 
#> Model cost at call  44 :  4218.591 
#> Model cost at call  45 :  3940.397 
#> Model cost at call  46 :  3690.395 
#> Model cost at call  48 :  3690.395 
#> Model cost at call  49 :  3690.385 
#> Model cost at call  50 :  3038.366 
#> Model cost at call  53 :  3038.365 
#> Model cost at call  54 :  2637.866 
#> Model cost at call  57 :  2637.865 
#> Model cost at call  59 :  2588.01 
#> Model cost at call  60 :  2588.009 
#> Model cost at call  63 :  2576.742 
#> Model cost at call  66 :  2576.742 
#> Model cost at call  67 :  2574 
#> Model cost at call  68 :  2574 
#> Model cost at call  69 :  2574 
#> Model cost at call  71 :  2569.76 
#> Model cost at call  73 :  2569.76 
#> Model cost at call  74 :  2569.76 
#> Model cost at call  75 :  2569.403 
#> Model cost at call  76 :  2569.403 
#> Model cost at call  79 :  2569.4 
#> Model cost at call  80 :  2569.4 
#> Model cost at call  81 :  2569.4 
#> Model cost at call  83 :  2569.4 
#> Model cost at call  86 :  2569.4 
#> Model cost at call  90 :  2569.4 
#> Model cost at call  99 :  2569.4 
#> Model cost at call  100 :  2569.4 
#> Optimisation by method Port successfully terminated.

summary(fit.ff.2, data = FALSE)

#> mkin version:    0.9.44.9000 
#> R version:       3.3.2 
#> Date of fit:     Fri Nov 18 15:20:52 2016 
#> Date of summary: Fri Nov 18 15:20:52 2016 
#> 
#> Equations:
#> d_parent/dt = - k_parent * parent
#> d_m1/dt = + k_parent * parent - k_m1 * m1
#> 
#> Model predictions using solution type deSolve 
#> 
#> Fitted with method Port using 104 model solutions performed in 0.44 s
#> 
#> Weighting: none
#> 
#> Starting values for parameters to be optimised:
#>             value   type
#> parent_0 100.7500  state
#> k_parent   0.1000 deparm
#> k_m1       0.1001 deparm
#> 
#> Starting values for the transformed parameters actually optimised:
#>                   value lower upper
#> parent_0     100.750000  -Inf   Inf
#> log_k_parent  -2.302585  -Inf   Inf
#> log_k_m1      -2.301586  -Inf   Inf
#> 
#> Fixed parameter values:
#>      value  type
#> m1_0     0 state
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>              Estimate Std. Error Lower  Upper
#> parent_0       84.790    2.96500 78.78 90.800
#> log_k_parent   -2.756    0.08088 -2.92 -2.593
#> log_k_m1       -4.214    0.11150 -4.44 -3.988
#> 
#> Parameter correlation:
#>              parent_0 log_k_parent log_k_m1
#> parent_0       1.0000      0.11059  0.46156
#> log_k_parent   0.1106      1.00000  0.06274
#> log_k_m1       0.4616      0.06274  1.00000
#> 
#> Residual standard error: 8.333 on 37 degrees of freedom
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>          Estimate t value    Pr(>t)    Lower    Upper
#> parent_0 84.79000  28.600 3.939e-27 78.78000 90.80000
#> k_parent  0.06352  12.360 5.237e-15  0.05392  0.07483
#> k_m1      0.01478   8.966 4.114e-11  0.01179  0.01853
#> 
#> Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   19.66       3 16
#> parent     17.56       2  7
#> m1         18.71       1  9
#> 
#> Estimated disappearance times:
#>         DT50   DT90
#> parent 10.91  36.25
#> m1     46.89 155.75