Functions to transform and backtransform kinetic parameters for fitting

The transformations are intended to map parameters that should only take on restricted values to the full scale of real numbers. For kinetic rate constants and other paramters that can only take on positive values, a simple log transformation is used. For compositional parameters, such as the formations fractions that should always sum up to 1 and can not be negative, the ilr transformation is used.

The transformation of sets of formation fractions is fragile, as it supposes the same ordering of the components in forward and backward transformation. This is no problem for the internal use in mkinfit.

transform_odeparms(parms, mkinmod,
                   transform_rates = TRUE, transform_fractions = TRUE)
backtransform_odeparms(transparms, mkinmod,
                       transform_rates = TRUE, transform_fractions = TRUE)

Arguments

parms
Parameters of kinetic models as used in the differential equations.

transparms
Transformed parameters of kinetic models as used in the fitting procedure.

mkinmod
The kinetic model of class mkinmod, containing the names of the model variables that are needed for grouping the formation fractions before ilr transformation, the parameter names and the information if the pathway to sink is included in the model.

transform_rates
Boolean specifying if kinetic rate constants should be transformed in the model specification used in the fitting for better compliance with the assumption of normal distribution of the estimator. If TRUE, also alpha and beta parameters of the FOMC model are log-transformed, as well as k1 and k2 rate constants for the DFOP and HS models and the break point tb of the HS model.

transform_fractions
Boolean specifying if formation fractions constants should be transformed in the model specification used in the fitting for better compliance with the assumption of normal distribution of the estimator. The default (TRUE) is to do transformations. The g parameter of the DFOP and HS models are also transformed, as they can also be seen as compositional data. The transformation used for these transformations is the ilr transformation.

Value

A vector of transformed or backtransformed parameters with the same names as the original parameters.

Examples

SFO_SFO <- mkinmod(
  parent = list(type = "SFO", to = "m1", sink = TRUE),
  m1 = list(type = "SFO"))

#> Successfully compiled differential equation model from auto-generated C code.

# Fit the model to the FOCUS example dataset D using defaults
fit <- mkinfit(SFO_SFO, FOCUS_2006_D, quiet = TRUE)
summary(fit, data=FALSE) # See transformed and backtransformed parameters

#> mkin version used for fitting:    0.9.47.6 
#> R version used for fitting:       3.5.1 
#> Date of fit:     Mon Nov 26 12:29:19 2018 
#> Date of summary: Mon Nov 26 12:29:19 2018 
#> 
#> Equations:
#> d_parent/dt = - k_parent_sink * parent - k_parent_m1 * parent
#> d_m1/dt = + k_parent_m1 * parent - k_m1_sink * m1
#> 
#> Model predictions using solution type deSolve 
#> 
#> Fitted with method Port using 153 model solutions performed in 0.701 s
#> 
#> Weighting: none
#> 
#> Starting values for parameters to be optimised:
#>                  value   type
#> parent_0      100.7500  state
#> k_parent_sink   0.1000 deparm
#> k_parent_m1     0.1001 deparm
#> k_m1_sink       0.1002 deparm
#> 
#> Starting values for the transformed parameters actually optimised:
#>                        value lower upper
#> parent_0          100.750000  -Inf   Inf
#> log_k_parent_sink  -2.302585  -Inf   Inf
#> log_k_parent_m1    -2.301586  -Inf   Inf
#> log_k_m1_sink      -2.300587  -Inf   Inf
#> 
#> Fixed parameter values:
#>      value  type
#> m1_0     0 state
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                   Estimate Std. Error  Lower   Upper
#> parent_0            99.600    1.61400 96.330 102.900
#> log_k_parent_sink   -3.038    0.07826 -3.197  -2.879
#> log_k_parent_m1     -2.980    0.04124 -3.064  -2.897
#> log_k_m1_sink       -5.248    0.13610 -5.523  -4.972
#> 
#> Parameter correlation:
#>                   parent_0 log_k_parent_sink log_k_parent_m1 log_k_m1_sink
#> parent_0           1.00000            0.6075        -0.06625       -0.1701
#> log_k_parent_sink  0.60752            1.0000        -0.08740       -0.6253
#> log_k_parent_m1   -0.06625           -0.0874         1.00000        0.4716
#> log_k_m1_sink     -0.17006           -0.6253         0.47164        1.0000
#> 
#> Residual standard error: 3.211 on 36 degrees of freedom
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                Estimate t value    Pr(>t)     Lower     Upper
#> parent_0      99.600000  61.720 2.024e-38 96.330000 1.029e+02
#> k_parent_sink  0.047920  12.780 3.050e-15  0.040890 5.616e-02
#> k_parent_m1    0.050780  24.250 3.407e-24  0.046700 5.521e-02
#> k_m1_sink      0.005261   7.349 5.758e-09  0.003992 6.933e-03
#> 
#> Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   6.398       4 15
#> parent     6.827       3  6
#> m1         4.490       1  9
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_sink 0.4855
#> parent_m1   0.5145
#> m1_sink     1.0000
#> 
#> Estimated disappearance times:
#>           DT50   DT90
#> parent   7.023  23.33
#> m1     131.761 437.70

fit.2 <- mkinfit(SFO_SFO, FOCUS_2006_D, transform_rates = FALSE, quiet = TRUE)
summary(fit.2, data=FALSE)

#> mkin version used for fitting:    0.9.47.6 
#> R version used for fitting:       3.5.1 
#> Date of fit:     Mon Nov 26 12:29:20 2018 
#> Date of summary: Mon Nov 26 12:29:20 2018 
#> 
#> Equations:
#> d_parent/dt = - k_parent_sink * parent - k_parent_m1 * parent
#> d_m1/dt = + k_parent_m1 * parent - k_m1_sink * m1
#> 
#> Model predictions using solution type deSolve 
#> 
#> Fitted with method Port using 350 model solutions performed in 1.597 s
#> 
#> Weighting: none
#> 
#> Starting values for parameters to be optimised:
#>                  value   type
#> parent_0      100.7500  state
#> k_parent_sink   0.1000 deparm
#> k_parent_m1     0.1001 deparm
#> k_m1_sink       0.1002 deparm
#> 
#> Starting values for the transformed parameters actually optimised:
#>                  value lower upper
#> parent_0      100.7500  -Inf   Inf
#> k_parent_sink   0.1000     0   Inf
#> k_parent_m1     0.1001     0   Inf
#> k_m1_sink       0.1002     0   Inf
#> 
#> Fixed parameter values:
#>      value  type
#> m1_0     0 state
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                Estimate Std. Error     Lower     Upper
#> parent_0      99.600000  1.6140000 96.330000 1.029e+02
#> k_parent_sink  0.047920  0.0037500  0.040310 5.553e-02
#> k_parent_m1    0.050780  0.0020940  0.046530 5.502e-02
#> k_m1_sink      0.005261  0.0007159  0.003809 6.713e-03
#> 
#> Parameter correlation:
#>               parent_0 k_parent_sink k_parent_m1 k_m1_sink
#> parent_0       1.00000        0.6075    -0.06625   -0.1701
#> k_parent_sink  0.60752        1.0000    -0.08740   -0.6253
#> k_parent_m1   -0.06625       -0.0874     1.00000    0.4716
#> k_m1_sink     -0.17006       -0.6253     0.47164    1.0000
#> 
#> Residual standard error: 3.211 on 36 degrees of freedom
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                Estimate t value    Pr(>t)     Lower     Upper
#> parent_0      99.600000  61.720 2.024e-38 96.330000 1.029e+02
#> k_parent_sink  0.047920  12.780 3.050e-15  0.040310 5.553e-02
#> k_parent_m1    0.050780  24.250 3.407e-24  0.046530 5.502e-02
#> k_m1_sink      0.005261   7.349 5.758e-09  0.003809 6.713e-03
#> 
#> Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   6.398       4 15
#> parent     6.827       3  6
#> m1         4.490       1  9
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_sink 0.4855
#> parent_m1   0.5145
#> m1_sink     1.0000
#> 
#> Estimated disappearance times:
#>           DT50   DT90
#> parent   7.023  23.33
#> m1     131.761 437.70

initials <- fit$start$value
names(initials) <- rownames(fit$start)
transformed <- fit$start_transformed$value
names(transformed) <- rownames(fit$start_transformed)
transform_odeparms(initials, SFO_SFO)

#>          parent_0 log_k_parent_sink   log_k_parent_m1     log_k_m1_sink 
#>        100.750000         -2.302585         -2.301586         -2.300587 

backtransform_odeparms(transformed, SFO_SFO)

#>      parent_0 k_parent_sink   k_parent_m1     k_m1_sink 
#>      100.7500        0.1000        0.1001        0.1002 

# The case of formation fractions
SFO_SFO.ff <- mkinmod(
  parent = list(type = "SFO", to = "m1", sink = TRUE),
  m1 = list(type = "SFO"),
  use_of_ff = "max")

#> Successfully compiled differential equation model from auto-generated C code.

fit.ff <- mkinfit(SFO_SFO.ff, FOCUS_2006_D, quiet = TRUE)
summary(fit.ff, data = FALSE)

#> mkin version used for fitting:    0.9.47.6 
#> R version used for fitting:       3.5.1 
#> Date of fit:     Mon Nov 26 12:29:22 2018 
#> Date of summary: Mon Nov 26 12:29:22 2018 
#> 
#> Equations:
#> d_parent/dt = - k_parent * parent
#> d_m1/dt = + f_parent_to_m1 * k_parent * parent - k_m1 * m1
#> 
#> Model predictions using solution type deSolve 
#> 
#> Fitted with method Port using 186 model solutions performed in 0.857 s
#> 
#> Weighting: none
#> 
#> Starting values for parameters to be optimised:
#>                   value   type
#> parent_0       100.7500  state
#> k_parent         0.1000 deparm
#> k_m1             0.1001 deparm
#> f_parent_to_m1   0.5000 deparm
#> 
#> Starting values for the transformed parameters actually optimised:
#>                     value lower upper
#> parent_0       100.750000  -Inf   Inf
#> log_k_parent    -2.302585  -Inf   Inf
#> log_k_m1        -2.301586  -Inf   Inf
#> f_parent_ilr_1   0.000000  -Inf   Inf
#> 
#> Fixed parameter values:
#>      value  type
#> m1_0     0 state
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                Estimate Std. Error   Lower    Upper
#> parent_0       99.60000    1.61400 96.3300 102.9000
#> log_k_parent   -2.31600    0.04187 -2.4010  -2.2310
#> log_k_m1       -5.24800    0.13610 -5.5230  -4.9720
#> f_parent_ilr_1  0.04096    0.06477 -0.0904   0.1723
#> 
#> Parameter correlation:
#>                parent_0 log_k_parent log_k_m1 f_parent_ilr_1
#> parent_0         1.0000       0.5178  -0.1701        -0.5489
#> log_k_parent     0.5178       1.0000  -0.3285        -0.5451
#> log_k_m1        -0.1701      -0.3285   1.0000         0.7466
#> f_parent_ilr_1  -0.5489      -0.5451   0.7466         1.0000
#> 
#> Residual standard error: 3.211 on 36 degrees of freedom
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                 Estimate t value    Pr(>t)     Lower     Upper
#> parent_0       99.600000  61.720 2.024e-38 96.330000 1.029e+02
#> k_parent        0.098700  23.880 5.700e-24  0.090660 1.074e-01
#> k_m1            0.005261   7.349 5.758e-09  0.003992 6.933e-03
#> f_parent_to_m1  0.514500  22.490 4.375e-23  0.468100 5.606e-01
#> 
#> Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   6.398       4 15
#> parent     6.459       2  7
#> m1         4.690       2  8
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_m1   0.5145
#> parent_sink 0.4855
#> 
#> Estimated disappearance times:
#>           DT50   DT90
#> parent   7.023  23.33
#> m1     131.761 437.70

initials <- c("f_parent_to_m1" = 0.5)
transformed <- transform_odeparms(initials, SFO_SFO.ff)
backtransform_odeparms(transformed, SFO_SFO.ff)

#> f_parent_to_m1 
#>            0.5 

# And without sink
SFO_SFO.ff.2 <- mkinmod(
  parent = list(type = "SFO", to = "m1", sink = FALSE),
  m1 = list(type = "SFO"),
  use_of_ff = "max")

#> Successfully compiled differential equation model from auto-generated C code.

fit.ff.2 <- mkinfit(SFO_SFO.ff.2, FOCUS_2006_D, quiet = TRUE)
summary(fit.ff.2, data = FALSE)

#> mkin version used for fitting:    0.9.47.6 
#> R version used for fitting:       3.5.1 
#> Date of fit:     Mon Nov 26 12:29:22 2018 
#> Date of summary: Mon Nov 26 12:29:22 2018 
#> 
#> Equations:
#> d_parent/dt = - k_parent * parent
#> d_m1/dt = + k_parent * parent - k_m1 * m1
#> 
#> Model predictions using solution type deSolve 
#> 
#> Fitted with method Port using 104 model solutions performed in 0.474 s
#> 
#> Weighting: none
#> 
#> Starting values for parameters to be optimised:
#>             value   type
#> parent_0 100.7500  state
#> k_parent   0.1000 deparm
#> k_m1       0.1001 deparm
#> 
#> Starting values for the transformed parameters actually optimised:
#>                   value lower upper
#> parent_0     100.750000  -Inf   Inf
#> log_k_parent  -2.302585  -Inf   Inf
#> log_k_m1      -2.301586  -Inf   Inf
#> 
#> Fixed parameter values:
#>      value  type
#> m1_0     0 state
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>              Estimate Std. Error Lower  Upper
#> parent_0       84.790    2.96500 78.78 90.800
#> log_k_parent   -2.756    0.08088 -2.92 -2.593
#> log_k_m1       -4.214    0.11150 -4.44 -3.988
#> 
#> Parameter correlation:
#>              parent_0 log_k_parent log_k_m1
#> parent_0       1.0000      0.11058  0.46156
#> log_k_parent   0.1106      1.00000  0.06274
#> log_k_m1       0.4616      0.06274  1.00000
#> 
#> Residual standard error: 8.333 on 37 degrees of freedom
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>          Estimate t value    Pr(>t)    Lower    Upper
#> parent_0 84.79000  28.600 3.939e-27 78.78000 90.80000
#> k_parent  0.06352  12.360 5.237e-15  0.05392  0.07483
#> k_m1      0.01478   8.966 4.114e-11  0.01179  0.01853
#> 
#> Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   19.66       3 16
#> parent     17.56       2  7
#> m1         18.71       1  9
#> 
#> Estimated disappearance times:
#>         DT50   DT90
#> parent 10.91  36.25
#> m1     46.89 155.75