The transformations are intended to map parameters that should only take on restricted values to the full scale of real numbers. For kinetic rate constants and other paramters that can only take on positive values, a simple log transformation is used. For compositional parameters, such as the formations fractions that should always sum up to 1 and can not be negative, the ilr transformation is used.

The transformation of sets of formation fractions is fragile, as it supposes the same ordering of the components in forward and backward transformation. This is no problem for the internal use in mkinfit.

transform_odeparms(parms, mkinmod,
                   transform_rates = TRUE, transform_fractions = TRUE)
backtransform_odeparms(transparms, mkinmod,
                       transform_rates = TRUE, transform_fractions = TRUE)

Arguments

parms

Parameters of kinetic models as used in the differential equations.

transparms

Transformed parameters of kinetic models as used in the fitting procedure.

mkinmod

The kinetic model of class mkinmod, containing the names of the model variables that are needed for grouping the formation fractions before ilr transformation, the parameter names and the information if the pathway to sink is included in the model.

transform_rates

Boolean specifying if kinetic rate constants should be transformed in the model specification used in the fitting for better compliance with the assumption of normal distribution of the estimator. If TRUE, also alpha and beta parameters of the FOMC model are log-transformed, as well as k1 and k2 rate constants for the DFOP and HS models and the break point tb of the HS model.

transform_fractions

Boolean specifying if formation fractions constants should be transformed in the model specification used in the fitting for better compliance with the assumption of normal distribution of the estimator. The default (TRUE) is to do transformations. The g parameter of the DFOP and HS models are also transformed, as they can also be seen as compositional data. The transformation used for these transformations is the ilr transformation.

Value

A vector of transformed or backtransformed parameters with the same names as the original parameters.

Examples

SFO_SFO <- mkinmod(
  parent = list(type = "SFO", to = "m1", sink = TRUE),
  m1 = list(type = "SFO"))

Successfully compiled differential equation model from auto-generated C code.

# Fit the model to the FOCUS example dataset D using defaults
fit <- mkinfit(SFO_SFO, FOCUS_2006_D, quiet = TRUE)
summary(fit, data=FALSE) # See transformed and backtransformed parameters

#> mkin version:    0.9.44.9000 
#> R version:       3.3.1 
#> Date of fit:     Wed Oct 26 23:16:59 2016 
#> Date of summary: Wed Oct 26 23:16:59 2016 
#> 
#> Equations:
#> d_parent = - k_parent_sink * parent - k_parent_m1 * parent
#> d_m1 = + k_parent_m1 * parent - k_m1_sink * m1
#> 
#> Model predictions using solution type deSolve 
#> 
#> Fitted with method Port using 153 model solutions performed in 0.721 s
#> 
#> Weighting: none
#> 
#> Starting values for parameters to be optimised:
#>                  value   type
#> parent_0      100.7500  state
#> k_parent_sink   0.1000 deparm
#> k_parent_m1     0.1001 deparm
#> k_m1_sink       0.1002 deparm
#> 
#> Starting values for the transformed parameters actually optimised:
#>                        value lower upper
#> parent_0          100.750000  -Inf   Inf
#> log_k_parent_sink  -2.302585  -Inf   Inf
#> log_k_parent_m1    -2.301586  -Inf   Inf
#> log_k_m1_sink      -2.300587  -Inf   Inf
#> 
#> Fixed parameter values:
#>      value  type
#> m1_0     0 state
#> 
#> Optimised, transformed parameters with symmetric confidence intervals:
#>                   Estimate Std. Error  Lower   Upper
#> parent_0            99.600    1.61400 96.330 102.900
#> log_k_parent_sink   -3.038    0.07826 -3.197  -2.879
#> log_k_parent_m1     -2.980    0.04124 -3.064  -2.897
#> log_k_m1_sink       -5.248    0.13610 -5.523  -4.972
#> 
#> Parameter correlation:
#>                   parent_0 log_k_parent_sink log_k_parent_m1 log_k_m1_sink
#> parent_0           1.00000            0.6075        -0.06625       -0.1701
#> log_k_parent_sink  0.60752            1.0000        -0.08740       -0.6253
#> log_k_parent_m1   -0.06625           -0.0874         1.00000        0.4716
#> log_k_m1_sink     -0.17006           -0.6253         0.47163        1.0000
#> 
#> Residual standard error: 3.211 on 36 degrees of freedom
#> 
#> Backtransformed parameters:
#> Confidence intervals for internally transformed parameters are asymmetric.
#> t-test (unrealistically) based on the assumption of normal distribution
#> for estimators of untransformed parameters.
#>                Estimate t value    Pr(>t)     Lower     Upper
#> parent_0      99.600000  61.720 2.024e-38 96.330000 1.029e+02
#> k_parent_sink  0.047920  12.780 3.050e-15  0.040890 5.616e-02
#> k_parent_m1    0.050780  24.250 3.407e-24  0.046700 5.521e-02
#> k_m1_sink      0.005261   7.349 5.758e-09  0.003992 6.933e-03
#> 
#> Chi2 error levels in percent:
#>          err.min n.optim df
#> All data   6.398       4 15
#> parent     6.827       3  6
#> m1         4.490       1  9
#> 
#> Resulting formation fractions:
#>                 ff
#> parent_sink 0.4855
#> parent_m1   0.5145
#> m1_sink     1.0000
#> 
#> Estimated disappearance times:
#>           DT50   DT90
#> parent   7.023  23.33
#> m1     131.761 437.70
#> 

## Not run: ------------------------------------
# fit.2 <- mkinfit(SFO_SFO, FOCUS_2006_D, transform_rates = FALSE)
# summary(fit.2, data=FALSE) 
## ---------------------------------------------

initials <- fit$start$value
names(initials) <- rownames(fit$start)
transformed <- fit$start_transformed$value
names(transformed) <- rownames(fit$start_transformed)
transform_odeparms(initials, SFO_SFO)

#>          parent_0 log_k_parent_sink   log_k_parent_m1     log_k_m1_sink 
#>        100.750000         -2.302585         -2.301586         -2.300587 
#> 

backtransform_odeparms(transformed, SFO_SFO)

#>      parent_0 k_parent_sink   k_parent_m1     k_m1_sink 
#>      100.7500        0.1000        0.1001        0.1002 
#> 

## Not run: ------------------------------------
# # The case of formation fractions
# SFO_SFO.ff <- mkinmod(
#   parent = list(type = "SFO", to = "m1", sink = TRUE),
#   m1 = list(type = "SFO"),
#   use_of_ff = "max")
# 
# fit.ff <- mkinfit(SFO_SFO.ff, FOCUS_2006_D)
# summary(fit.ff, data = FALSE)
# initials <- c("f_parent_to_m1" = 0.5)
# transformed <- transform_odeparms(initials, SFO_SFO.ff)
# backtransform_odeparms(transformed, SFO_SFO.ff)
# 
# # And without sink
# SFO_SFO.ff.2 <- mkinmod(
#   parent = list(type = "SFO", to = "m1", sink = FALSE),
#   m1 = list(type = "SFO"),
#   use_of_ff = "max")
# 
# 
# fit.ff.2 <- mkinfit(SFO_SFO.ff.2, FOCUS_2006_D)
# summary(fit.ff.2, data = FALSE)
## ---------------------------------------------